[starnix] Add stubs for timer functions.
Adds implementations of the timer syscalls that call into a new
TimerTable. These functions currently do nothing but will be implemented
in a followup to manage the timer mappings.

Change sigevent to be opaque for bindgen. The optional function pointers
could not be deserialized using the default deserialization.

Change-Id: I53df1d1aa30a7f9e3b901d6f3f0ce24a05a1f634
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782184
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>