[fidl][dart] Refactor message decoder
This change refactors the `decodeMessageWithCallback` function to be
more agnostic to its position in the message buffer. Previously, it
assumed that top-level members were always offset `kMessageHeaderSize`
into the buffer. For large messages, that assumption will no longer
hold, unless we want to decode the header twice (we don't).

Instead, we simply have the callback supply the offset as an argument
(as of this CL, always `kMessageHeaderSize`). This brings us closer to
the C++ and Rust decoder designs, which operate on the idea that they
are decoding two sequential FIDL messages (a struct followed by a
header, consecutively in the same buffer), rather than special-casing
the transactional message.

Test: fx test fidlgen-dart; fx test check-goldens
Bug: 114261
Change-Id: I43906330bb3c35f3933ecd92227a5f1fd80478e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781609
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Yifei Teng <yifeit@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>