[fshost] add generic config overrides
Instead of having a builder method per option we want to override, add a
generic config override method that lets tests pass in any config and
the value to set for it. This uses a trait to allow for the relevant
subset of supported config value types to be passed to the same override
function.

Bug: 87511
Test: fx test fshost_integration_tests
Change-Id: Ica4278004b335e612261b90c75aa857a02b4ce23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783343
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>