[build] Add support for dynamic golden file sets
This change introduces a mode of `golden_files()` in which a JSON file
of golden file comparisons may be passed instead of a literal GN list of
them. This is useful in situations in which the set of golden files
is not statically known in GN, as is the case with generating syscall
documentation pages from the FIDL syscall representations in library zx.

Since this support in `fx check-goldens` requires a building of these
manifests, this behaviour is put behind a flag

Bug: 118025
Change-Id: I0d88862d1abdacb6dff2fea4f102c7ff8d87785b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781863
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>