[check-licenses] Remove unused "diff" logic.
Previously, check-licenses had the ability to diff a given license
file against a previously released one.

It was only needed one time. If we have to go through that exercise
again, we should be able to directly diff the SPDX files instead
of needing special logic baked into check-licenses.

This CL deletes that logic, which will allow me to reorganize things
to be more flexible. If necessary, We can add back the functionality
in the future with tests for better readability and maintainability.

Change-Id: Ia79c50a12c792187fc7625d74e5966944b155146
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783843
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>