[check-licenses] Fix Package SPDX IDs.
I need to ensure that SPDX IDs are always unique.
Previously I was using project names as the SPDX package IDs,
but there is no guarantee that there aren't two projects with
the same name in the build.

This CL changes it to use an auto-incrementing integer in
the package SPDX ID string.

Bug: 117652

Change-Id: I9296f5e923c06304d8d7ec9a5b1a35cb1c20ed56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784082
Reviewed-by: Yaar Schnitman <yaar@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>