[fidl][fix] Add CompiledTransformer
This is a transformer that provides the flat AST node(s) associated with
some raw AST source node when visiting that node during transformation.
This allows for decisions about whether (and/or how) to do the transform
to access post-compile information about the visited node.

Test: fx test fidlc-test
Bug: 114357
Change-Id: Ie2e9e6140333e3c9b49e47aa9525ea3131548116
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769088
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>