[starnix] binder ioctl must not return EINTR when it handled some data
The client of the binder can request both to send commands and receive
commands at the same time. When it does this, the driver waits for some
command to be available before returning. If this wait is interrupted
when the driver has handled some command from the client, it should not
return an error to let the client knows that its commands have been
handled.

Change-Id: Ifb503d2cd15211e649efb491376a9351fb4b8712
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778945
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>