[av400][dsp] Modify the way SMC accesses the dsp firmware loading address.
In order to improve the compatibility of the DSP driver, modify the
method of obtaining the loading address of the dsp firmware. In the
original program, there was an error: the value of kDspSecPowerSrt was
incorrect, and this error was not found during verification.

Change-Id: I434aff8fedbefd15dd86c70de9c365f166e4f1f1
Signed-off-by: Dingxin Jin <dingxin.jin@amlogic.corp-partner.google.com>
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780943
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>