[pave] Change default to --run-once
The previous default for `fx pave` and `fx pave-remote` was
`--keep-running`, so as soon as the script identified a device in
zedboot it attempted to pave and it will keep doing that cycle
forever. This creates a problem if the device is set up with USB as
the first boot because when pave issues the reboot command it
reboots back into zedboot and starts paving all over again. The
problem could be prevented by passing the -1 argument to the script
so that it only attempts to pave once, but many users (especially
new ones) don't know about this argument and don't expect the
continuous paving behavior.

This CL changes the default to `--run-once`. It also changes the flag
from `-1` to `--run-once` in order to be a bit more dscriptive.

Change-Id: I0bc93adcca186a99f44b8810d73f620fba8a0610
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780646
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>