[fidl][screenshot] Add new TakeFile method to the screenshot FIDL API
When used, the screenshot will be sent as a |fuchsia.io.File| over
the using the file channel in the |Screenshot| response.
This flow typically useful if the client is on
the host and does not support VMOs, as is the case for ffx tools.
Both in gfx and in flatland a map is added to hold the vmofile and
the channel observer.
Once the channel is closed, the entry is removed from the map.

Test: UT
Bug: 109048

Change-Id: Ia1f3419967874befe3ded5db7ef5bce750d2bd82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762865
Reviewed-by: Felipe Archondo <felipearchondo@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
API-Review: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Henry-Michel Cantave <hmcantave@google.com>