[moniker] Make &str -> InstancedChildMoniker fallible
We implemented `From<&str>` for InstancedChildMoniker and
unwrapped the result implicitly. This was mostly used in
tests, but is a dangerous implementation because it is also
available for production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<&str>` for InstancedChildMoniker
and changes all tests to unwrap explicitly.

Bug: 118649
Change-Id: Ie547861836ff13918e9fe6564c1eb3fd80d1ffba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786082
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>