[tpm] Burst count should respect the max acceptable by FIDL.
There was a very subtle bug in the TPM driver where if the STS register
returned a burst_count greater than the max data transfer allowed by
FIDL it would fail the FIDL call before sending the command to the
tpm-lpc driver. This occured when testing TPM2_CreatePrimary which
was the first time this occured.

Change-Id: I82aa026771573d1186aa21043b1615292bc42682
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785946
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Ben Wright <benwright@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>