[elfldltl] Clean up VMO name handling in VmarLoader
The SetVmoName helper doesn't need to be a method and so doesn't
need to be declared in the public header.  Regardless, the
constants used as template parameters don't need to be public
since they're not themsleves used from code in the header.

Bug: 91206
Change-Id: Icb13c9bc704f2eb2da39bd052b994c3e377959a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785746
Reviewed-by: Alex Brachet <abrachet@google.com>
Reviewed-by: Luke Nicholson <lukenicholson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>