[docs] Include sections on ffx repository and package
We are in the process of migrating of `pm` over to `ffx package` and
`ffx repository`. This updates the documentation to include sections
that describe how to use the new tools and the legacy `pm` that will
stick around until we've completed the migration.

Bug: 116114
Change-Id: I36c0f6d718ed3c361a18acea311b374efdf208eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781582
Reviewed-by: Kyo Lee <kyol@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>