[component_manager] Fix leak in events synthesis
Events synthesis spawns a task with `detach` that performs the
depth-first exploration for event generation. Bind the task to the
target component's TaskScope so it gets released when the component is
destroyed.

This leak has existed for a long time, but it was exacerbated by
fxrev.dev/761894 since that it makes it inevitable that event routing
will complete.

Bug: 116722
Change-Id: Id0a8eef1f7f55df382338dbe1bd0a13a04581747
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780647
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Gary Bressler <geb@google.com>