[moniker] Make Vec<&str> -> AbsoluteMoniker fallible
We implemented `From<Vec<&str>>` for AbsoluteMoniker and
unwrapped the result implicitly. This was mostly used in
tests, but is a dangerous implementation because it is also
available for production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<Vec<&str>>` for AbsoluteMoniker
and changes all tests to unwrap explicitly.

Bug: 118649
Change-Id: I9a4f5d644a2f6f1ed9f2fcf79c2a1c1385289c98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786362
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>