[moniker] Make Vec<&str> -> InstancedAbsoluteMoniker fallible
We implemented `From<Vec<&str>>` for InstancedAbsoluteMoniker
and unwrapped the result implicitly. This was mostly used in
tests, but is a dangerous implementation because it is also
available for production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<Vec<&str>>` for
InstancedAbsoluteMoniker and changes all tests to unwrap
explicitly.

Fixed: 118649
Change-Id: I175dd46e810dca69bc6f2d6f507bbad967dc9787
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786502
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>