[wlan][policy] clean mod.rs test setup
This change adds the presaved networks in the fake saved networks
manager for tests as part of test values so they can be used directly in
tests. Previously the SSID's of networks saved in test setup were
directly used in tests and it was not obvious that they were saved
during test setup.

This test also changes the way FakeSavedNetworksManager is initialized
with some saved networks. The new way will allow tests in a following CL
be more straightforward.

This change also cleans up some tests that were not fully taking
advantage of the test_setup fn.

Test: ran unit tests; this is a refactor of the tests.
Change-Id: I43f5d32af61f1ada1fe19723e9557e98e4fc7821
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785765
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>