[assembly] Change WLAN assembly logic
Change the `include_wlan_aibs` flag so that it only gates the inclusion
of non-driver packages, whereas having "fuchsia::wlan_softmac" or
"fuchsia::wlan_fullmac" in the BoardInformation's provided_features list
would gate the inclusion of driver packages.

Being able to include non-driver and driver packages separately will
make the transition to Product Assembly easier, as the former packages
are included via product config and the latter packages via board
config. Because we sometimes combine in-tree products with out-of-tree
boards, or in-tree boards with out-of-tree products, with the previous
assembly implementation for WLAN, it was difficult to ensure that a
package gets included exactly once.

Other additional changes with this commit:
- x64 boards now include driver packages via Product Assembly
- Move the driver packages into base_driver_packages list from
  base_packages.

Bug: 117249
Test:
1. CQ
2. Verify that WLAN drivers bind on smart display and workstation

Change-Id: Ie87689e54d7bee4552e5ee48da8675396c31fab3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786285
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Kiet Tran <kiettran@google.com>