[blobfs] Handle multiple shutdown/stop requests to component.
Multiple requests to lifecycle stop or fs_admin shutdown (or any
combination of the two) will now only perform a single teardown and
return to all of them with the same teardown result.

Prior to this multiple calls to lifecycle stop would attempt to perform
teardown procedures twice. While this would only trip a debug assert
currently, this may not always be true, and the handlers should make it
safe.

Bug: 117992
Change-Id: I9e39e61a2ba85a15fe97e7213d46046f4a7537bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786162
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>