[fshost] Don't file crash reports on FDR
When Fxfs is using the legacy crypto format, we implement FDR by
creating a marker file in Fxfs which is detected on boot and which
results in fshost reformatting Fxfs.  This also creates a crash report
at the moment, which is not what we want.

Fix this to not file a crash report.  Also fixes the integration tests
to actually exercise this (the relevant tests were matching for
the "fxfs" feature, but the integration tests for the legacy format set
the "fxfs-legacy-crypto" feature instead), since they would have caught
this bug.

Fixed: 118909
Change-Id: Ia93c984a566292bc41bbf0be63956b7e6f6b45a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787090
Reviewed-by: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>