[fdf] Check rebind during IsUnbound
Check that `pending_rebind_` is false during the compat device's
`IsUnbound()`. The dev connector closes a connection that attempts to
connect to a device's controller if `IsUnbound()` is true.
`IsUnbound()` used to only check if the device is pending removal.  If a
DFv1 device in a DFv2 compat shim is rebound then the compat shim marks
the device for pending removal. Now if a DFv1 device in a DFv2 compat
shim is rebound, the dev connector will successfully connect to the
device's controller instead of closing the connection.

Set `pending_removal_` to false once a compat device has been rebound.

Bug 117669

Change-Id: If1251d8b4b8696ab02930ac7a4ca71ce1d0ea847
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786742
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>