[device-watcher] Fix request race with teardown
Previously, `IterateDirectoryTest` tests would create requests to the
fake filesystem via the `device-watcher` library's `IterateDirectory()`
function. The requests are asynchronous/pipelined and not confirmed to
be acknowledged by the server by the time it is passed to the test's
callback. Some tests would not verify that the connection is
acknowledged before finishing, which would cause the fake filesystem and
its async loop to teardown as well. If the destruction occurs before the
outgoing request is acknowledged, this would cause the request to be
dispatched on a dead async loop.

This change uses `ManagedVfs` instead, which can handle this case.

Fixed: 118686
Change-Id: Id0300607e3c0697750f653497a72d2806c3a9e3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782024
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Chris Tam <godtamit@google.com>