[virtualization] Route fuchsia.input.wayland.Keymap virtio_wl.
Now that the wayland server is inside the virtio_wl device, we need to
update the routing for this service. The addition of this service
conflicted with some of my other changes so this is just fixing up
routing for this service, both for static-child vmm's and also for vmms
launched using the newer vmm_launcher path.

Bug: 115651
Test: Observed the routing error for wayland.Keymap disappears when
      booting the linux terminal on workstation.
Change-Id: Ia16fd1b43ca806d6969fbf12e193511a3542d9c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787203
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>