[superproject] Roll third_party/antlion Fix DHCP options error in Dhcpv4InteropTest
The Dhcpv4InteropTest was updated to use generate_tests but does not
have an option for test_name which was added. This change fixes that bug
and refactors the test to improve the readability of the test
parameters.

Original-Fixed: b/232574848
Test: Dhcpv4InteropTest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/787282
Original-Revision: 73965b78a0f8f71ae56aa206b346e44e7e315f3d
GitOrigin-RevId: 5d1d1d5051b35ddc5a2e791939cb28613603dbc8
Roller-URL: https://ci.chromium.org/b/8792658415909918705
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I234b47de5d1da5823dca485c1bd4c31af57cbe8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787320