[kernel][ktrace] Migrate to interned categories.
This change is an intermediate migration to FXT interned categories in
the kernel and ktrace. To maintain compatibility with the existing
ktrace syscall interface, the defined ktrace filter groups are
statically mapped to interned categories with the same category names
that ktrace_provider uses.

- Update ktrace session logic to test categories instead of tags.
- Update ktrace fxt wrappers to accept categories instead of tags.
- Update trace points to supply categories instead of tags.
- Update category checks and remove redundancy.

Test: Observe traces with various settings
Change-Id: I1b0dd627f414048fdb0aefbf54045813fe1e6297
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780749
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>