[fidl][rust] Add on_closed() to ControlHandle
This CL adds on_closed() to ControlHandle, similar to the method on
Proxy types. This makes it possible to await PEER_CLOSED while doing
other things in the middle of a request handler.

I have used it to simplify the case given in fxbug.dev/101155.

Test: fx test archivist-tests --test-filter='*batch_iterator_terminates_on_client_disconnect*'
Fixed: 101155
Change-Id: Ib83cf7c616cf7d106165c3b06d7bafd1b93a5a14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787058
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>