[virtualization] Always use the embedded wayland server
Remove the code and plumbing to allow using the wayland_bridge
standalone component in production. The virtio_wl device retains the
ability to use a remote server for testing purposes only. In production
the wayland server for virtualization will be linked directly with the
virtio_wl device.

Change-Id: If8231958970e250d605dbe6ff73af2c0cf4e0169
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783504
Reviewed-by: Alexander Radchenko <alerad@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>