[wlan][policy] fix scoring log spam in match
There was a typo in a match statement that lead to a lot of log spam if
the value matched the intended bucket. This changes the match style of
the match statements to be less likely to have similar typos, and there is
no unexpected case that would cause logging anymore.

Change-Id: Ic1272333119125b9c37b2cf585c9d65ed94900a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787508
Fuchsia-Auto-Submit: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>