[superproject] Roll llvm-project/libc Reland: "[libc] Templatize str{,n}cmp"
This will be used to implement the case insensitive str{,n}casecmp

This was initially reverted because it broke tests on arm platforms.
Unfortunately, it didn't break on my arm machine, but I suspect the
problem was the old comparator returned char and not int.

Differential Revision: https://reviews.llvm.org/D141235

GitOrigin-RevId: 637f416409d311ed79800305d108160b0c0e0344
Original-Revision: 94236d3b45a0f241863bd35ee965eed1265fdf11
Roller-URL: https://ci.chromium.org/b/8792335538401443105
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ifcebad889174df305ea7a61fa0e42a11e3808822
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787863