[emu] Fix --reuse semantics
The attached bug called out an unanticipated change in the semantics of
--reuse; when the instance did not already exist, the new semantics
would fail the launcher instead of creating the new instance.

This CL adds a new EngineOption type as output for get_engine_by_name,
such that callers can distinguish between failing to retrieve an
existing instance and that instance simply not existing. Now, if the
instance didn't exist before the start call with --reuse, we notify the
user, revert the --reuse flag so that the instance can be properly
initialized, and then continue the start sequence.

Bug: 118006
Change-Id: I591743b2a4bd4a5ce53bbc5b5518941dd650a3d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786441
Reviewed-by: Matt Hamrick <mjhamrick@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>