[gnsdk] Fix output manifest name for coverage test components
Recent changes to component.gni and prepare_package_outputs.py
resulted in the output name of the cm file for tests with coverage
enabled not matching the test name (it had a prefix and suffix added).

This adjusts it so that the output name of the merged manifest is
the same as the base manifest, but places it in a subdir to avoid
output path collisions.

Bug: chromium:1406284
Change-Id: Iab3ca3b3087f11ac9d983bed94c1522777d181b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788034
Commit-Queue: Bryant Chandler <bryantchandler@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>