[ffx][product] add Fuchsia Tools client id and secret
This isn't a web app, so the "secret" is not secret. As is done for
gsutil and other client side executables, the client id and secret
are baked into the program in the source. This means that Refresh
tokens must be kept secret (they should be anyway, but for client
executables the security comes down to the refresh token).

Change-Id: Ic06d35db8f96cd562f1028b047d0f499df29bcbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787309
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>