[fidl] Add Fixable* helpers to Reporter
The |Reporter| now has helpers to report |Fixable[Error|Warn]|s, which
should be used to report errors that that should be reported as such
(ie, always build failing) in diagnostics, but can optionally be ignored
and disabled when fixing. This implies that compilation should be able
to fully recover from such errors - exiting out early on them is not
allowed, as there are scenarios where we want compilation to always
complete in spite of them.

Test: fx test fidlc-test
Bug: 114357
Change-Id: I624aa3f623d740a46c06242933d2e9b4c5ea43da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782896
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>