[hierarchy] Remove usage of regexes in select_from_hierarchy
`select_from_hierarchy` was previously executing a `filter_hierarchy`.
In the case of Sampler, this wasn't necessary since Sampler receives
hierarchies that have been filtered by the Archivist.

In the case of triage, there's no need to filter the whole hierarchy and
then collect the properties into a vector, we can instead get the
properties as we traverse the hierarchy in a DFS way.

This still doesn't remove usage of RegexSet in the Archivist but it does
for Sampler.

Additionally the new implementation removes cloning of data and returns
the selected properties in a hierarchy as references to the original
hierarchy. There's some cloning still in triage that might be possible
to optimize into dealing with references, but that's out of scope of
this CL as Sampler is the main target.

This refactor brings back ~300K in Sampler:
https://screenshot.googleplex.com/WXqGbYVzNzV5FwN.png.

Bug: 117157
Run-All-Tests: True

Change-Id: Icab6052937ead3fdb52f8cc541f4b212181b99e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777862
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>