[emu] Remove fx emu-remote
The `fx emu-remote` command has been broken for a while now, since it
had a dependency on `fx emu`. Even before that, it was going unused
simply because there's no demand for it; running over CRD works
perfectly well for all known use cases.

This CL removes the command, a test script that was meant to exercise
it, and removes a few references to it. There was also a test script for
`fx emu` that was previously missed that is included in this CL as well.

Bug: 116978
Change-Id: Ie2ee447ff0a7cc74e40daf57c9cbeac08c473ac9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788424
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>