[miscsvc][assembly] Migrate to a package
zxcrypt was updated to look for the config file in two locations - /pkg
and /boot. This is necessary because when miscsvc is packaged it no
longer is presented with /boot as if it were /pkg. In order to vary the
contents of it's /pkg per board, we would need to make it into a board
specific component. While switching to structured config would allow us
to remedy that, it is outside the scope of this change to switch miscsvc
over to structured config.

Change-Id: Ia9de1db61db7d855600b278d97ba1e9bca4e3d4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784125
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>