[doc-checker] Detecting brackets that are not links
This adds code to check for the situation where the
markdown contains [something]. This is ambiguous since
it could be an autolink to the reference [something],
or it could be simply text.

This change compares the contents of the link to the link_url
and if they are identical, it is flagged as an ignored autolink.

This generates an info message, but does not block the submission.

Note: This change is only visible if you use the
`--check-reference-links` flag.

Change-Id: Idc354165898e8c5a7faec582d3aed4bcf25dbc65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788605
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Matt Hamrick <mjhamrick@google.com>