[fidl][dynsuite] Server suite separate controller
Change the server suite to use a separate controller protocol instead of
the reporter protocol. In addition to sending reports on one-way methods
received by the server under test, the controller protocol is also used
by the harness to command the SUT to send events, and has the ability to
report the result of sending an event in addition to just commanding it
to be sent. This separation of the "send event" command from the SUT
also avoids some potential pitfalls of manually building the "send
event" command and sending it on the same channel as the SUT.

Bug: 102339
Test: fx test //src/tests/fidl/server_suite
Change-Id: I1286988b74e39df9589cbbe381ae4841ed2568b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786302
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>