[emu] Consolidate TestEngine impls
Some of the ffx emu code uses a minimal implementation of the
EmulatorEngine trait for testing. Previously, these impls have been
generated in the file where they're used, but this leads to a lot of
duplicated boilerplate code in those locations that must be adjusted
whenever that trait is changed.

This CL adds a default implementation for each function in the trait
that simply contain "unimplemented" statements which panic if executed
directly. The test locations that create TestEngines can now implement
only the functions they need to mock out, reducing duplicate code in
those areas. This also makes it much easier to add TestEngines to some
other parts of the code where they are needed for additional testing
use cases.

Change-Id: I6d07eab360af2f4d7bf2e85d2640c7f72aea59ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789177
Reviewed-by: Matt Hamrick <mjhamrick@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>