[usb][dwc2] Set clear-NAK and ep-enable bits on timeout
This sets the IN/OUT EP0 clear-NAK and ep-enable bits in preparation of
the tokens to follow a timeout interrupt. The tokens to follow will
either be a SETUP token if the host received the data but its ACK was
lost, or an IN token if the data sent to the host was lost in transit.

Bug: 105382
Change-Id: Ia4dd06d4e9b23d2b36e9bb85be7f8a0ddb4946f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789272
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>