[ffx][error] Downcast anyhow errors, not type check
When I changed over to the new error type I mistakenly thought doing a
`.is<T>` check on the anyhow error would give me a result in line with
the likely success of `.downcast<T>`, but it doesn't. Downcasting
actually *searches* the error for a matching type of error, so it'll
pull one up from deeper in the error stack.

This fixes the logic to use downcast instead, resulting in it reaching
down the error stack to find an ffx user error.

Fixed: 119301
Change-Id: I3ccac0fa53f07feb9f8a905edf482c2f1242d23d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789956
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>