[emu] Add testing around control sequence of start
This CL adds more tests to the emulator start subcommand, specifically
in the domain of the start command's control sequence. It verifies that
the dry-run, stage, and edit flags interact correctly with the config,
reuse, and verbose flags and that changes in each stage are correctly
propagated to the final start Command.

In writing these tests, we uncovered an as-yet undetected bug in the
--edit flag behavior that was actually losing the user's modifications
before launching the emulator instance. That has been addressed in this
CL as well.

Change-Id: I31d35851987e75ae5b605ef3455ef163f6046f4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789729
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Matt Hamrick <mjhamrick@google.com>