[ffx][metrics] Abstract metrics logic from ffx main
This moves the logic for metrics reporting out of the main ffx run
function, allowing it to be run by subtool processors and putting the
argument redaction logic in the right place. This change is intended to
be a code shuffle with no major logic changes, follow up changes will
actually do the work.

There is one logic change though:
This *does* make it so that when run in isolate mode, metrics
will be turned off.

Bug: 116489
Fixed: 114678
Change-Id: Ieedca59a50dc887a117f473f1f730c1d38400d8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789517
Reviewed-by: Bob Evans <bobevans@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>