[xhci] fix HubDeviceRemoved race condition
If HubDeviceRemoved was called during NormalRequestQueue when the lock
was released for the WaitForIsochronousReady call, the TR would have
allocated a TRB context already, but when HubDeviceRemoved completes, it
would've disconnected the device and dealt with all TRBs in
pending_trbs_. After the lock is released in HubDeviceRemoved,
NormalRequestQueue may continue running, but it doesn't check again if
the device is disconnected or not and commits the TRB, placing it into
pending_trbs_. This TRB context does not get cleaned up and causes a
crash when deallocating.

This is fixed by adding checks after we re-acquire the lock after the
WaitForIsochronousReady call.

Bug: 119477

Test: tested on Atlas

Change-Id: I7d45a4dc437b8d06806d1ffc0fca59b8a1a5ead9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790511
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>