[tools][sdk-tools]: adds functions to SDKProperties to accept a context
Adding a context.Context to common SDK commands that may be long running
(e.g. ssh, gsutil, ffx) to allow for cleaner process cancellation and
cleanup.

This also allows for a logger to be passed through the Context (to be done
in a later commit).

Change-Id: Ib673821c5f5893c0b728c6d3c4c21ed4a22a229c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788310
Reviewed-by: Theodore Siu <theosiu@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>