[fdio] Always initialize and rewind dir iterators
If a client were to create some `DIR*` via `dup()`ing another file
descriptor and then `fdopendir()`, the resulting directory iterator will
be uninitialized but the underlying connection may be shared with the
original descriptor. For remote connections, the state of the directory
stream pointer is held within the connection (since the connection is
stateful), so the `DIR*` will share the directory stream pointer with
the original file descriptor. Clients who call `rewinddir()` expect that
pointer to be rewound.

Previously, this would not occur because `rewinddir()` would act as a
no-op on an uninitialized directory iterator. However, given the
aforementioned scenario, this change makes `rewinddir()` always try to
initialize and rewind directory streams.

Fixed: 119323
Bug: 119968
Bug: 118601
Change-Id: Ida9b884b2f0783faf83fa5731af688d798c9f9c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790665
Commit-Queue: Chris Tam <godtamit@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>