[network/tests] Run NS2 tests with expectations
Rather than running all test targets with expectations, I've elected to
run only those targets for which we need to skip/fail a test. The
benefit is faster runtime and less configuration overhead; the cost is
that we support both formats of running tests. If this support becomes
cumbersome, we can elect to migrate the other targets later.

Bug: 115159
Change-Id: Ibab8eec4e94d52a6567992d51c25f93342fced43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790663
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>