[components] adjust regex used in testing URL validation
The URL validation portion of the unit tests for the cm_fidl_validator
library uses proptest to generate random input to the validator based on
a regex. This regex allowed invalid URLs to be generated, and thus needs
to be adjusted to prevent this scenario.

The new regex is constructed from the ABNF specification in the URI
standard (RFC 3986), but tailored to a subset of value that pass
URL constraints enforced by the `url::Url::parse()` parser.

Note that the special case "Invalid scheme" error is not actually
invalid, if the URL is valid as a relative URL, which is often the case,
so this error check has been removed and test cases updated. (Including
this extra error check would break proptest!() because the regular
expression for proptest!() could generate a legal URL that still failed
the scheme test.

MULTIPLY: cm_fidl_validator_url_proptests
Fixed: 118872
Change-Id: I0e477350914d686326e79228411d4c729b2a820b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785852
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Rich Kadel <richkadel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>