[fuchsia.io] Node2.Reopen takes a server_end:Node
Node2.Reopen should only be able to create connections to Nodes, so
the flexibility of zx.handle:CHANNEL is not needed and the type can
be narrowed to server_end:Node.

This is a breaking change; it cannot be expressed using platform
versioning because to do so would require conditional compilation
support in Dart (see https://fxbug.dev/105932).

Change-Id: I223893a73b662469218ef5deda9299aaf985704a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790825
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Adam Barth <abarth@google.com>