[ui] Migrate ft_device_test to safer teardown
This CL migrates focaltest-test.cm to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Bug: 118385
Change-Id: I2e4541fa6e6e0ae63ea0ab3fbbe4f7866356fff5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791471
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>