[zxio][1/3] Allow closing handles without waiting
This change introduces a new transitional function to allow for
`zxio_close()` to accept a `should_wait` parameter.  When `should_wait`
is set to `false`, this allows remote connections to drop the underlying
channel instead of waiting for a response from the remote server. This
can be useful when the result and errors of a close are not needed and
provides a significant performance boost in these scenarios.

Bug: 118942
Bug: 102657

Change-Id: Icaa2ee9288e08b96f0f9ebcc5bc38dd59820407b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785462
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Chris Tam <godtamit@google.com>