[iwlwifi] CancelSync should wait if task is executing
It's possible for the debug assert in TaskInternal's destructor to fail
if the destructor is called while the task is currently running.

CancelSync() should always return the task back to idle whether that
means waiting for the task to complete, doing nothing, or removing the
task from the queue. This commit changes CancelSync() to wait if the
task is currently executing.

Bug: 120054
Change-Id: I9f0e0a63ada742985926e3272fb22e02e16f21cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791662
Reviewed-by: Louis Lo <yjlou@google.com>
Commit-Queue: John Tan <johntan@google.com>