[netstack3] Downgrade "not supported" logs to warn
Due to limitations in the test framework (see https://fxbug.dev/65359),
an error log in a test case fails the whole suite, even when run with
expectations. Since the "not supported messages" are just informative,
downgrade them to warn; this makes it easier to run many of the socket
suites.

Bug: 65359, 115159
Change-Id: Ic63e5a4e8ab041f05878ee9755632a510543b0a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791646
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>