[kernel][dlog] Fix the calculation of remaining ring buffer space.
Fix an issue discovered by KASAN which only (sometimes) triggers when
rendering the most recent DLOG messages into a crashlog.

Basically, what was going on here was that the code was attempting to
compute the 1 or 2 regions which contained the payload portion of a
log record.  It correctly computed the offset in the ring buffer of
where the payload stated, but when it went to compute the distance
from the start of the payload to the end of the ring buffer, it ended
up using the start of the _record_ instead of the start of the
_payload_.  This meant that when the payload was split across the
break point in the ring buffer, the code could end up reporting a
single region which went off the end of the ring buffer, instead of
properly reporting it as two regions.

Fix the issue and add some new tests which explicitly check the
behavior of the internal method which was misbehaving.

Tests: new tests
Fixed: 120254
Change-Id: I88e1e9c1549e0e210c44b00fca18ad9b2a11fb21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791637
Commit-Queue: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>