[fx][ffx] Only warn on `ffx repository default` mutations
`ffx repository default get` will return the default repository that's set
up by the tools/devshell/contrib/ffx wrapper script, so it should be safe
for users to call this method. This changes the wrapper to only emit a
warning if the user tries to run a command that mutates state, which
wouldn't be visible to users of this wrapper script.

Change-Id: I3594876ed56b50f1202bd1f30d8ccd1796161dfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791847
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>