[sdk] Add missing entries to //sdk/fidl:sdk_fidl_targets
`//sdk/fidl:sdk_fidl_targets` contains the list of all the FIDL
libraries that should be exported in the GN SDK. Some of them aren't
currently listed and only end up being exported because another library
depends on them. This CL explicitly list all the partner libraries and a
future one will add some enforcement rules to prevent regressions. This
CL is a functional noop, the content of the SDK is unaffected.

In order to confirm that this is a no-op you can:
- Run `fx set core.x64`
- Build the `sdk:merged` target without this CL, archive
  out/default/sdk/archive/merged.tar.gz in a temp directory.
- Patch this CL, then build `sdk:merged` again.
- Compare the generated archive with the one saved earlier. The total
  number of files in both archives should be the same, as well as the
  list of FIDL libraries in the fidl/ subdirectories of these SDKs.

I'm not sure why the manifest change was required here. "fuchsia.auth"
should already have been added in the SDK as it's currently exported,
maybe the parser was fine with this because
"sdk://fidl/fuchsia.auth.oldtokens" is in the list already? We should
investigate, I've opened fxb/120473 to track this (it looks like an
overlook made in the CL that initially created this manifest file). It's
safe to proceed with this change regardless, this library IS included in
the SDK and should be listed in its manifest.

Bug: 120472
Change-Id: Idfc093569269b317f465ad37f6ffc42d4c583573
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792367
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Shai Barack <shayba@google.com>