[ffx][screenshot][plugin] New ffx plugin to take a screenshot
The plugin uses the [`TakeFile`] method as defined in the
screenshot.fidl file.
The default directory can be overwritten by using the `--dir` flag
The default behaviour of the plugin is to save the BGRA file data.
In order to save the file as a png the `--png` file must be passed.

To use the command, run `ffx target screenshot`.
To run the unit tests run `fx test ffx_target_screenshot_lib_test`

Tesst: UT
Bug: 109048

Change-Id: I8a93a8696686e74747a091760847617e8d593f6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757563
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Henry-Michel Cantave <hmcantave@google.com>