[sysmem2] sysmem2 tests
This ports all the sysmem(1) tests in sysmem_tests.cc to sysmem2.  For
the sysmem2 tests we use natural types everywhere.

This CL intentionally minimizes diffs to the sysmem(1) tests.  We can
potentially apply some of the syntactic simplification to the sysmem(1)
tests if we think that's worthwhile enough, but for this CL the intent
is to ensure the diffs are concentrated on creating the sysmem2 tests.

Token conversion (v1/v2 conversion) tests will be in a later CL.

Change-Id: Ia11b4c9aa61ac114028afe77b5f4e6979f1276f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763285
Reviewed-by: John Bauman <jbauman@google.com>
API-Review: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>