[fshost] add more context to errors in write_data_file
Add more context calls in the write_data_file path, so it's easier to
narrow down where errors are coming from. Also check the contents of the
written file in the tests, add a test for writing the file to the root
of the partition, and fix that behavior in the rust fshost.

Bug: 120377
Test: fx test fshost_integration_test
Change-Id: I87812cc8547cb9d4bb7fc156c81e2484215bae89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792262
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>