[superproject] Roll third_party/pigweed pw_protobuf: Codegen kConstantCase field enums
pw_protobuf generated code includes an "enum class Fields : uint32_t"
type in in message, with values equal to the IDs for each field in the
message.

Unfortunately, the existing generated code uses SNAKE_CASE for the
names, rather than kConstantCase. In addition to be against the
style guide, this makes collisions with legacy C macros much more
likely.

Indeed, I ran into this issue due to a naming collison on OUTPUT_TYPE
between stm32cube.h and descriptor.pwpb.h. Of course, ST deserves
significant blame here for using inadequately namespaced global
identifiers... it's not just pw_protobuf's fault for not using
kConstantCase.

This commit does not yet migrate pigweed code to use the kConstantCase
values, that will be in an immediate follow-on. However, this commit
DOES add a GN variable to control generation of the legacy
names (in addition to new names):
    pw_protobuf_compiler_GENERATE_LEGACY_ENUM_SNAKE_CASE_NAMES

This currently defaults to true, but will change to default to false
after the follow-on commit which migrates upstream pigweed away from
SNAKE_CASE. That will have to be done by an upstream pigweed developer.

Original-Bug: b/266298474
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126804
Original-Revision: 0da1c61e72fbaf8db658ae520097be146b1701c1
GitOrigin-RevId: 3092f61033ff937c49921d05f568d619956d95f7
Roller-URL: https://ci.chromium.org/b/8791015896393303489
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I9c683ab22e4f211ed38676f4c79c777151da6d20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793003