[gnsdk] Add _hlcpp suffix alias for fidl libraries
Fidl libraries in the GN SDK currently just use the target name
directly for hlcpp bindings, but the new pattern being used in-tree
is to add a suffix specifying which kind of binding it is. This
change adds an alias for the `_hlcpp` suffix in order to allow OOT
users to switch over before the unified bindings are added.

When the unified binding versions are added, the target without a
suffix will cease to exist.

Bug: 115033
Change-Id: Ic374683da404ed2ff6788684d005105f30b18cb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793421
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>