[cros-es] Migrate core-tests to safer teardown
This CL migrates chromiumos-ec-core-tests to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Bug: 118385
Change-Id: Ied069d63f83d8129da27ad9b5f64581c6550a21c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791472
Reviewed-by: Sarah Pham <smpham@google.com>
Commit-Queue: Gary Bressler <geb@google.com>