[fx] Add prebuilts to $PATH when running commands
To ensure that fx scripts which uses a python3 shebang use the prebuilt
Python included in the checkout instead of the system-installed Python,
which can't be assumed to be a specific version, or even to exist at
all.

This lets us remove a hack from `fx sync-to` that ensured that it always
used the vendored Python prebuilt.

The fx environment cannot be made completely hermetic because many
commands depend on random environment variables, but we can at least
ensure that the $PATH prioritizes prebuilts that are included in the
checkout.

I also merged the `prebuilt.sh` library into `platform.sh` because the
two files overlapped a lot and were used for similar purposes.

Bug: 79905
Bug: 120294
Change-Id: If628bae99dd0fe1e1d4ee1b2fd0bf051a2fc758d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793360
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Shai Barack <shayba@google.com>