[pb] Construct a PB with virtual devices
Product bundles are now constructed with the existing virtual device
specifications. There were some bugs in the virtual device logic that
needed to be addressed.

First, a VDv1 was being parsed then written to a new file but in
the wrong format. The 'data' was being written and the 'schema'  was
stripped. This led to errors parsing the VDv1 files. To fix this, I have
moved the writing logic into the VirtualDevice struct to ensure that it
always gets [de]serialized correctly.

Second, the paths in the VDv1 and the VD manifest had the wrong
relativity. We could consider a refactor that hides all the VD writing
logic away from the product bundle logic, but I would prefer to land
this and get emulators working now, and refactor later.

Tested with:
  fx set core.x64

  # specific device
  fx ffx emu start out/core.x64/obj/build/images/fuchsia/product_bundle
  --device virtual_device_recommended --headless

  # default device
  fx ffx emu start out/core.x64/obj/build/images/fuchsia/product_bundle
  --headless

Bug: 119190
Change-Id: I1e5314e0598b2e1c74b653466704269c0675ac9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793523
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>