[gnsdk] Fix visibility for fidl_library hlcpp alias
There were two problems with visibility for the hlcpp alias
as previously submitted. First was that it didn't have visibility
if an explicit visibility parameter was set on the fidl_library,
the second problem was that the hlcpp target had default visibility
instead of respecting what was set. This fixes both of those problems.

Bug: 115033
Change-Id: I1c38c45c02ccf1dae2f26360b517dda0153ea39b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793914
Commit-Queue: Bryant Chandler <bryantchandler@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>