[rust-3p] Update rustc-demangle to 0.1.21
The mirrored version of rustc-demangle will be used to produce Breakpad
tools with the ability to demangle Rust symbols. However, this requires
a version of the crate which can be called via its C API.

Note, the crate's licenses and dependencies remain unchanged between the
old and new versions.

Bug: 64315

Change-Id: I7ad13bcc83950fbf28222e64476c158b861ba471
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793348
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>