[doc-checker] Forcing line number correction on Headings
This adds correcting the line number being reported in errors
by matching headings to line numbers in the markdown file.

This is unfortunately needed since markdown is "loose" in its
parsing but the output of the parser is "correct". This means
it correctly parses lists without leading newlines, but the parsed
output gives no indication of it being missing.

Change-Id: I79ce9444d758ac2e50a6195ff7608b9d781993fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793141
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>