[block][abr] Add ABR shim driver
Add a driver that binds to the 'misc' partition on Pinecrest and sets
the one-shot recovery flag when rebooting to recovery or bootloader.
Block ops on this device are passed through to the parent until the
suspend hook is called, at which point the ABR logic takes over if
needed.

Bug: 115458
Bug: b/255495213
Change-Id: Icc755f398cb82fe176755a75c5dd83f2743ae218
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791820
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>