[pinecrest][emmc] Make abr-shim bind to misc partition and update paver
Add metadata to the eMMC device to make abr-shim bind to the misc
partition in place of block core. abr-shim intercepts calls to reboot to
recovery and uses libabr to set the one-shot recovery flag.

Update the paver so that it uses the mapped GUID to find the misc
partition instead of the GUID on disk.

Bug: 115458
Bug: b/255495213
Test: dm reboot-bootloader and dm reboot-recovery work on Pinecrest
Test: fx test paver-test
Change-Id: Iedfe3a772c364d008a59503cbde8e7b8fe0126f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791821
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>