[bt][avrcp-tg] Report no song position if playback rate is unavailable
Currently, the playback rate is initialized as a non-null, stopped
playback rate. It is valid for the rate to not exist, in which case the
song position should be None. Store the playback rate as an optional to
account for this.

Test: fx test bt-avrcp-target-tests
Change-Id: I96ed869dd8a7215769ee163eede7681a4224e189
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795512
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>