[sdk][driver][testing] Add testing library
Adds a driver unit testing library to the driver sdk. This unit test
library is composed of the following building blocks:

driver_lifecycle: can be used to start/stop the driver being tested.

test_environment: is used to host the environment that backs the start
args given to the driver. This includes the incoming namespace, outgoing
directory, and fdf node server.

test_node: a mock fdf::Node server to provide the driver.

start_args: a helper to create a driver's start args and return the
channels needed to support it.

Bug: 88576
Tests: Multiple driver_base  unit tests that uses this library in
different threading configurations.

Change-Id: I79eff4c08fa0bf348778acfeb1d5277868562025
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770412
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>