[testrunner][ffx] Set timeout on `target wait`.
The default timeout is 2 min, but the default test timeout is 5 min, so
if we retry `ffx target wait` 3 times, we'll hit the test timeout before
we can exit with a fatal error. We should set the timeout to something
smaller so that we can finish the connection attempts before hitting the
timeout.

Change-Id: I69204b48e1f54ada194d5193f96ce0519c12199f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795723
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>