[arm][gicv3] Fix mask/unmask behavior for SGI/PPIs on GICv3
ARM's GIC series of interrupt controllers defines a couple of types of
interrupts which are always targeted to specific CPUs.  These are SGIs
(software generated interrupts) and PPIs (private peripheral
interrupts) and occupy interrupt vectors [0, 31].  Note that any
instance of one of these interrupts shares a vector number, but there
is a seperate instance of each interrupt for each PE in the system,
and they may be masked/unmasked independently from each other.

One famous and Very Important example of an interrupt such as this is
the CPU timer interrupt which is required to be routed to a PPI by the
ARM ARM.  See section D10.2.4 Timers:

```
If the PE interfaces to a Generic Interrupt Controller (GIC), signals
a Private Peripheral Interrupt (PPI) to that GIC. In a multiprocessor
implementation, each PE must use the same interrupt number for each
timer.
```

GICv2 and before offer a set of distributor registers
(GICD_I[CS]ENABLERn) for setting and clearing the masked/unmasked
state of any interrupt register.  Each register handles 32 vectors,
and register 0 (vectors [0-31]) are required to be "banked" on
multi-core systems.  This means that each CPU has its own copy of
GICD_I[CS]ENABLER0, which control's the masked state of its own SGIs
and PPIs.  CPU X cannot change the state of the vectors for CPU Y (X
!= Y).  From Section 4.3.6 (docs for the clear register):

```
In a multiprocessor implementation, GICD_ICENABLER0 is banked for each
connected processor. This register holds the Clear-enable bits for
interrupts 0-31.
```

GICv3 extends this a bit.  While it still offers control of SGIs and
PPIs from the banked interrupt registers via GICD_ICENABLER0, it warns
that only the first 8 PEs are guaranteed to have banked registers, and
that

```
Accessing GICD_ICENABLER0 from a PE with GICR_TYPER.Processor_Number > 7 is
CONSTRAINED UNPREDICTABLE:
```

Instead, it seems like control of the SGIs and PPIs should be done via
the new "redistributor" registers instead, specifically GICR_ICENABLER0.
There is a separate redistributor instance per PE, getting around the
banking limitation of 8 when accessing this functionality from the
old distributor interface.

So, what is the bug?

The bug is that the GICv3 code was attempting to mask/unmask SGI/PPIs
for a given vector across all PEs using the new redistributor
registers when its `(mask|unmask)_interrupt` implementation was
called.  This can cause issues in many ways, but the specific way this
was manifesting was during shutdown/mexec.

During this sequence, CPU 0 makes a thread, gives it a deadline
profile, sets its affinity to a non-zero CPU and runs it.  The thread
handles the CPU shutdown, and then signals CPU 0 to keep going.  CPU 0
repeats this process for all the remaining CPUs.  When the first CPU
shuts down, it disables its timer interrupts in the process.  When it
comes time to spin the next thread for the next CPU to shut down, if
the thread is not _immediately_ eligible to run (perhaps there is an
even more important deadline thread which needs to run), the scheduler
will need to wake up in a short while in order to manage the switch to
the shutdown thread.

But; the first CPU to shut down disabled its timer interrupt, which
(because of the bug) actually disabled _all_ timer interrupts.  Now,
if the thread currently running on the target CPU fails to voluntarily
yield, it will never shutdown, and the entire shutdown process will
hang.

The Fix:

Align the GICv3 behavior with that of GICv2.  IOW - ensure that when
masking and unmasking SGIs and PPIs, processors only end up affecting
their local instances of these interrupt vectors, and not any others.
Maybe (in the far future) there will be a reason to allow CPUs to
affect other CPUs' SGI/PPI masked states, but if and when that day
comes, we will need to think carefully about how add that
functionality to the existing interrupt controller driver interface.

In the meantime, also add a comment which to mask/unmask which clearly
states that these routines only affect the local processor instance of
an interrupt, in the case of per-processor interrupts which share a
vector number.

Change-Id: I7b36b2417b42fdfca640776345be752cf9596ca4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795671
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: John Grossman <johngro@google.com>