[netemul][test] Add test for interface offline
Add an integration smoke test which ensures that in typical environments
smart displays would be in in production, an interface going offline
then online results in connectivity being restored.

The test has variants for IPv4-only, IPv6-only, and dual-stack
environments. IPv4 connectivity is provided by running a DHCPv4 server
on the gateway realm, and IPv6 connectivity is provided by faking
RA messages.

Note that it's relevant that the interface watcher API is used to
verify that addresses can be found after the interface becomes online
again, as this guards against regressions which cause clients to the
API to be mislead into thinking that connectivity hasn't been restored
when it in fact has.

Fixed: 117869
Test: fx test netstack-smoke-integration-test
Multiply: netstack-smoke-integration-test
Change-Id: I9dad808aa5f7568486fbbe9ad5db73c281ce7a3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790982
Reviewed-by: Amanda Tait <atait@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>