[superproject] Roll gwp-asan [GWP-ASan] Remove thread clamping in tests.
It's better and easier for us to just have threads contend against each
other in the tests if it's more than the maximum supported number of
hardware threads available.

Specifically, the recoverable test fails on Android because the
GTEST_SKIP in a called function, and it only properly works from the
TEST_* harness function. Android tests run on cuttlefish, which can be a
single core with two hyperthreads.

Reviewed By: fmayer

Differential Revision: https://reviews.llvm.org/D143221

GitOrigin-RevId: ff315dccee7d2fedb472a9e459614745bfe10eac
Original-Revision: f36df953a3bab8b8e54e1ce7214cb710a5605a1a
Roller-URL: https://ci.chromium.org/b/8790217354327120705
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I9d7268c11769827a03dd24d12ba708643e2b2a33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796684