[superproject] Roll llvm-project/libc [libc] Add `LIBC_GPU_TEST_ARCHITECTURE` option to set architecture
Currently, the plan is to support testing on a single GPU architecture.
We query the supported architectures from the user's system. However,
there are times when the user would want to override this. This patch
adds the `LIBC_GPU_TEST_ARCHITECTURE` option, which allows users to
specify which GPU architecture to build for.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D143400

GitOrigin-RevId: 025f0f19503a106f614465e4c49db6c6da212628
Original-Revision: 200ed59c651fa8e6111d46ef7f5b9667e548d108
Roller-URL: https://ci.chromium.org/b/8789882799803508513
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id437ecf8fcc7bbf7ca432fafbe51c2ad3fbe4b3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797827