[perfcompare] Remove non-functioning wrapper for running perfcompare_test.py
This wrapper no longer actually runs perfcompare_test.py on Infra,
because FUCHSIA_DIR isn't set when the wrapper is run.  This probably
stopped working when the builder/tester split was implemented in Infra
recipes.

The wrapper was originally added in this change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/305128

It was based on this wrapper:
https://fuchsia-review.googlesource.com/c/fuchsia/+/276943

Change-Id: I7674b8fef15b7a93f4f28f26d52471cc4cb829cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795680
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>