[reachability] Fix criteria for pinging internet
When there has been a neighbor discovered that is not a gateway,
and the gateway is not pingable, check to see if the interface
has any default routes prior to attempting to ping the internet.

There is no reason to attempt to ping the internet without default
routes because we don't expect it to succeed.

Add new test to cover case when we have discovered a neighbor and
have default routes but haven't discovered gateway and cannot ping
the gateway.

Change-Id: I5c0926e1d5b7822dd1147d9e000fe5514713ae0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797486
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>