[fxfs][storage] Add caching of object handles to FUSE-Fxfs
Add an in-memory cache to store the object handle when create/open is called on a file and remove the object handle from cache when release is called on a file. Every cached object handle is stored with a counter to ensure that each release matches a create/open.

Bug: 117461
Change-Id: I4ee9ae8db26a6c464a58c2300ca583ed64d41e5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795486
Reviewed-by: Rupert Shuttleworth <ruperts@google.com>
Commit-Queue: Kongzhang Hao <haokz@google.com>