[check-licenses] Gate license analysis behind a flag.
License analysis happens in CQ.
We don't need to also check them when generating license artifacts.

This CL introduces a new flag: runAnalysis. When true, license pattern
matching and result tests are enabled for the current run.

Currently, this flag is set to true by default. This needs to be
explicitly set to true in static-checks before I can make the default
behavior "false".

This CL also gets rid of the "World" package in favor of a simpler
strategy of using the packages directly.

Change-Id: I02b276e0edd1269877173b021007c6d5409234e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795041
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Yaar Schnitman <yaar@google.com>