[kernel][vm] Simplify traversal return code handling
The ForEveryPageAndContiguousRunInRange helper in turn calls
ForEveryPageAndGapInRange, supplying it with a per-page function and a
gap function. ForEveryPageAndGapInRange is capable of handling
ZX_ERR_STOP in these functions correctly, so we do not need to handle
it in the page and gap functions used here.

Also fixes a bug where contiguous_run_len would not get reset to 0 in
the case of early termination due to ZX_ERR_STOP, invoking
contiguous_run_func on the last range twice. The added test case covers
this.

Test: vmpl_unittest
Run-All-Tests: true
Change-Id: I704339ebed24ce50265c294620d992048b7c3540
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799047
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>