[versioning] Define FUCHSIA_HEAD as a decimal number
When I tried using FUCHSIA_HEAD, the compiler errored with

  ../../sdk/lib/fit/include/lib/fit/task_set.h:49:26: error: expected a version of the form 'major[.minor[.subminor]]'
  class ZX_AVAILABLE_SINCE(FUCHSIA_HEAD) task_set {
                           ^
  gen/zircon/public/sysroot/cpp/include/zircon/availability.h:12:22: note: expanded from macro 'FUCHSIA_HEAD'
  #define FUCHSIA_HEAD 0xffffffff

switching to a decimal number fixes it.
Also removes the mention to fxbug.dev/104513 since the linked bug
is closed with the decision to not upstream to clang.

Change-Id: I82e33df459f5abda02eef330d19eddc1475ce949
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799685
API-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>