[amlogic-display] Fix read_segment_length type
Each element of read_segment_length is the length of a segment that's
returned. According to the EDID spec this can be up to 256, so increase
the type to uin16 to account for that.

Fixed: 121409
Change-Id: I05bb8233a38a71a0e5f76bec24525d4ab300a324
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798070
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yilong Li <liyl@google.com>
API-Review: Christopher Anderson <cja@google.com>