[netstack] Factor out common exposes from shards
Netstack3 and Netstack2 are both used as the system netstack. That means
they need to have a common subset of exposed protocols. Factor that list
out into a single file so that it can be shared between the two, and so
updates can go in one place instead of two.

Fixed: 121408
Change-Id: I1d3c894e05366f48d3a4c9c1849fb341a7c06d94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798044
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>