[netconform] Mark IGMP 3.6 as flaky
We flipped the expectation from Flaky -> Fail once IGMPv3 landed, but it
turns out that this case runs soon enough after an ANVL-initiated IGMP
query that Netstack still runs in compatibility mode. Because of this,
we are still observing the original flaky behavior.

Bug: 104720
we're still seeing it pass occasionally.

Change-Id: I15a280d918e9ebd4571b75f0f129bc63fe0604f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800702
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>