[wlan][policy] don't retry short connections
When a disconnect is detected from SME, the state machine tries once to
connect to the same AP. After this change, the immediate connect
attempt will not happen if the connection uptime was short. Instead
it will fall back to network selection where another AP may be chosen,
and the previous AP might be avoided because of the short uptime.

The reason for this change is that we have seen some AP's repeatedly
disconnect us, and we keep successfully connecting to the AP. We should
go to network selection and choose another AP, which may not do the same
thing.

Test: added unit test
Bug: 120610
Change-Id: Iedd03b9420e032d817a0cd37a0ca41857069008a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800863
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Hayden Nix <haydennix@google.com>