[starnix] Escape file names in /proc/../maps
Newlines in the names of mapped files should be replaced by "\012" in
maps output so that the contents can be parsed unambiguously.

This change is mostly refactoring tests to permit running tests for
/proc behavior both on Starnix where /proc is not always mounted and on
the host where /proc is mounted and the tests are usually run without
root. This introduces a ProcTest fixture that will set up a proc
filesystem somewhere if needed.

Change-Id: I7975c9fd0251e7505a322556af9e8a57f6874df8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801709
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>