[bt][avrcp] Fix packet decode issues for media player feature bitmask.
The issue was occurring due to incorrect endian-ness interpretation
between the rust bits and decoded feature bitmask number.
AVRCP packets transfer multiple-octet fields in big endian. The lowest
octet/storage address feature flags are therefore the big end of the
feature number that's transferred. The FIDL bits is modified so that the
feature bitmask can be correctly interpreted into the feature flags. We
make the changes on the FIDL side and not the packet decoding side,
since it's
easier to compare the values against the spec if we keep the decoding to
match the endianness that's specified in the spec.

Test: fx test -v -o bt-avrcp-tests
Bug: 108372
Change-Id: I2e886e7d68d67c4e4016b39669b4b1937ec0142a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797862
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
API-Review: Marie Janssen <jamuraa@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>