[dlog] Remove dlog_force_panic.
There are two separate and orthogonal ways of representing a panic
within dlog. The first one consists of simply bypassing it, by
overriding the boot option. The overall effect is that printfs
will not attempt to write to dlog if we are bypassing it.

* dlog_force_panic(void).

The second form of panic, consists on a flag that is actually set in the
panic path. This flags causes writes to dlog to bail out early with
'ZX_ERR_BAD_STATE'. In this case, printf will not bypass dlog, but will
attempt to write it, unsuccessful of course. But because it failed,
it will continue and attempt to write directly to the console and
serial.

* dlog_panic_start(void).

The only user of dlog_force_panic is k wtd force, which relies
on this method for writes to stdout to go straight to the uart.

This is still the case, and is closer to what panic actually
does.

Bug: 89182
Test: k wtd force
Change-Id: Id49299ef9312c97c7e0d40c36fb55c2bb3d3c6c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799026
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>