Revert "Revert "[realm_builder] garbage collect manifests and local components""
This reverts commit 9f49e810e37f37dd056b6bca8f60dcd057493b44.

The original commit was reverted due to causing flakes in tests that did
not keep the `ComponentRunner` channel provided in the `Builder.Build`
call alive for the duration of the realm's lifetime.

There have been no significant changes to the approach taken in the
initial commit. The only way to identify which tests do not keep this
channel alive long enough is to land this and fix-forward any tests that
begin flaking due to this.

Fixed: 121196
Fixed: 121808
Run-All-Tests: true
Change-Id: I0f5d3c7c144c5461cf24fc3a95cc0c1410a4e4b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802709
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>