[starnix] Add ARM64 stat_t and trampoline file.
Define stat_t for ARM64 which is slightly different than x64. Change the
definition of the dev_t type in uapi.rs to be u64. Previously it was
__kernel_old_dev_t.

The generated uapi bindings don't capture the dev_t definition so I
assume __kernel_old_dev_t was used instead. But the structure is defined
using dev_t and not the "old" one. __kernel_old_dev_t is 64 bits on x64
so it matches dev_t, but it ends up being 32 bits on ARM64 which is
incorrect.

Adds a stub file for the restricted_enter() and restricted_return()
functions so it can link.

These functions are left unimplemented so that they can be tested at the
same time they are written. This patch adds detailed implementation
notes.

Bug: 121659

Change-Id: I63d22aaf5b8d72d63a3981b1456864f0049e7aec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801233
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>