[stanix] Explicitly drop Task on exception
Previously the task reference that was in the syscall dispatch loop
would not be decremented if the thread was killed in the exception
handling path.

This is a temporary hack that needs to be in place until restricted
mode exposes exceptions to normal mode.

This change explicitly decrements the Task reference count when the task
is killed due to an unhandled exception.

Change-Id: I5b526da35020725e9508c074f301a0eaf87d4210
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806347
Reviewed-by: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>