[image_format] Round up width for caller.
Existing callers sometimes pass a width value that hasn't yet been
rounded up to the next coded_width_divisor alignment, so instead of
returning failure in that case, round up the width for the caller.

If the caller is a producer, the caller is still expected to indicate
a width to other participants that's divisible by coded_width_divisor.

BUG: 269323365
Change-Id: Ibe7b225163c8665c9b4a93961a669cad27f35c3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807987
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>