[ramdevice_client] Add method to wait for removal
Add a method to the rust RamdiskClient that unbinds the ramdisk and
waits for the device and all of its descendants to be removed.

Only calling `schedule_unbind` in `destroy` leaves a small window where
the child block device can still be used after `destroy` returns. The
storage stress tests want to wait for that window to be over so
filesystems can be shutdown without being able to flush anything. The
current method of waiting for the directory connection to be closed
doesn't work because directory connections are not closed when a
directory is unlinked.

Change-Id: Idde375a7fe93cc5f187646a18b852ef521f98970
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/808891
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: Chris Tam <godtamit@google.com>