[flatland][display] Release BufferCollection when a test ends.
fuchsia.display.Controller FIDL interface requires that the
collection ID in ImportBufferCollection() must not be in use,
so in order to reuse a buffer collection ID, the old collection
must be released first.

Once fxrev.dev/803773, fxrev.dev/803774 and fxrev.dev/803775 are
landed, importing buffer collection with an ID in use will be
disallowed for fake display.

Flatland display tests and pixel tests use an injected display
controller (which can be either real or fake), and the controller
is reused across different test cases. If we don't release
imported buffer collection at the end of a test case, the buffer
collection will be still alive in the display driver and the
collection ID will be occupied in the following test cases, which
would cause a test failure.

This CL adds deferred callbacks to release buffer collections
which run at the end of the test, to all the test cases that
import buffer collections to display controller to solve this issue.

Test: flatland_display_compositor_pixeltests
Test: flatland_display_compositor_pixeltests_with_fake_display
Test: flatland_display_compositor_smoketests
Test: flatland_renderer_unittests
All the tests above are tested with and without the CL chain listed in
fxrev.dev/803775.
Bug: 122131

Change-Id: Icf566348e977ad163a3115e413fd5314fb2bb175
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/805348
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Yilong Li <liyl@google.com>