[starnix] Don't decommit brk memory.
This CL changes the behavior in Starnix's memory manager so that it
doesn't issue `ZX_VMO_OP_DECOMMIT` operations on its brk memory, as
those are not allowed on child vmos. Instead, it now issues
`ZX_VMO_ZERO` operations, that are deduplicated with the zero page.

Test: fx test syscalls_test
Fixed: 122447
Change-Id: Ib170dd1580772f80f8b7b331d05388cbc82437d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809388
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>