[testsharder] Partition by environment name
Testsharder was partitioning envs into shards considering
the values in the dimensions set, however when creating
the name of the shard not all values are considered, leading
to different shards having the same name causing some builders
to fail.
This CL fix that.
Bug: 122468

Change-Id: If6c9bf0cf32598b9e2acd492ae44959d42c09c72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810152
Commit-Queue: Andrés Villegas <andresvi@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>