[ffx test] Remove suite_start_timeout_seconds timeout
This timeout was added to address hangs due to ffx stability bugs. If
ffx lost connection to the device, the test used to hang. This timeout
has unintended consequences, as on device processing can take long time
for large packages and is unreliable. So remove the timeout.

In future if we see ffx stability issues, we should add a heartbeat API
to test manager and periodically use that to timeout on connection loss,
or overnet should do it internally.

BUG=122105
BUG=122144

Change-Id: Ie27a5adafd2455cb62dcb7ed232ee79c0b9f7cef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/805889
Commit-Queue: Ankur Mittal <anmittal@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>