[network] Add an AIB for the netstack package
Move the selection of the netstack package out of the legacy mechanism
for the minimal product and into product assembly. This will enable the
choice of netstack to be configurable at assembly time, though for now
it is a no-op.

Bug: 85450, 121811
Change-Id: Ib8d6e0beeed7bc1decd10b08c68d5f32c3209a14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807750
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>