[fxfs] Accounting for file size changes in project usage.
Tracks changes to match with updates to allocated_size on files. Avoids
locking issues by adding a ProjectUsageDelta mutation to make atomic the
changes to the size in the journal.

Adds more comprehensive fsck and test for project id related data.

There is a change in the serialization, but adding new enum options are
"free" so no migration types needed. Related version change in
https://fuchsia-review.googlesource.com/c/cobalt-registry/+/791171

Bug: 114744

Change-Id: I500e5865650e4f8c2a0ddc8d760e10efccd6bbe0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791959
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>