[flatbuffers] Update to a56f9ec50e908362e20254fcef28e62a2f148d91
BUILD.gn and flatbuffer.gni are adapted and copied from Chromium
respectively. The latter retains Chromium's license header because it is
used unmodified.

https://source.chromium.org/chromium/chromium/src/+/main:third_party/flatbuffers/

Remove the non-zero default value of Quat as flatc now produces:
  src/ui/scenic/lib/gfx/snapshot/snapshot.fbs:15: 14: error: default values are not supported for struct fields, table fields, or in structs.

This default value appears not to be load bearing:
  $ git grep snapshot::Quat
  src/ui/scenic/lib/gfx/snapshot/serializer.h:  snapshot::Quat rotation = {0.0, 0.0, 0.0, 1.0};
  src/ui/scenic/lib/gfx/snapshot/snapshotter.cc:    transform->rotation = snapshot::Quat(rotation.x, rotation.y, rotation.z, rotation.w);

Replace the deprecated `flatbuffers::Vector::Length` with
`flatbuffers::Vector::size`.

The UBSAN suppression is removed.

Fixed: 46773
Change-Id: Ia43cbe78f5aab80d51413e0b16528c747c6a7f56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813611
Reviewed-by: Emircan Uysaler <emircan@google.com>