[fshost] add zbi ramdisk in fvm_ramdisk tests
In tests most tests which use the fvm_ramdisk flag, the tested behavior
would have a zbi ramdisk launched in the real version. Update the tests
so they also provide a zbi ramdisk with a blobfs and potential data
volume, so fshost will set it up like normal.

To do this, we stop using the ramdisk_prefix to determine the zbi
ramdisk when matching. Instead, we collect the ramdisk topological path
when we launch the ramdisk and use the specific path during matching.
This tightens the behavior around "ramdisk prefix" matching to only
match against the launched zbi ramdisk. We also only launch the ramdisk
when configured with fvm_ramdisk, instead of doing it any time it's
available in boot items.

This gap has caused bugs in the past, and in fact adding coverage
catches two existing bugs, which are also fixed here.

- wipe_storage returns a handle to a fresh blobfs instance, with the
  intention that it can be written to, but it tried to launch that
  blobfs instance using the static blobfs child, which in reality would
  be actively used by the ramdisk blobfs. This change makes it launch a
  dynamic blobfs component instead.
- write_data_file, when being used with fxfs, would fail to start its
  fxfs because the crypt unlocking function always bound the unencrypted
  volume to a static namespace path and never unbound it, in effect
  limiting the number of fxfs instances fshost could launch to one when
  using the new encryption scheme. This changes the crypt functions to
  unbind the path when they are done with it.

The second bug fix also uncovered a bit of latent tech debt, where
shred_volume relies on the unencrypted volume to continue to be bound in
the namespace so it can delete the keys. This should be fixed somehow,
but is left for future work - for now we just keep the "main" fxfs
binding around in a special spot.

Fixed: 112163
Fixed: 112861
Fixed: 120377
Fixed: 122960
Bug: 122966
Test: fx test fshost_integration_tests
Change-Id: I68f83d3cde6a428dd45f758baf2f31bb4842a114
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813790
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>