[kernel][vm] Page list support for intervals
Add support to install zero range intervals in the VmPageList. New
intervals are only created if they cannot be coalesced with existing
intervals to the left or right. The ForEveryPageAndGap traversal helper
is altered to not treat intervals as gaps, but as pages.

No usage of intervals outside of tests yet.

Test: vmpl_unittest
Bug: 122842
Run-All-Tests: true
Change-Id: I99cad5f26dd17f8284a5258fe302fb529110eef4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794342
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>