[kernel][vm] Explicit paged property
Makes is_paged a property stored in the VmObject instead of a virtual
method. The VmObject has unused padding such that this does not
increase the size of the object.

Using a stateful property provides a potential performance improvement
as well as smaller code generation. These benefits are not realized at
the moment as is_paged is not presently on any critical path of a
benchmark.

Change-Id: I501a187a9b95c69bc851c32c441e42a1a387b30d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795484
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>