[web_runner_tests] Use hermetic debug netstack
Hermetic netstack to not be affect and be affected by the rest of
the (non-test) system and its netstack.

Debug netstack to debug a flake in web runner tests where it is unable
to open a socket.

Bug: 122169
Change-Id: I041a1442fbce3d84f4321740f3263a68bca447c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815575
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>