[netstack3] Introduce TX device queue
This change introduces a queue that sits between the stack and the
device to hold packets while the device is not ready to transmit
buffers. The queue is implemented fully in core and the bindings
is simply expected to support waking up a TX task which is
responsible for kicking the core to send packets to the device.

Note that this change changes the semantics of
`DeviceLayerEventDispatcher::send_frame` to always accept a buffer
and not a `Serializer`. This is because the queue will hold serialized
buffers for the device. Future changes (see https://fxbug.dev/105927)
will leverage a custom allocator to serialize frames directly into a
device's data VMO, reducing data copies in the TX path.

Later changes will update the implementation of
`DeviceLayerEventDispatcher::send_frame` to not drop frames when TX
buffers are unavailable. See https://fxbug.dev/105921.

Note that we do not yet use the TX device queues. The relevant changes
to bindings will come in a follow-up change.

Bug: 105615, 105921, 105927
Change-Id: I90325f662ee7e306b68daaa8f0c72dc1ee98e9f8

test

Change-Id: Ic05198922712195e5607a8d2b19135f5abeeb346
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814227
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>