[rfc] Package Sets
This RFC provides an official definition for the term "Package Set" and proposes
an updated framework for package sets. This framework expresses packages sets
in terms of the availability and version control they provide.

This was previously in work as fxrev.dev/761803 but as an external contributor
I'm not able to add new patchsets to that CL.

Bug: 110948
Change-Id: I121368f4677886d343eef2df3c973466b2b88099
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794079
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Marvin Paul <marvinpaul@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Seth Ladd <sethladd@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>