[fxtest] Use incremental publishing by default
Integrating the incremental publisher into `fx test` has been
challenging, since it runs in parallel with the rest of the build. That
led to races between the two, and caused issues where some packages were
not published.

To fix the race, we changed `fx test` to also trigger a publish after we
did a build, to make sure all the packages we built actually end up in
the repository. This ends up removing the need for having a file-watcher
based publisher.

Because of this, this patch removes the need to have the background
incremental publisher running, by adding a feature flag specifically for
`fx test` to opt into this new workflow.

Bug: 122023
Fixed: 123180
Change-Id: I81ccb601ec9bcbac7ab621533415cc1c66fd55e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815570
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>