[ffx][emu] Reland Removing manual target manipulation
This is a reland of fxr/810447. It was rolled back because
of failures in the Flutter Engine and Chromium projects.

The root cause was the waiting for the instance to start fully was relying on the old logic handling ephemeral targets had a 1 second delay built in. Removing this made the loop iterate for N times where N is the number of seconds in the timeout.

This change adds using a timer to keep track of the elapsed time explicitly.

After this change, the chromium presubmit test passed successfully.

Change-Id: I0f6a2df49fdd43f8aa91e8c9f35bedfb4c8f3776
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814450
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>