[device-watcher] Delete recursive_wait_and_open_node()
Instead of using recursive_wait_and_open_node() to get a NodeProxy and
then using the NodeProxy to open a different type of connection, we use
recursive_wait_and_open<P> to get the type of connection we want. In
the cases where we don't need the Proxy, we call recursive_wait() instead.

Bug: 120890
Change-Id: I4516471a73cdf9813c85f6d29ac5c1984bec7c04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814148
Commit-Queue: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>