[fxtest] Build `updates` if package not in publish list
The incremental publisher uses the `all_package_manifests.list` to know
which packages to incrementally publish. That file is created by `fx
build updates`, so it can get out of date if the user has changed the
build topology.

This changes `fx test` to check if the test we want to run is in the
`all_package_manifests.list`. If not, that file is probably out of date.
When that happens, run `fx build updates` so we know what to publish.

Fixed: 123398
Change-Id: Ia6790dd536d7ceaf69febb7659a0a8d42d638178
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817251
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>