[kernel][scheduler] Remove the leaf-lock annotation from queue_lock_
Remove the leaf-lock annotation from the scheduler's queue_lock_,
because it is not (strictly speaking) a leaf-lock.

A leaf-lock is defined as a lock which is only held at the very edge
of the lock DAG for a program (a leaf node).  In other words, when
code holds a leaf-lock, it will never attempt to obtain any other
lock.

lockdep allows users to annotate locks as leaf locks, and will enforce
checks of the invariant (above) when enabled.  If the invariant is
violated, it is considered a fatal violation and lockdep (in the
kernel) will panic/complain.

The queue_lock_ is _almost_ a leaf lock.  There is at least one place
where it actually isn't.  Specifically, in a method called
`DumpThreadLocked`.  `DumpThreadLocked` is called from exactly one
place; the LockupDetector.  When the LockupDetector is unhappy, and
things that a CPU has gone out to lunch, it starts to complain.

It will print a bunch of information about the apparently stalled CPU,
starting with the basics, like "Which CPU?" and "How long has it been
since the CPU has checked in".  Eventually, it will go to the
scheduler for the CPU in question, and ask it to dump information
about the scheduler's currently running thread (as well as the RunQ)
to the FILE* target passed by the LockupDetector.

The scheduler needs the Running/RunQ state state to hold still while
this printing happens, so it needs to hold the `queue_lock_`.  It
obtains the `queue_lock_`, and then starts `fprintf`ing to the target
passed by the LockupDetector.

Unfortunately, most FILE* targets require that we grab other locks in
order to print the information.  If lockdep is enabled when we attempt
to grab one of the locks required to print, it will declare a fatal
violation (Attempting to obtain a lock while holding a leaf lock), and
it will want to tell us all about it.  So, it tries to put the system
into a panic state and print out some info.  Printing tries to obtain
a lock, which makes lockdep angry, and it tries to tell us about it.
And so on...  Eventually, the infinite recursion takes us to a very
bad place, which (on x64 at least) results in a triple fault and the
system rebooting without any useful debug info having been printed.

In theory, this same mechanism should cause a similar failure in the
case of a failed ASSERT while the queue lock is held.  Most folks are
not running with lockdep enabled, however, so they would probably
never see it.

For now, remove the leaf-lock annotation on the `queue_lock_`.  If we
ever get back to the point where we can consider it to be an actual
leaf-lock, we can add it back in again.

Bug: 119371
Change-Id: Iae5167018b5615dd8811b1502853ac8d534fc54b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817254
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>