[benchmarks] Avoid using the SL4F server for tests that don't need it
Avoid starting the SL4F server for the simple test cases that just use
runTestComponent().

The benefits are described in the comment in helpers.dart.

Change runTestComponent() to be a function instead of a method so that
it can instantiate PerfTestHelper without starting the SL4F server.
Its use now involves a little less boilerplate code.

Bug: 122165
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/microbenchmarks_test
  with QEMU started without using TAP networking (with "--net user"):
    ffx emu start --net user --console --headless
Change-Id: Ic1de09bc60925e6ef284c2eda404a974a8c9bec5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813867
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>