[a11y] Instantiates Gesturehandler inside GestureManagerV2.
This change is part of the migration of a11y services to the new gesture disambiguation protocols.

The code here is just a copy of the original Gesturemanager, instantiating a different data structure (GesturehandlerV2).

This is necessary in order to support both gd protocols for a while, as we do the migration.

Bug: [a11y] Adds Gesture Handler v2

This change is part of the migration of a11y services to the new gesture disambiguation protocols.

This change is just a copy of the old Gesturehandler, changing a few data types to match the new expected format (mostly GestureContext).

Once the migration is done, this will be renamed to be back the Gesturehandler, and exists as it is for now so that we can support both gd protocols for a while.

Bug: 100030
Change-Id: Ia3823c6e1fd8d60557ad19c42386aac2d51935db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801597
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>