[a11y] Accessibility view instantiates a touch source.
This change is part of the a11y migration to use the new gesture
disambiguation protocols.

This change instantiates, as part of the GFX | Flatland Accessibility
View creation a touch source, which is immediately upgraded to an
augmented touch source.

In order to support both the new API and the old one for a11y pointer
events, this changes introduces the following notion: - a11y view
continues to be created only once; - A touch source is created together
with the a11y view; - The touch source can be taken from the a11y view,
and can be returned to it when it is no longer being used; 1. This is
necessary, because the old API creates / destroys gesture arenas as
needed; 2. Once we get rid off the old API, the new gesture arena will
only take ownership of the touch source and keep it.

Bug: 100030
Test: fx test a11y-lib-tests
Change-Id: Iff801a9cbb98bec4715b9636a8efde09b790f098
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801598
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>