[media/ffmpeg] remove sbc encode and vp9 decode from profiles
This CL removes the sbc encoder and the vp9 decoder from the max and
default profiles for ffmpeg. The sbc encoder was added in error, and
the vp9 decoder was previously in x64 builds only and is not needed.

Test: build issue (used in ffmpeg roll)
Change-Id: I0d60eef895c0bfbd2af2758a7f3e2fea705fb241
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817552
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>