[fvdl] Removing dependency on package_archive
The use of //build/gn:package_archive as an input to the test
is causing the creation of that archive on effectively all
incremental builds. This is a 40 second operation with a repository
of this size (this is an operation that doesn't scale with the
size of the incremental build, it's a constant-sized operation based
on the size of the overall TUF repo (base + cache + universe)

The package repository was not actually used in the test, so it is just
removed, and an empty directory is created instead.

The tool `fvdl` is being actively deprecated in favor of `ffx emu`
so the test is not actually being modified to improve coverage of
successfully starting package serving. This will be covered in
`ffx emu` e2e tests which should be written in the near future.

Bug: 123378
Change-Id: I28bf9130d9109b92302465c6c23613acbd1cfdec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817851
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>