[ffx][repo] Return target communication failure if target offline
If ffx loses connection to a target right when we're trying to register
a repository, we might get a `None` for the target nodename or ssh
address if ffx has only partially torn down the connection to the device.
If we see this happen, return an `TargetCommunicationFailures` instead
of `InternalError`, since this doesn't appear to be an internal bug.

Change-Id: Ibfb461fc126956c0c393a87169eafb4020d22d44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819187
Reviewed-by: Andrew Pollack <andrewpollack@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>