[pkgdir] Refactor tests to use a test realm proxy
- Replace test.fidl.pkg.Harness with fuchsia.testing.harness.RealmProxy
  which will eventually be partner-facing in the SDK.
- Reimplement the pkg-harness component as a test realm proxy.

The new test realm proxy doesn't support switching between the
"backing" of the package directory; it looks like this feature is
not required by the tests anymore anyway. If it is needed in the
future it can be accomplished by subpacking the test with a
different proxy implementation that uses the new backing, or a
test harness that spawns RealmProxy instances given some object
specifying which package_directory backing to use.

Bug: 121395
Change-Id: I928a0cbea5984707ff672bc0693f198c39ed330a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794462
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Kendal Harland <kjharland@google.com>