[kernel][phys] Plumb arch-specific data from physload
This moves some state from per-arch global variables into a
per-arch `struct ArchPhys`. Each ArchSetUp() implementation is
responsible for setting `gArchPhys` to point at an initialized
instance of this (if it's used at all, not so far on x86).  This
is plumbed through physload handoff to set `gArchPhys` in the
physload module to point to the one set up earlier.

On arm64, this includes the PSCI information which heretofore was
being lost on physload handoff so that physboot's uses of PSCI
didn't have the state set up properly. (This only affected the
panic path. ArchPhysHandoff doesn't use these globals.)

On riscv64, this provides a way to plumb the boot hart ID,
which is now included in ArchPhysHandoff.

Bug: 122871
Change-Id: I6cf74a8297633fcd190258f5a4348bbfeb606fb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819530
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>