[ffx] Clean up host ids in RCS
When a new connection comes up, the remote-control-runner calls
AddHost (legacy) or AddOvernetLink (CSO), adding a new u64 id which
gets appended to a list. After 8K of these or so, we can no
longer respond to IdentifyHost calls because the DISLmessage is
too big.

The fix is to have each client loop maintain an Rc to each of its
ids, and the handler only send back the ids that have not been dropped.
So when the RCR exits, the loop will exit and we will no longer report
its ids.

Fixed: fxb/123624
Change-Id: I9aff14069de0d222e19020e196892b3e9c3bbf4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819297
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>