[netstack/bench] Report lost packet %
Include the percentage of packets that were lost (as opposed to just the
total count) in the results of the iPerf benchmarks. Also, mark the
"lost_packets" metrics as "smaller is better" so that decreases are
correctly detected as improvements and vice versa.

Fixed: 123437
Test: catapult_converter_unittest
Change-Id: I84f51d8870de502f23fcf293c2993186dcf85d9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819788
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Tony Gong <gongt@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>