[rust] Add #![allow(clippy::redundant_async_block)]
This CL adds attributes to allow the new redundant_async_block clippy
lint to unblock the toolchain roll. It only adds it to affected crates.
It also adds #![allow(unknown_lints)] which can be removed as soon as
the toolchain rolls.

We are not fixing the lints right now because there are false positives:

https://github.com/rust-lang/rust-clippy/issues/10482
https://github.com/rust-lang/rust-clippy/issues/10509

Once those are resolved, we can remove these attributes and fix the
occurrences or add more narrowly scoped attributes.

Test: fx clippy --all --raw # tested with old and new toolchain
Bug: 123528
Bug: 123778
Change-Id: If874cae8621b9469b9fde0cb6220250710c8848c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820449
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>