[overnet] Fix race in ServiceMap
ServiceMap was trying to be clever and have different locks for
different data, but the data had to be kept in sync and it wasn't quite
clever enough to pull that part off. This removes a bunch of the
interior locks and replaces them with one big lock around everything.

A side effect of this is that a couple of functions that didn't have to
take that lock now do, so they become async.

Fixed: 110501
Fixed: 122211
Fixed: 113099
Fixed: 98883
Change-Id: Ib8e2d7b305d5450819d804fce409905d29d98e0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821128
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>