[ctf] Use subpackaged runner in fuchsia.ui test
Uses the subpackaged rust_test_runner in the fuchsia.ui.composition
CTF tests to ensure that the runner is always compatible with the test.
This CL also releases the test in CTF now that the test runner is
bundled with the test.

Fixed: 118010

Change-Id: I00f464c407caa8d458447062cf048c40a2b5e349
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/808327
Fuchsia-Auto-Submit: John Shamoon <johnshamoon@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kendal Harland <kjharland@google.com>