[starnix] Handle forking of writable pager backed VMOs
Prior to this CL, we'd always clone if the VMO was pager backed, but
that's the wrong thing to do if the VMO was writable.

Unfortunately, we can't (yet) snapshot pager-backed VMOs, so we
workaround that (for now) by copying the VMO.

This should have limited impact currently because there should be few,
if any, cases where we use pager-backed VMOs, but we will have them
soon.

Bug: 123742
Change-Id: I6fb7d3486011a01c46b46edcf208e05dd73341cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820331
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Robinson <jamesr@google.com>