[netstack3] Drop or use forgotten struct's fields
`Inner` is forgotten (`core::mem::forget`) so the destructor for its
fields will not be run. This is okay for primitive types but not for
types that may have custom destructors.

Before this change, `Inner::marked_for_destruction`'s destructor was
not explicitly called since it was an `AtomicBool` which does not do
any special work when it is dropped. However, to be safe and guard
against future changes that _may_ change `marked_for_destruction` to
a type that _does_ perform work on destruction, this change explicitly
drops `marked_for_destruction` when it is no longer used.

Change-Id: I42bc63c9081e84577debb100539251b500d11fa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821876
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>