[devshell][ffx] Build ffx subtools as fx host tools
This makes it so that if you have
`ffx_build_dual_mode_plugins_as_subtools` set in your build, if fx is
included in your build so will the subtools that are supposed to be
built. This gets `fx build` building them as it does with ffx.

Fixed: 123543
Change-Id: I3e002b04135c48da535c178d0a103f274f4d3e32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819188
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>