[timekeeper] Makes timesource child destruction conditional
`Realm::DestroyChild` isn't quite idempotent: it produces log spam
if we attempt to destroy a child that does not exist.

The old time source code would attempt to get to a 'clean slate' by
trying to destroy the time source component that it's about to start
first. It did so unconditionally, creating log spam in the process.

The new code first checks if there is a matching child component to
destroy. This is not error-proof, since the child component could
still go away between the time we check if it is there, and the
time we destroy it. But this should either be rare, or would point
to a bug.

Bug: 124043
Change-Id: Id3c8e7f3a1df453bb3d50cadc871be500504ebdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822778
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>