[starnix] Don't eagerly copy the whole VMO
It was a bad assumption that most of the VMO would be mapped. It turns
out that the mappings are often small compared to the size of the VMO,
so copying the whole VMO meant a performance and memory burden.

This change now means that we only copy the parts that are mapped.

Bug: 123742
Change-Id: I2c1ef6c1cca0c77a767de4134080251486cb377d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822478
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>