[fusb302][vim3] Add class responsible for conveying Sink power policy.
This CL adds a class that contains the logic for translating between a
power policy and low-level USB PD details. The implementation is fairly
specific to the VIM3 board that we're currently targeting, but the API
and testing approach should allow extending this to other cases we want
to support.

Upcoming CLs will plumb this class into the USB PD Sink Policy Engine
state machine implementation.

Bug: 122113
Test: fx test fusb302-test (no functional changes)
Change-Id: I26ab1679568ddafff5838002ee86cec4de928836
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823330
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Victor Costan <costan@google.com>