[sysmem] Do not clear NodeProperties::parent_.
NodeProperties::parent_ field tracks the Node on the tree,
and it is used to propagate the node counters to the Node's
ancestors when NodeProperties is destroyed. If parent_ is
cleared before its destructor is called, the node counter
will not work correctly, causing LogicalBufferCollection
leak in sysmem driver.

Test: flatland_pixeltest
      1. Verified in "fx shell memgraph -v" that there's no leaking
         Flatland VMOs.
      2. Printed debug logs (SetVerboseLogging()) and verified that
         the LogicalBufferCollection root has been destroyed.
Fixed: 124201
Multiply: flatland_pixeltest: 10
Change-Id: I81e1935837c179aec7a8ef5c5102cea638ecf332
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823977
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Yilong Li <liyl@google.com>