[rust-3p] Update polling to 2.4.0
This updates pollng to 2.4.0, the latest version that doesn't depend
on a more recent version of windows-sys. As part of this, polling switched
to a different crate for wepoll, so this switches our empty version of
wepoll-sys to wepoll-ffi.

Bug: 123569
Bug: 120661
Change-Id: I2e004cccba35e9cb9a988c58854a39fb2474435d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822968
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>