[gn_desc] Move file parsing into optimized library
This adds a function to the always-optimized gn_json rlib
so that all users can easily get the performance of a
release compilation of serde, without having to enable that
on their own binaries.

Tested by temporarily setting the binary back to not being
optimized, and verifying that there wasn't a performance
regression in a simple query to summarize a single label.

Change-Id: Ie428b71503acff5d01b4c0f55e4cce305863e285
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825072
Reviewed-by: David Koloski <dkoloski@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>