[a11y] Update responses only for held interactions.
This change ensures that a11y only tries to update a response of an
interaction if we previously put that interaction on hold.

Because of the nature of async code, it can be the case that an
interaction inside a11y becomes resolved before we send the initial
response to it, which caused an UpdateResponse call, trying to update a
value that was not committed.

Existing tests already cover this case, but can only be triggered in
specific conditions, so this change fixes potential flakes.

Change-Id: I9e3a5ad3872169eeb6dfac845180e7d8d0c1d92b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825069
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>