[arm64][crashlog] Try to dump memory around ELR when crashing.
Implement the feature request tracked in fxb/121735.  To summarize,
when we hit certain types of exceptions (esp. Unknown Exception), it
would be good to know (if possible) what the .text around the
exception link register address looked like.  This would help is to
either rule out or confirm that the exception might have been taken as
a result of a possibly corrupted instruction.

Every attempt is made to make sure that the memory should be readable
before we attempt to fetch, however as an added safety, there is a
chicken bit which is set and should suppress the feature in the case
that even attempting to read what should be valid memory results in a
double fault.

Fixed: 121735
Change-Id: I6c4b0673e623dc5b611c77a7d5ffef4d4bb9ad6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821178
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>