[ffx][emu] Using --length-is-lower-bound when resizing fvm
When launching the emulator, the product bundle metadata is read to
get the desired size of the fvm file. In some cases, such as building
the ASAN variant, the fvm is sized larger than specified in the
metadata. This would cause an error like

`Error resizing fvm: |length|(2147483648) must be greater or equal than
|disk_size|(3154280448 bytes)`

This change adds the -length_is_lower_bound flag to the fvm command
line so it will not be an error in this situation.

Long term, the meta data in the product should match the actual size
of the fvm.

Bug: 102866
Change-Id: I0aab3c9295f58eedcab44a796c52fe2df2402c05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825222
Reviewed-by: Dwayne Slater <dwayneslater@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>