[ui][tests] Use a hermetic font provider for web_runner_pixel_test
Instead of a config-data package, this change configures the
web_runner_pixel_test to use a hermetic font server, which has
all the fonts it needs bundled in the package.

This allows us to stop using config-data, an approach that should be
obsolete in a CFv2 world.

Bug: 123585
Change-Id: I3ee9bf20135b697af829adcd36eb57e1f17dc034
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825593
Reviewed-by: David Dorwin <ddorwin@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>