[ffx] Stop guessing if the current_exe is main ffx
The well-meaning attempt to figure out if `rerun_bin` was being run from
ffx or a subtool already had sharp edges, but since we distribute ffx as
a standalone binary called `ffx-<arch>` it causes very confusing errors
for certain users through no fault of their own.

Generally we know what kind of binary we're running from, though, so
just use that knowledge and add an extra field to `EnvironmentContext`
so it knows if the current binary is a 'real' ffx binary or not.

Fixed: 124387
Change-Id: I286d1a63d9e9e60ee1e8b65ea41d364cc74f30fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825818
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>