[fs_management] Address a race in find_partition_in.
In some rare case such as re-partitioning, a device may appear
(generating a `WatchEvent`) and then be removed (`Volume.Destroy`)
before the watcher loop in `find_partition_in` has a chance to
process the event. In such cases, we will fail to get the topological
path of the device.

The correct behaviour here is to ignore such failures and keep watching.

Change-Id: I6a2ef8a5280b46d5a51a9d14209a0d81b71ad84c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825003
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: Chris Suter <csuter@google.com>