[i915][audio][hdmi] Allow TigerLake HDMI HDA to be seen, probed, and configured.
This contains changes to the Intel i915 HDMI audio driver
which allow the HDA audio to be recognized and configured as an audio
stream on TigerLake.  There is still no audio output, but the monitor ELD is recognized and the audio stream is published.  Note that the TigerLake support is disabled by default, and needs a one-line
developer change to enable the support.

Tested: Verified (on NUC11TNHv5) that audio-driver-ctl info returns the monitor's audio info if enabled, and that it returns the appropriate error if not enabled.
Change-Id: I2d9bfd59f4ce7c65c4e47b09ee8349d1dcc99f69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824388
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Jack Brennen <jbrennen@google.com>