[netstack][zxio] Set dgram write threshold to UDP max
Currently, clients sending data using Fast UDP can end up busy-looping
when using blocking sockets or waiting for the socket to become writable
with an API like poll. This occurs when the zircon socket is not
completely full (and is therefore writable) but does not have enough
capacity to fit the outgoing payload.

Change zxio's implementation of datagram_socket::wait_{begin,end} to
internally use ZX_SOCKET_WRITE_THRESHOLD instead of ZX_SOCKET_WRITABLE,
and in the netstack, set the socket's write threshold to the maximum
size of a UDP payload + the size of the TX metadata. This ensures that
when a client waits for a socket to become writable, the next write will
succeed as long as the payload size is within the maximum.

Fixed: 123626
Test: fdio-zxtest, zxio-unittest-package
Change-Id: Ifaf562244c736dcfb099a03c42f18731bb8c6323
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821907
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tony Gong <gongt@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>