[zxdump] Fix and test Thread::read_state
This wasn't tested and so the dump-via-TaskHolder changes
accidentally removed the actual support for dumping thread state
altogether because it had never been implemented in the live-task
API.  This implements it in the live-task API, and also refines
the Thread API to be more useful for examining the actual data
types.  It's tested via dump tests, which covers both the
live-task API indirectly by requiring it for dumps to have the
data and the dump-reader version of that API for verification.

Change-Id: I88a2fad9010b4f19eaf3647c865ae2934400f1de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825359
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>