[devshell][ffx] Set default target in build config
This makes it so that we set the default target as a build level config
for ffx as well as setting it in a .device file adjacent to that. It
does not yet start using that configuration value, but it sets the stage
to later once this change has spread more widely.

It also adds some helpers to the fx code for manipulating a json config
file, and uses those rather than going through ffx to avoid
bootstrapping issues. It's a reasonable expectation that fx is "in
charge of" the build level configuration file, so it's acceptable for it
to modify it this way.

As part of this, various tests have been amended to deal with the fact
that using the default device or enumerating devices will call ffx a bit
more, mostly when a device isn't already set or passed to fx.

Bug: 74963
Change-Id: Idbbcd4ee154c5c820525bdba13aba7b374e017ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822031
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>