[dfv2] Fix OOM test and put back in buildbot
Fixes the OOM test on DFv2. This test looked for a specific log message
that was dropped when the dfv2 main was separated out from v1. This log
is not necessary since there is another one that gets looked for the
comes after it. This fixes the log mismatch. Also puts back the test in
the buildbot now that it passes.

Bug: 108569
Tests: fx test src/tests/oom

Change-Id: I96a3e91741ee3b593baa86fb0c43a8d9ea7865c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827881
Commit-Queue: Novin Changizi <novinc@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>