[assembly] Add 'userdebug' AIB set for assembled_system to use
There are some artifacts that are needed on both eng and userdebug
builds and so they need to be included in non-user contexts, but they
aren't as restricted as the eng-only artifacts are.

This allows their use in userdebug systems.

Change-Id: Iaf17c37d192fa3dcd77131095953c6308d522b7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827936
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>