[fasync] Add shutdown call for emulated handles
Because emulated handles live in the process using them, not the Zircon
kernel, writes buffered in emulated handles go away when the process
dies.

This patch adds a call that ensures all emulated handles are flushed.
It at least guarantees whatever subsystem is on the other end of the
handles (i.e. Overnet) has seen whatever was in the buffers.

Bug: 124614
Change-Id: Iec70a9036a58970e075ca41d59eb6f612798ed60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828159
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>