[kernel][vdso] Fix build nits for GCC
Regressions in the build system long ago prevented the vDSO from
being built with GCC when the kernel is, as it originally was.
Subsequent changes will resolve that build system regression, but in
the interim some problems snuck into the vDSO code so that it is no
longer usable with the GCC tools. This change fixes one C++ bug
preventing compilation of the vDSO C++ code and another linker
script bug preventing runtime use of the vDSO in some places.

Bug: 69697
Change-Id: I33a26960d74e301272357564f95860da9ec5b5e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829036
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>