[brcmfmac] Don't glom tx frames in FW recovery
When the driver reloads the firmware in recovery, the firmware does not
expect frames it receives to be glommed, which changes the structure of
the frame. The driver later enables glomming in the firmware by setting
an iovar, but when initializing the firmware in the recovery worker,
frames should not be glommed.

Bug: b/274966334

Test: Manually crash the firmware with the crash iovar, check that
firmware recovers and that user can connect to AP and ping.

Change-Id: Iaaee3b5753c56afc46325d6d604852ce30e655c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828137
Commit-Queue: John Tan <johntan@google.com>
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>