[perfetto-bridge] Rig up Shared Memory in PerfettoProducer
This sets up perfetto producer to write perfetto trace events to the
shared memory buffer it is given as well as implements enough of the
perfetto::Producer API to handle the ipcs to start/stop tracing.

This means that:
- Perfetto producer no longer needs to be a trace observer
- We can write perfetto events to a buffer and funnel them to the trace
  controller

This doesn't yet connect to the perfetto tracing api yet. This directly
writes protobufs to the buffer rather than through the perfetto TRACE
macros.

Next steps are to:
- Figure out how to have the perfetto trace macros use the trace writer
  we create
- Abstract out the perfetto::Producer implementation to be usable
  outside of this test

Bug: 111976

Change-Id: I97cf64f044ed2efbeab0f5b720e0ccc634dffc93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827879
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>