[radar][proxy] Add connect request queue to fix startup flake
If a connect request is received immediately after radar-proxy starts,
there is a chance that it will be handled before the existing radar
devices are enumerated and a connection is established. This will cause
the connect request to be completed with an error even though radar
devices are available.

To fix this, a queue is added to hold connect requests that arrive
when there is no connection to a device. Once a connection is made, any
requests in the queue are completed using the new device.

Also re-enable radar-integration-test now that the issue is fixed.

Fixed: 124621
Multiply: radar-proxy-test: 1000
Test: radar-integration-test passed on Nelson tryjobs
Change-Id: Id1a69d77d6cfa89eff3f288fb89917a81e9dbbde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827220
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>