[netstack3] Send ethernet frames over loopback
Support Linux-style packet sockets by treating loopback as an ethernet
device that can send frames with source and destination MAC addresses.
This is invisible to the rest of the netstack but will be exposed via
packet sockets. As part of this change, rename queue-related things to
be frame-oriented or ambivalent.

Bug: 106735
Change-Id: I0ebd23e5cd6e974dd2e87c60cca624d3027a5f3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828237
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>