[build] Add clean build fence to work-around Ninja incremental failures.
The change in [1] was submitted successfully, but it triggers
flaky incremental failures, due to the fact that it changes the
build graph in a way that Ninja cannot deal with properly
(see associated bug for underlying reason).

This CL introduces a clean build fence to force clean builds
after it is applied, to get rid of the issue completely.

Bug: 83397
Fixed: 124905

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/829176/1

Change-Id: Ib294cb0e13fe703c55fc60830effc858c8a05e35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830162
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>