[bt][hfp] Connect CallManager on reconnecting peer
When reconnecting a peer, the peer task would be started up, but this
would leave the peer in a state without being aware of the currently
active CallManager, causing calls to never be transmitted to the peer.

This most often was seen as a failure to use HFP, and instead fallback
to A2DP and the internal microphone.

Bug: 96493
Bug: 105082
Test: added unit test for PeerImpl, fx test bt-hfp-audio-gateway-tests

Change-Id: I7964850ea16ec8ffeee1077edd7c990ac9bf0230
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822235
Reviewed-by: Matthew Kehrt <kehrt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>