[kernel][riscv][user-copy] fix user copy routine
Previously this routine was trashing s2 register, which is now used for
the shadow call stack. However it was also relying on the memcpy routine
being compiled a particular way, which it no longer is, due to frame
pointers and shadow call stack.

Though it's not a complete fix, tighten up the code that wraps memcpy to
try to protect against most registers being trashed. The better solution
is to have a fully asm implemented memcpy/memset routine so that up
front we known what registers are touched.

Also clear the sstatus.sum bit on kernel entry and add some additional
validity checks on exception entry to try to catch things getting out of
place.

Do not attempt to process the page fault if it originated from kernel
space and isn't part of a user copy.

Bug: 122873
Change-Id: Id8eca53708c1fddb87d3a206f4cdceb970a6e2a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823109
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>