[fdf] DriverRunner removes orphaned node after binding
Update DriverRunner so it stores orphaned nodes in a
map. When the DriverRunner binds a node, it will remove
it from the orphaned nodes.

This is to address an edge case in which a node fails
to match to a driver, then binds to one after a
RequestBind() call.

Bug: 123392
Tests: Added a test to driver_runner_tests
Change-Id: I97ab1bacebb9fd36cdc0fa1c298e4e7e168b16aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828384
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>