[kernel][riscv][brwlock] disable the PI version of the brwlock for riscv
At the moment, the PI version of the brwlock requires the use of a
doubleword compare-and-swap operation which does not exist on the RISC-V
architecture. For the moment, simply disable using the PI version of the
lock for this architecture by defining BrwLockPi to actually be
BrwLockNoPi.

Bug: 122873
Bug: 124881
Change-Id: I289114f062fed30741eff23c2019c0f984163568
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824298
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>