[ffx][target] Loosening network changed detection
The effect of returning true for network_changed from
find_matching_target is to trigger reconnecting the host_pipe
to the target. The main reason to do this is for user mode
networking with an emulator, where the port mapped to SSH changes,
but the host pipe is using the old port mapped to an old instance.

A side-effect of this physical targets that respond to
mDNS on IPv4 and IPv6, the address will change quickly
as the target is coming up, this can cause a lot of
confusion and race conditions.

This is the probable cause of failures in the Chromium CI/CQ tests
which use physical devices.

To avoid that, only return network changed if the port
is specified as well as a change.

Change-Id: I4beff420a2d963adf8f06b3c3b025f627cdea686
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830316
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>