[ui][test] make `InputDevice` panic more liberally
The `input_helper` doesn't have a good way to surface errors
back to the FIDL client, or even from `InputDevice` back
to the library that calls into it.

That means that errors generated by `InputDevice::serve_reports()`
are silently ignored.

We could try to plumb errors from `InputDevice` to the
library that calls into it. But, that's counter to the "fail
fast" intent of this component as a whole.

So, instead, modify `InputDevice` code to panic as quickly
as possible when an error occurs.

Along the way, update tests:
1. For tests that _intend_ to cause an error in `InputDevice`,
   add the `should_panic` attribute.
2. For tests that did not intend to cause an error, but did
   so unintentionally, update their logic to avoid causing
   an error.

Bug: 124883
Test: fx format-code && fx test input-helper-tests
Change-Id: Ib2c6b46f22e0c9feeae6177314b415ca31247a8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830656
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>