[scripts] Run incremental publisher with time versioning
The ffx repository server has a TUF client, which tracks the versions of
the TUF metadata. We generally generate metadata where we use the latest
time as the version number, rather than monotonically increasing the
number. This allows us to do a `fx clean-build`, and the TUF metadata
to be correct.

Unfortunately we were running the incremental publisher without time
versioning. So if a user does `fx clean && fx test ...`, this can cause
a new TUF repository to be created, where the metadata version is 1.
TUF considers this a rollback attack, and will reject the metadata.

This avoids that case by enabling time versioning in all the places
where we use the incremental publisher.

Change-Id: Ib855e415f627f2b1622985b3e51fdede9ffb4225
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831298
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Pollack <andrewpollack@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>