[display][core] Enforce thread affinity of Fences/FenceReference creation and destruction.
Fence refcount implementation is not thread-safe (though thread
compatible). In today's display::Client implementation, all the
Fences and the References to Fences are created on the same
controller loop thread.

For the sake of simiplicity and to avoid data race, we add a
thread enforcement to Fence / FenceReference that they must be
created and destroyed on the same thread.

This change adds a fit::thread_checker which issues debug asserts
to make sure that methods of Fence and FenceReference always run
the same thread where the Fence is created. There's no behavior
change introduced by this test.

Test: display-core-unittest on core.x64-asan
Test: vkcube-on-fb on core.x64-asan
Change-Id: I3cf5127efd740736210be27b962754e683c471b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823687
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>