[netstack3] Make device BindingId nonzero
Make BindingId an alias for NonZeroU64 (instead of u64) and use it
instead of the ad-hoc u64 throughout test and non-test code in the
bindings.

Note that this changed the logic of the Devices struct to hold the next
ID to allocate instead of the last ID allocated, though the first ID to
be allocated is still 1.

Change-Id: Ia1ef31d30b6d127d311717c5dea29180c7eb1043
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831517
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>