[fragment] Move outgoing directory creation to DdkInit.
This potentially fixes a shutdown flake, where the driver
host is crashing on Driver destruction, which asserts that
no tasks are still pending.

The driver host creates a Driver object before calling the driver's
Bind operation. If DdkAdd fails, the driver host will destruct this
Driver object. The flake's logs suggest that adding the fragment
device is failing (due to the driver host shutting down).

It seems that the SynchronousVFS destruction in the OutgoingDirectory is the culprit.
I have moved the outgoing directory creation logic to DdkInit, which ensures that either
the task will not be posted (if DdkAdd fails), or the driver host
will know that DdkInit is still pending before destructing the driver.

Multiply: host_x64/obj/src/tests/shutdown/shutdown_tests.sh : 100
Bug: 118767
Change-Id: Ia188dad1eae78267d6c04e5840387f468e59a3a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820888
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>