[gigaboot] fix TCP stack corruption
It looks like the TCP socket struct, which was placed on the stack,
could still be partially in use by the TCP driver if we returned early
without properly closing it out.

This led to memory corruption during `fastboot continue` when the driver
continued flipping bits that now no longer belong to the TCP struct.

Bug: b/274973959
Test: `fx qemu ...` + `ffx target flash ...` no longer seems to fail
Change-Id: I3c70eb74afc69d7dc7337e9d96e2e286684e7071
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831481
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Dov Shlachter <dovs@google.com>