[bti-test] Wait for file removal
This test causes the driver host to crash and then waits for it
to be re-created.

Specifically wait for the file to be removed from devfs.
This fixes a flake where the driver host has crashed, closing the
client channels, but DriverManager hasn't removed the devfs entry
yet and the test connects to the old entry.

Fixed: 124764
Multiply: platform-bus-bti-test
Change-Id: I12d04247771842f27409dc0080f4241c9654766d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831364
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>