[flatland] Remove extra discard config display calls
Discard in CheckConfig is a costly call. It only discards pending
changes, so it is only necessary if we attempted to set layers on
display. This CL removes the unnecessary calls we make at the start and
when we fallback to gpu compositing due to config.

Bug: 123838, 84271

Change-Id: I55d82ee1ed3b15bf0040912605ba1eac1c3872b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831284
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Emircan Uysaler <emircan@google.com>