[vim3] Add LCD_RESET expander GPIO pin.
LCD_RESET is an expander GPIO pin connected to the
PIN "P0" of TCA6408 [1]. It's used to detect if there is a
MIPI DSI display plugged in.

We added this pin to vim3-gpio so that the board driver can
use it to detect built-in touchscreens.

[1] VIM3 Schematics, Page 6,
    https://dl.khadas.com/products/vim3/schematic/vim3_sch_v14.pdf

Bug: 125133
Test: "fx shell gpioutil read VIM3_LCD_RESET" on vim3,
      when touchscreen is plugged in, returns 1; otherwise returns 0
Change-Id: I85232cca62f043c0f632294383eac73ba876083f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832897
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>