[inspect] Exclusive ownership for Container
This refactor makes Container non-Clonable and instead requires a
exclusive reference to it to be able to mutate it (or shared reference
for reading), aligned with Rust ownership semantics.

This is a pre-req for the follow-up which brings a more efficient
Container.

Run-All-Tests: true
Bug: 123517

Change-Id: I6de127a3225ddb358a963a8a538443a692d9e817
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821007
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>