[ffx][emu] Resizing oversize Disk image files
The FVM or fxfs file that is part of the product bundle contains the blobs for
the image. The disk image file then needs to be extended to allow for the
minfs or fxfs file system to be created. If the original image file was larger
than the target size, the image file was unchanged, and the file system could
not be created, causing the emulator to not completely start Fuchsia.

This change checks the size of the image file, and if it is over the target
size, it is resized to 110% of the original size to allow for the
file system to be created. Longer term the product metadata in the
product bundle should be updated to reflect the contents of the FVM
so an appropriate target size can be specified.

Bug: 102866
Change-Id: I3649b5675749b0c4735fd1aa24bd81fb9e33d012
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831479
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>