[libc] Fix buffer overrun in test
The pthread_getname_np function's API permits it to write
to the whole buffer provided. It's not valid to assume that
it won't write as much as the size_t argument allows, even
if Zircon thread names are never that long.

The llvm-libc implementation of strlcpy writes NULs to the
entire buffer size given, which is suboptimal but valid for
it to do.

Bug: 122880
Change-Id: I9b0486c720756483f82d1b55dbea41de0f22dec5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834436
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>