[package-tool] Package Manifests Lists are relative to working dir
It turns out `pm publish` assumes package lists are relative to the
working directory, rather than the file. Scrutiny relies on this, so
we need to switch `package-tool repository publish` to be compatible.

We eventually want to migrate the package manifest list to json (or
some other structured format), so for future flexibility, this keeps
the `create-all-package-manifests-list.py` to merge lists rather than
reverting back to `cat.sh`. This should make it easier to migrate
to a different format in the future.

It also migrates scrutiny to use `create-all-package-manifests-list.py`
to make it easier to migrate formats in a single place.

Change-Id: I52122e9acd8282be53f685068fe382686b40327c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832245
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>