[fxfs] Allow extra space in generated images
Adds fxfs.reserved_space_bytes to the generated images config for Fxfs
which allows additional free space to be reserved in the Fxfs image.
Some extra space is required for the image to be very useful, since
without any extra space, the data volume can't be created and formatted
and the system will not boot.

The default value is arbitrarily 16MiB.

Bug: 122056
Change-Id: I5915fa265c56441072a2fc6d1b95a7cec75acc9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/833997
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>