[ffx] Make isolate names correspond to test name
When debugging a self-test failure in Infra, we need to be able
to find the logs for the failed test, which is difficult when
the isolate directory names don't correspond to the test names.

The rule in this commit is not super-strict, but roughly it means
that a test with path '<component>::...::test_<test_name>' will
be found under '<component-test-name>'.

Fixed: 125417
Change-Id: Iff3b2d9670b96ad48d87c342d07d7c2a884ca582
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835877
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>