[sys_driver_cpp] Explicitly teardown in destructor
On ZirconConnection teardown there are a few things that need to happen
in order:
1. Disable the notification handler
2. Shutdown the loop (cancels remaining async tasks)
3. Destroy the MagmaSystemConnection (calls into the driver to tear it
   down).

Currently steps 2 and 3 are run by destructors, ordered based on their
order in the ZirconConnection. This is a bit fragile, so it's safer to
run them explicitly in the destructor.

Change-Id: Ia3fac1b7efc5a0fb99e0d03e45bdce7803683779
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/834938
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Craig Stout <cstout@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>