[fidl] Improve error catalog docs and script
This CL simplifies the fidlc error catalog README and improves the
add_errcat_entry.py script used to create new errors. This is in
preparation for adding fidlc tests that ensure an error's "documented"
state in diagnostics.h is in sync with the docs.

Test: ./tools/fidl/scripts/add_errcat_entry.py 205
Test: ./tools/fidl/scripts/add_errcat_entry.py 203
Test: ./tools/fidl/scripts/add_errcat_entry.py 204 --good 3 --bad 2
Change-Id: I6c0434b97db45cabb3380d353e3e75d81777a662
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836257
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>