[fdf][amlogic-display] Use FIDL version of sysmem
Make the sysmem fragment of the `display` composite device match the
FIDL version of the sysmem protocol instead of the banjo version.

Make the `amlogic-display` driver that binds to the `display` device use
the FIDL version of the sysmem protocol instead of the banjo version.

Bug: 124903
Change-Id: Ie644c49c31f46c42cac9023cdd968dd30d11b3da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831358
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>