[flatland] Add cache flushing to MapHostPointer.
Tests may map VMOs to host memory addresses to write to texture
VMOs or read from capture / render target VMOs. Sysmem is designed
to not implement any special provisions for cache management except
for providing some indicators about the VMO's cache coherency domain,
and the client application should make sure cache flushing is done
at the correct time [1].

If the cache is not flushed before read or after write, the updated
contents may not get propagated to the RAM; the contents displayed
on screen and (or) captured by the display engine will be incorrect.

This change adds an enum argument HostPointerAccessMode to
flatland::MapHostPointer(). This specifies the client's behavior
on the host-mapped VMO, and MapHostPointer() will flush the cache
lines properly, and we can remove all the cache flushing logic in
our tests, which can be error prone when we add new tests.

This fixes the flatand_display_compositor_pixeltests flake we saw
in fxrev.dev/835421. That change fixes the display image comparison
logic so that all the pixels are being compared (previously only the
first row is compared), so it scaled the tests by 1024x which makes
it easy for the bug to occur.

[1] Sysmem: Design Overview, "Limitations", internal doc.

Test: multiplied flatand_display_compositor_pixeltests with
      fxrev.dev/835421 (see "Checks" in fxrev.dev/835421)
      on astro, sherlock and nelson
Multiply: flatand_display_compositor_pixeltests: 50
Multiply: flatland_renderer_unittests: 10
Bug: 124484
Change-Id: Icde88858e86422bea1fc007e234c3aa3b8fe488b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837081
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>