[radar][proxy] Add structured config entry to control reader proxying
Add a structured config entry to control which protocol (burst reader or
provider) is proxied. For now only the provider protocol is proxied, but
upcoming changes will make it follow the build type (proxy provider on
user builds, reader on others).

Also add a radar-integration-test package that proxies the reader
protocol so that both modes are tested.

Bug: 123714
Test: Radar features work on Nelson
Test: fx test //src/devices/radar/test/radar-integration-test
Change-Id: I3bd4e9efa206227418158cbeb3a4dfb67decd2b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835896
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>