Reland "[starnix] Implement POLLRDNORM and POLLWRNORM"
This is a reland of commit 11818c056255945ba78dc99fb66c55101ca9cf4b

This CL is the same as the previous landing, except with the flaky test skipped. This test is flaky because it's using a race condition to trigger an interesting error condition.  When the error condition doesn't trigger, the test passes.  When the error condition triggers, we fail the test because we don't implement the error condition correctly yet.

Previously, we always failed the test because we didn't behave correctly when the error did not occur.

Original change's description:
> [starnix] Implement POLLRDNORM and POLLWRNORM
>
> We need to set these FdEvents as well as POLLIN and POLLOUT.
>
> Fixed: 79315
> Change-Id: I7a30c05f33b9304ab987b5f43e95fd1d1ca92313
> Reviewed-on:
https://fuchsia-review.googlesource.com/c/fuchsia/+/836416
> Reviewed-by: Theodore Dubois <tbodt@google.com>
> Commit-Queue: Adam Barth <abarth@google.com>

Change-Id: I826bca8f0ef15104b35922c5abeaf6d4d4888d77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839678
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>