[compat][test] Add StartDriverArgs
Update tests so they pass configuration in via a struct instead
of function parameters. This makes it easier to have defaults
for tests that don't want to change behavior.

Update the TestFile object so it can only be configurated on
creation.

Both of these changes are a setup to let the test infrastructure
run on a separate thread.

Change-Id: I7f971448c6f253cf7b80ccd3fdd8212890a19f61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840056
Reviewed-by: Novin Changizi <novinc@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>