[compat][test] Add IncomingNamespace class
Pull all of the mocks out into their own class. This should not change
the threading behavior, but it makes it possible to put this class into
a DispatcherBound object in a followup CL.

The IncomingNamespace class is entirely configured when it is started,
so TestProfileProvider must be updated to also be configured on
startup.

Multiply: compat_test
Change-Id: If2920d5b8958df98595d8f890d3acc55bc0dcbb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840057
Reviewed-by: Novin Changizi <novinc@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>