[starnix] Delete Credentials::from_passwd
from_passwd ignores almost everything you give it, which is confusing
and makes it seem like the username and password specifications are
doing something when they aren't.

For containers, it seems simplest to get rid of the init_user
specification. Everyone is just setting it to root.

For components, introduce a "uid" key. Let the gid default to the same
as the uid, as that's what every single current use of "user" does.

Change-Id: Ia0af1964d5c9d1ec0bca25bcb1ea6b40c7a308b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839458
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Theodore Dubois <tbodt@google.com>