[dfv2_v1_lifecycle] Don't use DevfsExporter
Use the new node devfs API instead of DevfsExporter.

This requires adding a "dfv2" node in the topology which is awkward, but
this code is in the process of being re-worked and there are no real
clients for DFv2-in-DFv1 at the moment.

This is the last user of DevfsExporter, so it can now be deleted!

Bug: 125288
Bug: 117760
Change-Id: I2afcf54c487b69b7a27ade2e05836e8efb354891
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839063
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>