[kernel][vm] LookupOrAllocate API cleanup
Callers of LookupOrAllocate also pass in whether they expect any
intervals in the page list, and if so, how to handle them. This can
prevent accidental misuse where a caller ends up directly populating a
slot in the middle of an interval; each caller should now be aware of
whether or not intervals can exist in the page list they are
manipulating.

Bug: 122842
Run-All-Tests: true
Change-Id: I22c50620ed30df2c14e28ba9068bb728ee9cc4c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839396
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>