[sdk][ffx] De-dup ffx_tool atom files entries
Based on feedback from the initial sdk metadata, change the files
section to be an object with three fields:

- executable: the binary, previously at the top level
- executable_metadata: the fho metadata file for the executable
- support (optional): other files the subtool uses

Previously the executable* items were at the top level, and then
duplicated in files. This was somewhat convenient for writing
distribution logic for the tools, since it's the same as host tools, but
this makes things a bit clearer and we can make the logic down the line
handle it.

This also updates the build and ffx sdk tests to correct for changes in the
metadata formats.

Change-Id: Ibf39c89014d0747453cd4126a56877db8f4399df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837997
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>