[audio] Include configs in pipeline-benchmark packages
In the vast majority of cases, tests do not need to modify the system
image simply to provide a config file to the code under test.  The
pipeline benchmarks appear to currently rely on passing configs through
the system image; this patch causes the tests to resolve the configs
from the test package itself instead, which is likely better for
developer iteration speed as well as simplifying product assembly.

Change-Id: I661615d03374ab26d38ccc8b60e7a99a7246eed5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841963
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>