[dfv2] Split Runner into separate class
There is an API boundary for loading the driver components through
the component runner API. Pull this code out into its own class so it
can maintain its own state and be encapsulated.

The main reason for doing this is because DFv1 wants to start loading
DFv1 drivers through the driver runner. This is much easier to do when
the driver runner is unaware of DFv2 objects like Nodes. This will be
handled in a followup CL.

Bug: 125288
TEST: driver-runner-test
Change-Id: I3fcadefccc43aa71ac586be95d319426e1441a23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836756
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>