[rbe][rust] remote cross compile with host sysroot
Tell remote rustc to use the location of the host sysroot,
which is where the host compiler expects to find the per-target
std libs by default.

This is an alternative to the proposal to move the rust sysroot to
a host-independent location.

Consequentially, the command will differ from that of a native (linux-64)
remote compile (cache-isolation), but least it works.

Tested: successful remote rust cross compile
$ build/rbe/remotetool.sh --operation show_action --digest 26700af61185c3b17f50aa8b6f68f4a02a3f8093b36b46a17ee4a0c196e76261/147
(Observe that command contains: "--sysroot=../../prebuilt/third_party/rust/mac-x64")

Bug: 125064
Bug: 105799

Change-Id: Id12de6a770d59165c2f9bfca5ca98e7ae8a8eb51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843596
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>