[rfc] Clarify scope of RFC-0154 and RFC-0208.
The emphasis on the testing-related motivating use cases can lead the
reader to the impression that these mechanisms are _only_ intended to
be used in testing scenarios. I've added a paragraph to each,
indicating that production use cases are in scope, but there will be a
higher quality bar to use these in prod.

I'm sending this amendment around to many of the original
authors/reviewers to confirm that this was the RFC's original intent
(or an acceptable extension thereof).

Change-Id: I73f47aaf0c44af45f92a1ea0976097f69bfc3030
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831556
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>