[ffx][fho] Properly support legacy machine output
The previous implementation of all of this didn't properly support
machine output from legacy plugins. It turns out that fixing this
required a fairly deep refactoring of fho because of assumptions about
how machine output state can be determined.

Many of those problems came from severing the `FfxTool` and `FfxMain`
traits from each other, which was done to make it so that FhoEnvironment
could be stored properly between them. We made that not necessary by
switching to using an Arc<dyn Injector> which let FhoEnvironment become
Clone, so this separation wasn't needed anymore and is undone here.

We also move building the 'main' object to happen earlier, simplifying a
bunch of things about the lifecycle of the object.

And last but not least, we break the `ToolRunner` implementation into
two separate ones, one that implements the tool and one that prints
metadata, rather than doing the switch between them late and awkwardly.

Fixed: 125869
Change-Id: Iaebcf3733342b368b2ffbf98ad32f03cdd032f09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841458
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>