[dhcp] SubnetMask type uses net_types PrefixLength
The DHCP server configuration defines a SubnetMask type that overlaps
substantially in functionality with the (later-introduced)
net_types::ip::PrefixLength type. This change updates SubnetMask to wrap
PrefixLength and delegate all overlapping logic to the latter, but
preserves the existence of SubnetMask in order to be able to implement
serde and FIDL conversions on it.

The serde formats are preserved by deriving serde::{Deserialize,
Serialize} on the old version of the type and then manually implementing
them on the new version to forward to the old version.

Change-Id: I89b12504b1dbd04df714c88991b004731160c681
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843897
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>