[dhcp/client] Bundle component with Netstack3
Route the DHCP client provider capability from the netstack component to
netcfg, and include the still-under-construction DHCP client component
in the Netstack3 package used for production.

Routing the capability from the netstack package itself makes the choice
of whether to use the built-in or external DHCP client configurable
per-netstack instead of requiring an orthogonal flag in netcfg or a
setting in the network realm.

Netstack2 does not expose the capability, and so will continue using its
built-in DHCP client, while Netstack3 will use the new component.
Because of this, scrutiny verification fails. Since the RFC-0155
mechanism for supporting this use case isn't implemented, add an entry
to the routes exception allowlist.

Bug: 125773
Change-Id: I1d3ce25023391f768a3288da69aad7b9d6ec1a16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840519
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>