[cobalt] Use test config data from package, not system
Tests changing the underlying system image by including config_data that
are only consumed by the test component are more of a pain to iterate on
(and for product assembly to construct system images for) than properly
hermetic tests.

This moves the config_data resources into the package, and then updates
the cobalt testapp to look in the test package for that config data,
rather than in a system-global location.  Since it appears the tests
were otherwise hermetic, we can also take them out of the system-tests
realm.

Change-Id: Id5c0d0de10606e6fbbc67e03d2c02d6f562baa8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842116
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Cameron Dale <camrdale@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>