[fxtest] Add helpful flag suggestions.
Many people have missed the '--output' option to print all output to
stdout. This change ensures that the configured behavior of output is
reflected in information at the top of the invocation, and it provides
suggestions on how to change that configuration if desired.

Fixed: 126087
Change-Id: I36b2664fb98ca894de180d9aafcf87d22fd215ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/844120
Reviewed-by: Chris Phoenix <cphoenix@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>