[kernel][riscv] Fix bug in Map's returned count
The Map function currently does not include existing entries in the
count of the number of pages mapped. This contradicts the behavior
documented in
https://cs.opensource.google/fuchsia/fuchsia/+/main:zircon/kernel/vm/include/vm/arch_vm_aspace.h;l=72
and causes the failure in the attached bug. This CL adds the existing
pages back to the count.

Fixed: 125951
Change-Id: Ic4447b4473e5c8f0296626b9eef748015acbbbf0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842719
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>