[netstack3] add fuchsia.net.routes/State.Resolve
Implements the basic functionality needed for `Resolve`, including
looking up the destination in the routing table, identifying the
preferred source address for communicating with the destination, and
querying the neighbor table for the remote's MAC address.

A follow up CL will be needed to handle the case where the neighbor is
not present (or incomplete) in the neighbor table. Core will need to
initiate neighbor discovery, and bindings will need to wait for it to
finish.

Bug: 120878
Change-Id: I8a51ad3f87d2637e0d74778a9eaed63df2f4ee75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822787
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>