[kernel] Extend core test timeouts on RISC-V
The Bti.DecommitRace test case takes 15-17 minutes to run on RISC-V. To
get the bot green, this CL extends the timeout of the tests to 30
minutes. We chose 30 minutes as it is the default timeout (10 minutes)
plus the 17 minutes needed for the DecommitRace test case, along with a
few minutes of buffer.

Bug: 126120
Change-Id: I511142c3410558c61ba6aad10637bb0c4a4892d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/846853
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>