[kernel][vm] Remove gen count optimization
The purpose of the gen count is to determine when attribution is
changed and is ambiguous then whether code paths that do not imapct
attribution should increment the gen count. Failure of those paths
to increment would cause this optimiziation to be bugger.

Due to the way the new LookupCursor works, re-performing the page
lookup is actually quite cheap now and this optimization is not
providing any value.

Change-Id: I5d57c5e9d1176c517467ed5a09b3f18834eabb88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847276
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>