[fidl][rust] Replace table consts with fn or let
This CL replaces each const that contains a FIDL table with a function
returning the table or with a let binding, depending on the situation.
This is necessary for the migration from EMPTY to default(), because
default() is not a const fn. Such consts were fragile anyway, e.g. they
did not work with String fields for example.

Bug: 102876
Change-Id: I1e21bd940b48823e8ab946f8099e6237c2d35eba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845818
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>