[check-licenses] Updates to readme package.
Previously, the tool would recursively search for license files
in go/dart/rust packages to add to the project. This strategy
disregards any skiplists or other configuration settings in other
parts of the tool, which is not good.

This CL changes it to only add any license files at the top-level
of those project types. If other licenses are found deeper in
a given project, they'll be added using a separate process.

Updated the tests to reflect this new behavior.

Change-Id: I835aba25757c77369d44c9a81b03a51237418a4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848157
Reviewed-by: Yaar Schnitman <yaar@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>