[03/14][fidl][display] Remove ImageConfig.pixel_format.
This field is not used in any in-tree or out-of-tree code.

Since it modifies the struct, this is an API and ABI-breaking change.

Currently the only out-of-tree usage of the display API is through
VkLayer_image_pipe_swapchain_fb; it's not automatically tested so
we'll need to manually rollout the layer binary.

Test: fx build
Bug: 126113
Change-Id: I2760457c7319b9f2ad09d2a4a2d3d344a04d20e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843241
API-Review: John Bauman <jbauman@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: John Bauman <jbauman@google.com>