[superproject] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Add support for global ctors / dtors for AMDGPU
This patch makes the necessary changes to support calling global
constructors and destructors on the GPU. The patch in D149340 allows the
`lld` linker to create the symbols pointing us to these globals. These
should be executed by a single thread, which is more difficult on the
GPU because all threads are active. I chose to use an atomic counter to
sync every thread on the GPU. This is very slow if you use more than a
few thousand threads, but for testing purposes it should be sufficient.

Depends on D149340 D149363

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D149398

GitOrigin-RevId: d193b2dca72dd4429db3f0c03e657234d6d794b7
Original-Revision: ec040b7f0382eb47cda869f69d4bca02fc872b25
Roller-URL: https://ci.chromium.org/b/8782468262605328817
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849023
Original-Revision: 349abecb907dc03134110143b26462adba6bad45
Roller-URL: https://ci.chromium.org/b/8782467935266055697
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I1bc0a40ae3832d62d11ffc1e01a79b1e40cf44dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849024