[virtualizaton] Build test_util binaries in CQ
These are prone to build regressions as a result of FIDL refactorings.
Without this CL, the virtualization builders would need to be added to a
CL to discover that it regressed the builds of these binaries.

This has been a frequent cause of failures in the virtualization
builders and this CL seeks to surface these errors in CQ without adding
substantial load on infra bots or build image size.

Change-Id: I43848d4873b295d96fd0deb2b00397727e68b311
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850479
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>