[rbe] Write download stubs, download later
When --download_outputs=false, write out download stubs.
Stubs are set up as .dl-stub files, symlinked from the
file's actual path.

A stubbed download for 'foo/bar.o' will look like:
  foo/ (dir)
    \- bar.o (symlink -> bar.o.dl-stub)
    \- bar.o.dl-stub (text file)

On download, the link to the stub file is followed,
the blob is downloaded to a temp file.
If the download succeeds, the link is replaced with the full file.
The stub file (.dl-stub) is kept around, to make it
easier to inspect the origin.

Bug: 123178

Change-Id: Id5a694064ab133665d818566ce0b87351ad674a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841919
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: David Fang <fangism@google.com>