[dhcp] Test NS3 client + NS2 server
Amends the integration tests to be able to choose the netstack versions
for the DHCP server and client independently, and adds expectations
for the newly-run combinations.

Also introduces an `annotate` utility to `netstack_testing_common`
to make it easy to add debug prints to a long-running future to aid in
determining where a test is getting stuck.

Multiply: dhcp-integration
Change-Id: I0f9c43225267049d70f06b4d013585179cf91ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849838
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>