[zither] Forward both enum member value and expression
This change fixes a bug in which the expression in FIDL used to define
an enum member value was forwarded to the backend as a raw numeric
value: in the case of an OR'ed expression, this would give a bogus
binding (not being able to account for and transform any the other
elements making up the expression). As we do for constant bindings, here
we forward both the reduced numeric value as well as the original
expression (when it differs meaningfully) to the backends: the value is
used to generate the binding, while the expression is giving as a
trailing comment for documentation.

Change-Id: I35375abcdb01aca33f0022abc3404a90cd445ea4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850198
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>