[Honeydew] Use tempdir for tracing.terminate_and_download() unit test
This change is a follow-up to fxr/841098. It makes the following changes:

* The trace file name is now lower case.
* The unit test now uses a temporary directory.
The previous code used the /tmp/ working directory for the unit test. This could lead to problems if the unit test was run multiple times in a row, as the trace files would be overwritten. The new code uses a temporary directory for the unit test. This ensures that the trace files are not overwritten, and that the file will be deleted after unit tests execution.

Change-Id: Ic67185abb5d09db9a428060cd0aea5d6bb903371
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851285
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Fuchsia-Auto-Submit: Mahesh Saripalli <maheshsr@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>