[ffx][ssh] Do not write public SSH key file unless needed
This changes the ssh create if needed function to not write
to the public key file unless the private key is new.

This is important for situations where the public key is read-only,
or on a non-writable file system.

This also correctly checks the file permissions for the private key
to make sure it is 0o0600. This change only happens when running
`ffx doctor --repair-keys`.

Bug: 126684
Bug: 126675
Change-Id: I7496594bea56b4422a645ea73b78e8bceea2db58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851997
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>