[media/audioconsumer] add audio_consumer component
This CL adds an implementation of fuchsia.media.AudioConsumer to
replace mediaplayer in non-smart-display builds and tests. This
implementation supports LPCM streams only (not compressed streams)
for now. This will be adequate for SBC streams, which are decoded
before they are submitted to the consumer. AAC streams, which are
delivered directly to the consumer, will be rejected for now.
Current end-to-end tests use SBC.

Test: new unit tests for all implemented features

Future CLs will include this component in BT end-to-end tests.

Change-Id: I2732ebdeb8cba916f263be63e5a9f9113e699e59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851221
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>