[zircon] Add mechanism to "kick" a thread out of restricted mode
This adds a system call to tell Zircon to raise the "restricted kick"
state on a target thread. A thread with the "restricted kick" state
raised will exit restricted mode soon if it is currently in restricted
more or on the next attempt to enter restricted mode if it is not with
the reason code ZX_RESTRICTED_REASON_KICK.

This is implemented by two pieces of state on the Thread object. First,
a bit on Thread::signals_ is used to signal and potentially interrupt a
running thread on another CPU. Thread::ProcessPendingSignals() observes
this bit when the thread is running and either immediately enters normal
mode if the thread was running in restricted mode or lowers the signal
and raises a bit on Thread::flags_ if it was not. The next attempt to
enter restricted mode will observe this flag and immediately exit
through the vector table with the reason ZX_RESTRICTED_REASON_KICK.

Bug: 121511
Change-Id: I942e071a071b8fb59966f7e2b8a54b6c129991b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842601
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
API-Review: Carlos Pizano Uribe <cpu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>