[superproject] Roll llvm-project/libc [libc] Use Linux errno and signal strings for Fuchsia
The exact set of supported values is determined by the <errno.h>
and <signal.h> headers, which don't (yet) come from llvm-libc on
Fuchsia. The mappings of SIG* and E* codes to psignal/strsignal
and perror/strerror text used in Fuchsia libc today is the same
as for Linux.

Reviewed By: abrachet

Differential Revision: https://reviews.llvm.org/D150026

GitOrigin-RevId: b4e9a7c2514030845f5c892289c3bb09be8aaefe
Original-Revision: 7119166424a7127ddeb6f70b8a4ddff3f3221eca
Roller-URL: https://ci.chromium.org/b/8781703854890066113
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I45b6d3e604b20f62f68454b2fed6d92963f3f7db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853419