[ffx][sdk] Separate target files in metadata
The earlier versions of this replicated the mistakes of the way
host_tool atoms are built by making the architecture implicit and hidden
in the path. For ffx tools, use the target_files to separate out the
architecture of the binaries.

This still builds two separate atoms for x64 and arm64, just because
the change is already quite large and that would go off into the build
system weeds but it should also support merging them into one atom
later, without causing backwards compatibility headaches.

Change-Id: If31eec50206ae8feebeb24d1a9ccda8f405f7b60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851523
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>