[product] wrap lookup and download in get; disable old get
The prior `ffx product get` is renamed to  `ffx product download`.
The `get` command is now essentially removed.

Adds `ffx product get` which does the function of both lookup and
download. This is behind an --experimental flag which bypasses
the message about the tool being renamed.

The experimental version is not for production use and is a proof
of concept example.

Change-Id: I3fa2b6b713116877a7745127412aa3a31be61d88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841896
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>