[audio] Remove incorrect driver API documentation comment
Per the driver SignalProcessing API's GetTopology API method:
If GetElements returns one or more elements, GetTopologies must return
one or more topologies.
Hence it was incorrect to say that it is ok to not expose a topology
if topologies are static. We don't repeat the requirement in the
documentation since it is already in the FIDL file.

Bug: 64877

Change-Id: Ied6604d13dcad6aecc9d076f338ddf5639e34a18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853121
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>