[pbv2] Fix canonicalization of virtual_devices_path
In some enviroments, virtual_devices_path is a symlink, and
canonicalizing the entire path causes problems with finding sibiling
json files that are in the same product bundle directory. This change
makes it so we only canonicalize the parent directory of the
virtual_devices_path file.

Change-Id: I8faf320970f2c1c43016874e44dc1bb077fb7ae7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851181
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Michael Brase <mbrase@google.com>