[lowpan-ot-driver] Restore work-around from <fxrev.dev/715287>.
This CL repairs the work-around from <fxrev.dev/715287> that was
inadvertently reverted by <fxrev.dev/755648>. Specifically, this change:

https://fuchsia-review.git.corp.google.com/c/fuchsia/+/755648/10/src/connectivity/lowpan/drivers/lowpan-ot-driver/src/driver/srp_proxy/discovery_proxy.rs

By removing the call to `.clone()`, the assignment was rendered trivial
and optimized away by the compiler. This short-circuited the intended
behavior of tying the lifetime of `subscriber` to the lifetime of the
stream.

Bug: b/281638502
Bug: b/281740970
Bug: b/281881247
Change-Id: Ic6458a877e7c386f2cf3186418b24f732f227eb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855012
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>