[ffx][config] Basics of fuchsia_env file parsing
Adds a new library that will hold the fundamental logic of fuchsia_env
files written in toml or json5 and the mechanisms to find and read them.

This is just a starting point so is definitely incomplete. More tests
and logic to come.

Bug: 122608
Change-Id: Idfc9b8ac417e71e8ec0f920132a1562f5e2a5512
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854481
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>