[dfv2] Use restart_url in FinishRestart
If the restart_url is set during restart, we want to look for a new
driver to attach to the node, so go through the full BindToUrl flow
instead of re-starting the existing driver.

Tests: fx test \
fuchsia-pkg://fuchsia.com/driver-runner-tests#meta/driver-runner-test.cm
--test-filter "DriverRunnerTest.BindAndRestartThroughRequest"

Tests: run msd-vsi-vip-integration test to see the test version of the
driver get binded when the Rebind was requested.

Change-Id: Ie951947ca6fed888cca9c6469ea4328ae57174a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852938
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>