[kernel][uart] Fix TX irq mask/unmask ordering
There is a potential race in some of the uart drivers where a log
writer thread is waiting on a TX interrupt, but the irq thread masks
the interrupt. This can happen when -
1. Log writer thread is waiting on TX interrupt
2. TX interrupt occurs and the interrupt handler signals blocked threads
3. Before the interrupt handler could set the TX mask, the log writer
   thread loops back, unmask's TX interrupt and block again on the
   event.
4. The interrupt handler now masks the TX interrupt
5. TX interrupt will no longer be triggered and hence the logger thread
   will be blocked forever

This CL fixes the issue by first masking the interrupt in the irq
handler and then signaling the blocked thread.

Bug: 126785
Change-Id: I34a0e6dea8f70c8353fbdb6817653075e1fd996f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853877
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>