[starnix] Use interruptible locks for append
Before this CL, a write would take a RwLock to ensure that O_APPEND
write were atomics. It means that if a append write was blocked, and a second
write was started, it could not be interrupted before the first one
has finish, even if it should have been interrupte because of signals.

This CL introduce locks that can be interrupted, and uses it for the
append lock.

Change-Id: Ic38abea1b30ba0f1d5181065344967e9153ecfd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855296
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>