[audio][fdf] Convert tests for audio banjo migration
As part of banjo migration of the audio codec
protocol, this CL changes all audio tests from obtaining
a fidl channel through the banjo protocol
to calling the class directly.

This also removes the move constructor of the
SimpleCodecClient, which was only used in tests.

Bug: 124658
Testing: this CL updates tests.
Change-Id: I701241c46686b95c44c186f5e5c6487fedbaf59f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852252
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Garratt Gallagher <garratt@google.com>