[gigaboot] Fastboot flash maps between new and legacy partition names
Some devices use different naming conventions for the expected
partitions.
The 'old-style' convention expects the following partitions:
* zircon-{a,b,r}
* vbmeta-{a,b,r}
* misc
* fvm-fuchsia

While the 'new-style' expects the following:
* zircon_{a,b,r}
* vbmeta_{a,b,r}
* durable_boot
* fvm

This presents a coordination problem among tools, infrastructure, and
individuals: which naming convention do we expect on a particular
device, and how do we encourage the new style while allowing the old style?

During the transition to the new style, fastboot flash adopts a
lenient strategy: accept either the old name or the new name and
dynamically map that to the actual name on the device.

Change-Id: If078667555f85be769f233c9b42d027583ee0b4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856681
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Dov Shlachter <dovs@google.com>