[bazel] Add --publish_only flag to publish_packages
In-tree repository management can sometimes conflict with bazel repo
management, particularly when the remote workflow is used or when the
version of ffx in the SDK is different from the one in the Fuchsia tree.

Add a --publish_only option to publish_packages.py to allow publishing
to an existing amber repo, so the other steps in package publishing can
be handled manually or through fx.

Change-Id: I2951d288fcdc216cbeb823977e75adefb7eaa3ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854997
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: John Bauman <jbauman@google.com>