[product] avoid creating empty out dir in pbms lib.
The set up for download created an empty output directory. We've
since moved on to a strategy where the data is downloaded to a
temp location and renamed/moved into position.

The empty directory was mostly harmless, but proved to be an issue
in specific mounted devices. Not creating the empty directory fixes
the issue.

Bug: 127332
Change-Id: Ic221b5ffc982f54cc67123a51546909ae5a46531
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856917
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>