[build] Fix hermetic_inputs_action.py depfile generation.
Ensure that the depfile's parent directory exists before writing it.

This is necessary to remove the use of stamp files in our GN-generated
Ninja build plans. As of now, the parent directory is created because
of the depfile. This will no longer be the case once [1] is submitted
to GN and `no_stamp_files = true` is set in our `.gn` file.

+ minor cleanups and properly handle depfile quoting.

[1] https://gn-review.googlesource.com/c/gn/+/12866

Bug: None
Change-Id: Ided2a6154cc1b12b0092beeaf3b5219a390559ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857183
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>