[time][security] remove timekeeper from scrutiny allowlist
The FIDL protocols named in the allowlist are not essential
for RealmBuilder as used by the timekeeper component.

We remove it for this reason.

However, this requires modifying the `offer` for the unneeded
protocols by making them `from: "void"`. Since they are unused
this should make no matter to the correct operation of
timekeeper, but will remove the need for allowlisting it.

Fixed: 115634
Change-Id: I13ff3f9c177efad998dd9fe14b13ad79bb48e665
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857317
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Marina Ciocea <marinaciocea@google.com>