[dash-launcher][fidl] Rename LaunchWith[Pty|Socket] methods
The protocol will be extended with a method for exploring packages,
this change establishes a naming scheme convenient for the new method
(ExplorePackageOverSocket).

Even though the FIDL is not in the SDK, it is called by ffx and so
the renaming needs to be done with a soft migration. The old names can
be deleted once ffx binaries built with the old names no longer need
to be supported.

Bug: 127374
Bug: 126945
Bug: 127317
Change-Id: I71d658b0484e0e2d206dcecc14c6c52c928dd2ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857291
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>