[starnix] Add POSIX_WRITABLE for RDONLY
The Fuchsia open flags and the Linux open flags do not line up
perfectly. On Fuchsia, flags are hierarchical, which means you need
RIGHT_WRITABLE on a directory in order to open files within that
directory as writable. In Linux, you always open directories as RDONLY.

To resolve this mismatch, Fuchsia has POSIX_WRITABLE which is ignored
for most files but inflated to RIGHT_WRITABLE for directories. When
translating a RDONLY or RDWR flag to Fuchsia, other clients, such as
FDIO, add in POSIX_WRITABLE. We need to do the same thing.

Change-Id: I7836afe8680666185e0ce9c13c42515a893a9533
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858037
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>