[audio][dsp][fdf] Use FIDL version of clock protocol
Make the as370-tdm-output, as370-pdm-input, and aml-dsp drivers use the
FIDL version of the fuchsia.hardware.clock protocol instead of the banjo
version. Update the audio and dsp composite devices to bind to the FIDL
version of the clock protocol instead of the banjo version.

Bug 12606

Change-Id: I0a887f38fd7114fdd6115067bfc7573ce1b2c280
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857279
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>