[kernel][riscv64] Revamp syscall dispatch using a lookup table
The old method of syscall dispatch assumed that each trampoline
sequence could fit into 8 bytes. In fact, the trampoiline
sequences were always larger (12 bytes) and then getting relaxed
to smaller (6 bytes) when the jump targets were within 1MiB
distance. When they're not that close, each sequence is too large
and the alignment causes each one to actually be 16 bytes, so
that the indexing calculation is wrong and the dispatch goes
to the wrong trampoline for the syscall number.

This instead uses a lookup table of addresses, stored as 32-bit
offsets to halve the number of cache lines occupied by the table
compared to storing full addresses (that would get relocated).

Change-Id: I5c7d8d24e2cda998e0155851e889edfeb2cff24a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858556
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>