[ui][testing] ui_test_realm no longer supports config.use_input
use_input was useful for determining whether to expose input-specific protocols in the
test realm. Given that scene_manager is the only possible scene owner,
and it is also always the input owner when in use, this change removes
use_input under the assumption that use_scene_owner suffices for
expressing the same outcomes. We therefore read use_scene_owner in all
the places where use_input would have mattered.

In reality, this means that tests that used to not expose input
protocols now always will. This should be a reasonable outcome for
scene manager-managed realms, where the input protocols are already provided via the scene_manager_scene_with_input manifest, and the fact that scene_manager sans input is not a real production use case. This should also be a reasonable outcome for scenic-only realms, where the only protocol now being exposed is the pointerinjector protocol.

Fixed: 127469
Test: fx test //src/ui/tests
Test: fx lsc presubmit fxr/858616
Change-Id: I892f47047bad2db76f50c8d36aba36c9fd497fde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858616
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>