[driver] TestSyncDispatcher no default ctor
Removes the default constructor of the TestSynchronizedDispatcher. This
helps us keep usages of this consistent.

Also move the Start/Stop functions to be private since they are no
longer called manually but instead through the ctor/dtor.

This change also improve a lot of the documentation provided in the
header so users are less confused between the two options.

Change-Id: I89f1a7f1ecbd4b646b31992a931ab9c9c1d58b25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859967
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>