[ffx][config] Don't force config files to exist
The code changed here took a bit of a shortcut when opening the config
files by opening them for create+read+write, which makes it atomically
create the file if it doesn't exist. There are two reasons for doing
this, potentially:

- it ensures the path is writeable at all, and makes sure a file exists
  there.
- It avoids having to figure out *why* we couldn't open the file.

The former is an extreme solution to something that's probably not
really a problem in practice, and the latter is not a very good reason.

This makes it so that it just tries to open the file for read, and
returns errors only if the error was something other than "not found".

This is a partial fix for fxbug.dev/127147 -- it changes the unexpected
write to an unexpected read. Another change will make it so that we just
don't access this at all for the build action in question.

Bug: 127147
Change-Id: Idc28202de190581ccce0ae170e43721c0637a989
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860577
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dwayne Slater <dwayneslater@google.com>