[component_manager] Update CPU stats inspect format
We now store the data in arrays. This leads to a much smaller
representation in both the Inspect VMO and the JSON format.

See go/fuchsia-cm-diagnostics-encoding. But as TLDR:

- ~76% less bytes for the measurements in the Inspect VMO
   in Component manager or 50% if we round to the page size
- ~55% less bytes for CPU stats in the inspect.json file
  in snapshots

In preparation for the addition of handle/port counts the data
moves to be under stats instead of cpu_stats.

Fixed: 126219
Bug: b/280805706

Change-Id: I062a39d2111d961dfb0ab7d2f78648c01dec7e13
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857181
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>