[netlink] Send RTMGRP_IPV{4,6}_ROUTE events
Update the Netlink Async Worker to collect clients of the NETLINK_ROUTE
protocol family into a `ClientTable`. The set of clients is then handed
to the "Routes Worker" allowing it to publish events to the
`RTMGRP_IPV{4,6}_ROUTE` multicast groups as it observes routing changes
reported by the Netstack.

Bug: https://issuetracker.google.com/283827082

Change-Id: If50a3c2b0573f4c7eb08bfb07070d8bd53002bec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859858
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>