[starnix] Refactor sysfs
Use kobject and ktype to represent sysfs objects:
- A kobject is the fundamental unit of sysfs. It is embedded in each
sysfs object. A kobject has a name, a ktype, pointers to its children,
and a pointer to its parent, which allows it to be organized into
hierarchies.
- ktype is the type of object that embeds a kobject. It controls what
happens to the kobject when being created and destroyed.

Refactor sysfs:
- Use the root kobject to create a dynamic directory.
- All implementations of FsNodeOps in sysfs should hold a reference
to its kobject.

Refactor uevent:
- registry.rs is where the dispatch happens for now. It should be
migrated to SysFs in the upcoming CLs.
- All device information should be acquired via the kobject. Other
env states should be kept in the registry.
- More uevent actions are introduced.

Bug: 119437, 121327
Change-Id: I79c3995d285666ee17b0aae96699d3b96a5cad0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854028
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>