[msd-arm-mali] Remove use of fbl library
fbl isn't available in the SDK and likely never will be, so copy the
only files from it that we use into msd-arm-mali. Eventually we'll want
to replace it with std::format, but that's not available yet.

Change-Id: I8955e8c2cc9232dd3a2c46b6079c46b8df0bb2a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861517
Reviewed-by: Craig Stout <cstout@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>