[netstack3] Move state out of BoundSocketMap
Move the storage for socket state out of BoundSocketMap and into the
datagram and TCP socket code that depends on BoundSocketMap. The result
is that the BoundSocketMap is used only as a lookup table for
demultiplexing, not as the source-of-truth for socket state. The calling
code needs to be careful to make sure that the address and sharing state
stored for a socket match what's used in the BoundSocketMap.

This will make it easier to store socket state behind references in the
future.

Bug: 126141

Change-Id: I6bf4b03d5ca5fc44ffdd18a970c8f8689c99b62d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855957
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>