[bazel_sdk] Rewrite test suite launcher in Python
This is a rewrite of scripts/bazel_test.sh in Python.

This is done to support the generation of a Ninja depfile
to ensure that `fx build bazel_sdk_tests` is always
incrementally correct.

The //build/bazel/bazel_sdk:bazel_sdk_tests GN action
is updated to use the new --depfile option as well.

+ Update the README.md file to list all possible ways
  of running the test suite.

Bug: None
Change-Id: I65003cf42b1329eeaca7a83ac2a3bb66bae4e852
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851643
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>