[driver][compat] Add logging library to ease migration
Most clients want an easy way to migrate existing DFv1 code to DFv2. By
providing a library which provides the existing APIs with new
implementations we lower the cost of migration. Note that while this
code relies on global, it is no worse than what we do in DFv1 already
and what drivers seem to be doing themselves anyways. As long as the
symbol is not exported into the process namespace, we do not need to
worry about drivers resolving to the same symbol.

We may want to consider switching the main driver library to similarly
install a global singleton for improved ergonomics (while continuing to
retain ability to specify the logger).

Change-Id: I05873b62bcfcda6d75576054ecdf56c62f96a7fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858308
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>