[fxfs] Add fsck check for "zombie" files
A zombie file is one which is present in the graveyard (i.e. will be
cleaned up on mount), but there are still links to the file.

Note that these would already emit a RefCountMismatch error, but this
error should be much clearer and easier to debug.

Change-Id: Icda4e0a2a96fd57f19d32c8a5a0065773f17493b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863256
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>