[fxfs] Expand and rename NextKey trait to support merge options.
Replaces the NextKey trait with the LayerKey trait which provides
options to support different necessary merge strategies as well as the
`next_key()` call.

During `find()` we call `push_iterators` which will stop when no
obviously same or adjacent key is found in the current layer. So if
something is not modified long enough adding it to the mutable layer
will drop the old value. By making merge strategies more explicit now
this behaviour will be specifically enabled for the keys that use it
with the default being the cumbersome FullMerge option.

Change-Id: I19c1ea33e47cb83ce602077391a9a427fb1f10c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853919
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>