[fshost] Remove child matchers for data and blobfs
Remove the data and blobfs PartitionMatchers and instead create an
FvmMatcher that explicitly mounts the blob and data partitions. If the
blob partition does not exist, fshost fails. If the data partition does
not exist, reset_fvm_partition will create it.

This approach makes fshost more resilient as it can now handle states in
which the data partition does not already exist. This might happen, for
example, if a device were to lose power in the middle of
reset_fvm_partition after the old data partition had been destroyed but
before the new data partition was allocated. In the old matcher system,
fshost waits for partitions to match against but cannot detect when an
expected partition does not show up. Now, fshost is more explicit about
which partitions it expects to appear in fvm.

Bug: 125255
Change-Id: I63a9972b240e10d0c5a994ef67b27153666ed3f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854177
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>