[gigaboot] Drain tx queue and free pkt buffers on mDNS broadcast
The UEFI simple network protocol is non-blocking, which means that it
is difficult to know when a packet buffer is no longer in use and can
be freed. Blocking behavior can be implemented by following every call
to transmit with a loop that checks the transmit queue until it is
empty. After the loop terminates, the packet buffer can be safely
freed.

bug: b/281883111

Change-Id: I3842fae81cc2e6c5bfd4821ecafb85bda1a7418e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861525
Reviewed-by: Sergii Parubochyi <sergiip@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Dov Shlachter <dovs@google.com>