[sysmem][svchost] Move sysmem-connector to miscsvc
This breaks a cycle between drivers and svchost, allow svchost to
provide a strong dep to the driver framework, ensuring it outlives it.
This means that we can continue to process stack traces for crashes that
occur in the driver framework during shutdown.

Change-Id: Iadf610b31464455871ab400a4aa54d5293837bcf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863724
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>