[starnix] Add /sys/devices/system/cpu/active
This resolves a flake in tcp_socket_test. tcmalloc would sometimes
look at this file, expecting it to be present. If this happened
between a syscall inside of a test assertion and the point at which
the assertion checked the errno, the errno would be overwritten to
ENOENT (since the file was missing).

Multiply: tcp_socket_test

Fixed: 127944, 128001
Bug: 121327
Change-Id: Ib0843bcd7f0ca3f1cb7629f27f694f928743325e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864136
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>