[libc] Disable function type mismatch check for MmapProtExecTest
This particular test copies code onto an executable page then runs it.
However, when checking for function type mismatches, UBSan will check
for a function signature associated with a function that would be
located 8 bytes before the function. Since we only copy the function
itself and not anything before it, this check will lead to a page fault.
This particular copied function would not have been instrumented anyway and
shouldn't have a signature. I think it should be safe to just disable
the check for this one test.

Bug: 128274
Change-Id: I5f88badaf8156c6cfa7e15bebcce52f64fdb8fef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864343
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>