[starnix] Move linux_root to container
Previously, the `linux_root` mirror of the root file system was exposed
on the kernel component. This design is awkward now that we delay
creating the container until the client actually starts the container.

This CL moves this mirror to be exposed on the container component,
which makes more sense, logically. The CL also renames the mirror to
`fs_root` to avoid confusion.

Change-Id: I58e2d28acdc7336a304fa38c3a8224ec5b7a32c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864359
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>