[netstack3] Add loom infrastructure and test
Add a host-only test suite that uses loom to exhaustively explore the
possible interleaving of threads that are modifying the netstack. This
enables writing tests for atomicity of operations without relying on a
large-scale integration test.

Add a test that asserts the atomicity of packet socket bind operations
as a proof of concept. The test passes, but correctly failed due to a
bug in an earlier version of change
Iadf12d2d9544191cdc680576620951880ab07663.

Bug: 125654
Change-Id: I4c570a485c6780f234461af38c99c55614e6d24d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857419
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>