[botanist] Move flash to ffx experiment level 0 and emu to 1.
When prod builders are switched to enabling ffx with experiment level 0,
flashing will be switched on by the migration of prod controllers to
idling in fastboot. For emu, we'll put it back to level 1 until we're
ready to enable in prod.

Bug: 91040, 95938
Change-Id: Iea07e554400f8d32dea4450baee4f28c156cf867
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864941
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>