[vfs] Refactor VFS to support node connections in one place
Implementing this for fuchsia.io2 will be simplified if there is one
place that handles node connections. This refactors the Rust VFS library
so that is the case.

This also changes the semantics of NODE_REFERENCE slightly to match how
O_PATH works on Linux: the F_SETFL fcntl should fail with EBADF on a
file opened with O_PATH.

This change also consolidates OpenFile and OpenDirectory into
OpenNode. The two differed previously: OpenFile would support async
close. To reconcile this, we run with the synchronous version of close
which forces Fxfs to spawn a task when it needs to. This actually
uncovered an existing issue where Fxfs would not always flush when
dropping references.

Bug: 77623
Change-Id: I74ad2b5684feca99285ad563c115627214ac4eee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862583
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Size-Review: David Iserovich <diserovich@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>