[display-coordinator] Replace manual memory management with fbl::Vector.
This CL brings a few memory safety improvements to display coordinator's
Controller::DisplayControllerInterfaceOnDisplaysChanged(). Previously,
the code relied on allocating C-style arrays on the heap, and passed the
arrays as unmanaged pointers to an async task closure. This CL migrates
the code to use fbl::Vector.

This CL does not impact the number of heap allocations and the use of
AllocChecker.

Test: fx test display-coordinator-unittests
Change-Id: I4beb820e7fbd26d2ca9a90266e38fef5dfc3880e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865025
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>