[fuchsia.io] Add a node protocol
Make requesting the node protocol explicit.

Previously, the intention was that an empty table would indicate this
state, but checking for an empty table in FIDL is prone to error, and we
also need to be able to indicate that only a directory is acceptable (in
order to support Posix O_PATH | O_DIRECTORY).

Bug: 77623
Change-Id: I2d858922e168c9da218964bed73fa013d6a2bddb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863936
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
API-Review: Adam Barth <abarth@google.com>