[fuchsia.paver] Add Controller to UseBlockDevice
Pass the Controller in as well, in preparation for when
fuchsia.hardware.block/Block will stop multiplexing this
protocol.

The controller is not load bearing yet, as there's one
place that still needs to be updated to pass it through.
Most uses of this API can be easily updated to pass the
controller though.

The paver-test needed to be updated to a new
GetNewConnectionsMultiplexed function as there's
still some reliance on multiplexing, but this will
be removed in a followup CL.

Bug: 127870
Change-Id: I42aac87ec09b3a91b8a90ab0fc2bf1584cee5530
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863558
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>