[build][clang_toolchain] Remove duplicate computations.
Remove some duplications in the clang_toolchain() GN template
implementation by defining a `_link_common` scope that is used
to hold results shared by `_rust_link_common` and `_cxx_link_common`.

In particular, this ensures that the strip / gsym / buildidtool
steps are only computed once and in a consistent way. This
introduces a new substitution expression: {{REMOTE_STRIP_PREFIX}}
that is expanded for Rust and C++ differently based on the values
of `rust_rbe_enable` and `cxx_rbe_enable`.

Manual inspection shows that this does not change the generated
Ninja build plan, except for the non-RBE case where a superfluous
space is removed between the link command and the stripping one
(i.e. "&&  ../path/to/llvm-objcopy" becomes
"&& ../path/to/llvm-objcopy")

Bug: 74100
Change-Id: Id5f7c5f49cda009e242d5c6176d045705772a3a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866478
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: David Turner <digit@google.com>