[riscv] Fix gn gen for minimal.riscv64
We are using zedboot_qemu because we _have_ to have a ZBI, but we do not
want a vbmeta. Even though we are building a zedboot, we can ignore it.
It is not currently possible to _not_ build a recovery slot.

Bug: 127665
Change-Id: I1d54867ddd6751c584559c40daa485143bb1cfb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868016
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>