[audio] Remove is_input from the composite drivers API
Removed in favor of the signal processing topology describing the one
or many input and output pipelines that could not be expressed with
this one field for the whole driver. If the driver exposes only one
input or output that is also described by the signal processing
topology.

Bug: 124865

Change-Id: I6b9f527be0fbb7b393fa8bb966e745e4fbc58f26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866686
Reviewed-by: Dale Sather <dalesat@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>