[boot-shim] Introduce DevicetreeBootShim
A DevicetreeItem is a BootShim item where |GetInitMatcher| would return
an instance of a matcher. After this call, it is expected that items are
able to return the number of bytes required to store the zbi item
generated.

The DevicetreeBootShim provides InitDevicetreeItems such that any
DevicetreeItem::GetInitMatcher instance is part of the devictree::Match
called during |IniteDevicetreeItems|.

Test: fx test boot-shim-unittests
Bug: 78675
Change-Id: I66c0456c17952e8250fcf9d4459bbe0f08fbb1b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831538
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>