[concurrent] Rename inc -> inc.h
libconcurrent contains some files which are technically header files,
but are meant to be used in a single specific translation unit in the
environment that libconcurrent is being used.

Generally speaking, these files contain the implementation of a class
which is templated on an adapter class specific for the environment
(the OSAL).  They are meant to be include in a single translation
unit, where the implementation can be manually expanded into the
actual implementations of the functions with the OSAL included.

Until now, these files have ended in a `.inc` extension in an attempt
to distinguish them from normal header files.  Unfortunately, using a
`.inc` extension can prevent things like auto-formatter utilities, and
editor syntax highlighting from working correctly.  So, rename these
files from `*.inc -> *.inc.h`, hopefully preserving the indication
that "This is not a normal header", while restoring the formatting and
syntax highlighting behaviors.

Change-Id: Ie34eb3049603f709e8500a7d422c8a01a9076b4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868881
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>