[archivist] Support sending logs to serial
This change adds support to write to serial in the Archivist.
At the moment, the change is a no-op, since the structured config is
empty. A follow-up will fill the structured config accordingly and
also remove the equivalent now-deprecated structured config for
console.

This change has a small problem I'll fix in a follow-up.
Given that the Archivist is multithreaded, writing to serial when
ingesting a message can lead to interleaving
zx_debug_write calls when a message has to be split across multiple
lines. Resulting in:

```
00009.317 19891:00000 [netstack] INFO: I'm message one, I'm long and I
00009.319 01103:01171 [mdns] INFO: Starting mDNS
have continuation that ended up in a new line
```

To keep this change a bit smaller, I'm tackling that in a follow-up.

Bug: 100486

Change-Id: I72750e0a670a73830180ada8a1f3531c6ee1a681
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867196
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>