[starnix] Add SyscallArg
Instead of using a raw u64 to represent syscall arguments, this CL
introduces a SyscallArg type for that purpose. In a future CL, we will
use this type for the third parameter of sys_ioctl, whose type is not
known until we decide which ioctl operation we need to perform.

Bug: 128697
Change-Id: Ic5cc4841eedc949a74d4dbb616d3eb525878011b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869218
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>