[audio_core] Wrap perftest::ResultsSet in unique_ptr
The smart display hwasan builder is detecting this as a leak. Since this
is used throughout the lifetime of the program and the mem is technically
reclaimed once we exit main, let's just wrap it in a smart pointer to
keep lsan happy.

Change-Id: Ic5119db1426ad83fa55efd84a48202ec6c7de677
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869485
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>
Fuchsia-Auto-Submit: Leonard Chan <leonardchan@google.com>