[fidl] Require explicit unknown interaction keywords
This changes the fuchsia.git build to require all protocols to be marked
open, closed or ajar and requires all methods to be marked strict or
flexible. This is just in-tree and won't affect the behavior of fidlc in
the SDK.

To allow this change all in-tree FIDL files have been updated to have
those required keywords. The current default or closed / strict has been
made explicit so that the API and ABI won't change.

This will allow us to change the defaults later, keeping the ABI and API
unchanged.

Bug: 88366
Change-Id: I76add0f75a792d6d1c0b4c41ed63c3cff083bfbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868166
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>