[netstack] Increase send-ready timeout in test
Increase the timeout on waiting for send-ready after calling connect
used in the TestBridge unit test from 1s to 3s. The test expects
send-ready to be asserted when the bridge is up within the timeout,
and relies on the timeout occurring after the bridge is torn down to
determine that the connect fails. When the test is run in infra,
sometimes it takes more than 1s for the connect to succeed and
send-ready to be true, which causes the test to spuriously fail.

Multiply: netstack-bridge-gotests
Fixed: 128191
Change-Id: I1b1b4678c85ff2c81cf489bdea1aef90597cdd36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869676
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>