[devicetree] Generalize PropEncodedArray
Initially introduced RegProperty, which is an instance of a
PropEncodedArray.

A PropEncodedArray is an opaque array of uint32 Big endian cells. Each
element of the array is a fixed number of cells. We introduce
PropEncodedArrayElement as well to model these elements.

Each element of the array is a N-Tuple, where each element of the tuple
has a fixed number of cells. The PropEncodedArrayElement<N> represents
this N-tuple, where N is the number of elements in the tuple.

This generalization is introduced to support other PropEncodedArrays
such as 'ranges', 'interrupts', etc.

Bug: 786875
Test: fx test devicetree-tests
Change-Id: I0a12e96fc7c1985033257eb110a2afd230fb4056
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836641
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>