[intl][tests] Create a fake intl_services package
Sometimes it may be useful to have a fake intl_services
which is self-contained for testing.

The real service depends on the `setui` service, which can be
heavyweight to set up in tests.

With this change, we expect that tests that 'just' need
to have some locale available for the entire lifetime
of the test, should find that locale to be always
available.

Change-Id: I025a69e20b326f1ff88b4d5fb985abb43fbdca22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870127
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>