[starnix] Remove eager vmo copy in mm snapshots.
This CL modifies memory manager to avoid doing eager vmo copies during
snapshots. Instead, the new code replaces both parent and child vmos by
snapshots.

During `clone()` and `fork()`, starnix's Memory Manager performs a
`snapshot_to` operation, which replicates the memory mappings into a
different memory manager.

For shared mappings, we keep the same VMO. For private, non pager-backed
mappings, we can take a snapshot of the vmo, for read-only pager-backed
vmos, we can keep the same VMO.

For writable, pager-backed, private vmos, we need both vmos to behave
like a snapshot, but currently that's not supported in Zircon. We need
changes on each vmo to not be reflected on the other.

The previous code used a `ZX_VMO_CHILD_SNAPSHOT_AT_LEAST_ON_WRITE` child
and eagerly write over all its pages, to force a copy-on-write copy.

This causes extra memory consumption, as each of the mapped pages would
be needed to be copied, and that happens from child to child, on each
fork (so if we nest two forks, we end up with 3 copies of those
mappings).

This CL changes things around: the reason why we need to do an eager
copy is because `ZX_VMO_CHILD_SNAPSHOT_AT_LEAST_ON_WRITE` on
pager-backed vmos will show changes that happen on the parent vmo, until
we force a copy-on-write operation. If we also replaced the parent vmo
by a `ZX_VMO_CHILD_SNAPSHOT_AT_LEAST_ON_WRITE`, both processes will have
a copy-on-write copy of the underlying vmo. This behavior is similar
to vmo snapshots, but maybe less efficient.

Bug: 123742
Change-Id: Ibb0cf287f17eb27f0abf69deae207dfaa6b83c49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868296
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>