[fake-display] Add a config parameter to allow/disallow image readback.
For fake-display driver, if it supports capture, all the framebuffers
and capture buffers must be of linear format, and inaccessible heap
must not be used. This is not supported on some of the GPU platforms,
for example, goldfish-vulkan (AEMU).

This adds a new feature flag to FakeDisplay which can allow / disallow
image readback (capture). For fake displays without capture feature,
they should accept any format of images and images allocated on any
heap. While for fake displays with capture feature they should only
accept image format that is friendly for composition within the display
driver.

Note that this is a short-to-mid-term workaround; instead of offloading
the GPU complexity to the display stack, in the long-term future, tests
should consider using software-based vulkan implementations so that
there'll be no platform or device capability differences. Tests will be
always allowed (and preferred) to use captures in that case.

Bug: 128891
Test: fake-display-test
Change-Id: Ie685378f4cb7877c6bcbbd93e1af9aa64b7343e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869480
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>