[fxblob] Make target image size optional
Allow setting target image size to None, which means "make it as large
as needed".  In practice, use a similar heuristic that was used for FVM,
which is to make the image 2x larger than the base contents.

This is useful for emulators where there isn't practically a limit to
enforce and it is tedious to maintain an artificial limit.

Change-Id: Iba5eaf3fa93d0771eacf01cbbef8b057c4933407
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871996
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>