[storage-benchmarks] Support Fxblob-based systems
When there is no FVM available, the storage benchmarks will attempt to
create an FVM in the system's GPT and use that FVM for benchmarking.
This supports the Fxblob use case (since there is no FVM there).

The test will not fail if both the FVM and GPT are absent, but no test
cases will run.  An error message is printed here explaining how to run
these tests on an emulator (which is useful for local development and is
typically the cause of neither an FVM nor GPT being available).

Fixed: 126938
Change-Id: I2356d93b7e0b1e423177d8f00470166ccaf15d71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868941
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>