[boot-shim] Template UartItem and UART_ALL_DRIVERS
This change is meant to allow boot-shim-tests to be compiled with
UART_ALL_DRIVERS=1.

Before this change, when the test would be compiled with
UART_ALL_DRIVERS=1 it would crash due to ABI incompatibility between
what the test think uart::all::Driver is and what the library sees.

By templating over uart::all::Drivers, the mangling (even as the alias
uart::all::Driver is the same, the underlying type is different) is
different, so the generated functions are different and the template is
instantiated with the correct types.

Bug: 129047
Bug: 129046
Test: fx test -v -o boot-shim-unittests
Change-Id: Ib49cf8e7462171e33e2e9b01e6947053dcbf310a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870897
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>