[starnix] Refactor execve arg reading.
This CL adds a method to memory manager to read a C string into a
vector. This is intended to be used for reading strings that can be very
large.

The other alternative, `read_c_string`, required the entire buffer to be
preallocated all at once, and copied all that memory each time.

The `execve` `argv` and `envp` parameters can take up to 32 pages each,
which was not a good fit for `read_c_string`.

There is still some room for improvement to avoid extra copies and
accept uninitialized data, but out of scope for this CL.

Change-Id: If347557e2dcba93969dcccbffef0e5c4d659205a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871261
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>