[starnix] Add vdso tests
These tests match the vdso behavior in Linux: almost all memory mapping
operations are allowed on the vdso, as long as they don't split the vdso
mapping.

Currently in Starnix the vdso is only one page, so that test will be
skipped. In Linux it passes.

The CL also adds tests that make sure that the vdso is not writable
across process boundaries (vdso modifications are not visible after
fork or after execve).

Bug: 128842
Change-Id: Ic57dc52a9367d775bae579afcd2c887b4b85be7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870718
Reviewed-by: Maria Glukhova <mariagl@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>