[starnix] Refactor async dispatcher in remote_binder
1. Read the fidl command concurrently instead of sequentially.
2. Await all FIDL related command

This ensures that all fidl related tasks are awaited on the local
executor and allows to remove the cleanup method from the state.

This also solves a crash in the executor happening when the fidl task
was sent to the client task thread to execute the binder command and not
awaited.

Fix: 129211
Bug: 129183

Change-Id: I2a0bf77771eb1337e0c8a8dead61f3d2cb5281ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872496
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yifei Teng <yifeit@google.com>