[starnix] Delay service requests until UTC clock starts
This waits for the UTC clock to start running before Starnix starts
serving requests and instantiating containers. Many Linux waiting and
timing APIs default or only provide operations in terms of the realtime
clock and do not function in a reasonable way when it is stopped. The
UTC clock is typically started by the timekeeper service early in boot,
often before Starnix itself can start, but in some situations the
startup order can be reversed leading to surprising behavior. To meet
the general expectations of the Linux world this just avoids doing
anything at all until the clock is going.

When Starnix is waiting for the UTC clock to start it emits a WARN level
log. Most of the time Starnix will not block at all and not log
anything.

Fixed: 126111
Change-Id: Ic5e5e6c0506752df6e06f4d78cd2e61bc702ce1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871561
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>