[fxblob] blob_writer stops splitting writes if remaining bytes fit
The blob_writer client used to split all writes so that none would be
larger than half the VMO to prevent filling the entire VMO with a
single write (in which case the server would need to wait for a FIDL
round trip in between processing writes).

The blob_writer client no longer splits writes if the VMO is large
enough to contain all the remaining unwritten bytes.

Change-Id: I74c3a8fe0bc5910bdc90198a1a50e9d978fba87b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872725
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Nikita Jindal <nikitajindal@google.com>