[starnix] Fix remote binder task creation
This fix remote binder task creation.
Before this CL, if a request was received while a task was considered
unassigned, and an unassigned request was available, it was possible to
take over the unassigned task, but no request a new task for the already
present unassigned request. This lead to deadlock where some unassigned
task where never handled.

This CL fixes the issue by not adding a task to the unassigned pool is
there is an unassigned request. Instead the task is assigned to the
request. This ensures that if a new task is added and an unassigned task
is present, there cannot be any already present unassigned request.

Fix: 129183
Change-Id: Iff1c7385ae7591ef4c25a4f4c01b0ddc1ccc6bb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873237
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>