[fidl][rust] Stop using macros in bindings
This CL moves FIDL encoding/decoding logic out of Rust macros and puts
it directly in the fidlgen_rust Go templates instead. This should make
it easier for people to understand and contribute to the bindings, since
there is only one layer of code generation instead of two, and you only
need to learn one templating language instead of two.

One downside is that encoding/decoding clutters up the generated file
and people usually don't care about it relative to the public API. To
mitigate that, all this code is emitted separately at the bottom.

Test: fx check-goldens fidlgen_rust
Test: fx test fidl_rust_conformance_tests
Fixed: 124339
Change-Id: Ief6f0b70e5a77da417e03598a91f3ada2975ed40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873941
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>