[starnix][kernel] Move most capabilities to optional besides Vmex
Concretely, what this does is that the parent of the starnix kernel
component could route void to these optional capabilities. This would be
useful if a test doesn't exercise that capability, and would like to
avoid routing warnings.

If starnix did use a capability and it was offered from `void`, that
will result in a DEBUG level log which is hidden in tests most of the
time. The operation that depended on that capability will most likely
fail.

Change-Id: If688322e8b9dcbfd6376088c2333333e2c816e58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875463
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>