[net/mdns] fix cache flush implementation in instance requestor
This CL fixes the cache flush implementation in instance requestor,
which previously cleared all A/AAAA resources for a given name from
its cache when it receives an A or AAAA resource with the cache
flush bit set.

This change causes the requestor to flush all *matching* resources
(A's for and A resource, AAAA's for an AAAA resource) less than one
second old, as specified in RFC 6762.

Fixed: 129021

Change-Id: I86380e25005e8406e61ef980908c9bc1a659288f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872186
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Michael Spang <spang@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>