[ffx] extend delay when killing daemon
If the daemon doesn't die cleanly in tests, we kill it then make sure
it actually died. But the delay of 50ms apparently isn't enough when
racing with systemd to reap the child process, (causing spurious flakes
in some tests) so let's increase it to 250ms.

Bug: 125459
Change-Id: Iab86b70f91690cacc34bccc5f4a85fd6643aa7cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875542
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>