[wlantap] Ensure child classes are destroyed in order
WlanPhyImplDevice now ensures that WlantapMac is destroyed before
WlantapPhy is destroyed because WlantapMac keeps a pointer to
WlantapPhy, which may result in a use-after-free if they are not
destroyed in the correct order.

This also uses a ServerBindingGroup in WlantapMac to ensure that FIDL
connections are closed when the class is destroyed.

Bug: 129563
Multiply: hw-sim
Change-Id: I4c49b051d85acca65c46bc4e031522709afe2c3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875836
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: John Tan <johntan@google.com>