[netlink] Cache interface `Control` handle
...instead of getting a new handle for each interface update request
(like adding/removing addresses).

This change holds the `fnet_interfaces_ext::admin::Control` handle with
the rest of the interface's properties when the control handle is first
obtained. This avoid handle thrash.

A side-effect of this change is that requests modifying interface state
will now fail if the interfaces eventloop is unaware of the interface
(has not received a `fuchsia.net.interfaces/Event` for the interface).
This is okay because if the interface was actually installed, the
eventloop will eventually learn about the device and until then, it will
act like the device has not yet been installed.

Change-Id: I17f17353f495e540f6328f941168dacad2ead448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875956
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>