[gvisor_syscall_tests] Skip variants of TCPListenCloseDuringConnect
The two variants are flaky in the absence of dual-stack sockets. The
following can happen:
Assume that the listener binds to a v6 address with port p. Then the
connector connects to v4 address with port p. Now this normally triggers
a RST since no one is listening on port p on the v4 address, rendering
the test to pass (though not intended). However, chances are that port p
is used by some other listener in other tests running, then our test
won't pass as the test is expecting a RST. Besides, the fact that the
test iterates 100 times makes it easier to trigger the flaky case.

Bug: 129660
Change-Id: I86bef5a40d529ecda82ca06281699e5bf4f7b7bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877583
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Brown <nickbrow@google.com>