[paver] Add BlockDeviceClient
Add BlockDeviceClient which represents a block device.
Having `block_fd` on PartitionClient was not helpful, as
most PartitionClients were returning invalid file descriptors.

Getting a BlockDeviceClient from a PartitionClient can return an
error if it's not the right type of client. This is better than
the asserts and casting that was happening before this CL.

Bug: 112484
Change-Id: I759e1fa94d676ad60f483bb9724827103a3995fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877776
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>