[fshost] Remove time-out from find_data_partition
We shouldn't need the time-out for finding the FVM data partition. By
using VolumeManager's GetInfo method, we can be sure that all the
partitions will be present in the topological tree.

This now means that if no data partition is present, we don't have to
wait for the timeout to expire to find out.

Fixed: 129098
Fixed: 129699
Change-Id: Iceb8ac997e1a5dc16e7f0222acf60d43267981e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877236
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Commit-Queue: Chris Suter <csuter@google.com>