[sysmem] Fix buffer collection async trace events.
Fix the labels of the async end events to match the begin events so that
the async events terminate and accurately reflect the allocation times.

Also add a trace duration for `OnBuffersAllocated` to make it easier to
locate and measure the activity related to allocation completion.

Change-Id: I45b3aef0d597c665354f9f72fb0fdb3a04ff5402
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875505
Commit-Queue: Corey Tabaka <eieio@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>