[bedrock] Delete many unused types and libraries
We have built a lot of capability types and execution traits as an
learning exercise. However, they're not used in the short term when
refactoring ELF runner to use bedrock, so keeping them alive and working
is just a drag force.

`multishot::{Sender,Receiver}` will be dealt with in the next CL because
it's used extensively in serve_processargs tests. But we'll be replacing
it with `Open` at least in the short term.

We can always resurrect them if we find that we do need some later.

Bug: 122024

Change-Id: Ib4833893ba87203ac47689466104ab449ef19681
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878915
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>