[sdk] API Review for driver mirror libraries
API Review to put the following libraries into the partner sdk:
 - driver_incoming_cpp
 - driver_outgoing_cpp
 - driver_logging_cpp

These libraries are the driver counterpart to existing component
libraries. They are mirrored to be as similar as possible to the
counterparts.

Due to current limitations in how drivers are run, these libraries
are not a part of the component versiones of the libraries.

For incoming/outgoing we provide driver transport variants of
everything and regular components do not need to bring in the
driver transport fidl dependency. They also do not use globals like
their component counterparts as drivers can be in the same process.

For logging, the same problem of globals not being possible
currently apply.

Change-Id: I93cfdca72ea9f4356a8b882ceeb659afa167fd0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861332
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Size-Review: Aidan Wolter <awolter@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>