[assembly] Only rebase ui.sensor_config path if not already rebased
This is a temporary check for a transition.

If the ui.sensor_config does not start with '../../', then
rebase the path.  If it does start with that string, then
assume that it was already rebased.  This is so that the
usages of this (in vendor repos) can switch to passing the
file in as an already-rebased path.

Bug: b/289396131
Change-Id: I657ee1ac96e7f6778d4a56def9b0065ede0b3795
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879711
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>