[starnix] Implement `ioctl(_, SIOCGIFHWADDR, _)`
This `ioctl` returns the link-address for an interface.

The implementation simply translates the requests to the NETLINK_ROUTE
implementation with a temporary socket so that there is only a single
entity in starnix interacting with the Fuchsia netstack.

Fixed: 129987
Bug: b/286855878
Change-Id: I7aece9e640573584b488f441db10cee60bd07e26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879691
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>