[bsdsocket_test] Longer ReadAfterShutdown timeout
Increase the receive timeout on the read-after-shutdown from 1s to 2s
because the call to `AssertBlocked` which asserts that the recv call
is blocked but not yet timed out for 50ms can sometimes take so long
to run in infra that the receive timeout is hit.

The increase is conservative because all test cases where the socket
is still readable after shutdown will take the full timeout duration
to complete.

Fixed: 130000
Multiply: netstack2_fastudp_bsdsocket_test
Change-Id: I0192657dc262e34413206cdfd8cb16206be6c55c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880975
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>