[assembly] Remove bundles:debugging
This target is moved into assembly for all eng/userdebug builds.

pistress got moved into bootfs instead of being a shell package, so that
it can still be used for kernel development. Previously, it was added to
//bundles:debugging, which in turn was added to bootfs_labels, but
because it was a shell package it was never added to any products. We
add it as a dep to a platform AIB to ensure it does not bit-rot, but we
do not include it in any products by default.

Bug: b/288301936, b/288301471
Change-Id: Ia153c49b005444a16ffac93641857b4a1467a501
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873816
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Size-Review: Aidan Wolter <awolter@google.com>