[netlink] enable/disable links via RTM_SETLINK
Allow NETLINK_ROUTE sockets to enable and disable links by sending
`RTM_SETLINK` requests.

At the moment, only disabling links is synchronized with the events
observed on the interface watcher, as it's not currently possible to
synchronize enable requests. This is because the interface watcher only
reports whether the interface is online, which is a combination of if
it's enabled, and if the underlying link is UP. This will be addressed
down the road if it's determined to be an issue for clients of
NETLINK_ROUTE.

Fixed: b/283136220
Change-Id: I04b5ed4425dfe0895b2c03fff038f5e4f86a333d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880052
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>