[netstack3] Move handshake status tracking from Bindings to Core
Core is now the source of truth for the status of a connection, we no
longer need the `on_connection_status_update` method on the non-sync
context. This will help merge TCP socket IDs since the bindings will
not need to remember the type of the socket.

Fixed: 130212
Bug: 126141
Change-Id: I0808db16af0060edf7bf4a612911cc9c8efd48be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881819
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>