[ffx][fastboot] consider boot timeout a success
When issuing a "boot", "reboot", or "reboot-bootloader", it's
possible that the "okay" response gets lost in a buffer. So
inside of consdering a lack of an "okay" in those cases as a
a timeout failure, consider them successes.

Also, since we are now considering a timeout to be an expected
result, let's timeout in only a few seconds instead of waiting
30 seconds.

This change includes a fix to fastboot::read_with_timeout() to
actually use a timeout, rather than blocking indefinitely
and checking the timestamp after the fact.

Fixed: b/290218326
Change-Id: I32f0fc5cb423dfd20fed3ee569bc63ba5e3a8db8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882214
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>