[component][fuzzing] Make realmfuzzer target library optional
This CL makes realmfuzzer's target library treat a failure to connect to
`fuchsia.fuzzing.CoverageDataCollector` as a non-error. If an executable
is linked against the library but run outside the fuzzing environment,
it will be unable to find a destination for its coverage data but should
still operate normally.

Change-Id: I6fedf3027a0b765934a411657d73e5c2de480704
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810867
Reviewed-by: Dzmitry Yatsushkevich <dmitryya@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>