[security] Rename "minfs" directory policy to "data"
The "minfs" directory was renamed to "data" in change
Ic958469b022ea86877ddf464937a6544b54ca913

This disabled the allowlist for `data`, so any component that had
a route for the capability could use it without restriction. Renaming
the directory to `data` in the allowlist fixes the enforcement.

* Allow `/core/session-manager` to use `data`, required as of
  change Icd9c40d876cd4ef443abbef32adf64a9052daa99

Bug: 130281
Change-Id: I434552723bbd08302b9bef6066fddbca54827a60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882164
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>