[fidl] fixing fidl-fix compilation
This fixes the fidl-fix compilation by adding the version_selection
parameter to the call of fidl::fix::Fix constructor.

This also refactors the processing of the command line so that
tests can be written to exercise the command line processing.

The fidl-fix tests are part of the //tools/fidl/fidlc/tests label.

Bug: 130019
Change-Id: Ifb6c8e46883d2056269ae61ac756a15b9e974652
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882892
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>