[kernel] InitInterrupt take enable_interrupt_callback
To match previous implementations order of operations,
InitInterrupt takes a callback that is in charge of unmasking the irq
the device is assigned to.

The main purpose is to match the legacy driver operation sequence
as much as possible. See fxbug.dev/129378 for details on a tracking
bug for revisiting this decision after migration is complete.

Bug: 126785
Change-Id: I3b078c89f983a69acb91c7e0aaab7f4a8e85b0f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854018
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>