[kernel][vm] Handle zx_process_{read,write}_memory crossing the end of a VMO smaller than its mapping
Zircon allows mapping a VMO into a virtual range which is larger than
the VMO itself. This CL changes the way we handle
process_{read,write}_memory requests that cross the end of the VMO in
such a mapping: instead of returning an error, we now partially fulfill
the request until the end of the VMO.

Trying to operate on a memory range which lies completely past the end
of the VMO still results in a ZX_ERR_OUT_OF_RANGE error.

Change-Id: I310517f00a3fa5c14877cee8c030b8a339a93099
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881941
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Fabio D'Urso <fdurso@google.com>