[installer] move core installer logic into a lib
We want to add the installer logic to userspace fastboot, so we're going
to extract it into a library that both the installer UI and fastboot can
call.

No functional changes here, just creating the new library and a few
small fixups to pass build and test.

Bug: b/282797139
Test: installer USB still works on NUC11
Change-Id: I8ac90c5729787122be18129916cf8a7eb8dd7baa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868185
Reviewed-by: Sergii Parubochyi <sergiip@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: David Pursell <dpursell@google.com>