[kernel][vm] Update gen count in VmCowPages::fbl_recycle
Move the hierarchy generation count update from ~VmObjectPaged to
VmCowPages::fbl_recycle. We drop the VMO lock between the two and the
actual page list updates happen in VmCowPages::fbl_recycle, which is
what affects the attribution counts. So it's possible for a caller to
sneak in an attribution count query between the two lock acquisitions
and cache a stale attribution count, which will never be invalidated
since the generation count won't change.

Added a test that reliably fails without the fix.

Test: core-vmo-clone2
Bug: 129800
Run-All-Tests: true
Change-Id: I50de8823734c480b5f443a8534444eac2c2fca42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882220
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>