[wlan][policy] Fix add_phy e2e test flake
When the test adds PHY, most of the time it can expect a
`DeviceMonitorRequest::GetSupportedMacRoles` message right away.
Occasionally, however, the call to add PHY may wait to acquire a lock on
PhyManager before any request can be made to DeviceMonitor, causing a
flake because the test expects a `DeviceMonitorRequest` message when
there's none.

This commit adjusts the test expectation to try up to 5 times if no
`DeviceMonitorRequest` is found right away, with a sleep of 1 second
in between.

Fixed: 130436
Test: CQ
Multiply: wlancfg-tests
Change-Id: If710df194617eb54033eb8120045339c273005fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883122
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Commit-Queue: Kiet Tran <kiettran@google.com>