[docs] Replace symbolizer markup spec with a link to LLVM's
LLVM's copy of the spec should be considered normative. I've kept the
original spec document around. It contains only a hyperlink for now, but
the format is extensible, and this would be a good place to document any
extensions that are intrinsically Fuchsia-specific or that haven't yet
been incorporated into the upstream implementation.

Fixed: 129373
Change-Id: I3a5132dfda82ac9a26098edefea15c7a13caeee5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883111
Reviewed-by: Dangyi Liu <dangyi@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Daniel Thornburgh <dthorn@google.com>