[bedrock] Implement Remote for Open and use it to serve the namespace
After converting Dict to Open, now we implement Remote to get a
fuchsia.io/Directory connection.

The meaning of Remote for Open is to open the capability and return a
client endpoint.

The primary advantage of this design is that both Remote and TryIntoOpen
now has a purpose:

- TryIntoOpen tries to convert a capability into Open, which integrates
  with fuchsia.io. This is a place for us to put a variety of fuchsia.io
  logic, such as node name validation.
- Remote convert a capability into a handle. In case of Open
  capabilities, it will always open the capability with semantics
  determined by fuchsia.io

Change-Id: I2016914fa509323407efb883741994dc111851ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881351
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>