[inspect] Route InspectSink to connectivity
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

The change touches so many files as the offer shard uses `offer InspectSink to "all"` syntax, which is viral when in included in shards. Anything that includes one of the connectivity shards at any depth will end up with an offer of InspectSink from parent to each child and collection.

As a general rule, that would be an SDK compatibility issue for OOT users (eg with RealmBuilder shards), and make it impossible to include many shards in CMLs that need to do custom InspectSink routing (eg in some Archivist stuff), so the automated tooling did not insert offer shards into other shards.

Bug: 93344
Change-Id: I3a310a64e5c20449bc8fe20b265a8663e6aea28f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881452
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Clayton McCray <claytonmccray@google.com>