[fdf][ddk] Add zx_device_t* parameter to get_root_resource
This allows us to lazily fetch the root resource and avoid using a
global to store the cached version of it in case the driver requests it
multiple times. This is overall more efficient and allows us to avoid CF
logging warnings every time a failure occurs due to a proactive request
that inevitably fails.

A new global was added to the x86 acpi board driver to route the root
resource to code which requires it. Piping it through the acpica code
would make it more challenging to keep up with upstream whenever we sync
to newer revisions of the library.

Test: CQ
Change-Id: Iba934621ab51c0eb07923142ed5b779e989ae273
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862216
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>