[installer] transfer ownership to firmware
None of the current installer owners work on Fuchsia anymore; for the
time being transfer ownership to the firmware team.

I was planning on using the higher-level OWNERS for a while longer, but
it turns out there are a handful of other projects that include this
OWNERS file, and that type of indirect ownership does not fall back to
parent directories, leaving those other projects essentially ownerless:
* //build/images/installer
* //src/sys/live_usb
* //src/sys/lib/payload_streamer

So we might as well just transfer ownership now.

Bug: b/282797139
Change-Id: Icf182f8dd32ca35e1fe0bc1194a941f53ddfb62f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884051
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>