[installer] move do_install() to library
do_install() performs the top-level workflow of copying images from one
disk (USB) to another (local storage). Moving this into the library will
allow both the installer USB and userspace fastboot to use this
workflow.

The big trick here was that the installer supports status callbacks so
that the caller can be notified as various actions are taken and report
them back to the user (e.g. display them on screen or send fastboot INFO
messages). Previously these callbacks were owned and moved around, but
this didn't work in the library because it needs to send the callback
multiple times, once for each partition copied.

To fix this, the API now borrows a callback reference only during flash,
rather than owning it and keeping it around. This matches the current
usage better anyway; the current user (just the installer) gives a new
short-lived callback each time so there's no need to store it long-term
internally.

Bug: b/282797139
Test: NUC11 installer USB still seems to work
Change-Id: Ibca1b253e6b24e54f40c2338bbdf510a643a675f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883720
Reviewed-by: Sergii Parubochyi <sergiip@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: David Pursell <dpursell@google.com>