[elfldltl] Do not zero beyond the VMO size
Fix a bug in VmarLoader::MapWritable which would zero past the last page
of the VMO if the size was page-aligned. Without this fix, the
elf-search test fails with fxr/883224 that enforces out-of-bounds check
on the specified range instead of trimming it to fit within the VMO
size.

Bug: 130470
Run-All-Tests: true
Change-Id: Id2a41bb79853572ba40cec97348a5669d3968f56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883542
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>