[assembly] Move default mdns config to assembly
The products are still adding private mdns configs using GN, but those
configs are added to a separate destination and merged at runtime. See
the bug for more details.

An AIB is added to add the mdns service for wired udp so that the device
can be discovered by ffx.

This AIB is not enabled on user builds. It is added by default on
userdebug/eng builds, but can be disabled explicitly by the product
config.

Bug: b/290984622
Change-Id: Ic8ecf4cbe423095023c12ffb67aa07133f13b2d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883632
Reviewed-by: Dale Sather <dalesat@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>