[gn][sdk] Use the prebuilt name in sdk_pkg
This fixes an issue with sdk shared_libraries where the name of the
shared library .so was different from the sdk atom. This is the case
when the sdk_shared_library has explicitly stated a sdk_name that is
different from the gn target_name. This change forwards the .so name
into the generated fuchsia_sdk_pkg from the meta.json.

Change-Id: Idf095a690194fed5aeb398f1b54e2fbf191e0261
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883580
Reviewed-by: Jerry Belton <jcecil@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>