[lowpan-ot-driver] Address race condition when updating meshcop DNS-SD.
When updating the DNS-SD records for meshcop, there is a chance that
the updated service information could be registered before the previous
information was fully removed. This commit addresses this by waiting
for the previous registration to be fully removed before adding the
new one.

Bug: b/289240101
Change-Id: I19fe8df0b6e25afd4660a38da8d47b07a6f3a1bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884331
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>