[boot-shim] Resolve UART IRQ
Requires one extra scan if there is no interrupt controller within the
uart's ancestry. In that case it will store the phandle of the interrupt
parent in the ancestry, and look that one up.

Then the specific translation is performed based on the compatible
string of the interrupt parent.

Added GIC and PLIC interrupt translation.

Bug: 129729
Change-Id: I57f7b795fc4507549ddcb6c3c89a4585bb5108df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881631
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>