[codec][aac] Add encoder test hashes for new encoder
After the newer encoder roll, the encoder will generate the same output
on both x64 and arm64, so we'll then be able to remove 2/3 of these
hashes.

Generally lossy media encoders are not required by spec to generate
any particular encoded data, being free to optimize as the encoder sees
fit, so it's not surprising that a newer encoder generates different
encoded data.

BUG: fxbug.dev/128466
Change-Id: Ief2d9515be01b72dee8ede27030a6de182793e12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884871
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>