[ld] Plumb processargs protocol
This makes the Fuchsia dynamic linker startup interpret the
actual (legacy) processargs protocol.  It uses this to wire up a
debuglog or socket (from fd 2 in the message) for logging.

This includes preparing two new pieces of test harness.  First,
TestLogSocket is a simple collector for the socket given to a
test for its diagnostic logging, so it can be matched against
expectations.  Second, TestProcessArgs builds and sends the
processargs message, letting the test precisely control the
details of what goes into it.

Bug: 130483
Change-Id: Ic657090b0a81dc7bb95ca81176c2c535382999b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884247
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>