[starnix] Loosen what waitpid test records
The test involves checking to see that a SIGUSR1 was
appropriately delivered when the child exited. We see a flake that
looks like the handler got more than one signal, and that
one of the signals was a SIGCHLD. I hypothesize that SIGCHLD
happened because of a suspend, rather than an exit.  This CL
ignores signals that happen for a reason other than the child
exiting.

Bug: 130640
Change-Id: Id7ff1833c42d0dd880f001945b2b511f5b798517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884701
Commit-Queue: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>