[fake-display] Add CaptureImageInfo class; Use DriverCaptureImageId to identify capture images.
FakeDisplay used to use ImageInfo for both display images and capture
images, and refers to imported images using raw pointers. This is
prone to use-after-free errors and could cause images of incorrect
type to be used.

This change adds a new class CaptureImageInfo to store metadata for
capture images. Also it adds an `id_` field with type
DriverCaptureImageId to identify the image, which will be the value
of returned capture_handle in ImportImageForCapture().

This change doesn't change any invariants (explicit or implicit) on
validity of imported images.

Test: fake-display-tests
Bug: 128706
Bug: 128653
Change-Id: I41a46eab28cff6585a01ad3afd2233928c074fbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882471
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>