[fxfs] Add dirent caching to LSM tree.
Wraps both dirent objects to reduce the memory from the full
ObjectValue enum. This required changing the type signature, but
won't affect the actual serialization.

The cache lives on top of the whole tree, which means that it may
duplicate items in the mutable layer but also that there is no
resolution required when sealing.

The cache is only used for volumes, other lsm tree instances in Fxfs are
given a `NullCache` that essentially does nothing.

Bug: 128188

Change-Id: I585a00e558a2ca1e8425343e68abb3c8deb35d68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875681
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>