[fxblob] BlobDirectory::lookup now checks for CREATE before WRITABLE
This change is in anticipation of removing support for the old
fuchsia.io write API in fxblob. Pkg-cache relies on create_blob()
returning ALREADY_EXISTS to know whether or not to write a blob so when
create_blob() calls lookup with the CREATE and WRITABLE flags, we want
lookup to fail with AlreadyExists and not AccessDenied from the check on
the WRITABLE flag.

Change-Id: I1ee8593fec3a80d4ac5dde16abf6d5751c43bd22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886711
Commit-Queue: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>