[vm][test] Rewrite Pager.VmarRemapTest
This test overwrites a VM mapping from one pager-backed VMO to another,
and verifies that blocked threads see the new contents when unblocked.
Rewrite the test to use a mapping that the test itself creates, instead
of relying on the one created by the Vmo instance. This avoids using
UserPager::ReplaceVmo().

Remove UserPager::ReplaceVmo() as this was the only test using it.
ReplaceVmo() would replace Vmo::vmo_ under the Vmo's lock, but a caller
might have a const reference to the vmo_ outstanding at the time, which
could lead to data races.

Test: core-pager
Bug: 129315
Change-Id: Ie4088388c6b62ee3fe6665153fc2d73bf5aed303
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886413
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>