[ffx][config] Add flag to ffx for config domain root
Implements `ffx --env-root <path>` which points at a path to search for
a configuration domain root in. It also adds support for using a
configuration domain with an isolation root, though the exact parameters
of that are subject to change.

`--env-root` is currently hidden as the interface is subject to change,
and this is being implemented mostly for dogfooding to find the right
tradeoffs here. The flag may be merged into `--env` later.

Bug: b/290946533
Change-Id: Ibbdbc144c73d6c96638164634904f7e93f90647b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887227
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>