[network/tests] Run dhcp client against NS3
https://fxrev.dev/843898 retained the hardcoded `Netstack2` when
creating a server realm, because NS3 didn't support receiving packets
with the unspecified source address. Now that it does support this, run
the tests with NS3 in both the server and client realms.

Since NS3 doesn't currently let us bind to the broadcast address, amend
the test source to bind to the unspecified address.

Bug: 81593
Bug: 122609
Change-Id: I7592d1c56d8c2eed42cee8f6758f7af1141a77ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887831
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>