[netlink] Synchronize route sink multicast
Refactor FakeSenderSink to be backed by a Sender/Receiver pair instead
of a vector. This allows clients of FakeSenderSink to fetch all
available messages, even if they are not available immediately, as long
as they will be ready.

Thanks to @martinjeffrey for helping to come up with a synchronous
approach to this problem so all the callsites of `take_messages` did not
require an update.

Fixed a race in the RTM_NEWROUTE test that occurred when attempting to
check for a new message immediately after adding a new route. The test
now blocks while waiting for the expected messages to arrive.

Fixed: 130548
Change-Id: Ib2951427d9eda08812b46e05cef7693e3eb702f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884054
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>