[storage][watchdog] Allow longer timeouts on watchdog tests.
Occasionally we get flakes on the watchdog tests since it sets a timeout
of 1 second then waits 2 seconds to see if the expiry call ran. This
adds a looping poller to give up to 10 seconds for the expiry call to
schedule and run rather than just 1, but polls every 100ms to keep the
runtime of the test as low as possible.

Bug: 130876
Change-Id: I74ba6303acbc45f238033ac473d3d66106faf101
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888231
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>