[driver_manager][v1] Resolve packaged boot-drivers
Driver manager (v1) attempts to resolve a boot component url from the
boot component resolver and return a proxy to the boot component's
package directory.

If the resolver returns a "Not Found" error, then Driver manager will
fallback to the legacy code path of returning a proxy to the `/boot` dir
for the boot component.

Change-Id: I476c2fdd5a4d888a36e41009f3a65accdd5d0894
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875976
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>