[vfs][journal] Fail replay on incorrect payload prefix
Previously this would cause a runtime assertion which would crash Minfs.
Now we detect this condition and stop replaying the journal.

This should already be guarded against in production by the entry
checksum, which is verified before decoding the entries. However, if a
bit flipped before we calculate the checksum, we may persist a journal
entry with this condition. This was found by the fuzzer, which skips
verifying the checksums.

Fixed: 130131
Test: fx test journal-unit-tests
Change-Id: I3251acd47e3e85cee998d4a94d319f28216592de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888331
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>