[netlink] Support RTM_NEWROUTE e2e
This change allows netlink sockets/clients to perform the `RTM_NEWROUTE`
request.

For now, we assume that if the route in the request is valid, we can
attempt to add the route to the Netstack. In other CLs, we will
early-exit route requests if it is not unique from the routes already in
the routing table.

Support for using the `NLM_F_REPLACE` flag will come in an additional Cl
and implemented if/when it is needed. b/290803327

Bug: b/283136222
Change-Id: I0b77b370b046a7164cf9fc3c43d40dd07de82aa0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878533
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>