[debugger] Do not assume gTest during `run-test`
This flag is gTest specific, and causes other types of tests like gUnit
tests to not execute correctly in the debugger.

Replace the corresponding `run-test` case with a script. We no longer
expect to hit implicit breakpoints from a test case failure in the
inferior, but this does ensure that we can start a test and intercept
the output.

Bug: 130870
Change-Id: I8e4146946f3873566b1eb9aea79f3a476982e43c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887792
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>