[profiler] Restructure Target Management
The method of adding targets during a run was... bad to put it lightly.
This change the approach to something much safer. It forces a requery
into the tree of targets each time instead of trying to save a reference
to the target a watcher is watching.

Since this adding is always done on a single threaded async dispatcher,
we don't need a lock protecting it.

Change-Id: Id46fc228a00934f0b2d356d187c5e4dc05559366
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884348
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>