[build][sdk] Deprecate sdk() template.
The sdk() GN template is now deprecated, since it
has a misleading name (it does not build an SDK or IDK,
just an SDK collection), and individual archives like
core.tar.gz should not be distributed outside of the
platform build.

generated_fuchsia_idk() is the template to use to
generate a proper IDK archive, since it ensures that
it supports all target CPU architectures, and may in
the future support building the atoms for several API
levels and maybe even build variants.

This changes //sdk/BUILD.gn to use sdk_collection()
instead of sdk(), properly documenting the cases where
creating an archive is still needed.

Note that sdk() still exists since some //vendor/ IDK
definitions still use it. When this is fixed, it will
be removed from the build.

+ Update //src/firmware/README.md with updated
  instructions to generate sdk/archive/firmware.tar.gz

+ Properly document why //sdk:zircon_sysroot exists.
  A future CL may remove it from the _fuchsia_sdk_deps
  list, while keeping the //sdk:zircon_sysroot target,
  once all use cases for it will have their instructions
  updated, to include an explicit call to
  `fx build sdk:zircon_sysroot`.

BUG=b/291989349
Change-Id: I251ca05bc52aa6fb35b72072eab229d46f386b1e

Change-Id: Ib7165f8cec30f1989b9f76c73a1beb5dac540523
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889834
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>