[superproject] Roll scudo [scudo] Make Options a reference for functions.
Modify all places that use the Options structure to be a const
reference. The underlying structure is a u32 so making it a
reference doesn't really do anything. However, if the structure
changes in the future it already works and avoids future coders
wondering why a structure is being passed by value. This also
makes it clear that the Options should not be modified in those functions.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D156372

GitOrigin-RevId: b157c3ab72273fbe76fc40ea80f8968c40716811
Original-Revision: 597231560e99e00c115239bde5613889a035346f
Roller-URL: https://ci.chromium.org/b/8774451345183747329
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ia85cb7e4a0851ce168a858b42ed59955d0ba2ccd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891014