[netstack3] Remove FakeBufferIpSocketCtx type
This is redundant with the FakeIpSocketCtx type, which it was just a
thin wrapper around. Replace all usages of FakeBufferIpSocketCtx with
FakeIpSocketCtx, implement any required traits on the latter, and remove
the now-unused type.

As part of this cleanup, remove the Default impls for some
UDP-specialized test types and replace usages with `new_xyz()`
functions. In addition to not causing trait impl conflicts, this is more
correct since the functions create types with devices and addresses,
which does not fulfill not the "empty" contract that Default implies.

Change-Id: I8e7b6ff6072bff7b52c672b4505858ec6942ffe0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891174
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>