[netlink] Properly sync route test requests
Follow up to fxr/884054. That CL correctly blocked while waiting for
messages, but the messages may not have been sent at all due to
breaking from the `select` macro before 1) the route_set_fut could
send the message to the Routes Watcher and 2) the EventLoop could
process the event from the Routes Watcher, sending the multicast
message.

In the Netlink Route tests, combine the future that sends the
`AddRoute` message with the future that waits for X messages to be
received. Ensuring that both of these futures complete prior to
breaking out of the `select` macro will mean that the RouteSet handler
is able to send the event to the Watcher and the EventLoop is able to
observe the event from the Watcher.

Confirmed with @csuter and this seems to fix the issue with the
unification of the single and multi-threaded executor.

Multiply: netlink-lib-test-no-err-logs
Bug: 130548
Change-Id: I8d7133873a0675bfebdcb9eb05f0b2720a02016e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890669
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>