[fxtest][rewrite] Terminal output library.
The termout library provides support for repeatedly overwriting the
contents of a terminal window to create animated status updates, without
scrolling the window.

We use this instead of curses because curses switches to an entirely new
buffer, and we want to preserve the context of the user's terminal
history.

This library uses ANSI escape codes to move the cursor, delete
appropriate lines, and overwrite them to create the desired effect.

It includes a termsim module that simulates a small subset of ANSI
escape codes for testing.

Change-Id: I38e2aaadb32c5d2cee7e1acb5e3b02bb700ca9dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884706
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>