[netstack3] Provide summary for socketmap addr tag
Instead of making the entire address vector available to socket
implementations for the purpose of computing the per-address tag values,
only provide a summary of the information. The tag computation doesn't
need to know, for example, whether a connected socket with a device
bound is bound to a loopback vs ethernet device, just that there is one.

This has the happy follow-on effect of letting us get rid of the
`Bound<S> : Tagged<AddrVec<I, D, A>>` bounds that were annoyingly viral
and therefore everywhere. Since there is a blanket impl in terms of `S:
SocketMapStateSpec`, that bound can now just be inferred.

Bug: 21198
Change-Id: I9cbc0e79c208293eca88ff71589683da8481340c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891172
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>