[profiler] Migrate ffx profiler to a subtool
ffx profiler depends entirely on fuchsia.cpu.profiler which is
currently an internal API, so it shouldn't be distributed as part of
the ffx sdk.

This migrates it to solely be an ffx tool, which will also benefit its
iteration time.

Change-Id: I18d5b15ea4db8e200e086996224174603d067094
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890856
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>