[check-licenses] Skip ".cache" directories by default.
This is causing issues in some environments.

For the most part, check-licenses shouldn't be traversing any directory
that starts with a dot, but the skip logic doesn't support wildcards at
the moment, so this CL just adds ".cache" to the list of skips.

Fixed: 130958
Change-Id: I319c0972f96ae73dd9b39af2ecc655adb9530a4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892659
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Yaar Schnitman <yaar@google.com>