[starnix][tests] Fix flake in test_clock_nanosleep_interrupted_relative_to_fast_utc_clock
This test was attempting to interrupt a relative sleep from another
thread roughly halfway through the sleep. Since the test can't see the
time value used by the relative sleep it is not possible to reliably
know when the interruption will occur. This updates the test to avoid
doing two relative sleeps and relaxes the invariant so that if the
interruption happens earlier than expected the test still passes so long
as all other invariants are maintained. This also improves the assertion
output on failure to aid future debugging.

Fixed: 131337
Change-Id: I9af257e75bc7796bd8bdaf37343685f49af85023
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892736
Reviewed-by: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>