[amlogic-display] Fix a naming confusion in canvas endianness.
Canvases are default to little endian for each 128-bit chunks,
and captures are configured to big endian for each 64-bit chunks.
In order to read the captured bytes correctly, we need to set the
canvas endianness to match the capture interface.

This fixes an incorrect naming of canvas endianness used for captures.
The capture should be 64-bit big-endian instead of 64-bit
little-endian.
This change also replaced the magic numbers with bitfield names.

This change doesn't change the behavior of the driver.
Test: no-op.
Bug: 131232

Change-Id: I4768ea78d796cf8db49547979828ed70620012b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894290
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>