[py] Use argparse.BooleanOptionalAction to remove --foo/--no-foo
This commit updates the argparse parser to use the BooleanOptionalAction
action, which was introduced in Python 3.9. It was possible due to the
Python upgrade we made to 3.11. This allows users to specify the foo
flag with a single flag, rather than having to specify both --foo and
--no-foo.

Change-Id: I2a6893b2334d6bb18ed029dabdfe86a294a0081b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894388
Fuchsia-Auto-Submit: Mahesh Saripalli <maheshsr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>