[netlink] Support RTM_DELROUTE e2e
This change allows netlink sockets/clients to perform the `RTM_DELROUTE`
request.

Re-arranged some error handling, allowing for the unicast field
extraction function to be used for NEW and DEL route requests. This
required updating some Err values in existing tests due to hitting the
Err flow for EINVAL with an inadequate destination subnet rather than a
missing Oif NLA. This new implementation is more appropriate because if
an Oif is not present, we will need the subnet to be appropriate so we
can resolve the destination address (see b/292103361). Doing this
resolution is a concern for NEW requests, but not DEL requests.

Extracted implementation and helper code into generic over IP functions.

Bug: b/283136222
Change-Id: Ice89cba8bf3286705bb4c6373a4571b89f0a3cd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890656
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>