[starnix] Assert SO_SNDBUF/RCVBUF is doubled
Add a test which asserts that SO_SNDBUF and SO_RCVBUF is doubled on
get compared to the value that is set, which is the desired Linux
behavior.

Currently NS2 exhibits this behavior, but NS3 doesn't, so this test
acts as a change detector for when we switch over to NS3.

Change-Id: I0bafda9407348cb97369b412fea9a5f377579488
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893378
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Tony Gong <gongt@google.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>