[brcmfmac][sim] Fix flakiness in start_ap_while_scanning_test
Root cause:
The test case "ScanApStartInterference" starts an AP interface during an
ongoing scan operation. SimInterface simulates SME to handle the
OnScanEnd event from brcmfmac driver. In the overrided version of
OnScanEnd which is defined in ScanTestIfc, it calls the inherited
version of the same function from SimInterface first, and then call into
the test handler of this event. Since the inherited OnScanEnd replies to
the brcmfmac FIDL call, it unblocks the dispatcher on that end, so that
the test kept running after this and cleared out the ap_start_pending
state before the check line in the test handler of this event.

Solution:
Exchange the two operations in ScanTestIfc::OnScanEnd, so that
completer.buffer(arena).Reply will be called after all the operations
finish.

Bug: 131096
Multiply: start_ap_while_scanning_test: 500

Change-Id: Ib0bed15948ac81417cf3baecc25fd7883ef2035b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893386
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>