[ui][a11y-manager] use `optional` in FakeColorTransformHandler
Using raw `bool` and raw `ColorCorrectionMode` makes the
tests hard to understand, because it's not obvious whether
an EXPECT-ation would fail in the absence of a call to the
handler.

E.g., is EXPECT_FALSE passing because the handler was
called in a way that caused color inversion to be disabled?
Or is it passing because the ctor initialized the bool
to false?

As it turns out, it is _not_ the latter, because
1. The handler uses the default ctor
2. The default ctor default initializes the `bool`, and
3. Default initialization leaves a bool uninitialized

That makes it too easy to write a flaky test (since
reading uninitialized memory yields undefined behavior),
and/or trip up the UBSAN bot.

Instead of using raw types, wrap these fields in
`optional`, so that the values are always initialized,
and tests can distinguish the "handler-not-called"
case from the "handler called but wrong value seen"
case.

Test: fx test configuration-tests-component
Change-Id: I3c52ac17c3c786daa837498b26f39dc906e09929
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892707
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>