[testsharder] Check that test deps are built
Currently, it's the recipe that enforces that every file listed in a
shard's deps exists on-disk. However, from the recipe's point of view,
deps are aggregated by shard, so when a dep doesn't exist we can't know
which test the dep belongs to, which would be helpful information for
debugging (as the test likely failed to hook up its deps properly in the
build graph).

Now we'll do that validation in testsharder instead, where we're able to
know which test each dep belongs to and print the test's name.

Change-Id: I9538c8da747a757b47f0cebd36424c904c6839a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896778
Reviewed-by: Andrés Villegas <andresvi@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>