[trace-engine] Use release/acquire for rolling_buffer_current_
When tracing in streaming mode, the buffer number being used is
stored in rolling_buffer_current_, but the status of a buffer being
filled is stored separately. At a buffer swap, the filled status is
cleared and then rolling_buffer_current_ is updated to point to the new
buffer, but with relaxed memory ordering another thread could observe
the second write without the first.

That would cause the second thread to believe the new buffer is already
full and trigger an unwanted buffer transfer, which would result in the
previous contents being transferred a second time along with likely many
dropped records.

Using release/acquire ordering ensures that the threads that observe the
new value of rolling_buffer_current_ also observe the new value of the
filled status, i.e. that the buffer is not immediately full.

FIXES: b/294082351
TEST: Starnix boot test with fxrev.dev/891233 included, observe that
trace does not include duplicated events (though depending on buffer
size it may still have some dropped events).

Change-Id: I477f51636943762fb1966a83bcc9953e39e9baf5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895984
Commit-Queue: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>