[vm][test] Fix the vmo_remap_test
The VMO remap test is meant to work as follows.

1) Create a VMO.
2) Map it somewhere in the kernel.
3) Fill it with a random test pattern, then read back to verify that
   the pattern stuck.
4) Unmap the VMO.
5) Map it again, somewhere else.
6) Verify that the random test pattern is present in the new location.

The common `fill_and_test` routine uses the virtual address of the
test pattern to compute its "random seed" and perform its fill.  In
this case, it just xors the high and low 32 bit words of the seed and
fills the region with this value.

Side Note:  This is really not the best for these tests.  Every test
page gets filled with an identical test pattern, meaning that if the
pages were mapped, but somehow in the wrong order, we would never
detect this.  But I digress.

Later on, after unmapping and mapping again, we have to pass the seed
needed to verify to our call to `test_region`.  Unfortunately, we had
been passing `ptr` as our seed, which is the _new_ location of the
mapping, not the original seed used to generate the fill pattern.

This "works" and has gone undetected for a long time because we pretty
much always get the same address back the second time we perform the
mapping.  Given enough trials on a sufficiently loaded system (or,
just enabling ASLR) will result in a different address coming back
(sometimes) and the test failing.

Fix is simple; just remember our original seed and use that during
validation of the buffer.

Fixed: 131560
Change-Id: I915010ea324fc0f1e051e18f985dd8668251ab75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895907
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>