[early-boot-instrumentation] Flexible kernel files
No longer hard code the path, instead use a convention where
instrumentation data is under:
 * kernel/i/

Each entry under that path is a directory whose name is the sink name of
the debug data.
 * kernel/i/sink-name/

Each sink directory has up to two subdirectories:
 * kernel/i/sink-name/s/ -> static data
 * kernel/i/sink-name/d/ -> dynamic data.

Each of this directories has a named file representing the module.suffix
the kernel wish to expose. For now we can use the 31 characters
available for the object name to encode this information. If this
becomes a problem it can be revisited. For now, name is just truncated.

Then early boot component exports each of the corresponding files with
the respective file name.

Side-effect filenames are no longer remapped, but that is fine, since
test manager just copies over static/dynamic directories as is.

Bug: 120396 # Done
Change-Id: I971b481de6026ade63712833e55e948108fd1622
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822975
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>