[ffx][emu] add --reuse-with-check
This option is similar to the existing `--reuse` flag, which
will reuse a staged instance of the emulator, avoiding the
pre-processing of the .zbi and file volume files.

The "with-check" part calculates a hash of the original
zbi and fvm/fxfs file and compares them to the ones calculated
from the product bundle. If the hashes match, the instance is
reused. If they do not match, the files are replaced with the
ones from the product bundle.

Since this works with staged instances, the emulator should be stopped
with `ffx emu stop --persist` in order to persist the instance data.

This change also involved a considerable amount of refactoring to
make it easier to understand the state of the instance before staging.

Tests were added to make sure the product bundle is only loaded once,
and that no new engine instances are created unnecessarily.

The command line processing was also collected into a single method
instead of lazily resolving the values from configuration data or the
command line.

BYPASS_INCLUSIVE_LANGUAGE_REASON="dumb" is a terminal type name

Fixes: b/293897098

Change-Id: I56e83e0e81e82fbb09245acba563de2f51c8aecb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894803
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>