[boot-shim] Introduce SetBootOptionsWithoutEntropy
SetBootOptions when used in the context of shims must hace the option to
not consume the entropy of the command line, since in many cases is the sole
source of entropy and is meant to be used as a seed for kernel's
entropy.

Other boot shims also manually restore the entropy bits when parsing
the serial command line.

Bug: 129401
Change-Id: I0cddbe277933d20d38e82a6306929dc8df2cd4fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895879
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>