[fxfs] Fix fuzzer issue (not panic when close)
The fuzzer catches the panic when we close the filesystem in the fuzzing
test. Remove the panic here. `fs.close()` can fail in a situation where
we have objects to be tombstoned in the journal and that object in the
tree is corrupt. When compaction is called, this will fail (as expected).

Fixed: b/293619473
Change-Id: I4ab5b27544e23a23fce3417f43cf91bfd7b811c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899099
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>