[build] fix libc toolchain selection for kasan.
The previous code was setting `sysroot_libc_toolchain` wrong when the
kasan variant was selected. It was setting it to:
`user.libc_x64-kasan/libc.so.debug` instead of
`user.libc_x64-novariant/libc.so.debug`

Note that when kasan is enabled, only the zircon kernel is instrumented
with asan, nothing in userspace should use the asan variant.

Change-Id: I71c39e5349a7dc7074805e1745ff393ac0a4436b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899720
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>