[wlan] Extend test timeout.
watch_phys tests the ability of the wlandevicemonitor
internals to watch a directory and report the presence of a
new PHY device.  This test relies on driver manager and the
wlantap driver.  It seems like in infra, this test sometimes
takes a while to run, resulting in flakes due to timeouts.
This change extends the timeout to reduce flakiness.

Bug: 131839
Test: fx test wlandevicemonitor-tests
Multiply: wlandevicemonitor-tests
Change-Id: I3258ab9d79f87c5f9bb6848beae173ca4a2d8710
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897398
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>