[build] Remove unused //build/config/zircon:freestanding
This config is not used at all today, and we don't really want
to use -ffreestanding for compiling even our "standalone" things
like kernel binaries, so we shouldn't need this in the future.

Note the kernel separately using -ffreestanding (but should not)
and this change does not affect that.

Change-Id: I95c532273faa0221e0c56ea44b513ac723b366f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898558
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>