[superproject] Roll antlion Manually define enums for Fuchsia constants
Revert part of https://fxrev.dev/890135 that uses auto() for defining
enum values for Fuchsia-defined constants. These constants should be
manually defined to ensure they are kept in sync with the Fuchsia
equivalent; there is less room for error and is more discoverable when
defined manually due to auto() lowercasing only in the case of StrEnum.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/897640
Original-Revision: 42405daaee09fb3e09695d55ebe5907975c004e4
GitOrigin-RevId: d193587f01cf726061788c5d665d35126eae38c3
Roller-URL: https://ci.chromium.org/b/8773100691704083857
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I6ad845f5da8869d68b4e4e2bd82333f939fb29d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899831