[ffx][testing] Skip target test if image is too big.
This CL modifies the `ffx target_tests` to disable them if the zircon
image is too big. The test uses gigaboot with UEFI to boot, which places
a hard limit on the zircon kernel image size of 7MiB.

The test is now disabled by build logic if the kernel image is too big.

Note that after CL 874540, physboot will now panic in this same
situation.

Some common reasons for the kernel image to grow too big are compiling
it with extra instrumentation, like `kasan` or `kubsan`.

Bug: 129386
Change-Id: I4463414465acfe79920023d6a692204ead515acb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882094
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>