[netstack3] Combine datagram socket state and addr
Hold the state and address in the same field. This is a functional no-op
but storing these together will allow them to vary together, which makes
it possible to replace, e.g., ListenerState with an enum that encodes
whether it is bound to allow dual-stack operation.

Bug: 21198
Change-Id: I2baa985b1a74ac22ee622efb5c34219383667caa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899723
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>