[audio][drivers] Add VIM3 support
We configure the aml-g12-tdm driver for buses A and B in the VIM3.
Bus A goes into the BT chip and bus B goes into the GPIO header where
we can probe the signals.
The aml-g12-tdm driver is slightly modified to support the A311D SoC,
which is very similar to the A905D2 already supported. The driver's
GPIO usage is made optional to support the VIM3 case.

Fixed: 132037

Test: aml-g12-tdm-test
Test: device-enumeration-test on VIM3
Test: Measurement of signals on the VIM3 GPIO headers with oscilloscope.
Test: Loopback of signals on the VIM3 GPIO header by shorting output and
input. Signal generated via audio-driver-ctl as in:
$ audio-driver-ctl -d 17f0e62d tone
and captured in another fx shell with:
$ audio-driver-ctl -d 9a74ca44 record /tmp/t

Change-Id: Ica7a4a2603534ff43d5044f61ca532ff19bd6e20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888515
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>