[zxdb][starnix] Disable e2e test on LTO builds.
This test fails if the execute_syscall function gets inlined which
is prevented with an #[inline(never)] annotation in most builds.
It seems that the annotation is not sufficient to prevent inlining with
LTO when the size of the function crosses LLVM's heuristic bounds, so
for now we'll disable this test on LTO builders.

Bug: b/293716995
Change-Id: Idc9fdadd09d38b0221aa1b5d1922cfbee1d190f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901362
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>