[storage][inspect] Replace volume in lazy node callbacks.
Currently when a volume is unmounted the lazy node stays unless or until
the related volume is deleted. The data in it is just blank. After
remount despite re-registering the volume, this state continues.

Instead of re-inserting the LazyNode which doesn't replace the callback
on the existing lazy node, nor replace it for the caller, we just
replace the volume in a map that is looked up at call time.

Bug: 130564
Change-Id: I1f9e3ad66b051ff897aa84c259d9fe24217a5d5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889954
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>