[assembly] Verify recovery image outputs
The bazel image outputs for recovery are now compared with the
equivalent ones in GN. We also re-enable verifications for bazel-only
assemblies so that we can compare the zbis for the fshost transition
into assembly. After the transition it can be removed again.

Bug: b/295034062
Change-Id: Ia00a4545ae7cfa8a08cb001913b40074adc385ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901683
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>