[elfldltl] Introduce loadinfo-mapped-memory API
`LoadInfoMappedMemory` implements the Memory API, encapsulating a memory
object and the `LoadInfo` associated with it.

`LoadInfoMappedMemory.ReadArray` will find the loaded segment that
corresponds to the passed in `ptr`. The offset in the memory object is
calculated using the segment's information, and is forwarded to the
`ReadArray` call on the underlying memory object.

Bug: 118060
Change-Id: I8e3f142096e581e22f02754c1d6a042463acdb37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886698
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>