[starnix] Use UserAddress as key for private futexes
When we added support for shared futexes, we changes private futexes to
use the VMO koid + offset as a key. That change was incorrect because if
you have the same file mapped twice into your address space, the private
futexes on those two mappings are separate. This CL makes FutexTable
have two modes, one for shared futexes and one for private futexes. The
shared mode uses koid + offset and the private mode uses UserAddress as
keys.

It's difficult (if not technically impossible) to create a robust test
for this change because we would need to test that the futexes in one
mapping are not woken up when we direct a wake towards the futexes in
another mapping. However, a futex can always have a spurious wakeup, so
the test cannot actually fail if we wake up both sets of futexes.

However, a non-robust test case (see
https://fuchsia-review.googlesource.com/c/fuchsia/+/901558) shows that
this new behavior is correct.

Fixed: b/295928846
Change-Id: Id19cc8968f7e772c9c28b890241e5d8aafec2047
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901658
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>