[storage][benchmarks] Flush journal for Fxfs cold benchmarks.
We clear cache by remounting the filesystems, but in Fxfs case
remounting includes replaying the journal back into the in-memory layer,
so the items appear to be recently modified and "hot" again anyways.
This overrides the cache clearing process to include forcing a flush of
the journal by filling it up until inspect reflects that a flush has
occurred.

Bug: 130564

Change-Id: I1fc7b917fa5b7bd78fa84f95ca34c1e4abcc9ade
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888250
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>