[ld] Add Zircon version of LdStartupSpawnProcessTests
This uses fdio_spawn to launch the test process in the normal way
via the system program loader (i.e. the fuchsia.process.Launcher
FIDL srevice).

This is not ready to be used yet, so for now it's just compiled
and not tested.  Before it can be used, it will need plumbing for
a fuchsia.ldsvc.Loader service to deliver the dynamic linker VMO,
as well as for the startup dynamic linker to correctly link the
vDSO references in the test modules.  Manual local testing has
verified that it works as far as component_manager logging that
it couldn't launch the PT_INTERP because there was no ldsvc
handle supplied.

Bug: 130483
Change-Id: I1a2730be4d376fd16f01003bdd407b9a62333e22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900579
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>