[sys] Start of a bedrock program type
The purpose of a bedrock program type is to describe which operations
are supported on a program, that corresponds to the invariants we want
to maintain.

With the exception of "start" that will be addressed later, each
operation is safe to use by both component manager or bedrock users in
general. There's no component manager only backdoor or play favorites.

Right now Program is a minimal wrapper over ComponentController but we
should be able to move other program-y API here over time, such as an
API to get the outgoing directory of the program.

Bug: 122024
Multiply: fuchsia-pkg://fuchsia.com/component-manager-tests#meta/component_manager_tests.cm
Multiply: component-manager-stress-tests

Change-Id: I50d5dff85582a52a3dc0e1aa2857632e3953399c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897620
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>