[components] Create logger without going through namespace
StartAction creates a logger attributed to the started component.
This requires a connection to the LogSink protocol routed to the
component.

Previously, StartAction got a connection in a round-about way by
constructing the component's namespace, finding the appropriate entry
with a directory that contains LogSink, and opening the protocol
through that directory.

Now, StartAction issues a route request for the LogSink protocol use,
without connecting through the namespace.

This simplifies StartAction and removes the need to write a "connect
to protocol" method in Namespace, at least for now. ELF runner still
connects to LogSink in the component's namespace in the same way.

Bug: 131717
Change-Id: I4468da5812a033aae761749613438cec0f55153e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897618
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>