[netstack3] BidirectionalConverter generic output
Hoist the `Output` from an associated type to a generic parameter on the
`BidirectionalConverter` trait. This allows us to implement the trait
over any-input/any-output for a given type (as opposed to a
any-input/fixed-output).

In a follow up CL, this will enable converting connection state using the
same type (i.e. `UninstantiableConverter`) as is used for listener
state. It also allows us to support conversions where the inputs/outputs
references.

Bug: 132333
Change-Id: I101dfca1c33c564b987fc82e62c67fa274477aa4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903418
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>