[boot-shim] Enable devicetree-shim-core-tests for arm64
Temporary workaround for GIC driver relying on physical address of
different memory banks to mapped in a contiguous region (and same order)
introduced for ARM.

The workaround, for QEMU, is just marking the range [0, 1G]. To achieve
this a specialized PoolMemConfig item for ARM64 shim was added that
appends an extra range to the item generated by the base
PoolMemConfigItem. To support this temporary workaround PoolMemConfig
now provides two protected methods to calculate its payload size (offset
from the payload start) AND AppendItems with 'extra_ranges' that allows
the base class to allocate enough space for additional ranges.

Bug: 130384
Change-Id: I19fb9926c1f9c7ef433dc504ebf78f9bf2f53c88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892900
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>