[cm_rust] runner is not optional
If a program declaration is present, it must specify the runner.
This property is enforced by cm_fidl_validator.
So there's no value in making it optional in cm_rust.

This way we'll be able to say that a component has a runner if and only
if the program block is present.

Change-Id: Ied25e1cce71b9ff338cda49de3259dea182bc1af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905398
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>