[assembly] board configuration now supports file-relative paths for vbmeta keys
The vbmeta signing keys can now be file-relative, and the GN
template copies them to a location that's a subdir of where
the board_configuration.json file is created.

The validator has been changed to also support the file-
relative paths for vbmeta keys.

Change-Id: I557dd1bd45674dd6b60544a292b75b90f1e938ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905059
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>