[zircon][tests] Move thread-initial-state-test into core-tests
This is a low-level test of basic kernel semantics and should be
included in the core-tests suite.  This gives it bringup-only and
boot-test versions of the test to catch kernel issues earlier in
simpler circumstances, in addition to the component version of
the test that is easier to use and debug in userland.

Bug: 132495
Change-Id: Ifcdcc994197020144d876c8362bcb8fadfeca601
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906483
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>