[sys] Delete NullRunner
In the existing system, starting a component always means starting its
program, except that sometimes the runner is NullRunner, a special
built-in runner that's not a capability, and does nothing.

In this CL, starting a component only starts the program if the program
has a runner declaration, from which we can locate a runner capability.

This lets us delete NullRunner.

This CL changes the exit status of a program-less component slightly.
Previously, stopping a program-less component results in a PEER_CLOSED
exit status, which gets interpreted as a crash. That's just because the
NullRunner closes the control channel without sending any epitaph.

While the exit status of a non-existent program is an abstract question,
we'll have to produce a value for backwards compatibility. In this CL it
becomes ZX_OK thanks to a previously un-exercised code path in
stop_component by jmatt@.

Run-All-Tests: True

Change-Id: Id73b22344ac98b21ba2b040cacd5af90edabee40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905280
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>