[trace][rust] Option to block provider init until actually initialized.
Because the trace bindings spawn a separate thread to run trace provider
logic it's possible for the trace events early in a Rust program's
execution to be missed even if a trace is already running when it starts.

This change offers an API that blocks the provider creation bindings
from returning until the spawned thread has had a chance to process any
trace start/stop messages that were received while synchronously
registering the provider.

We can't make this the unconditional behavior because some Rust programs
are dependencies of trace_manager and deadlock when trying to
register themselves, it's mostly useful for integration tests anyways.

Bug: b/293716995

Change-Id: I6411ac7bc7566da319254cb3c86a50d0d251df6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906422
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>