[dart] Remove more dart infrastructure
Continue tearing down dart/flutter infrastructure as the only remaining
users in-tree are SL4F and a few host tools (which are in the process of
being rewritten in python).

API verification and strict deps are no longer needed - the only
remaining SDK library using dart is the SL4F client, and it is frozen.

Bug: b/240154259, b/240154518, b/240154434, b/240153813, b/240153915
Change-Id: Ice2b2c93d0c9c756468b0dced891370f0810c2bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906463
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Jason Campbell <jasoncampbell@google.com>
Commit-Queue: David Worsham <dworsham@google.com>