[zbi][rust] Ensure that zither's rust ZBI bindings are built
This change begins to generate the crate of zither's rust bindings for
the ZBI data types - and does so as a first step toward replacing the
hand-maintained //src/sys/lib/fuchsia-zbi/abi with them.

Fixed in this change is a backend bug that generated rust 'modules' with
improper names (i.e., those that feature an hyphen). The ZBI library
was the first occurrence of that bug, as it attempted to map
driver-config.fidl to driver-config.rs. The fix is just to map
hyphens to underscores.

Bug: 132599
Change-Id: I3938bb1b89b2f4df197ff446bdf0904a4de3f43f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906433
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Sergii Parubochyi <sergiip@google.com>