[net/mdns] send correct cache_flush bits
This CL addresses issues in the mDNS implementation with cache_flush
values used when transmitting DNS resources. In particular, it sets the
cache flags properly for A and AAAA resources so that obsolete addresses
are flushed from remote resource caches.

In addition, all resource types other than PTR, A and AAAA have their
cache_flush bits set in compliance with the mDNS RFC. Those resources
are single-valued for a given resource name, so that change is unlikely
to have end-user impact.

Bug: b/296239102

Test: new and updated unit tests

Change-Id: Ic6f89f5b0f39d66f77aad8d11f1bd70262e660d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904817
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>