[build] Relocate drivers-build-only and create_all_drivers_doc to //bundles/drivers/BUILD.gn
This moves these targets out of the root //bundles/BUILD.gn, so that
they are not included in the build graph unless added by a recipe,
developer, or tests.

This reduces the size of the GN build graph that is created for
most developers that are not adding all tests.

Change-Id: I60d9f61be46c286b7c56ef545c50b67e76cad26b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907734
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>