[bazel-tool] Add 'actions' command.
This command can print the actions generated for a target, or
set of targets. This is similar to doing a plain `bazel aquery`
except that inputs will be filtered to remove the _very_ long
lists of sysroot and toolchain related input files, which pollute
the output and make it completely unreadable.

Bug: None
Change-Id: I3f2e2445b1610a26802191f4609483a8230e06a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871157
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>