[session] Route `dev-bt-hci` to session
The `dev-bt-hci` directory capability is used to publish bt-hci
devices. This allows the Bluetooth stack to communicate with the
Bluetooth controller.

Route this directory capability to session so that it can be accessed
from Starnix.

Route this directory to the tiles-session which is the default element
for the session.

Bug: b/278287238
Test: fx reboot and verify capability is accessible from starnix
Change-Id: I1d6bdef2d442992d62c5f03db2c641922cd11871
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903123
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>