[zircon][clock] Add ZX_RIGHTS_SIGNAL to ZX_DEFAULT_CLOCK_RIGHTS
Allows a clock creator to create clocks with signaling allowed. It is the
clock creator's responsibility to attentuate this right appropriately.

This changes the outcome of attempting to signal on a clock. Before this
change, you would get ZX_ERR_ACCESS_DENIED if you try to signal
ZX_CLOCK_STARTED on a clock. Now, it is ZX_ERR_INVALID_ARGS, since now
signals are allowed in general but ZX_CLOCK_STARTED is not allowed in
particular.

Modified the documentation for `zx_clock_create()` to match.

Bug: b/297433088
Change-Id: I86f3b143be936bc66d94904f4e23f8bbb2badce9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908656
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
API-Review: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>