[shac] Add shac.star and shac.textproto
SHAC is a new tool and framework for integrating new and existing static
analysis tools with local and infrastructure workflows.

The contents of `fx format-code` and `fx lint` will eventually be moved
into SHAC Starlark code, for now we just set up a skeleton to make SHAC
runnable in fuchsia.git.

For now we only run `gn format` with SHAC as an example check, other
checks will be moved into SHAC incrementally.

The SHAC prebuilt already exists in-tree, it was approved in OSRB bug
https://fxbug.dev/131750.

Bug: 82386
Change-Id: Icfc2ebe095c1f9ce9650c72fbc4cda9d61778be8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877498
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>