[mdns] Properly clean up ResponderPublisher on error from callback.
While investigating b/295351837, it was noticed that multiple queries
for a subtype of a service (for example) would eventually cause the
services that were not a part of that subtype to no longer appear to be
available on the network. This was because `GetPublicationNow` method
was terminating early if the `OnPublication` callback indicated
anything other than success.

This would prevent the `OnGetPublicationComplete()` from executing,
which was necessary to ensure that the accounting state was being
updated properly.

This change makes sure that the `OnGetPublicationComplete()` method
gets called even in the event of a non-success response from
`OnPublication`.

Bug: b/295351837

Test: unit test added
Change-Id: I7c6c2d5383f6e5be15a5591abe6cd0dc4a317141
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909500
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>