[netstack_iperf_test] Include iperf3 server logs
Include iperf3 server output as test artifacts to facilitate
troubleshooting.

Also includes changes to the logic for waiting for the iperf3 servers
to be running on Fuchsia target by waiting for listening sockets to be
more robust, and to look for all servers (multiple servers are spawned
for multi-flow) in each inspect fetch.

Bug: 131856
Change-Id: Ibb5c280d842d1fef772c87eab8593a750db7ae7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905708
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>