[flatland] ObjectLinker cleanup.
For DX convenience, we previously gave ObjectLinker the ability to
dispatch link resolution/invalidation to a particular thread.  However,
we found that in practice, this resulted in race conditions that made it
difficult to deterministically test the code; we wanted only *part* of
the closure to run on a separate thread.  Consequently, we stopped using
ObjectLinker's built-in redispatch, and added our own in LinkSystem.

This CL removes ObjectLinker's redispatch ability: enough time has
passed that we are confident that it won't be useful in the future.

Test: gfx_unittests, flatland_unittests
Change-Id: Iea5d0df8622494e531f7cd03395f0111683f5566
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735283
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>