[fxtest][rewrite] Create new process groups for job handling
Without this change, sending a signal to the processes only affects the
initial processes created, but does not affect their children. This is
fine in many cases, but specifically when running an `fx` command it
results in killing only `fx`, and not the command being run by `fx`.

The result is that AsyncCommand ends up hanging until the process stops
emitting output, which is typically when it decides to terminate
anyway.

This change instead creates a new process group for all commands, and
when terminating or killing it signals to the entire process group. The
result is that both the calling script (`fx`) and the subprocess are
sent the same signal.

Fixed: b/297439693

Change-Id: I5e1d990bbcd0cda9c089942d27a165747342a295
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908658
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>