[audio] Move audio tools to AudioCore protocol
Consolidate the audio tools onto the AudioCore protocol
(as opposed to a mixture of Audio and AudioCore), for
consistency and easier management.

All the audio-related examples use the simpler Audio
protocol, which is what we want. One small adjustment to
'simple_sine' clarifies the SetReferenceClock call.

Test: manually ran these tools and examples

Change-Id: I14f6262940391ddd5d47002b07f04701c4ca8b50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909834
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>