[trace_processing] Preserve ordering of events with same start timestamp
Before this change, perftest_trace_events_test is flaky on VIM3 and
sometimes fails with an error like this:

  Expected: [
              'test_group',
              'test_setup',
              'test_run',
              'test_run',
              'test_run',
              'test_run',
              'test_teardown'
            ]
    Actual: MappedIterable<DurationEvent, String>:[
              'test_group',
              'test_setup',
              'test_run',
              'test_run',
              'test_run',
              'test_teardown',
              'test_run'
            ]

That occurs because (in this case) the "test_run" and "test_teardown"
events share the same start timestamp, because of the low timer
resolution on VIM3, and so can get reordered by the use of a
non-stable sort in the trace importer.  Fix that by using a stable
sort.

Bug: 131863
Test: host_x64/sl4f_client_tests (unit test),
  + perftest_trace_events_test on VIM3
Cq-Include-Trybots: luci.fuchsia.try:terminal.vim3-release
Multiply: host_x64/perftest_trace_events_test
Change-Id: I6e8156fca8969569fc7370bbb4954c8bb11e59b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910128
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>