[fxtest][rewrite] Consistent handling of --count and --limit
We would miscount in the status display when the user specifies --count
or --limit. With this change we now display the number of test
executions that will be expected.

Additionally, the concept of "test was selected but is not run due to
user preference" is now exposed in the event data model for logs and
display purposes.

Fixed: b/298207598
Change-Id: If6ea2b52121d7b60b6139c3f48a0c3184c0069e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909781
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Kevin Cho <chok@google.com>