[blobfs] Remove fuchsia.blobfs.Blobfs
This API was never used and there are no plans to use it, thus there is
no reason to port it to Fxblob.  If we need an API for corrupt blob
handling in the future, we can create a new one.

Change-Id: I2f579fafd12993a2e9b708ac073f5593247fc1ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910024
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>