[build][tests] Validator for test types
This is based on fxr/906930, but makes substantial-enough changes that
it has been created as a new CL.

The 'validate_test_type' tool takes a pair of files with the same
metadata as:
  - tests.json
  - test_components.json

and then using those files, can validate that all tests match some
criteria.

This initial version only supports the verification that all tests
are hermetic.

In order to do that without encountering its own errors, it contains
substantial bits of code required to understand how host tests are
represented, as well.

This is meant to be used with a GN template (in a following CL), which
can validate that all tests in a group are hermetic tests.

Bug: b/297364940

Change-Id: I098a0701e4d6f1dc4c862bc4f1bb9602f0c6e6f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910320
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>