[kernel][vm] Move split bits automatically
Changes the way split bits are managed when swapping references for
pages, and vice-versa, in the page list. Previously callers were
expected to set the bits on a page so they matched the reference they
were being swapped for. However, since references are passed in without
a way to state the split bits, they would get auto applied from the
page being swapped with.

This strategy is not only inconsistent, but the only correct thing a
caller can do when swapping in a page is to query the current split
bits and set them. Given this, it makes no sense to require every
caller to remember to do this, when there is no other strategy.

Aside from making the API more consistent and less error prone, this
fixes some existing bugs in places such as MergeContentWithChildLocked
where the temporary reference gets turned back into a page. In these
locations the page used to resolve the temporary reference has no split
bit information, as its a fresh page, and we were failing to copy the
split bits from the temporary reference into the page. Other locations
involving the temporary reference being replaced with a page are also
problematic, although a failure was never observed at them.

Bug: 60238
Change-Id: Ia422896c3e24cf87bdea6897237fe9be7cf1e6ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910394
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>