[audio] TRACE events for packet timing
Emit TRACE_INSTANT events when tracing is enabled for the
"audio" category, in certain per-packet circumstances that
are valid behavior but may not be the client's intention.
These checks are only active in tracing-enabled builds, so
their cost is only incurred when a developer effectively
opts-in on an engineering build intended purely for
diagnostics purposes.

Discrete events are emitted for each packet that is:
1) Timestamped, gap >= continuity threshold (pkt unchanged)
2) Timestamped, gap < continuity threshold (gap eliminated)
3) Timestamped, overlap < continuity threshold (no overlap)
4) NO_TIMESTAMP+DISCONTINUITY, received late (delay packet)
5) NO_TIMESTAMP+DISCONTINUITY, received on time (unchanged)

No change on non-tracing builds.

Test: caused these behaviors on TRACE build, observed events
Change-Id: I116f1a485e8122fee68050a6bf3b9ad426c97d9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910833
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>