[ffx] Close Overnet connection when peer goes away
This fixes a bug in Overnet where the connection is kept
open even after the peer is lost. This causes issues
with ffx log trying to automatically reconnect to the device.
The daemon should wait for the device to come back up
before trying to reconnect to it.

Bug: 132955
Change-Id: I16e52b7a9178e1f1a79d73a8120c311dbd94d270
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910960
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>