[starnix] Fix toctou issue on futex wait.
This CL fixes a time-of-check/time-of-use issue in the starnix futex
implementation.

The `FUTEX_WAIT` operation takes a futex word and a value, and before
deciding whether to put the thread to sleep, it checks that the futex
word and the value match. This check is done while holding the futex
state lock, making sure that we don't receive another futex operation
while that check is in progress (for example, a `FUTEX_WAIT`).

The previous implementation would do a lookup to get the vmo that was
located in that address, before taking the futex lock.

If another thread remapped the futex word memory location, a new vmo
will be created. If this happened after starnix got the vmo for the
check, the check would be performed on a stale vmo.

That is not a big issue because it is equivalent to changing the memory
after the kernel performs the check. The thread that modified the futex
word normally issues a `FUTEX_WAKE` operation to wake the sleeping
thread, and that is where the race condition can happen.

If the second thread issues the remap + wake operation *after* the first
thread got the vmo, but *before* the first thread acquired the futex
state lock, that `FUTEX_WAKE` operation will be lost, and the first
thread will end up sleeping forever.

The fix is to move the operation to get the vmo after starnix acquires
the futex lock, this way it will not miss any possible `FUTEX_WAKE`
operations.

Bug: 298664027
Test: fx test "syscalls_test" --test-filter "FutexTest.FutexWaitOnRemappedMemory" --count 100
Change-Id: Ie255c787d16bba1ed05d342d18656b514f7e799a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911355
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Adam Barth <abarth@google.com>