[fxfs] Update ctime for replace_child_with_object
Replacing a child in a directory will update change time (ctime). For
example, moving src0/src1 to dst1/dst2 will update ctime of src0, dst1
and dst2.

This change includes to test when a child object is moved to a different
directory, when the child is unlinked, and also when it replaces another
child.

Bug: b/298128836
Change-Id: I25d8a9523024fb5c2ebaff0034f63be24eaf3ce4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909633
Commit-Queue: Rui Qi Sim <rqsim@google.com>
Reviewed-by: Chris Suter <csuter@google.com>