[libc] Add unmangled symbol in abi
This starts the transition period to use the unmangled symbol
name rather than the mangled which was accidentally not marked
`extern "C"`. For now, the old symbol is kept in the abi, but
we've excluded it from the ifs file so our link stub will not
have it. There should be an error if this symbol is used at
link time, though while we wait for the 6 week window it will
continue to be included in the libc.so used at runtime.

Bug: 98888
Change-Id: Ia8f0012258b4413e96b5859846520f86cd8b0ec0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669654
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Alex Brachet <abrachet@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>