[starnix] Check that last robust list entry points to head.
According to the ABI for robust list, the last element of the robust
list has to point back at head, and that's how you determine whether you
need to stop.

The previous test was testing against the NULL address, which is also a
stop condition checked right after (after trying to read). This CL also
adds a test to make sure that is enforced.

TEST=fx test "syscalls_test" --test-filter "*RobustFutexTest*" --count 1

Change-Id: Ia8ff2816163748f2689a21442e109cb124ac4c10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911953
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>