[trace] Add rust bindings for TraceObserver and ProlongedContext
These two APIs in tandem allow for programs to initialize a data source
when tracing starts, then collect data and write it to the trace buffer
when tracing ends. TraceObserver is used to know when tracing starts and
stops, while ProlongedContext is necessary to prevent the trace session
from terminating while the provider is in the process of collecting
the data to write.

BUG: b/271467732
BUG: b/291094354
Change-Id: I8739870de19f99ff4acdddfe7907e5cdb5e7afd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891187
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>