[net-types] Decouple `SpecifiedAddr` from `ZonedAddr`
Zoned addresses are not implicitly "specified"; it would be valid
(though maybe not useful) to express
`ZonedAddr::Unzoned(I::UNSPECIFIED_ADDRESS)`.

By decoupling this relationship, we make it easier clients to use
`ZonedAddr<A, Z>` where `A` is a witness type of more properties than
just `SpecifiedAddr`.

A follow up CL will leverage this in NS3 to pass around
`ZonedAddr<A, Z>` whose `A` is both `SpecifiedAddr` and `NonMappedAddr`.

Bug: 132092
Change-Id: Ia7866ecef159321f0b3f2aec210cef6a5e5d8096
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911353
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>