[netstack3] Garbage collection for neighbor table
When a neighbor entry transitions into a state that is considered
"unused" and therefore discardable (STALE or UNREACHABLE), if the size
of the neighbor table is above some maximum, schedule a timer to do a
garbage collection pass. The GC will attempt to remove enough neighbor
entries to bring the size of the table within the maximum. Static
neighbor entries and entries that are in use by the netstack (REACHABLE,
INCOMPLETE, DELAY, and PROBE entries) are never removed. UNREACHABLE
entries are removed first, then STALE entries.

Bug: 35185
Fixed: 132095
Test: netstack3-core-test
Change-Id: Iea51bb08f736272e76d54580d445c19c276dcdd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910673
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>