[aml-gpu][vim3] Move aml-gpu driver to base
aml-gpu is currently in bootfs. It's only used by the Mali MSD, which is
in base, so it can be moved to base. This should make it easier to
develop the driver and assemble images using it.

Change-Id: I6f46c2c0731bb0f32564815bd76d14ecff75d944
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907983
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>