[overnet][compatibility] abi-revision flag for overnet
This adds an optional flag `--abi-revision` to the
remote-control runner program which is used to start
the connection between a host tool and the target platform.

This revision is checked against the version info for the
platform and the result of the compatibilty check is returned.

It also merges in the returning of the $SSH_CONNECTION value to
the host tool.

If the new `--abi-revision` flag is not present, the legacy behavior
is done, in order to be backwards compatible.

Bug: b/295899889
Change-Id: Ia21b047e811a7ec1294c574eb49baf9e52d816df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912616
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>