[memory] Use memory maps to infer Starnix process memory
With this change, Starnix processes, identified by running in the same
job as a process named starnix-kernel.cm, will be attributed VMOs if,
and only if, this VMO appears in the restricted part of their address
space.

This brings the performance of Memory Monitor and ffx profile memory in
line with other systems when Starnix is used.

This change also removes the --include-starnix flag, as it no longer has
any meaning.

BUG=b/285191677

Change-Id: I6b7b7b2ba37d04f1af0a46f541fc27881473dc4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903860
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Reviewed-by: Dave Moore <davemoore@google.com>
Commit-Queue: Étienne J. Membrives <etiennej@google.com>