[debroot] Include x11proto-scrnsaver-dev package
This is needed by SDL which is a dependency of QEMU. In newer releases,
this package was folded into x11proto-dev so this change only applies to
jessie and trusty.

In addition, we also switch the source URL to a faster mirror.

Change-Id: Ic9b4210e38b2e807803c9d36b0e46a13bb88791d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913372
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Petr Hosek <phosek@google.com>
Fuchsia-Auto-Submit: Petr Hosek <phosek@google.com>