[bt][snoop] Define an eager core shard
Currently, bt-snoop is started by bt-init when the device starts up.
This can be dynamically configured.
In some cases, only bt-snoop is required and not the rest of the
Bluetooth subsystem.

Define an eager core shard so that product integrators can use this in
cases where the rest of the BT subsystem is not included.

While here, update README with build instructions.

Change-Id: I37f0bef275afb893b1a6d8e0ed1f64372b91f990
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911254
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>