[starnix] Split process stop state
Previously, a thread group would enter a stop state as soon as
a stop signal was sent.  This CL changes that.

Thread groups are now considered in |stopping| state when a stop
signal is sent, and in |stopped| state when at least one of their
tasks has stopped.

The test test_stop_cond depended on the original stop behavior. It
was easier to replace it with another test than try to fix it. See
WaitpidFlagsTest.waitPidStopContinue.

Change-Id: If0e55accad7ad70338afe4efb25ebb604dc4c209
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914117
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>