[kernel][assert] Refactor the kernel assert macros.
Refactor the kernel assert macros so that they will catch accidental
assignments.

In the kernel, we generally run with pretty much all warnings turned
on, and warnings set to error.  In particular, we have -Wparentheses
enabled, which will catch situations where the result of an assignment
is used directly as the predicate for a conditional.  For example, it
would fire if someone wrote something like:

```
if (ptr = nullptr) { ... }
// or
while (value = 5) { ... }
```

Users who actually meant to write something like this can suppress the
warning by adding additional `()`s around the statement.  So, one
could suppress a warning from the examples above by writing:

```
// You really don't want to do this; this is just an example
if ((ptr = nullptr)) { ... }
```

This brings us to the current state of our ASSERT/DEBUG_ASSERT macros
in the kernel.  Unfortunately, they look something like:

```
  do {                                                \
    if (DEBUG_ASSERT_IMPLEMENTED && unlikely(!(x))) { \
      assert_fail(__FILE__, __LINE__, #x);            \
    }                                                 \
  } while (0)
```

This is pretty standard stuff when dancing with the devil which is the
C preprocessor.  We have no idea what is inside of the expansion of
`x`, so we are pretty much forced to wrap things in at least one set
of `()`s.

```
if (!my_ptr == nullptr) { ... }
// is not the same thing as
if (!(my_ptr == nullptr)) { ... }
```

The addition of the `unlikely` macro (expands to a `__builtin_expect`)
adds even more layers of ()s.  Unfortunately, this allows me to write
the following incorrect statements, without receiving any warnings.

```
DEBUG_ASSERT(bad_things_happened = false);
DEBUG_ASSERT(ptr = nullptr);
DEBUG_ASSERT(how_much_I_hate_the_preprocessor = 1000000);
```

So, instead of writing it as above, change the structure to be more
like the following instead.

```
  do {                                        \
    if (DEBUG_ASSERT_IMPLEMENTED) {           \
      if (x) {} else {                        \
        assert_fail(__FILE__, __LINE__, #x);  \
      }                                       \
    }                                         \
  } while (0)
```

It looks a bit odd, and it loses the `unlikely` annotation, but it
will catch errors like the ones shown above.  The evaluation of `x` as
a predicate now takes place entirely in its own, top-level, dedicated
`()`s, avoiding any potentially bad interactions with another
operators (like `&&` or `!`).

Change-Id: I9c603f4e49e8d62db064a8616b2113b4490e601b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912259
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>