[realm][fuzzing] Non-production fuzz targets
This CL replaces `testonly = true` with a `dep` on
"//build/validate:non_production_tag" for source sets that may be linked
into fuzzing targets. These targets are real code that likely is not
test-only, but fuzzable versions of them should not be used in
production.

Change-Id: Ib37da400376ef2eea398b377725e26d19f80d63e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907725
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Dzmitry Yatsushkevich <dmitryya@google.com>