[sampler] Simplify Inspect selectors in RealmProxy-driven test
We can exploit the fact that we have an archivist running in the test
realm that receives scoped events for the test case. That archivist
just sees "sampler" as the child, so we can get that directly and remove
the leaky abstraction that reveals internal test topology details.

Additionally, we now configure the sampler component name instead of
the realm builder instance name as that's the topology bit that is
visible and relevant to the test.

Change-Id: Icc05e8d465c343e458133d2429485c5626d32de2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915375
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>