[kernel][arm64] Annotate the read of the percpu pointer with volatile.
Annotate the inline assembly read of the ARM64 percpu pointer (x20)
with the volatile keyword.

Consider the following sequence:

```
void* p1 = arm64_read_percpu_ptr();
FunctionWhichMayContextSwitch();
void* p2 = arm64_read_percpu_ptr();
Log(p1, p2);
```

Without the volatile annotation, the compiler is free to stash the
first read of the percpu pointer on the stack, and omit the second
read.  If we _do_ context switch, and we start running on a different
CPU when we manage to run again (say we moved from CPU X to CPU Y),
then `p1` in the example above will point to the percpu structure for
CPU X, but so will `p2` (when it should point to CPU Y's structure
instead).

Change-Id: I543cdbe9712d32c825238d4d9b59bbb9d25736cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915696
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: John Grossman <johngro@google.com>