[debugger] Improve process to component mapping
Component instances and processes are not necessarily a 1:1 mapping,
although this is common for components which use the ELF runner. Any
component instance may spawn many processes or, many component instances
may map to a single process (i.e. driver-hosts). It is also possible for
a component to purely have routing information and/or host children and
be related to no process at all.

debug_agent's understanding of components and processes assumed that a
job would have a single component associated with each child process.
This allowed users to attach to components by moniker, URL, or any
subset of the two. However; when a particular job houses many
components, this 1:1 assumption caused filters to not match against all
components in a job or associated with a process, making certain systems
(drivers, in particular) challenging to attach to.

debug_agent now has a more correct mapping of N components for a
particular job, so new filters will be able to search across all
components for the given job.

Change-Id: I1205c47202690a5918eda85c5401b12cfce324da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915612
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>