[assembly][time] Allow timekeeper configuration
Some tests require that the UTC clock is available and running,
while there is no network available.

Since the Fuchsia UTC clock can not function in that case,
we configure timekeeper with a feature flag that starts the
UTC clock regardless.

Most of this change is plumbing the configuration bits to the
right place.

Bug: b/295537827
Change-Id: I36ba7a761497c2c6c454f0ed43f58dd7781ad79f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915695
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>