[bt][avrcp] Remove duplicate type definition
There are 3 definitions of PlayerApplicationSettingAttributeId. The
first is a FIDL type definition in the `fuchsia.bluetooth.avrcp` FIDL
library. The remaining 2 are defined locally in the bt-avrcp component
and both are identical. It is valid to convert from the FIDL type to the
crate local type.

Remove the erroneous duplicate definition.

Bug: 133584
Test: fx build && fx test bt-avrcp-tests
Change-Id: Ib7f70c2f3fd0851db8c12142b92b4b50ee80478b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917021
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>