[fidl][hlcpp] Remove transport error from fpromise::result constructors
This CL changes HLCPP result unions (i.e. compiler-generated unions for
two-way methods that are flexible, have error syntax, or both), removing
the ability to construct from fpromise::result of the transport error.

There are two reasons for this:

* Users should not be constructing results with transport errors anyway.
  They can still do so via WithTransportError or set_transport_error
  (like the bindings do internally), but there's no reason to make it
  convenient with fpromise::result.

* Without doing this, a `flexible Foo() -> () error int32;` handler
  can't return fpromise::error(123) like it can for a strict method;
  fpromise::error<std::variant<int32_t, fidl::TransportErr>>(123) would
  be required to disambiguate (https://fxbug.dev/133553).

For flexible two-way methods without error syntax, handlers will only be
able to return fpromise::ok (or the raw union).

Note that the conversion in the other direction, from result union to
fpromise::result, still includes the transport error, because the client
needs to be able to handle that.

Test: fx check-goldens; fx build tools/fidl/fidlgen_hlcpp:goldens
Fixed: 133553
Change-Id: Ie7e8c50f42a810845fd0735dc49159104ead38d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917138
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>