[starnix] Add Perfetto consumer
This perfetto consumer communicates with a Perfetto daemon running
inside the Starnix container via its unix domain socket interface.
When Fuchsia tracing starts, the consumer indicates to the Perfetto
daemon to start tracing inside the container as well, and when Fuchsia
tracing ends it collects the data and writes it to the Fuchsia trace.

Some additional files are added to the tracefs implementation so that
the Perfetto daemon will believe it is running with a functional
tracefs, which allows us to use its ability to control ATrace on images
that use it.

Tested: See tqr/760470
BUG: b/291094354
BUG: b/271467732
Change-Id: Ifb550624cece628b44aaf30621b0a9c702435cb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891188
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>