[Starnix] Keep zombies in the PID table
Previously zombie processes were not kept in the PID table. As result
this kill() was failing for zombies, which is unexpected. Now a
reference to ZombieProcess is stored in PID table. This allows to
fix kill() to handle this case correctly. Also in the future it will
allow to show zombie process status in /proc

This fixes flake in sigwait01 test, where kill() was racing with
process exiting.

Change-Id: I4686fe8d7aa2a0d1c54b313d6534a803227cb8f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907711
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>