[assembly] Allow de-duping for config-data
This can be turned on using a constant at the top of
assembly_builder.rs: DEDUP_CONFIG_DATA.

Allow the same config-data file to be added in both platform_configuration
and the legacy bundle as long as the contents are identical.

This allows us to carefully migrate config-data files from the legacy
bundle into platform_configuration.

We do this by replacing the usage of NamedMap with NamedFileMap which
gives us access to the function 'add_entry_as_blob' which finds the
merkle of the file and de-dupes as necessary.

Bug: b/300329753
Change-Id: I9ac8e3fc805584e999a4fdf92f9bc377a50c20b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917273
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>