[kernel][vm] Tighten aspace lifetime statements
Adds some additional commentary on the ArchVmAspace object lifetimes,
and how users are supposed to coordinate calls to Init/Destroy along
with other methods.

This is not a behavior change, and these comments reflect existing
assumptions (and assertions) in the code.

The x86 Destroy path is slightly changed to make misuse of this
contract cause more obvious bugs, although there are no known bugs.

Change-Id: Ie2ede528110276027c906f76108d99ebebc9c496
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917352
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>