[ffx][overnet]Adding ABI check to remote_control_runner
This updates the runner script on the target to check a provided ABI
revision to the one used to compile the runner.

This result is then returned in a JSON object before starting the
Overnet communication.

Support is also included to fall back to invoking the legacy
runner if the abi revision flag is not known.

This change updates the FIDL library to include an enum for the status
of the compatibility check, and then exposes them via the Serializable
CompatibilityInfo struct.

Bug: b/295899889
Change-Id: Ib42f19dbd42b69dbd860575b5385296b0c794acc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916357
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>