[vsock] Fix CreditInfo counting
Fix the credit values in CreditInfo.
Originally, we were setting fwd_cnt to how many bytes were in the
socket, but fwd_count is supposed to be the number of bytes consumed.

The incorrect fwd_count would lead to deadlocks because when the bytes
received equalled the buffers size we would report fwd_cnt = 0. This
meant the device thought the buffer was full when it was actually empty.

From the spec, for calculating sizes:
    /* tx_cnt is the sender's free-running bytes transmitted counter */
    u32 peer_free = peer_buf_alloc - (tx_cnt - peer_fwd_cnt);

This is now unit tested. The internal classes of SocketDevice are made
public for testing.

Bug: b/300123548
Change-Id: Ica37fb2cf2b629a2586997780a33d029777cb1f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918296
Reviewed-by: David Iserovich <diserovich@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>