[vsock] Fix rx-credit deadlock
Fix a deadlock that can occur when the driver receives too much data.

During periods of high amounts of data coming in, the device will not be
receiving any credit update messages for it to know that the socket
buffer is being drained. The solution to this in this CL is very
imperfect in that we just send a credit update after processing an
incoming RW message. This does not guarantee deadlock won't occur. We
should instead be using eventing off the `ZX_SOCKET_WRITE_THRESHOLD` so
that we can send credit updates only once the socket buffer drops below
some low-water mark instead.

Bug: b/300123548
Change-Id: Ifea18eb8b9d1e605f89b9c445abb1a42b4cdbf48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917555
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: David Iserovich <diserovich@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: David Iserovich <diserovich@google.com>