[fdf] Fix uaf in driver_server start
This use-after-free was occurring when the completion of the
StartCompleter was running before the dispatching of the Start from the
WireServerDispatcher (internal FIDL machinery) had the chance to fully
run. At the end of the dispatch handler, the request which lives in a
RAII type fidl::DecodedValue is released and the destructor is run. By
the time that this was running, in the error case, the arena had already
been freed up at the end of StartCompleter lambda, and the underlying
data (which the DecodedValue was wrapping) was already gone. This led to
a use-after-free.

The fix is to post a task to run the Start so that the callback is
guaranteed to run asynchronously and allow the cleanup on the
WireServerDispatcher to take place first.

Fixed: 133776
Change-Id: I26da47377cb004b7c76a7654ace717856b5cf702
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918178
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>