[brcmfmac] Forward boot metadata
Add boot metadata to the `wifi` composite device on astro, nelson, and
sherlock devices then forward that metadata to the `brcmfmac-wlanphy`
device. The `brcmfmac-wlanphy` device needs access to that metadata and
will not get it when DFv2 is enabled unless that metadata is forwarded
to it.

Bug: 133538
Change-Id: Ib140ec9fed4788ffb86974301fccc1df0c1ed1ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916392
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>