[sys][core] Deduplicate debug-dash-launcher
This component is declared in two places, remove the
declaration in core.cml so it is only declared in a
core shard.

This allows us to move the offer routes into the core
shard.

This also allows us to optionally define the offer
capability up to the root CML which allows us to
remove the security exception.

Fixed: b/295357307
Change-Id: I6b20a54931a5a0ea47d43aba83d37030315fa801
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907362
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Cameron Finucane <eep@google.com>