[ffx][fastboot]: add Fastboot struct
Fastboot struct works by taking a generic argument that is AsyncRead +
AsyncWrite + Unpin and sends fastboot commands to it.

Subsequent CLs will add generator functions for the TCP and UDP
transport as well as the usb_bulk::AsyncInterface

This is largely taken from the Daemon's codebase but handles the
UploadProgressListener and VariableListener slightly differently.

Change-Id: Ic88442a3fe51fecc989776b7863a848dba8105eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917177
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>