[ELF] Require filesz <= memsz only for PT_LOAD segments
Previously ELF parser was checking that filesz <= memsz for all segment.
This check was failing for the PT_RISCV_ATTRIBUTES section in ELF files
produced by GCC, where memsz is set to 0 and filesz is 87. ELF spec
doesn't require that filesz <= memsz for all sections. This change moves
the check so it's applied only to PT_LOAD sections. That's the only
case when memsz is used.

Bug: b/297312145
Change-Id: Ifed351b2fe24b87a42baffce2efb1ec313a88633
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918840
Reviewed-by: Alex Brachet <abrachet@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>