[async_patterns] Add a way to monitor Callback and Function completion
Callback and Function were fire-and-forget, meaning there is no
convenient way for the caller to wait for them to complete. This CL adds
an adaptor type that lets one chain a `callback(42).promise()`, and get
a `fpromise::promise<R>` object holding the return value of running the
callback/function.

This is largely backwards compatible except that Callback/Function can
no longer be implicitly converted to fit::callback/function, because the
call operator returns an adaptor.

Bug: 125790
Change-Id: I1158aa3aace5d5af470740a363e5d66ebf870401
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918933
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>