[ffx] Remove automatic connection to targets
Instead of connecting a host-pipe to a target on discovery, just note
the target exists, and connect on demand  (i.e. after a request to
the daemon such as `OpenTarget` or `OpenRemoteControl`). Behavior is
basically transparent, except that undiscovered targets will remain
unmolested, and the initial connection will be slightly slower.

(When the daemon starts, it will connect to manual targets in order
to learn their nodenames. Eventually, we will change the daemon to
store the nodename when it is first added -- b/300986641.)

Bug: b/297896623
Change-Id: Ia8f9cff1958b6c507e046273147a8cabac98f337
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908842
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Dwayne Slater <dwayneslater@google.com>