[ffx] Add debugging for ssh/emulator flake
fxb/133234 sometimes shows a failure of ssh to communicate to
the emulator.  Perhaps a timeout? Perhaps a failure in ssh?
Perhaps a failure in the Overnet logic?  This improves some
debugging, increases some timeouts, and changes a failed
ssh connection to be an actual error.

Bug: fxb/133234
Change-Id: Ied8439abd8b43086293849ac6d597baa8114aa06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919674
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>