[fidl][rust] Allow source-breaking opt-in
Rename the __Unknown flexible enum and union variant so it clearly
reflects that it is source breaking. Remove the non_exhaustive
prohibition and deprecated nudge.

For tables, rename the field to __source_breaking and type it to
the newly introduced fidl::marker::SourceBreaking marker.

That allows users to willingly opt-in for source-breakage, which is
particularly useful for in-tree code when used with care and
responsibility.

Change-Id: Iba2bad6787b044c6370d214ddc0181b33572f2f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917715
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>