[connectivity] hw-sim should include realm_builder_server
The hw-sim tests were trying to use realm-builder, but it wasn't
included in the package, which caused it to fail in infra (although it
doesn't appear to be a fatal error, since the tests passed).

This adds realm-builder as a subpackage dependency, since that's the new
way we expect this package to be used.

Change-Id: Ic87c49f9821d8820440656b1d5ca0ee0db71eacc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919259
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Louis Lo <yjlou@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>