[superproject] Roll third_party/pigweed pw_libc: Pull in 'abort'
abort is platform specific and is found in baremetal/,
generally we would prefer making a seperate source set
which uses stdlib/baremetal as the dir and pull in abort
from the `functions` list. Here it isn't necessary because
it is the only function we need from baremetal/ and it has
no associated test anyway.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/138518
Original-Revision: 7970ac98c81a9a00dfba6242d0b607b633fc2522
GitOrigin-RevId: 62ddb8b1934dbca3f07480f0e364c5bdee48dbb0
Roller-URL: https://ci.chromium.org/b/8769377144759485521
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5ad9da47e4df2138173a7577cf88cb0e820593f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919758