[shac] Integrate with doc-checker
Previously doc-checker integration was duplicated in `fx format-code`
and in the static-checks infrastructure recipe, but now it's
consolidated in shac.

We still continue to run it from `fx format-code` even though it's not a
formatter to preserve the behavior, but I'm hoping we can remove that at
some point.

Bug: 82386
Change-Id: I100f16cd2ac2498bb9f1dfa4e65cfecd4890cff6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919555
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>