[fxtest] Do not suppress color output of fx build
fx build detects if its output is not to a TTY and suppresses color
output. This change uses the appropriate env var to force color output.

This required changing the semantics of the env argument to AsyncCommand
to be to augment the incoming environment, which may require further
tweaking in the future if removal from the environment is needed.

Fixed: b/299145560
Change-Id: I14b6006aa7e315ca94bf0dd87c3a02f50dbd8b6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918188
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>