[paver] move error logs to actual errors
ExtractZbiPayload() is used to check whether a blob is a ZBI or not, so
we shouldn't be logging errors when it's not; this is an expected use
case.

Error logs are still printed if a blob does look like a ZBI but is
malformed. Also adds one error log to IsValidKernelZbi(), which does
always expect to see a valid ZBI image so wants to log if the blob
doesn't look like a ZBI at all.

Bug: b/300655645
Change-Id: I149f53ac86d3b2621efb1d4be7e57395491acc49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921135
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>