[cmc] Improve dep tracking for capability sources
Some capability types (`storage`, and soon `dictionary`) themselves have
a source in `from`. For example, storage depends on a backing dir, and
dictionaries can depend on another dictionary which they're initialized
from.

The dependency tracking strategy for this today involves replacing the
dependency on the capability with a dependency on its source. This may lead to
spurious cycles. Instead of replacing the capability with its source
dependency, represent the capability as a node in the dependency graph,
and add an edge between it and the source.

As a bonus, this should make the error more useful.

Bug: b/300500098
Change-Id: I509cfadf9d1b306b667a2e14bd83cf29c56d018a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921041
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Gary Bressler <geb@google.com>