[fuchsia-crate] Ensure that the main() Result isn't logged twice
The previous state of this macro would log the Result::Err to syslog,
but then return the same result which would trigger
Termination/report on the result, which would print the Result::Err
to stderr too.

This fixes this problem and guarantees that we only log to syslog.

Fixed: b/300155961
Change-Id: I77931354fa5ff35ad80a36f41f541be6df44c6e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919913
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>