[sys/component_manager] support use runner
This CL adds the ability to 'use' runners in CML files. Runners may be
used from the component's parent, a child, or from the framework. FIDL
and internal Rust representations of CML documents can also use runners
'from environment', but this source is not allowed in CML files nor the
Rust representation used for deserialization.

Specifying a runner in the 'program' section is still supported.
Contrary to the design specification for this feature, the translation
from the internal Rust representation to FIDL (and therefore CM) of
manifests that contain program/runner does not add a 'use from
environment' entry. This feature currently provides no tangible value
and invalidates test goldens in dependent repos. This change can be
made in a separate set of CLs, if desired.

Bug: b/296274455

Test: new and updated unit tests

Change-Id: If8737f5af7ca2a2aaa4a898019e4716427280f82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915433
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>