[starnix] Keep mount information in overlayfs
The mount information is required because if the upper directory was
mounted read-only, it should not be possible to write to it through a
overlayfs mount.

This CL doesn't yet fix this issue, but it lays the ground to allow it
once access to the node will requires the mount information.

Change-Id: I5c008c99506063629fdaf1deb5f122b18cacbf51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919239
Reviewed-by: Sergey Ulanov <sergeyu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>