[fidl] Require explicit unknown interaction keywords OOT
This changes the fidlc and replated tools to require all protocols to be
marked open, closed or ajar and requires all methods to be marked strict
or flexible. This was changed in-tree in June but this change will
extend it to petals. Before this can land I'll have to update all FIDL
sources in petals to include the keywords.

This will allow us to change the defaults later, keeping the ABI and API
unchanged.

Bug: 88366
Change-Id: I9aa0890e9f174c61cc46107b7e1bb4e7fdcb4183
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911314
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>