[honeydew] Change expected device type for qemu-x64
We're moving qemu-x64 builders to use x64 boards under the
hood. However, that requires a change in the infra repo.

In this change, we update the expected value and comment out the
device_type test.

Then we'll switch the actual boards in the infra repo

Finally, in a followup change, we'll uncomment the device_type test.

Bug: b/293640613
Change-Id: If8212f975030810f54f88ad139d1a05fbc4d85ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922874
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>