[audio] Inspect data for dynamic delay
Change 'internal delay' and 'external delay' from single
static properties to initial values, plus current value
and the time of that change. If needed in the future, this
could be changed to an ordered map, with the key being
the monotonic time of change and the value being the pair
of delays (both internal and external).

As a result, Inspect can show whether a driver's delays
have changed.

Test: audio_core_shared_unittests
      Unittests added, plus Inspect shows the changes.

Change-Id: I8e4701ca560eb446e8e2d448fcbe41e342bcad92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921192
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Cam Lloyd <camlloyd@google.com>