[fx format-code] Fix vendor/ repo formatting bug
I recently introduced a not-yet-functional shac.star file to a vendor/
repository, not intended for it to take effect yet, but it got picked up
by shac's default recursion strategy and broke `fx format-code` for the
vendor/ repo.

Setting --no-recurse isn't ideal because it means that more deeply
nested shac.star files will never be run. There aren't any such files
for now, so it's not the end of the world, but it should be fixed ASAP.

Bug: 82386
Change-Id: Ia1f428e07e6ff8f170d049f995967da85c4b4c0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923592
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>