[element_manager] drop view provider after providing view creation token
As a result, the element will no longer be stopped if it closes the
ViewProvider connection. This is reasonable behavior because the
protocol makes no promises about being connected to a component's
lifetime.

Context: See Milestone 3 of the "Refactoring Session and Window Manager Components" doc.

Bug: 128504
Test: fx test element_manager_tests
Change-Id: I6a63b09ecbf78f02f38b0230f2e7928aabc60689
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922881
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>