[archivist] Fix bug that was returning escaped monikers
Colons in monikers shouldn't be escaped. This is only done in component
selectors to remove ambiguity with selector segment separators.

Fixing this led to the need to update some goldens in particular in
iquery as well as ensuring that the escaping remained present when
we were dealing with selectors.

Run-All-Tests: true
Fixed: b/302150818
Change-Id: I15863f08e3baedbbe4d25ee5d333ac9c82f2908b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/920352
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Chris Phoenix <cphoenix@google.com>