[tools] Add `fuchsia-sdk-run` tool
This adds the `fuchsia-sdk-run` tool.

`fuchsia-sdk-run` is a command line tool to help with switching between
different project contexts where sdks may differ. It will
try to figure out through filesystem and environment context where the
appropriate sdk for the current project lives and search that SDK for
the host tool you're trying to run and run it with the remainder of
arguments.

Fixed: b/292283010
Change-Id: I69d8db89b2dd0892585c35e41480de002b79f202
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921142
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>