[recovery-util] Extract src/block.rs to a separate rlib
There's a few things that depend on just this one file from
`recovery-util` but as currently written, they wind up having to pull in
all of the dependencies of recovery-util, including the whole OTA stack,
which brings in rustls, webpki, and ring.  ring is noteworthy because a
newer version which we are trying to roll to includes assembly code
which must be compiled with -fPIC if included in a shared library.  The
firmware fastboot-tcp code creates such a shared library, and also
depends on another rust staticlib which depends on recovery-util, which
ultimately produces link failures upon trying to take the new version of
ring.

But the fastboot-tcp code and the fastboot-shared-lib and
//src/firmware/lib/fastboot/rust:fastboot-c all do not actually make use
of anything other than the block device signature from recovery-util.
By extracting just the one file they do make use of into a separate
target and restructuring the dependency graph, we avoid pulling all of
rustls and ring into the dep graph of a shared library named
"fastboot-shared-lib" that has no use for them, and we unblock the ring
roll.

Bug: b/283475308
Change-Id: I4456a05d389d12f09d31050833a54d1e34fe8e93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924936
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>