[ffx] unbuffer reading of ssh stderr
Diagnosis of some bugs requires accurate timing information from the
"ssh -v" output. When the output is buffered, the timing information
is lost.  This change removes the BufReader from the stderr line-
reader, replacing it with one that reads bytes directly. (It may be a
bit slower, but its use is limited to stderr, as well as getting the
first line of stdout. It should not affect the steady-state of a working
connection.)

Bug: 301312339
Change-Id: I5c9aff3b6e92cae662ff099fda1eb2c69a15750f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924399
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>