[starnix] Fix remote binder interruption mechanism
Remote binder needs to be able to interrupt a process waiting on a
binder call when a task needs to be freed to execute a taskless
operation (like setting the VMO, or creating a new thread). There was a
race in the current mechanism which could create deadlocks.

This update binder to let client interrupt the wait reliably.

Bug: b/296893420

Change-Id: Ibb91487726c1feede5b18b108b0a98df0bad9385
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925112
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>