[netstack3] Add instantiate_ip_impl_block macro
Introduce macro that repeats an impl block for IPv4 and IPv6. This
gently sidesteps the lack of specialization features while still
allowing effectively complete implementations for traits that are
generic on Ip.

This provides a meaningful and useful workaround to reduce code
duplication in Netstack3 due to the virality of non-Ip-complete
implementations as we walk up the dependency chain.

Change-Id: If76c774804f761476cd0a62cde8ceecf059e349b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925472
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>