[netstack3] Fix error code for changing multicast membership
Before this change, we return a single error code if the membership
is not changed. However, we should differentiate between joining an
existing group and leaving an unjoined group and return different
errors. This problem was not exposed previously because the NTR test
mistakenly launched NS2 even if the intention was to launch NS3.

Change-Id: I5e0371b8239682295a3dc5893448619026c1dcf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926016
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>