Revert "[cmxtinction][scenic] Delete old flatland benchmark"
This reverts commit 93775d3bf54787f1b15c8692b933eda00c559238.

Reason for revert:

The test should not have been deleted.  It was still being used to
produce performance metrics.  The replacement Python version does not
produce performance metrics yet.  It was deleted on the mistaken
assumption that it depends on a .cmx component.

Original change's description:
> [cmxtinction][scenic] Delete old flatland benchmark
>
> Bug: b/300132587
> Change-Id: I66ef22de116b4b7f02466b5e78b5a5541084ff81
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916798
> Commit-Queue: Gabe Schine <thatguy@google.com>
> Reviewed-by: Emircan Uysaler <emircan@google.com>

Bug: b/300132587
Change-Id: I27e2579a1cf044764a3ab9fa8efa07ff39396207
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927052
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>