[ffx][goldens] Don't check when building separated tools
This makes it so that the cli golden tests aren't done when the
`ffx_build_dual_mode_plugins_as_subtools flag is set, since the SDK
built when this flag is incomplete (does not contain the tools that are
built separately).

Also bumps the number of "internal" tools expected in the help-json test
down to 5 because once we migrate things out of the sdk that's about how
many there will be, and for now that helps the test pass with the flag
on.

Fixed: b/303319158
Change-Id: Ie8770f96691e7988e170fbd1c8dae1782b9e4449
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926736
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>