[ctf] Remove the allowlist
This CL removes the allowlist and the GN templates and scripts
that implement it. The allowlist is being removed because it
does not prevent tests from depending on things that aren't in
the partner-sdk category and adds toil to test authorship. We
will need to replace it with a different mechanism sometime
later. Additionally, this allowlist was a large part of the
reason the ctf_* templates exist. Most of them now only forward
arguments to their non-ctf counterparts and will be deleted in
a later change.

What is removed:

* ctf_element.gni
* ctf_version.gni
* write_ctf_file.gni
* verify_deps_in_sdk.{gni,py}

Change-Id: Ic13efbe669fa98d75d53a577a5c337ac4a813e97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925192
Fuchsia-Auto-Submit: Kendal Harland <kjharland@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>