[superproject] Roll scudo [scudo] Fix the use of ASSERT_CAPABILITY in TSD (#68273)
In getCache()/getQuarantineCache(), they return a reference to variable
guarded by a mutex. After #67776, thread-safey analysis checks if a
variable return by reference has the lock held. The ASSERT_CAPABILITY
only claims after calling that function, the lock will be held. But not
asserting that the lock is held *before* calling that function.

In the patch, we switch to use REQUIRES() and assertLocked() to mark the
code paths. Also remove the misused ASSERT_CAPABILITY.

Fixes #67795, #67796

GitOrigin-RevId: a4cd6cd3a668fd9f9fa416a98085d7a9e8412344
Original-Revision: 910aafcfc5fda11e1286ffd1435cb1b690ac68a1
Roller-URL: https://ci.chromium.org/b/8768034911770597665
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I06a136d96cb8ef4dc27617b4e10ea53fe3f080c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927847