[ffx] Reap ssh processes on failure to connect
A recent change caused the code that reaps ssh processes to be
skipped when there was an error connecting to the network. The
result caused unlimited numbers of ssh connections to be created,
then left in a defunct state, eventually leading to a system failure
as the process table filled.

We now kill the process when the connection cannot be made.

Fixes: b/303721424
Change-Id: Ib32ae78ea602293e118d034344559c2b26a4193b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928672
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>