[x86][phys][boot-shim] Fix build of tests on arm64 hosts
This test depends on a UART that is only defined on x86/x64 or if
UART_ALL_DRIVERS is defined.  The latter appears to be an affordance for
testing, so let's enable it on this test which is now being built on an
arm64 host for the first time.

Bug: b/283475308
Change-Id: Ifb9695d026608a511fd590b41bb22ea08e9b4717
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928468
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>