[bazel-sdk] add component info to fuchsia_package provider
The FuchsiaPackageInfo did not carry the FuchsiaComponentInfo along which
means that our workflows need to depend directly on a component if it
needs to get any specific information about the component like if it is
a driver or not. This causes problems because it forces the component to
be analyzed in a different configuration from the package because we
handle transitions at the package level. This forced us to put
a transition on the workflows which can cause us to run analysis even
more than we should. This does not fix the transition problem but it
sets it up to being fixed later.

BUG: b/303681889
Change-Id: Ie39a6e0f0ba76dd7f73158bdbdf42180fe0bc50f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929115
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>