[wlan] Prefer MacAddr and Bssid types over [u8; 6] in WLAN code
WLAN recently began depending on redaction of unhashed MACs logged in
the System Log. This change aims to accomplish some overdue cleanup
around general printing of MACs in the System Log.

Primarily, this change aims to significantly reduce the usage of
[u8; 6] in WLAN code which accomplishes the following goals:

- Reduced chance of debug logging a [u8; 6] containing a MAC address.
- Reduced chance of using a MAC instead of a BSSID. (MACs and BSSIDs
are morally the same, but the BSSID is typically special because its
tied to the AP a client connects to.)
- Discouraged creation of new helpers to print MAC-like data.

Bug: b/297365339
Change-Id: I7d1db96be4e2c0598002d9ea0259007799a6501d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924793
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Sam Balana <sbalana@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>