[fxfs] Precommit pages for blob decompression buffer.
We do a lot of inline page faulting while moving data into the buffer.
We should fault them ahead in parallel with the read here to make it
essentially free. This is handy for things like meta.far files that will
decompress a huge amount of zeroes which will write out really quickly
and the time will be dominated by page faults.

This is harder to do in other places and actually harmful, as buffers
that are reading will be committed elsewhere in the stack anyways, but
you also can't easily call read into a buffer while you commit on it due
to borrowing issues.

Change-Id: I97e0faa3ddd3a201f5f182413088932496a047d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926997
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>