[docs] Fixing reference links
This fixes markdown reference links (links with [] that are defined
later in the file).

This is in preparation to turn on doc-checker checking of reference
links, so it will be harder to write updates that have broken links.

Bug: b/289246762
Change-Id: I1cfa3bad451b3501b80816e826cdae47d7aa9bd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930273
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>