[policy] Separate out base policy
This base policy is platform policy that can be shared between
products.

The resulting component manager policy should be identical
before and after this change. This was checked by having the
config tool emit the full list of policies after the merge, and
by diffing the output before and after this change.

Bug: b/300506167
Change-Id: I507954076a2a1b3cd39134a2535e07ea2d4f0894
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928655
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>