[system-updater] Use resolved update package hash when necessary
If the update package URL is unpinned, system-updater will now use the
hash from the resolved package when setting the retained package index
instead of just not adding the update package to the retained index.

This bug has not been an issue so far because in production the update
URL is always pinned and on eng builds (which use
system-update-checker which uses the unpinned
"fuchsia-pkg://fuchsia.com/update" URL) the package would be protected
by the dynamic index.

Upcoming changes for open package tracking will migrate system-updater
to a PackageResolver capability that no longer adds packages to the
dynamic index, regardless of the state of the retained index.

Fixed: b/304607597
Change-Id: Iea52c773a71f2e54d0cd7805cd6cca8b7ec19e36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929933
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>