[coverage] Introduce src/test/early-boot-coverage
These are a set of end to end tests whose purpose is verify that early
boot coverage works as intended. There are two pieces:

(1) Extracting coverage from the emulator.
(2) Validating the that the coverage is as expected.

Additionally some changes were made to emulator.go to support a couple
of different configurations. In particular, supporting fxfs images,
which define the 'fxfs-blk' image type.

Bug: 109878
Change-Id: I8fc9234b5b4d64684414a251f27c091ba033d59a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916521
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>