[ktrace] Fix Streaming Mode in Ktrace
We copy all the data from the kernel buffer to the trace buffer in a
loop in an async loop task which blocks the buffer saving mechanism that
also runs on the async loop. Allowing the copy to release the loop and
reschedule itself allows the streaming mechanism to work properly.

Fixed: b/304349111
Change-Id: I0ae1ed6e00e0ef88fec8d48f77906cee347fe3d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929674
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>