[bedrock] Router routing prototype
This CL introduces a lego block for asynchronous/lazy routing called
Router. A Router takes a request to obtain a capability and may either
resolve the request, transform it, or forward to another Router. So it
is similar to fuchsia.io/Open or CapabilityProvider in spirit, but
adapted to support both pipelined and un-peered capabilities.

This CL shows that component outputs, program outputs, dictionaries, and
availability transformations can all be modeled as a Router. The precise
implementation inside the Router is of course specific to the capability
type, but you can plug these together to achieve flexible routing.

Router itself is also a capability so you can put it in a dictionary
etc.

Fixed: b/299994258

Change-Id: Ibf8bd6a875bf8a79807ed59665ddebb57ee72f38
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/920254
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>