[kernel][debug] Move mdscr from iframe to arch_thread
Remove the mdscr from the iframe_t on arm and move it into arch_thread.
Keeping it in the iframe_t makes it difficult to manage the value across
restricted/normal-mode thread transitions.

As implemented this will share the mdscr_el1 value for the thread in
both normal and restricted mode. We can mitigate this by managing this
value as part of Restricted/Normal state in r-mode, however this CL
doesn't implement that.

To test this change, I wrote a new zxdb e2e test that sets a hardware
breakpoint on `hello_starnix`, and enabled running this test on vim3.

Bug: 131121
Change-Id: I01bc65c05df359e7d2a52e3c89f1ae8ae1a745d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895026
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>