[netstack3] Remove non-FIDL addresses from core
Currently, fuchsia.net.interfaces.admin/Control.RemoveAddress can only
remove addresses that are present in the bindings HashMap, i.e.
addresses that have a corresponding FIDL worker / AddressStateProvider.
When an address to be removed is not present in that HashMap, we should
fall back to removing it from core (as it is a SLAAC or loopback
address).

Bug: 135102
Multiply: interfaces-admin
Change-Id: I73626a91f152c41a931525c891ff2a57345da243
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930232
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>