[fshost] add fxblob support to wipe_storage
This does not add support to the recovery workflows which call
wipe_storage. An additional argument is added to wipe_storage which lets
the caller connect to the new blob creator protocol, which when passed
in on a system configured for fxblob will be connected.

This required some changes to the disk builder to make sure fxblob could
be written on a gpt partition, which is a requirement for wipe_storage.

Fixed: b/293970561
Test: fx test fshost_integration_tests
Change-Id: I65507584dde50efdc4b0f09b8b454f4b95cb2120
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929352
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>