[uart][ns8250] Check the LSR.THRE bit instead of TEMT for determining if TX is ready
LSR.THRE (Transmitter Holding Register Empty) tracks that the FIFO
is empty and thus a full FIFO sized payload of characters can be
dumped into it. More importantly it follows precisely the definition
of the THRE interrupt so can be used safely to determine when to enable
or disable the transmit interrupt.

LSR.TEMT (Transmitter Empty) tracks that in addition to the FIFO being
empty, the shift register holding the current character being shifted is
empty.

This seems like a generally more strict signal for TxReady() to follow,
but in the driver this could cause the THRE interrupt to be armed while
TxReady() is still saying it's not ready, causing an interrupt storm
while the last character is being shifted out.

Found on a real device with a dw8250 core. Emulators will probably never
simulate at this level since characters will likely instantly dissapear
from the fifo, or at least never spend time in a shift register that
isn't present.

Change-Id: I28a8fe1328acd81eab760f4bad0765d8e31cccac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929686
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>