[netstack3] Avoid panic in UDP receive_icmp_error
This CL addresses a TODO to remove the `assert_is_ip_socket` function.
This was a runtime assert verifying that a dual-stack IPv6 socket never
receives an ICMPv4 error. When this assert was added, dual-stack IPv6
sockets couldn't yet be created; however, that is no longer the case.

At the moment it's not obvious how ICMPv4 errors should be handled by
IPv6 sockets, so I'm leaving that as a TODO for now.

Bug: 135202
Change-Id: I5f6e1da52a9061f7d7ead31a8fc867b9100c335e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931372
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>