[bedrock] Move open_flags from Open to Directory
This is a refinement to what goes into an Open capability.

In order to mint a capability that can e.g. open into an outgoing
directory of a program, we'll want to turn a fuchsia.io/Directory handle
into an Open. However, Open today has an extra baggage. It requires the
open_flags argument. There's no good value for this argument when
creating an Open capability to open within an outgoing directory, since
we don't know what are the rights on the outgoing directory.

The open_flags argument is required in order to remote the Open into a
Directory FIDL, because each fuchsia.io/Directory connection is
associated with some rights.

In this CL, the remoted protocol for Open is instead
fuchsia.io/Openable. An Openable FIDL doesn't have directory rights. It
only lets you open more items relative to it. This should be a better
fit.

The flags are added when converting an Open into a Directory capability,
at which point the remoted protocol changes from Openable to Directory
as well.

Multiply: attributed-logging-test

Change-Id: I7585b4dabb444721277ab9a74049ca574f5202a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930855
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>