[driver][index] No return for AddCompositeNodeSpec
Makes a change to the driver index's AddCompositeNodeSpec FIDL protocol
to not return a matched driver for the added composite spec. This is
because the driver matching process is how the driver flows to the
driver manager, the matched driver here has no effect. Previously we
were returning a not_found error when a match did not happen, but this
error is not important for the driver manager so now we return
successfully even when no match happens. This is because the composite
has still been successfully added.

Bug: 111076
Change-Id: If789bfc7f6a702eb6a2c9a98609299f6dce396e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925373
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>