[assembly][zxcrypt] Move the configuration of zxcrypt to assembly
This moves the zxcrypt configuration into assembly, and removes the
GN targets for the config files, with empty groups to aid in the
soft transition.

fshost tests use their own, private, target for configuring the
use of a null key with zxcrypt.

Run-All-Tests: true

Bug: b/304824454
Change-Id: I770d925f17c111dcf90bec7da87dcd968888ee44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929853
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>