[ffx] use env variables for target default
Instead of having no target default, set the default to be a fallback
list of environment variables:
  FUCHSIA_DEVICE_ADDR
  FUCHSIA_NODENAME
(But of course these will be overridden as always by "ffx -t" or by
"ffx target default set".)

This should provide more consistent behavior across different sets
of users (including automated systems which often set the target
via one of the above environment variables).

Note that semantics of ADDR vs NODENAME is different: the former
is an address, the later is a name. But when used in ffx, they
are interchangeable since a target query looks at both the address
and the name to find a target.

Bug: b/297896809
Change-Id: Ic250e7761cc508871dddbede4188fa90e83f3399
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932575
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>