[amlogic-display][hdmi] Allow lower HDMI PLL DCO frequency if the pixel clock is very low.
When the pixel clock is very low (e.g. on Surenoo SUR480480Y021A
the pixel clock can be as low as 16.96 MHz), the frequency of the
HDMI PLL DCO has to be lowered to 2.7 GHz in order to keep the
correct aspect ratio.

This change adds a new method
GetHdmiPllValidDcoFrequencyRange(pixel_clock_khz), so that the driver
can check and set the DCO frequency based on the desired pixel clock.

Test: "display-tool squares 1" on VIM3 with Surenoo SUR480480Y021A (low
      pixel clock), the aspect ratio is 1:1.
Test: "display-tool squares 1" on VIM3 with MIMO (high pixel clock)
Bug: b/301330744
Bug: b/306236155
Change-Id: If5c2f338efac870996bfa04f87d670f4572fb9e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934312
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>