[ffx][tc] Implement TargetUpdate to describe mutations
This struct captures the various ways targets are updated due to
discovery and RCS. Connection states are represented in TargetUpdate in
a simpler way.

NOTE: For backwards compatibility there are no changes to the target struct.

Discovery and RCS changes will be in a follow-up CL.

Bug: b/301988834
Change-Id: I2a9fade5e1478a725f7ce549279dcf8c3f636320
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927334
Commit-Queue: Dwayne Slater <dwayneslater@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>