[fdf] Check ancestor nodes when deciding node collection
Set driver framework nodes' collections to be equal to or higher ranked
than all of their ancestors' collections instead of just their parents'
collection. This is to avoid node components from having cyclic
dependencies on each other.

Sometimes a node's parents may not have a collection because the parent
nodes don't have drivers bound to them that require a component. In that
case, a node would need to check their parents' parents in order to
determine the which collection to be put in.

Bug: 134719
Change-Id: I3a58990052ce43d2c7c7b43154c0e3b0ac9aeb6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934136
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>