[kernel][x86][reset] refactor the 0xcf9 reset mechanism to allow AMD cores to use it
IO port 0xcf9 is actually a fairly standard way to reset PCs with PCI
busses, and according to the 15h and above PPRs it's also implemented on
AMD cores from Bulldozer and above (and probably much earlier as well).

Refactor the previously intel specific mechanism to be a bit more
generic and include the AMD cores as well.

This should allow for rebooting with a reason on anything reasonably
modern AMD wise.

Bug: 135095
Change-Id: I3e1891ae86051df36ce0c37c8f900b2c8c96ac37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931529
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>