[kernel][arm64][mmu] When querying a page, tighten asserts for invalid PTEs
While drilling through the page tables, assert that if the page is
marked invalid (bit 0 == 0) that the rest of the pte is also zero, since
currently nothing is stored in unused entries. This should help catch
random corruptions.

Previously the assert would happen later but not print the full value.

Bug: 135207
Change-Id: I693a0fc19d2a248c0e7d06e78b4c0c59bcf8c475
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931524
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>