[routing] Remove mapper from Aggregates
The mapper is cloned to be placed in the aggregate classes.
This means the mapper can't actually be used by scrunity
because the routes are being added to a new mapper.

Remove the mappers from these classes so we can remove
Clone from the Mapper class.

Change-Id: I63fe749e4e80a265610ced57e25e9f816d887c4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934912
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>