[wlan] Migrate SoftMAC `SetChannel` to `WlanSoftmacBridge`.
This change migrates `SetChannel` from the `WlanSoftmac` protocol to the
`WlanSoftmacBridge` protocol and migrates from Banjo to FIDL in MLME and
the generic `wlansoftmac` driver.

As part of the migration, the `get_channel` binding and associated
device state has been removed, as no clients (i.e., MLME) use this
information and the channel was only cached in `set_channel` and did not
query the vendor driver.

Note that this change does **not** migrate `WlanRxInfo` and so Banjo and
FIDL conversions for channel types remain.

Change-Id: Id744292f18433fdcbcded3fda2d925fd78c3cf54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929957
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>