[uart] Allow more UART drivers on more machines
We've now seen the motmot and dw8250 drivers in RISC-V contexts, so we
allow their use there. In the case of dw8250, we further make it
available for all machines (we might as well for all of the ns8250
variants given their prevalence).

Bug: b/299499151
Change-Id: I8c7ccb6a2ecdf4fe60b4690ca55516c16ea6d733
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936115
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>