[boot-shim][devicetree] Do not automatically overwrite UART driver
We update the devicetree boot shim's definition of InitMemory() to not
overwrite any pre-registered UART driver in the case where the chosen
matcher does not make a selection. This is useful in debugging and
bring-up scenarios where the UART driver might be hard-coded before
this function is called (and before we can even rely on the UART
matching logic to work).

Bug: b/299499151
Change-Id: I9567f932d05ab3a8f4b4add15715bc5323d9732f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936116
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>