[kernel][arch] PagingTraits::Set() should imply accessed and dirty
Where the qualifiers make sense (which is implementation-specific), a
call to Set() should result in a present entry being updated as
accessed. There is no case in which we would want to create a new entry
but not mark it as such. If hardware is managing these bits, then there
is no penalty to doing so; if software-managed, then we avoid an
unnecessary trap that we would only handle later on to set these bits.
Similarly, if an entry is Set() as writable, then it should reflect as
'dirty' when that makes sense.

(Since we do not actually handle such traps in the phys environment,
this change also addresses the bug in which we end up in this situation
(seen recently on hardware) and needlessly fail to boot.)

Along these lines, we also remove `bool accessed` field from
PagingSettings and MapSettings.

Bug: 91187
Change-Id: I64f88e081dd3e3d89d22aab37fd7f35325877243
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936213
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>