[netcfg] Add DeviceClass MatchingRule
Use the DeviceInfo to determine which DeviceClass to assign the device.
The DeviceClass from `lib.rs` will be used for now and then for the
naming convention we can use the InterfaceType.

Add tests for Ethernet/Wlan/Ap cases where the correct device class
is specified in the list provided to the MatchingRule or
not provided.

A separate CL will create the infrastructure for checking for an
interface match across multiple `MatchingRule`s.

Bug: 135094
Bug: 135087
Change-Id: Ifee2e5de5d3b780ab9c528a4a7564ccab5556132
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932207
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>