[starnix] Return EMSGSIZE on sendmsg
When calling sendmsg on a zxio-backed socket, a return status of
ZX_ERR_OUT_OF_RANGE should be mapped to the EMSGSIZE errno, rather than
EINVAL, because it indicates that the provided input buffer was too
large to fit in the zircon object backing the socket (a channel in the
case of synchronous UDP).

This incorrect behavior was discovered when running the starnix LTP
tests against Netstack3, which happens to only support synchronous UDP
currently.

Test: starnix_ltp_tests
Change-Id: Id4cbdf6ac5a022f3bfb0581a16db850a73afbd8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936332
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>