[kernel][vm] Skip redundant mark accessed
Skips the rest of the mark accessed logic if the page is already in the
target queue. The added kcounters show this short circuit can happen up
to 70% of the time.

MarkAccessedDeferredCount is also moved out of the header as there is
no benefit to it being there since it is too large to be inlined at
call sites anyway.

Change-Id: Iff5d834517071d9adaae93314922c9559afb0353
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936552
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>