[netcfg] Add TopologicalPath MatchingRule
Intake a pattern from the provided configuration and determine if
that matches against the device's topological path. If the string
is a match, this will constitute a match.

Add tests for patterns that match the provided path and do not.

I found out a way to make the pattern verified for correct formatting on
consumption of the configuration, which is very helpful for avoiding
run-time formatting problems. This functionality is tested in the
configuration consumption step.

Bug: 135094
Bug: 135087
Change-Id: Ie24653fe1147c1692999819cf317fd9ec67a1eac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933436
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>