[installer] Add missing test dependency
This test was missing a dependency on test data. We did not notice this
initially, as this test was run amongst other tests that depended on the
data, indirectly ensuring that it would be available to our test. When
run in a more isolated set of tests though, this test would consistently
fail.

Change-Id: I395865743191eba4c27d32b36b1244ca680deccc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937686
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>