[assembly] Product configs can contain lists of files
Adds support for lists of product-provided files in the product config.

The GNI change is necessary so that we list each file as a GN input and
rebuild when one of those files change.

The BZL change is necessary so that we copy the input files into the
bazel sandbox before running assembly.

Bug: b/299808720
Change-Id: Ic238f1e23bc5f48bc28ca6ff7d326ac19d052c80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937661
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>