[hierarchy] Do not return empty hierarchies when the properties weren't matched
I'm honestly not sure on the history of why we built it this way, but
looks like it was intentionally built. However, this is unexpected and
messes up with fetch-retry approaches which expect a request to
succeed only when the full selector was matched, not on partial matches.

Additionally this should result in smaller snapshots when the data
we want isn't there. Rather than including all the path to the data,
we simply don't include the data. This can also translate into using
less memory, but we'd need to measure to see the impact in practice.

Fixed: b/303732863
Run-All-Tests: True
Change-Id: I185f9f879f5f77a75ad0528489bf1acbf4ac0162
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938233
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>