[trace] Improve rust trace library performance
Benchmarks find that compiled in but disabled tracing events take ~50ns.
For reference, the c api events benchmark at ~1ns for a disabled tracing
event. Most of this unexpected cost comes from constructing the argument
array before we check if tracing is enabled. This changes the library to
have the trace macros check if tracing is enabled in the exposed macros
instead of the exposed function allowing us to skip the creation of the
argument array. It also updates the callsites.

Fixed: b/306420587
Change-Id: I5fa0b7a55e89700b58de0dfe0b61c455841a4f70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937164
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Size-Review: Aidan Wolter <awolter@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>