[netstack3] Make "bound in both stacks" explicit
A dual-stack enabled socket listening to the IPv6 any address is
considered to be bound in both stacks. Encode this explicitly as a
third-variant in the `DualStackListenerIpAddr` enum, which forces the
`BothStacks` case to be handled when interacting with listener sockets.

While here, adjust the `DualStackListenerIpAddr` to hold
`ListenerIpAddr` rather than it's own custom `DualStackIpAddr`. This
brings it more inline with other dualstack data types.

Bug: 135204
Change-Id: I0d0038ac93323eeedc9750b2edd5b3aa85809cf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938827
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>