[time] Set SIGNAL_UTC_CLOCK_LOGGING_QUALITY
This change starts signaling that the UTC clock is of logging
quality. This is signaled when we either read something from
the RTC, or something from an external source.

The resulting UTC clock is adequate for debug logging, but nothing else.

Bug: b/308214128
Change-Id: Ib3d1980750634ef5b3b0d97a94480fe70c984007
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938941
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>