[kernel] Introduce `kernel_extra_deps` GN arg
This arg is a list of GN labels comprising additional dependencies of
the kernel proper. This can be useful - in a prototyping or 'vendor'
capacity - for injecting new instances of subsystems that the kernel has
defined modularly (e.g., pdev drivers or k commands).

Change-Id: I24ed601390f01236bad3893cf8fd4f237ca7aa8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939987
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>