[starnix] Usercopy routines using exceptions
This implements a usercopy library for copying memory in/out of
restricted address ranges that translates faults into returns with
errors. This is intended to be used in Starnix in direct memory mapping
mode to catch faults when accessing restricted memory in normal mode
execution such as syscall implementations.

This library uses an exception handler on the current process to catch
page faults within the usercopy routine with a faulting address in the
configured restricted range. Only x86_64 is supported for now.

Change-Id: I568f4e973436de4e75cb3a545c56f3794610df8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917566
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>