[kernel][uart][motmot] Some tightening up of the kernel driver
-Enable a bit to force the clock to stay on permanently instead of
de-clocking after some period of activity. This was causing receive to
stop after a few milliseconds unless the transmitter was continually
working.
-Set a bit to force level triggering.
-Explicitly set the main control register to make sure TX and RX are
both enabled, instead of relying on existing settings.
-Move more of the uart setup to the early init phase and just leave
interrupt enablement and unmasking to the final stage.
-Add a safety check in the dputs path where the available fifo count may
be higher than the amount of fifo the driver thinks is available.

Change-Id: I64d292e7299fd1a3270507d6dc30c7d54f026a71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937781
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>