[testrunner][seriallistener] Use LineWriter for test output.
Test output and emulator output are both written to stdout and can
interleave with each other. We use a LineWriter to try to at least write
one complete line from each writer to stdout before writing from the
next writer. However, the test output writer was only using the
LockedWriter associated with stdout instead of the LineWriter. This fix
may or may not help with the attached bug, but it seems an issue with
what's getting written to stdout, so we can submit this and see if it
helps.

Bug: 135386
Change-Id: Ia405dfd0d7c3c04b67b168fe301861ab06bf9e6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940253
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Andrés Villegas <andresvi@google.com>