[sdmmc] Control SDMMC protocol type (Banjo vs. FIDL) via metadata.
The motivation for this change is to allow finer-grain control of
the use of SDMMC protocol types (Banjo vs. FIDL) in the presence of
separate DFv1 and DFv2 drivers that serve or consume the protocol.

The logic concerning the above is kept common amongst the DFv1 and
DFv2 core sdmmc drivers for simplicity rather than necessity.

Bug: b/301004279
Bug: fxbug.dev/134787
Test: fx test sdmmc-test
Test: fx test sdmmc-dfv1-test
Change-Id: Icc3adfbc3a88c3cb1afc441ddbfd25b48e95acdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939161
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
API-Review: Christopher Anderson <cja@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>