[fidl][fuchsia-controller] Simplify FIDL lookup.
When looking up FIDL files programmatically, it is easier to look them up based
on the library name. This removes the need for all_fidl_json.txt from
fuchsia-controller to find FIDL files. So far a future drawback will be when
users eventually try to use a library that will have a duplicate name to another
library. This scenario is not supported.

Future work could be to handle setting up FIDL IR links on a per-target
basis, but this merits further discussion (beyond the scope of this
change).

Bug: b/301178571

Change-Id: Ibcaf912147bed9b5d0495704e174aa5215acb843
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940058
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>