[swd] System tests do not depend on bin/clock
Instead of using bin/clock to get the monotonic time, we now use ffx
target get-time. In order to ensure that ffx is compatible with the
older builds, we use ffx from the build itself. To ensure that we are
always using the correct ffx for the specific device, the Reconnect()
method now takes a 'build' so that after a reboot we reinitialize our
connection using the ffx from the 'build'.

Because older ffx version are unable to restart the daemon for newer ffx
versions, the Reconnect() is a bit more complicated because it has to
manually stop the daemon using the newer ffx before starting the daemon
with the older ffx. Because infra does not support mdns, we have to
explicitly add the ffx target after starting the daemon.

Bug: b/304592348
Change-Id: I73806e1dd646cb6311432b375e623609920298ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938633
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>