[debugger] Make trivial debug_agent_unit_tests compile on Linux.
This only includes the trivial page alignment unit tests, there is no
working functionality. This patch just makes the build work for the
cross-platform code.

Several small updates throughout the code base to make certain
Fuchsia-isms behind ifdefs or to use cross-platform primitives.

There is a new LinuxTask class that represents a thread or process on
Linux (they're basically the same from the system's perspective). The
native thread handles exposed by ProcessHandle and ThreadHandle are
defined to be references to this class. Currently, the class is empty
and does nothing.

Change-Id: I4cc6040ff23177e46789ee5cffc257d90a852b8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941460
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>