[zxdb] Speed up module loading when debugging tests
When `ffx test` runs tests, we end up running the same process many,
many times, once for each test case. This process was painfully slow
because we were spending a lot of time reload the same modules over and
over again.

We have a mechanism for caching modules, but we were hitting a
pathological behavior and evicting the modules from the cache
immediately before they were needed because we discover the modules for
a process in two steps: first we get notified of two modules (presumably
the vDSO and the loader) and then we get notified of a dozen more
modules (presumably the actual modules for the test). The first
notification was causing us to dump the remaining dozen from the cache.

This CL increases the lifetime of modules in the cache by preserving all
the modules from the previous instance of the process. With this change,
we are able to keep all the modules alive in the cache long enough for
the subsequent process to fully load, making the whole test run
massively faster.

Change-Id: I65fc36d3bd8c75b92a49ff2980db5a5802579cd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941596
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>