[swd-e2e-tests] Fix adding blobs to system image and update images
This makes sure we always include the rest of the blobs referenced by
the system image or update images packages. Previously we would only
include them for consideration on the first pass, but would not
afterwards if the generated package was too large. That would ultimately
lead to us exceeding the physical capacity of the device.

This patch fixes the issue by making sure we always pass down the
additional blobs to `addRandomFiles` so they're included when we compare
the total size of the blobs to `maxSize`.

Change-Id: Iec653f749727b6e4ba5ae0f4bcad3ae2eaa70a86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941674
Reviewed-by: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>