[ld] Introduce RemoteLoadModule
Introduce the RemoteLoadModule class used in out-of-process dynamic
linking.

This CL defines the RemoteLoadModule data type and introduces the
`Decode` method, which will take ownership of the passed in VMO
and decode it to instantiate the module representation of the binary or
shared object.

Following CLs will implement dependency parsing, relocation, and
loading.

Bug: 134211
Change-Id: I15ed443435e04b40a570243adec6f2798f946350
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930549
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>