[zircon][input] Migrate hid and hid-parser library to //src/ui/input/lib.
This CL moves the libraries above away from //zircon as part of the
Layer cake deprecation work.

This CL replaces the zx_library() definition with a library_headers()
(with "lib/" prefix added to comply with the Fuchsia C++ style guide)
and a source_set().

Bug: 36063
Test: no-op. hid-parser-test passed.
Change-Id: I9a322c0bf5a1d72675ffd3b1ccc134d433a00d20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941286
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>