[kernel][riscv] fix potential issue with TLB flushing routines
When passing a 16bit value to inline asm, it seems the compiler doesn't
make sure the register is properly trimmed to a 16bit value if the logic
leading up to it computed it from a 64bit value. Solution is to cast it
to a 64bit value before passing into the inline asm so it knows that all
64bit bits of the resister are 'live'.

A condensed offending code sequence:

uint64_t get_satp();

void flush_current_asid() {
    uint16_t asid = (get_satp() >> 44) & 0xffff;
   __asm__ __volatile__("sfence.vma  zero, %0" ::"r"(asid) : "memory");
}

   call    get_satp()
   srli    a0, a0, 44
   sfence.vma      zero, a0

Since the shift of the 64bit value from get_satp() is only 44 bits and
not 48, it generates a 20 bit value. The compiler, however, does not
mask out the top 4 bits of the resulting value sitting in a 64bit
register.

If the asid variable is kept as 64bit or the 16bit asid variable is cast
back to a 64bit value before passing it in, the compiler will then
properly mask it out.

Change-Id: I753ca1555dc2ba04364fe9bbf9bc16809b4cae64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942985
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>