[starnix] Add ptrace getsiginfo/setsiginfo
Also fixes a bug where the signal injected by PTRACE_CONT went to the
end of the signal queue instead of being delivered next.

Also corrects the siginfo code sent in the case of signals that are
sent with kill() and tgkill().  This fixed a detection mechanism in
bionic timers that triggers a clean thread exit, fixing one of their
tests.

Bug: b/299348370
Change-Id: I451237deb153fbf2d761c44d8b4590ae388d739c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/920800
Reviewed-by: Adam Barth <abarth@google.com>