[include_bytes_from_working_dir] Add a new proc macro that replaces include_bytes with one that uses cwd-relative paths
The 'include_bytes' macro uses source-file relative paths, which
when used with files generated by the build, will include the
path to the outdir in the env vars used to pass the path to the
compiler.

This is a replacement macro that will instead use paths that
are relative to the cwd, which makes them the same across
build environments, and therefore result in better cache hits
on remote builders.

Bug: 94355
Change-Id: I9415a34e30d5336e8f383765fbb344b104190a5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942982
Reviewed-by: Alice Neels <neelsa@google.com>
Owners-Override: Alice Neels <neelsa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Fang <fangism@google.com>