[netstack2] implement root.Routes
This is built on top of the existing routes.admin implementation.  The
underlying function calls already support global route sets, so the only
changes required were to teach `routeSet` about global route set IDs
(which are nil).

In order to make the ID inside `routeSet` nillable, we had to convert it
to a pointer.  This added a surprising amount of complication because
`routeSetId` is an empty struct.  See the changes to the struct for
background.

Because of the change of `id` to a pointer, this change introduces two
minor memory regressions:

- Struct size increase from 9 to 16 bytes.  The previous struct could be
  reduced to 8 bytes by moving `id` to be at the start of the struct.
- Increased allocations and GC pressure.  The `id` pointer requires a
  new allocation, and every pointer requires tracing from the GC.

Because this API isn't performance-critical (only a small number of
users, and not a lot of churn in connections), these tradeoffs were
demed okay.

Bug: 130863
Change-Id: I831f0e64ef863917804d4c3d7004f548d657970e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941952
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>