[netstack3] Connect Dual Stack Datagram Sockets
Allow dualstack enabled datagram sockets to connect to remote IP
addresses in the other stack (e.g. an IPv6 socket can connect to an IPv4
remote). Notably, this should be disallowed if the socket was previously
bound in the current stack (e.g listening on an IPv6 address or
connected to an IPv6 address).

While here I also updated the `close` operation to allow removing
connected sockets in the other stack. This functionality made it
substantially easier to write tests that exercise dual-stack connect.

Bug: 135204

Change-Id: I361805e353d81c464051f2382d789df5804b8fbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931432
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>