[designware-hdmi] Rename "hdmi_dw" namespace to "designware_hdmi".
"dw" is not a common abbreviation and it's not obvious that it
stands for "DesignWare". This change renames the namespace from
"hdmi_dw" to "designware_hdmi" so that it uses the full name of
the IP core name, and it follows the existing pattern of other
driver names ("manufacturer-product", for example, "amlogic-display"
and "intel-i915").

Test: no binary change.
Change-Id: I8d85a75baf46b80d9037f8e93b2b9b46c2e323fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945632
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>