[compat] Name DFv1 driver VMOs
The driver host names each of the DFv2 driver VMOs as they are
created. Do the same in the compat driver, so that it is easier
to read memory usage in a driver host.

In order to name the vmo, it had to be opened with kPrivateClone,
which is what DFv2 driver vmos are opened with.

`ffx memory profile` was taken before and after this change on
QEMU to check that memory usage was the same.

Bug: 309972262
Change-Id: I2738296ac7e25f3192f00ab7a2246ae2c1c86297
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944694
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>