[starnix] Add /sys/class and /sys/block subsystems
Class and block subsystems are places for classification of devices.
They does not contain any device directories themselves, but only
flat lists of symlinks pointing to the unified /sys/devices hierarchy
tree.

/sys/class groups devices by their functionality, such as input.
/sys/block subsystem groups all disks and not partition devices.

A new ktype value, Collection, is introduced to group devices. When
a new device is added, besides adding it to the tree, it should also
be included in the corresponding Collections.

Bug: b/297372033, b/297379905
Change-Id: I6b14cfd093062334c4fa828e44e11f3ba3a6cf0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921558
Commit-Queue: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>