[component_manager] Availability is required
Remove the Option from here and update the default to
Availability::Required. This is the de-facto default.

We can see that the Option is not being used by the fact
that it is always unwrapped with 'expected'. Updating
this removes the chances of Component Manager panicking
at runtime on a bad expected.

Change-Id: I6ea814666f412144aa76bfd8006f5950fcb12c77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946112
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>