[bazel_sdk] use fuchsia_sdk_common in fuchsia_sdk
This allows us to wire up fuchsia_sdk_common to share logic between the
bazel sdk and the in-tree build. This CL just uses the shared sanitizers
to show that it can be used. In future CLs we will use more logic to
clean up the toolchain definitions.

Change-Id: Ifd006dfeae14258a73b107884d95a832dd1979fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902158
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>