[kernel][vm] Fix unified_accessed unittest
Fixes the vmaspace_unified_accessed_test to work on ARM. Note that this
change does not enable the unit test on ARM yet, it just modifies the
test to take into account the fact that ARM sets the accessed flag on
Map operations. This was not a problem in the past because x86, which is
the only architecture on which this test runs currently, does not set
the accessed flag on map operations.

Bug: 132980
Change-Id: I06468f89891597c448a8650227e505605a91b4c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945269
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>