[audio] Correctly translate driver formats
With this CL, the ADR expressly fails to load an audio
driver, if that driver declares that it supports a driver
audio format (in fuchsia_hardware_audio) that cannot be
translated into a valid framework audio format
(fuchsia_audio).

Previously, this would have succeeded, then later failed in
a way that was misunderstood by the developer.

This CL also adds a very significant chunk of unittesting
of the external validation functions. The functions that
are not yet unittested are labelled with TODO placeholders
in the unittest sources. I will continue to add to these,
and they may be complete by the time this CL's
dependencies have landed.

Bug: b/307787998

Change-Id: I804191594fee38df2bfccf5b8406de3c7eccb4ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945562
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>