[sdk/fuchsia.test.manager] reorder definitions
This CL reorders the definitions in test_manager.fidl so they are in
top-down rather than bottom-up order. It makes no other changes.

In a subsequent CL, the same file will be updated for conformance to
rubrics and the FIDL style guide. The updated definitions will be in a
similar order, and this CL provides a better diff base for the new
definitions.

Test: no behavior changes

Bug: b/309995479
Change-Id: I8506284463d4c99c47273c42d670f28c8341b766
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946744
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
API-Review: Dale Sather <dalesat@google.com>