[fuchsia-controller] Update IR path finding.
Uses the default config env to look for the IR path. Also adds a build
declaration to point to the default IR path within the config.

Will now not rely on using the build dir files or looking for specific
file markers to find the IR.

Bug: b/309489520

Change-Id: I20beb6cb5d105cd28c38afde64ba768fbf48e07f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942034
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>