[driver][async][sdk] Add async_helpers lib
Adds a new library that can be used to help with simplifying async
cancellations using a cancellation token and work tracker. This is
useful as we have a lot of async code in drivers and currently they
have a lot of client lifetime extension mechanisms that don't allow
for cancellation.

Change-Id: Ibeaec603e70c56d9f1bbd6920178405b8e6476c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945457
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>