[fuchsia-url] Support fuchsia-builtin:// scheme
The fuchsia-builtin:// URL scheme is a simple URL scheme that
identifies a component declaration contained within component_manager.
This CL adds support for parsing those URL to fuchsia-url, such that
we can reuse some of the validation logic.

At a high level it could be thought of as fuchsia-boot:// but where the
path must be "/". It is intended to resolve components such as the ELF
runner.

Bug: b/305862055
Bug: b/309471663
Change-Id: I911ac87b59edea9e78dcaaae5b4b3428d7e83610
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944824
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>