[netcfg] Name interfaces based on multiple rules
Determine whether an interface matches ALL of the matchers prior to
generating a name based on that selector's naming scheme.

Create a helper fn to check whether all matchers for a given
`NamingRule` match the provided `DeviceInfo`. When the BusType cannot be
parsed from the topological path, assume there was not a match instead
of surfacing the error. The error for not finding a BusType will be
surfaced during the naming portion if that dynamic rule is used.

Bug: 135097
Bug: 135087
Bug: b/309010485
Change-Id: I121f8b4ae92d7b135f1043ac27fe02080aebf535
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943936
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>