[bedrock] Replace Sandbox with Dict
The Sandbox type is a component_manager-specific version of Dict
that provides methods to get and insert "CapabilityDict" entries,
which contain a Sender/Receiver pair and availability.

To reduce confusion between the sandbox library, and the Dict type,
Sandbox is now a plain dict, and DictExt is an extension trait
that provides methods that operate on CapabilityDict entries.

Change-Id: I9278b4fdfc28e0c7282f1f41941d5de98ad1b983
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941580
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>