[ffx] Move fuchsia.developer.ffx out of /sdk
All of this is only used by ffx to communicate with itself.

We leave a couple of the daemon methods in /sdk in the new
fuchsia.developer.ffxdaemonlifecycle protocol so that we can put them in
the SDK later, and thus maintain daemon compatibility for that small
portion of the interface.

Bug: b/291987067
Change-Id: I4048acbe4e8c5f8eea153a665925a36f7fb5fcc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937217
API-Review: Ian McKellar <ianloic@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>