[designware-hdmi] Split the interface type and the implementation type.
In order to test aml-hdmi independently, previously we created a test
double for HdmiDw (FakeHdmiDw, which is actually a mock). Some of the
methods that may touch the HDMITX MMIO region were defined as virtual
methods being overridden in the FakeHdmiDw class.

That mock class inherited parts of the HdmiDw functions (EdidTransfer /
InitHw) and mocks the other methods, which can be very confusing for
the test developers. The existing FakeHdmiDw was also highly coupled
with the AmlHdmiTest fixture, making tests hard to change.

This change splits the HdmiDw to a pure abstract base class
(HdmiTransmitterController) and an implemnentation class
(HdmiTransmitterControllerImpl). Test doubles inherit the pure
interface instead of the implementation class.

This change also renames FakeHdmiDw to
MockHdmiTransmitterController which matches its behavior. It also
decouples the mock class from the test class.

Test: aml-hdmi-test and hdmi20-transmitter-controller-impl-test passed.
Bug: 136258
Change-Id: I3742c3c5d5e0646130f895eb9f0f6dc76dda5bbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945633
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>