[fbl] Add utility for load/stores to unaligned addresses
Loading or storing a value to/from an unaligned pointer such as an
offset into a char array or a packed struct is tricky to do correctly in
C++ and is needed in various places in kernel and userspace. This adds
a pair of fbl utilities for this that invoke the arcane rituals.

Change-Id: Ie8d945b8d34f76e3da0db0ea1970c11f8f5f4e19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947540
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>
Commit-Queue: James Robinson <jamesr@google.com>