[netcfg] Initial interface provisioning policy
Add in a new field to be read for configuration.

Reuse the matching rules and create a new enum for storing whether the
configuration has specified local or delegated provisioning.

Ensure that the initial deserialization works for an interface
provisioning rule that has empty fields.

Bug: 135087
Bug: 135161
Bug: b/309010997
Change-Id: I87b0ba698839a9249d00ec432382288687659212
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945420
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>