[HoneyDew]Re-enable Fastboot functional test in infra
fastboot test was failing sometimes because of a known platform issue causing the
device unhealthy and thus causing all subsequent tests to fails. So we disabled runnng this test.

Now that FuchsiaBaseTest has device recovery, we can reneable the
fastboot test

Re-enabled the test and ran 16 times in patchset#3 and was able to reproduce the
fastboot failure and made sure recovery logic kicks in and recovers the
device and subsequent test are now running.

Change-Id: If47ae582945e2eee9902da28045e5a938e308b92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946618
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Pavan Kumar Juturu <jpavankumar@google.com>