[bt][hfp-hf] Add delay handling search results.
Previously, we would immediately attempt to connect to a peer on a
search result.  However, this could race with the peer attempting to
connect to us.  Since it's more likely a peer will want to connect
first, this CL adds a delay to allow this before attempting to connect
to the peer.

Change-Id: I3548240e06e842ab9659327941f2513f212b1c8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864062
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>