[scrutiny] Add pre_signing scrutiny verifier
This plugin is intended to run an assortment of checks from the
perspective of a downstream signing service to ensure that userdebug and
user builds are safe to sign. It will support basic functions such as
searching for key-value pairs, strings, presence or absence of specific
files, or golden file comparisons.

Bug: b/308980335, b/303922650
Test: 'ffx scrutiny verify pre-signing' with a local policy file
Change-Id: I1d351af0819432718b7971b814c2c7d42fc2c025
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941199
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>