[debugger] Add Linux helper routines.
Adds a utility that will convert a block memory writes to a series of
aligned 64-bit writes. This is expressed in a generic way so that it can
be unit tested. It will be used for ptrace memory writing which can only
write 64-bit aligned memory words.

Adds a BufferedStream implementation for a Posix bidirectional pipe.
This will be used to channel debug_ipc over a pipe.

Change-Id: I850ea994305625c201b08b42d2afbf818e5ade1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945419
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>