[netstack] Remove GetDnsServerWatcher
There are no longer any users of the method.

Normally, we'd want to split off the Netstack2 and Netstack3 changes
from from the removal of the method, but that leads to compiler errors
due to unhandled messages.

Removed the TODO on dnsServerWatcherCollection.Bind() without removing
the method because it's load-bearing in tests.

Fixed: 134162
Change-Id: I19b8617b5416ff5f2de3d25a065dd9b6d6de972a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937198
Reviewed-by: Peter Johnston <peterjohnston@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>