[power-manager] Prevent busy-loop on closed activity provider
If the watch state stream closes, watch_activity attempts to reconnect
by calling fuchsia.ui.activity/Provider.WatchState. However, this FIDL
method is one-way, so succeeds even if the underlying channel is
closed. This results in an infinite loop in watch_activity which spams
reconnect logs and also, I believe, prevents the component's single-
threaded async executor from making progress on any other tasks (
because the only .await in the loop is on the stream which I think
is guaranteed to be closed at that point and so poll_next always
returns Ready(None) which I'm guessing prevents the executor from
polling any other tasks).

This change checks if the fuchsia.ui.activity/Provider connection is
closed before calling watch_state, which should prevent the infinite
loop and hopefully prevent some flakes in the OTA tests.

Later work should determine why the Provider connection is closed
and possibly move the closed check earlier in the process or make it
blocking and possibly add reconnect logic to the Provider connection
itself, given that the code is interested in reconnecting to
WatchState.

Bug: b/311783026
Change-Id: I6a13db3b0ff0a5566950ca1de124da76a84a34de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948626
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>