[netcfg] Correctly classify AlreadyExists error
When an interface name is generated and it is the same name that already
exists, it was being classified as a Fatal error for the
`handle_device_instance` workflow in netcfg and crashing the component.
This should instead be treated the same way as if an interface is
rejected from the Netstack with an AlreadyExists error, where temporary
names are attempted if the first one is unsuccessful.

Change the NameGenerationError::AlreadyExists to map to the
devices::AlreadyExists error instead of a fatal error so that we can try
alternative names instead of crashing.

Bug: 136406
Change-Id: I59c7f8208aed4b7105d6966d80be5295bcbab690
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947612
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>