[netcfg] Enable non-default interface names
Prior to this change, there was not enough information for interfaces to
call all other `NameCompositionRule`s and produce a functional name. For
example, a device that was named with a MAC persistent identifier would
cause a crash if the configuration specified something such as the
BusPath because there would be no topological path present to parse.

Pass through the fields directly when the device is discovered, so
that the name can be produced using all of the available information.

Create a struct that mimics the `DeviceInfo` struct and uses references
instead of owned structs. Also, it confirms the MAC is present, since
devices without MAC are not supported by netcfg.

Refactor some tests to pass in `DeviceInfoRef` instead of a
`PersistentIdentifier`.

Bug: 135109
Bug: 135087
Bug: b/309010488

Change-Id: Ibb9982f8363a348913feefb6e106b6dc392349d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945355
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>