[zxdb] Limit support for "run" to Linux
This command runs the specified binary in the debugger's namespace,
which is very unlikely to be what the user wants. Instead, users are
much more like to want to use the `attach` command to attach to an
existing process.

This CL limits support for "run" to Linux targets. Also, remove the
incorrect warning when using "run" on Linux.

Change-Id: If0838d7b1f41555e8cfd1209125e5dcd7ca8e170
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939364
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>