[security] Add pre-signing policy files
These policy files are intended to define the checks that run
prior to signing a user or userdebug build. This allows for potential
signing failures to be detected at build time by the Scrutiny
pre-signing verifier.

This change initially ports the additional_boot_args checks, which
ensure that netsvc is disabled in userdebug and user builds. It also
checks that virtcon is disabled in recovery.

Bug: b/308980335, b/308980335, b/303923352
Test: fx build with tqr/779300
Change-Id: Ie883b5ab06fc27488fbecbc7ccdcc217b6317f16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947830
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>