[build] Extend product bundle support for boot tests with EFI disks
This aspect of boot test support was missing.

If the boot test specifies an EFI disk image, then this should be
regarded as the contents of a bootloader partition - and in this case we
synthesize a basic config file; otherwise, we default to the
board-specific config.

Bug: b/304820366
Change-Id: I5718d37d524950db4504ad7faea98835a948344e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947907
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>