[debugger] Allow local Linux debugging in zxdb.
Enables the functional but incomplete version of zxdb. It is enabled by
passing the --local switch.

This uses a built-in debug_agent that is automatically forked on
startup. There is not currently a separate Linux implementation of
debug_agent for remote debugging but this is easily possibly by setting
up the right socket connections.

Restructures some of the connection initiation code to share code
between the various implementations (local, remote, minidump). This also
hooks up more state like the architecture for the minidump "agent".

Change-Id: I73e85bd6858508cfb4874824f56ec8462e80da5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412296
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Adam Barth <abarth@google.com>