[netcfg] Add InterfaceName MatchingRule
Intake a pattern from the provided configuration and determine if that
matches against the interface's name. This matching rule only
available to be used in the matchers for the provisioning policy - the
configuration will not allow this option in the naming scheme for
naming rules.

Tested: Added tests for matching an interface name with various glob
expressions.

Bug: 135108
Bug: 135087
Bug: b/309011118
Change-Id: Id0ebef82e4db894b942a420602286b6bd3d3077b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943337
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>