[debugger] Don't require slash for local symbols.
Previously, loading local symbols required that the module name begins
with a slash. This was done when the local fallback was kind of a hack.

Now that we have a strong notion of whether the connection is to the
local system, we don't need this extra gate before checking for local
symbols. This allows debugging Linux processes without specifying an
absolute path to "run".

Change-Id: I4abd07c2983af7f9a2b9dd66f78a3fab738db7ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950014
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>