[bt][hfp-hf] Make SCLI take one AT resp at a time
Currently procedures take a vector of AT responses.  However, this
doesn't work, as we can't guarantee that the underlying RFCOMM channel
will produce all the required AT responses in a single data transfer--
in fact we expect it to not do so.  This means we need to correlate
which AT responses go together into one vector to make the pattern
matching in the procedure transition methods work correctly, which is
hard to do outside the procedure.

To fix this, I am changing the procedure trait transition method to take
a single AT response.  This requires changing all the procedures that
pattern match against multiple AT commands, before changing the
transition method type.

Fortunately, the only currently implemented procedure that needs to be
changed is SLCI.  Several states in the procedure state machine for SLCI
expect an incoming response followed by an OK.  I've broken these
states into two states, the original and another for waiting for the OK.

I've updated the tests to match.

Bug: fxb/128997
Test: fx test bt-hfp-hands-free-tests

Change-Id: Ib0b2547426746aa2d530fa74eaa5b42f999dc80a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878814
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>