[bt][hfp-hf] Remove ProcedureMarker
The ProcedureMarker type is left over from copying the HF design from
the AG design.  In the AG, multiple procedures can run concurrently, so
the component must keep a hashmap of currently running procedures; this
was keyed by the ProcedureMarker.  In the HF, only one procedure can run
simultaneously, so ProcedureMarker was not being used for this.  Instead
it was being used solely as an extra step in converting ProcedureInputs
to Procedures.  This change removes this extra step.

The immediate impetus for removing ProcedureMarker is that its presence
makes testing the ProcedureManager difficult.  Rather than being able to
simply mock ProcedureInputs and Procedures, we would have to also add
any mocks to the ProcedureMarker intialization match expression, or
add yet another mock, this time for ProcedureMarker.  Since we were
planning on removing ProcedureMarker anyway, this makes it a good time
to do that.

Change-Id: Ibd4d5ae3f5cc54f85f5996487d4ef045ada6ca30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888891
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>