[swd-e2e-tests] Support the new log listener
(This is a re-attempt at http://fxrev.dev/948776).

This can use the new or old log listener to check for unknown
firmware error messages.

We read the syslogs to see if we installed unknown firmware. However
with the omaha updater we don't have a way to stage the update but
not reboot, so we have to handle the device restarting out from under
us. While we did catch the ErrExitMissing, the connection could still
be in the process of shutting down, which could cause an error.

This refactors the code to also wait for us to receive a disconnection
event before continuing.

Change-Id: Ic473713a0e8eb3e908ad6981e73361cde24a50c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950633
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>