[ffx][sdk] Don't resolve symlinks in sdk search
When run in google3 or otherwise through symlinks (such as in the bazel
sdk built in-tree), if we're searching for the SDK from the executable's
location, we should not resolve symlinks which means we can't use
`std::env::current_exe()` (which resolves them on linux but not other
platforms) or `Path::canonicalize()` (which always resolves symlinks).

In google3 this is especially problematic because the actual location of
the binary is in a CAS and not adjacent to the SDK at all, so it just
fails.

Instead, use argv[0] joined to the current working directory and search
from there.

Fixed: b/312491133
Change-Id: I5de02ee8163f346d6f3a51eff44ec37b555ee58e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950018
Reviewed-by: Michael Brase <mbrase@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>