[ffx][ssh] Adding sshconfig object API
This creates a struct that holds the sshconfig that is used
when building commandlines for ffx and subtools.

The struct is initialized from a sshconfig file which is compiled
into the library, but is usable directly by ssh using the `-F`
option. This has the advantage of being able to debug and test
configuration changes directly. The file can also be used by
non-ffx programs which are connecting to a fuchsia device via ssh.

The configuration can also be overridden programmically if the
need arises to change a default value.

The default configuration contained in this CL contains the
various configuration values that have been proven to be
an effective and efficient default configuration for connecting
to a device running Fuchsia.

Change-Id: I6d0ab34de03edd8949244217c85fd6df1fa5f8ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950031
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>