[assembly] Zedboot can be built when unused
We are moving in the direction of building the platform
once-per-architecture and then the product consuming ALL the platform
artifacts and building whichever product variant they wish. We need to
be able to build and export Zedboot to Bazel even when it is unused.

Change-Id: Ibcde21b5724c88b27cb2b8c9ffae9bbacbbedb2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/951955
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>