[netstack3] Swallow ICMP errors in core
Prior to this CL, core notified bindings of ICMP errors received on ICMP
and UDP sockets. However, bindings does not have a need to receive these
errors (yet) so it was simply logging a warning.

This CL moves the log line to core, which allows for the removal of some
unnecessary abstractions.

The following bugs are being closed as they are made obsolete by this
change.

Fixed: 135413, 135202
Change-Id: I8c170e259b3cddbd7d506328f7451d5b6663ce40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952074
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>