[trace][sdk] Export trace-vthread in sdk
magma is moving out of tree and needs the tracing libraries it depends
on to be exposed to the sdk to facilitate that. trace-vthread is a
library that augments the existing tracing api with "vthreads", a
grouping of events that represent work done off the cpu, e.g. on the
gpu.

Magma uses the trace libraries in a slightly abnormal way in compared
to a normal trace-provider. It optionally links in a static version of
trace-engine based on if it is compiled with tracing or not. This means
it needs to independently depend on the static library and the headers.

To support this, we also expose to the sdk the static library and header
only versions of trace-vthread as well as header only versions of its
already sdk exported dependencies trace and trace-engine.

See also fxrev.dev/926130 for calibration cl.

Fixed: 126078
Fixed: 133962
Fixed: 133615

Change-Id: I49e8bf1abf48d232a927f3debe59550816e1d1fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926015
API-Review: John Bauman <jbauman@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Reviewed-by: John Bauman <jbauman@google.com>