[fidlc] Use parameterized tests for span_tests.cc
This CL changes span_tests.cc to use gtest's value parametrized tests
instead of hand-rolling a test reporting system. The latter always
printed a lot of output even on success, which was annoying because it
made it harder to find the failure output when tests failed.

Test: fx test fidlc-test
Change-Id: Icd9470addd3a503f536e29ac03e244b0739435df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950762
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>