[fxfs] add data checksum journal record
Instead of snooping the checksum out of the extent record at replay
time, store all the data checksums in the journal so they can be
evaluated. This lets us evaluate the checksum even if the mutation
doesn't have checksum information in it, or if that information is not
available, for instance if it's encrypted.

Test: fx test fxfs-test
Fixed: b/309357938
Change-Id: Ia4eb9a1e3f8c8a7099586fb66e87370ddcebd771
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946114
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>