[starnix] clone vvar instead of snapshotting it
This CL changes the logic of starnix's vm cloning (`snapshot_to`), by
adding a special case for the `vvar` mapping to treat is as a shared
mapping.

The behavior before this commit was that the `vvar` gets snapshotted
during a fork, rendering it useless for the child process, as it would
not get updated by the starnix kernel. By making a clone, writes to the
writable vvar mapping in the starnix kernel will be reflected as well.

Bug: fxbug.dev/308831372
Change-Id: I780656af33628705996c28aee9291df32334bde6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/949536
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: James Robinson <jamesr@google.com>