[netcfg] Remove temporary interface name policy
Remove the behavior where fallback temporary names would be used for
interface names if the first attempt to add an interface to the Netstack
failed. This led to names like "etht0" .. "etht1".

When a device is discovered by the Netstack, get the
PersistentIdentifier for the device, and determine whether Netcfg
already knows about that device. If it does, then we will remove the
existing interface prior to installing the new one. If Netcfg does not
know about the device, then we do nothing and install the new one.

When a device fails to be added to the Netstack with an AlreadyExists
error, this indicates the interface was installed from another source or
an interface was already installed by Netcfg with a different identifier
that has the same name. In either case, we will reject interface
installation and log it as an error.

Multiply:netstack-management-integration-test.cm

Fixed: b/309010918
Fixed: 56559
Fixed: 135087

Change-Id: I5bdf7d49d26291656208e6558078d1df356d98ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950632
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>