[scrutiny] Refactor static package collector
Move the logic for extracting the system image hash to util in
preparation for re-use of this logic by the pre-signing verifier.

Move the logic for verifying package merkle to util to enable re-use
between system image and static pkg list verification. This also allows
for re-use of logic for pre-signing.

Bug: b/310265223
Test: fx test //src/security/lib/scrutiny/utils:utils_test
Test: fx test //src/security/lib/scrutiny/plugins:plugins_test
Change-Id: I902572d60d8d844343b7763a1f151d3dc54cdf02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952873
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>