[debugger] DebugAgent component overhaul
This change modifies the way that DebugAgent is launched and discovered
by clients. DebugAgent itself is now started as a dynamic component
instance by way of the fuchsia.debugger.Launcher protocol. The launcher
implementation holds no state about any of the running instances of
DebugAgent, but is the broker for the DebugAgent collection.

Running instances of DebugAgent can be found in the "agents"
collection, and will be terminated when a debug_ipc client closes the
connection.

Change-Id: I9a2bf023608d55b7ea996c7ac5cf22d43525aa68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953860
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>