[audio][drivers][aml-g12] Add reset/properties to composite driver
The implementation is similar to the DAI driver, but it resets all 6
TDM engines (3 for input and 3 for output). The driver properties only
reports the clock domain and no is_input since this composite driver
supports both input and output.

Since the driver is not receiving configuration via metadata,
its get properties API does not list manufacturer/product names and
metadata in the board driver is removed.

Bug: b/300991607
Bug: 132252
Test: fx test -uov aml-g12-composite-test
Test: fx test -ouv audio_driver_tests -- --admin

Change-Id: I6d0272b97c4204fa3988939df1a11978623daf76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/951800
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>