[memalloc] Introduce MarkAsPeripheral
This method allowed marking a range as peripheral after initialization
of a pool. Requires that it doesnt overlap with non peripheral ranges.

It will coalesce and merge nodes were it can. This is required for
properly supporting peripheral ranges from the devicetree boot shim for
arm.

Change-Id: I20826efcbaa68f53ddf646735ccfd4f70749bac5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945337
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>