[boot-shim] Add devicetree-test-fixture to boot-shim-test library.
Move into directory and update build definition and references to this
testonly library. This is necessary for later patches when it needs to
be included outside of the boot shim library directory.

For consistency, test-helper has been moved to a testing subdir as well.

Bug: 136174
Change-Id: Icb62d82a151de3d96b43a37e6cd8811996211288
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941862
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>