[doc-checker] Improving error messages
This changes the error for a missing yaml file into
a message about a reference to a missing file so that the
file referenced in the error message exists.

This also cleans up the heuristics around unescaped reference links
in order to reduce the noise.

Change-Id: I92e11f314486e0e10191dde9e1c37bef62188999
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954194
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>