[virtio-guest] Rename driver class to GpuDeviceDriver.
The name is intended to help clarify that the class is a Driver instance
for the VIRTIO GPU device. We didn't use VirtioGpuDeviceDriver because
"Virtio" seems redundant with the namespace. This CL also adds a class
comment that explains this role.

Bug: 137145
Test: fx set core.qemu-x64 --args 'use_virtio_gpu_v2=true' && fx build
Change-Id: I1afbd9bf38761fcc6b5fd86d7dad342690d788b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955752
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>