[golang] Only the go_build() invocation should dep on the library, not the host test
As the dependency on the go_library() is handled by the go_build()
target, there's no need to create a duplicate dependency edge from
the host_test() as well.  Especially since the go_build() target
will restat its output and host_test() cannot.

The result here is that if a change to a file in a go_library()
results in no change to the test binary, the build will not
dirty any downstream targets after ninja restats the test binary
and finds that it's unchanged.

This can truncate the build after only a few actions, instead of
thousands.

This also shifts how deps which are host_test_data() need to be
handled in the case of variant toolchains.  Those may need to
be explicitly included using the base toolchain instead of the
variant toolchain (the implicitly default), if those test deps
are coming through the go_binary() or go_library() targets and
not through the test target itself.

Change-Id: Id48a46bead7f619f358da3f28658b2d50bddf9e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954427
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>