[kernel][vm] Fix bug in SupplyPagesLocked
The current code for `SupplyPagesLocked` does not correctly handle the
case where the PMM returns a `ZX_ERR_SHOULD_WAIT`. Consider the
following series of events:
1. `TakePagesLocked` is called on a VMO that has been compressed.
2. The resulting splice list is passed into `SupplyPagesLocked`.
3. `SupplyPagesLocked` starts iterating through the splice list. In
   doing so, it pops the first page P off the list.
4. Page P is compressed, so `SupplyPagesLocked` invokes
   `MakePageFromReference`, which attempts to allocate a page from the
   PMM to hold the decompressed contents.
5. The PMM returns `ZX_ERR_SHOULD_WAIT`, causing `SupplyPagesLocked`
   to return. Notice that page P is no longer in the splice list.
6. The caller of `SupplyPagesLocked` (`VmObjectPaged::SupplyPages`)
   notices that `ZX_ERR_SHOULD_WAIT` was returned and therefore waits on
   the page request populated by the PMM.
7. Once the page request has succeeded, `VmObjectPaged::SupplyPages`
   calls `VmCowPages::SupplyPagesLocked` with the same splice list
   again, hoping to resume the operation where it left off. However,
   page P is no longer in the splice list, so it is skipped.
This means that the contents of page P are lost, and all successive
pages are inserted at the wrong offset (off by exactly one page).

This CL fixes the issue by leaving page P in the splice list until
`MakePageFromReference` has succeeded. It does so by "peeking" at the
first page in the splice list, determining if it is a reference, and
then decompressing it prior to actually popping the page from the list.

Bug: 126153
Change-Id: I475905912eaa932d4b9924cb0f5654d2c035eb7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950652
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>