[wlan][policy] log roam decision metric
Record a metric for whether a better network is found to roam to.
Proactive roaming does not actually happen; for now the metric is logged
for when a roam would happen.

Currently the logic for whether to roam is very simple, and a roam
happens if connection quality is low enough to consider roaming and
bss selection would choose another bss instead of the current one.
Improvements are in progress and will be implemented before roaming is
turned on.

Test: added metric check in unit tests
Change-Id: Ib716514d0a03e777b2180506119f194ee17f0d9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955475
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>