[netstack3] Reduce NUD error log severity to debug
Currently we log an error when we receive an upper-layer confirmation
for a neighbor that is not in the neighbor table. However, the way the
code is written, TCP provides upper-layer confirmations to NUD
regardless of whether the reachable IP address is owned by the netstack,
and we never create a neighbor entry for ourself. So this is more of a
red herring as an error log.

Change-Id: I85e6fe9dd7ccf6f818064707f47e01519156faed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956752
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>