[netcfg] Differentiate WLAN from AP interface name
When device_class is specified for an interface naming rule, as it is
for the current default rules, include "ap" in the interface name
instead of treating it the same as "wlan". This will make it easier to
differentiate from the two devices and make it so the interface names
will not conflict.

Fixed: 137185
Bug: 135087
Bug: b/309010918
Change-Id: I0a8d02fda0e30fe906da7ba130213f0edf500297
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956573
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>