[kernel][vm] Refactor ArchVmAspaceInterface and X86 implementation
This CL adds an `InitRestricted` initializer to the
`ArchVmAspaceInterface` and refactors the x86 code to make use of it.
This allows us to view the restricted status of an aspace as notionally
const; it cannot be modified between `InitRestricted` and `Destroy`.
This will allow us to be a bit more precise when implementing the ARM
unified aspace code (which will be published in follow-up CLs).

This CL also renames `InitPrepopulated` to `InitShared` to match the
terminology used in the unified aspace documents. The distinction was
previously made to highlight the idea that a shared aspace may not need
to be prepopulated in the future, thereby saving 1MiB of overhead per
shared process. This distinction no longer seems significant, and we can
simplify the code by assuming that all shared aspaces are prepopulated.

Finally, since `is_restricted_`, `is_shared`, and `is_unified` are all
set by `Init` methods and not modified until `Destroy*` is called, they
are notionally constant for the lifetime of an aspace. Therefore, we can
use a single enum to store all of them, instead of using separate
booleans with separate locking rules; this CL makes this change for the
x86 code.

There should be no functional changes from this CL, and I have verified
this to be the case by running the stress tests locally.

Bug: 132980
Change-Id: I38bbf5efc837f6fb6c77f25e54721b48d893ed97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956332
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>