[netdevice] Make Init and AddPort asynchronous
In the current API NetworkDeviceImpl.Init and NetworkDeviceIfc.AddPort
are both synchronous. In the migration to FIDL this presents some
difficulties in the Banjo to FIDL shim because many implementations of
Init immediately call AddPort and wait for it to complete. This
introduces a re-entrancy that imposes stricter requirement on the
dispatchers used by netdevice and prevents in-lining of calls because of
the dispatcher.

Solve this issue by marking both NetworkDeviceImpl.Init and
NetworkDeviceIfc.AddPort as @async. For the FIDL this has no impact but
they are always asynchronous but it allows the Banjo to FIDL shim to
work asynchronously thereby relaxing the requirements on the dispatcher
used for the shim.

In this change implementations remain synchronous, using sync
completions or other mechanisms to maintain existing behavior.

Test: All available automated tests for affected drivers.
Change-Id: I946998f2128e3c5d1876f788a6b8d73d6614c592
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933220
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Owners-Override: Adam Barth <abarth@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>