[component_manager] include exposed protocols in sandbox-based routing
This reverts commit 7b339649d8be2e281efd7db575fe16e92e0751e1.

This commit makes senders fallible, so that they will return an error
instead of panicking if they receive a `TrySendError`. This error can
happen when a component is unresolved or destroyed in between when a
route to the component is completed and a message is sent to that
component. That error case is now handled gracefully.

Fixed: 308407836
Multiply: power-manager-integration-test-nelson, web_engine_integration-tests
Change-Id: I603a77b781add10ad8c30522443efe1431b19c96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957413
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>