[ui][test] conformance tests use subpackage build
In this cl, we add package for ui/testing/context
and (touch / mouse)-test-suite so that (touch /
mouse)-conformance-test build with subpackages.

This is a step stone to CTF build. We want to
land this first to ensure following changes is
valid in subpackages build.

Bug: b/315475197
Change-Id: Ib31ecdc66f9ef1a816f08bb80f6839058535727f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957855
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Jianpeng Chao <chaopeng@google.com>