[netcfg] Use MAC to identify named interfaces
It is possible for two interfaces to be active concurrently at the same
topological path, but have different MACs. Therefore, we should identify
interfaces by what the device is that is connected, not where it is
connected.

This change is a no-op to the default naming policy to name devices by
MAC or topological path.

Bug: b/309010918
Bug: 56559
Change-Id: Ibf6f6370d56601e94e55bff4814f520c834640e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955832
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>