[network/tests] delegated provisioning test base
Outline a framework for running delegated provisioning tests on fuchsia.

This test makes use of netemul-runner to setup two netstacks in a
hermetic network. On one side is a DHCP server, while on the other side
a non-fuchsia agent may run a DHCP client. The realm includes a test
component that waits for the client side interface to be assigned a DHCP
address.

Bug: b/287292428
Change-Id: I92a0a1277718cb0d03c37348363cf0da33db0ee3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954432
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>