[zircon_boot] support Android boot image on-disk
We previously supported Android boot image format when RAM-booting via
`fastboot boot`, this CL extends support so that we can also handle this
format on-disk.

Unlike `fastboot boot`, on-disk Android boot image does not support ZBI
+ vbmeta concatenation; vbmeta must still be loaded from its own
partition. This may change in the future, but we don't need it right
now and it seems like it might cause confusion (e.g. if both a
concatenated and on-disk vbmeta are given).

Bug: b/306194835
Change-Id: Ia78331883cf868fea7ff84ec6ae7d3c90b20c469
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957940
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>