[structured_config] Support transitional availability
If a config is marked as transitional, it's not an error if it's
not routed from void. This is important to allow for soft-transitions
for adding new configuration capabilities.

Add an integration test that checks that a transitional value
with no route will get the value from the CVF file.

Fixed: b/315170479
Change-Id: I1d9a87458b0107fed5be09bfe03ec27ef3aee498
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956880
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>