[build] Require approval for setting per test timeout
Historically infrastructure has not supported per-test timeout
configuration because it is hard to correctly maintain (e.g. knowing
when to raise or lower the timeout) and because it allows people to
write very slow tests that run in presubmit and slow down presubmit for
everyone. The status quo is to set test timeouts on the per-builder
level, so every test run by a builder has the same timeout, which blocks
people from landing extremely slow tests in presubmit.

We added support for per-test timeout configuration feature because it
was required for boot tests, but we want to limit its adoption beyond
boot tests until we have a better story for managing per-test timeouts
without impacting presubmit latencies.

Change-Id: I608442f236ce75ba359325a9eca2b91ed6019c68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954433
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>