[sdk][sys] Remove test for channel rights
We are updating the SDK VFS library to use the in-tree VFS library
as a backend, similar to //sdk/lib/svc. The latter is already used by
//sdk/lib/component and does not check for ZX_RIGHT_WAIT.

Remove the test that verifies this in //sdk/lib/sys, as we would like to
swap out the backend for the in-tree VFS soon. This will bring behavior
of sys::OutgoingDirectory in-line with component::OutgoingDirectory.

Bug: b/293936429
Test: No functional change.
Change-Id: I2bef889b31c2f83136226241cdc2aeaac4009bdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959758
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>