[sandbox] Validate that Path is a valid fuchsia.io path
After each Path transformation, we check that the path is still a valid
fuchsia.io path. This is intended to catch simple cases of specifying
the wrong path. In the long run we'd need a fallable Path API and also
an efficient way to validate parts of a path.

Change-Id: I8e2dfcd716748eba364d671786974c89f4bdbcee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959252
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>