[devshell][ssh] avoid recursion
At some point the devshell directory ended up in my path. This script
called itself with the symptoms of ssh hanging and RCS being N for
all targets.  A `ps` call showed a considerable number of 'ssh'
processes.

This change attempts to detect if the ssh in the PATH is the same as
the script that has been called.  Additionally, this change adds
checks for all external binary calls to fail with a message if they
do not exist.

TEST=MANUAL: fails noisily with devshell in path (ffx target ssh tells the user to remove devshell from their PATH) and runs normally otherwise
BUG=none

Change-Id: I257300bc5a845f1d2ea521a1c07c68bb25773736
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959914
Commit-Queue: Will Drewry <drewry@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>