[kernel][jtrace] Don't disable preemption if interrupts are disabled.
Don't attempt to disable and re-enable interrupts when recording a
JTRACE entry if interrupts are already disabled.  This can actually
trip an assert if we are holding a spinlock when we attempt to
re-enable preemption (if we were the ones to drop the last preempt
disable count and the code actually wanted to re-enable preemption).

We can avoid this hazard by skipping the preempt disable/enable if
interrupts are currently disabled, as they always will be when holding
a spinlock.  Additionally, there is no need to disable preemption if
interrupts are already disabled (irq disable is even stronger than
preempt disable).

Change-Id: I4c3444450930d9fe15cbb558624662660baa1243
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960997
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>