[assembly][bazel_sdk] Omaha OTA configuration rules for the Bazel SDK
Add a rule that defines Omaha OTA configurations for use with product
assembly, as well as helper functions for defining the OTA realms and
TUF repos that are used with those OTA configurations.

There are tests that validate that the generated OTA config files are
the correct schema, using a new rule that validates them with goldens.

Bug: b/315352756
Change-Id: I371f37958cbb2028e1591dc2c76c39fbda692166
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960025
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>