[fuchsia-sdk-run] Reload context after bootstrap
Reorganizes things in the bootstrap logic so that the configuration
changes that are made during bootstrap are preserved in the context when
the sdk loading logic happens. Previously, the sdk loading would happen
with an outdated environmentcontext resulting in a failure to load the
SDK at all.

Fixed: b/316403479
Change-Id: I02a7e266e40bda491e880b641dfe6c58fecde3c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961052
Reviewed-by: Colin Nelson <colnnelson@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>