[sdk][work-on] Set sdk.root in work-on script
This adds a setting for the sdk root in the work-on script used in out
of tree repositories, making it so that ffx can load the correct sdk
without relying on magic "where am I?" detection that's fallible.

This is not a full fix for everything in the referenced bug, but at
least gets things so that people don't have to manually set it anymore
and helps the fuchsia-sdk-run stub work better.

Bug: b/313680992
Change-Id: Ie66d081d954c80d65560a615246aaeec78c0d517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960312
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>