[build][bazel] Trigger Bazel SDK tests on toolchain changes.
Ensure that the Bazel SDK tests are invoked from Ninja when the
content of the Python and Clang prebuilt toolchain directories
change.

This is already the case for our main Bazel workspace, but not
for the Bazel SDK tests which run in a different workspace and
with separate output bases.

In more details:

- Move the content hash generation logic to its own script
  under //build/bazel/scripts/ + a new GN template to
  invoke it named generate_prebuilt_dir_content_hash().

- Invoke it both from update_workspace.py (which is specific
  to the Fuchsia build's internal workspace) and in
  //build/bazel/bazel_sdk/BUILD.gn (which sets up builds for
  a different workspace under //build/bazel_sdk/tests).

- Add missing local_version_file to @fuchsia_clang repository
  in the internal WORKSPACE.bazel file.

Bug: b/295575834
Change-Id: Iac97f8a73100f530533625a83fedf536e19e7a1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959033
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>