[starnix] Remove KType in KObject
# Problem
KType was an enum representing the different types of objects that
embed a kobject. This approach became problematic as more metadata was
needed for each type, making it difficult to match and assert against
specific types.

# Solution
KType is removed entirely in this CL. Instead, Bus, Class, and Device
structs are introduced. Each struct owns a weak reference to its
corresponding kobject. A new trait KObjectBased and macro
impl_kobject_based are added to simplify the definition of these
structs.

Overall, this CL aims to make the kobject framework more robust and
flexible by addressing the limitations of the KType enum and
providing a cleaner way to manage different types of embedded objects.

Bug: b/297438732

Change-Id: I1a54ba3e0e8eb360a7028c8637d941b8558f1f5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959860
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>