[storage] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I84d2e3f8a302b81c5dc8e2a95e2464c167911154
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963118
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>