[performance] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ie8b6a502f9d02b24264b43ad35d069effa18c157
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962820
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>