[forensics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I1e2aece7a6a937bea3f1308ad40d801e80c8a2e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963112
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>