[networking] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I84b1ecb7a4cb4c119fbd9e7d0ae26ec97cb933d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963208
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>