[debug_agent] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Iba5ecc3917f04f59139b51a390f0092fbf220408
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963279
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>