[battery_manager] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: If4bc37e150da907fae2b1dff94d8680f52d94b91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963278
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>