[audio] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I0986a73158caad3f6ceb3c241e1418c679b1c70e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963113
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>