[audio] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I31293823b1e240b9c419964bc197aa75e0b4aa82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963204
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>