[fxfs] Stop tracking byte counts for deleted object stores.
The inmemory tracking of the object owners was getting copied into the
AllocatorInfo on every flush begin, and then copied back out into the
inmemory tracking on flush end. So the allocation counts would stay in
the tree forever. Now we remove the inmemory version when the mutation
is applied to mark the object store for deletion, and ensure that it
gets cleaned up again on journal replay.

Bug: b/316984515
Change-Id: I448f6706fc3c8d04002f6b1943ba1797101314fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963473
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>