[bt-hci-emu] Work with DFv2 unbind
In DFv2, we teardown devices in reverse from DFv1. As a result, the
child hci_dev is destroyed prior the parent device. In order to deal
with this reversal, a few changes are necessary in the emulator such as
taking note of when the child goes away.

Change-Id: Ife19b13ece661968021312bea03d110952b5eb1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961199
Reviewed-by: Matthew Kehrt <kehrt@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>