[fuchsia-controller] Prepare for isolation.
Adds some extra debugging logs to the e2e tests. When supplying a
non-ip-address target, enables mDNS (for local testing), if a target is
an IP address, keeps mDNS disabled for isolation (should be the
de-facto case for testing in infra).

In addition, this adds `ffx target add` functionality to the fuchsia
controller context. In doing this some code was refactored in order for
common paths to be used between ffx and fuchsia-controller.

Change-Id: Id1aa755070577a4bbc0ae57b75002f2a65db0346
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961881
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>