[starnix] Get general purpose regs with ptrace
This includes ptrace_peekuser / getregset / getregs.

A note: The data structures to do this are:

a) Outside of the UAPI for x64
b) Partly in the UAPI for arm64
c) Entirely in the UAPI for riscv

The data structures that were added for x64 and arm64 came from
running bindgen on the appropriate userspace headers and then
copying the results into the starnix-uapi/src directory.

Change-Id: I5c79f07d3b73c30882e5b5a5230da37077a0f262
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917702
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>