[scrutiny] Support the new update package format
The logic for parsing the update package to find the ZBI has been
updated to accomodate teh new update package format. I also found that
this logic had been duplicated in a few different plugins, so I took the
initiative to address this and move the logic to a single place in the
ZbiPlugin. The rest of the plugins depend on ZbiPlugin to get the
relevant information.

I am confident this continues to work properly because the scrutiny
verifiers pass, all the tests pass, and I am able to manually query for
information in the scrutiny shell.

Bug: b/316915399, b/42180322
Change-Id: I89e06ee129ab4b472918d2734608e449569814a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962705
Reviewed-by: Laura Peskin <pesk@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>