[logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I2d1138102c9516cc9572d415c8666396d335d48f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964086
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>