[radar] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Id7726304b10c942b3ab6baff6cfca9ff1c50e9e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964085
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>