[graphics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I7a95146c1835e9d5142d5ad4f7c54c73b19252cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964255
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: John Rosasco <rosasco@google.com>