[fuzzer] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: If3ecb38786665e24e4ea3ef218eac1fac2c94fe7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964083
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>