[idk] Delete `generate_final_idk`.
For a little while now it's been an alias for a few rules defined in
terms of `idk` and `idk_archive`.  This unrolls the `generate_final_idk`
template and removes any of the rules that don't seem particularly
useful.

The meaning of the `final_fuchsia_idk.*` rules has been left the same,
just because it's hard to track down all the places that depend on them.
They might not all need to do the validation or build an archive, but
whatever.

Bug: 316598209
Change-Id: I478e7eb264833d888df3754fbcdd88c5a84d7ec0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964672
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>