[audio] virtual_audio Codec driver
This CL enables the virtual_audio driver to spawn instances
of the `Codec` audio driver type. It also adds an instance
of this driver to the set that is automatically tested by
audio_driver_tests. It also adds support to the CLI utility
virtual_audio_util, so that virtual instances of Codec can
be interactively customized and added/removed.

Bug: fxb/83972
Bug: b/309860869 (ADT codec cases)

Test: fx test -ouv audio_driver_tests --devfs-only
Test: virtual_audio --codec --in --add --wait

Change-Id: I73a040e712dae6c9984dcf06ecdf0e5d21a065df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953575
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>