[graphics] Remove syslog dependency
Drivers shouldn't be using syslog, or any libraries
that transitively include syslog.
This CL removes the direct dependency on syslog and
refactors the dependencies such that syslog is only
used from the test binary and not from the binary.
FSL transitively depends on syslog which is why
that dependency was removed from the driver.

Bug: b/300157652
Change-Id: I6bfb1e27dc67831a53b2948aab24161e5cf7536e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964938
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>