[superproject] Roll llvm-project/libc [libc][NFC] Unify `FPBits` implementations (#76033)
`FPBits` is currently implemented as a general case and a specialization
for `long double` when `long double` is x86 80-bit extended precision.
This patch is a first of a series to provide an implementation based on
`FPType` instead of the C++ float type (which implementation is
architecture dependent).

GitOrigin-RevId: 39f808d9c351149faf90fed9d3a2f84a50c077d8
Original-Revision: c5f71d79e9ca6d140cbc1044f146c9eae2bd64a4
Roller-URL: https://ci.chromium.org/b/8761059765137300913
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0a1e920d04b3124720ee698bc772284c904761de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965155