[debugger] Restore debugging state from existing DebugAgent
...when reconnecting.

This change adds the ability for zxdb to connect to already existing
instances of DebugAgent, which may be attached to and/or have existing
configuration (e.g. breakpoints installed, filters set, etc).
Information is restored on a best-case basis. Some data, such as a
breakpoint installed at file.cc:3 will only be seen as an address in the
new zxdb session, since DebugAgent doesn't know about the location
information that produced an address.

Change-Id: I184e4d4548e5fb9cd43c8e952cad377d7cce864b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961863
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>