[fx sync-to] Improve UX for non-GI checkouts
Previously, trying to run `fx repro <bid>` where the build referenced by
`<bid>` used a non-GI integration remote produced the following error:

  ERROR: found no project/import/package named <project>

It also asked the user for a y/n confirmation and changed the checkout
to point to JIRI_HEAD even though the operation was bound to fail, which
is confusing and unnecessary.

Now it will produce the error message:

  ERROR: Integration remote '<build-remote>' does not match local checkout remote
'<local-remote>'.

The error message is also produced much faster and without changing the
checkout state or requiring user confirmation.

Change-Id: I00a1abc5f048c69840a32d178ab7c08501dcfced
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965953
Reviewed-by: Aidan Wolter <awolter@google.com>