[ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ie7afeb4d0ff734727c4158710aaba639ea31d211
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966642
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yilong Li <liyl@google.com>