[pkg] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I2c5392b2165872f5bad35be5a89e081a98867106
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966813
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>