[amlogic-display][viu] Do not set up VIU layers on initialization.
Previoiusly in AmlogicDisplay::DisplayInit(), the VideoInputUnit
instance needed to be created using a valid layer/display configuration,
and the layer was set up at initialization time.

However, this layer setup was unnecessary, because every ApplyConfig()
will call VideoInputUnit::FlipOnVsync() which can update the VIU layer
setup.

This makes it possible to remove the layer image size and display
contents size arguments in the VideoInputUnit ctor and factory
functions, and thus removes the dependency of VideoInputUnit on Vout.

Test: core.vim3 (display engine is reset on Fuchsia boot)
Test: core.nelson (display engine is not reset on Fuchsia boot)

Change-Id: Ib5e8d92f9d598bda1192f9e7b9edb3c646a46cfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965585
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>