[system-update-checker] Use the new image package manifest
The system-update-checker compares the active system to the system
described by the update package to see if a new update is available.

The update package format was recently migrated from storing images
(e.g. zbi, vbmeta) directly to storing a manifest that lists their
hashes and URLs from which they can be resolved. The
system-update-checker was not migrated, so it was still looking for
the images in the update package, which broke the check that
determined if the zbi or vbmeta were out-of-date, so the
system-update-checker would erroneously think the system was
up-to-date if only the zbi or vbmeta had changed. This would commonly
surface as kernel devs making a zbi-only change but then `fx ota`
would not update their device.

With this change, the system-update-checker uses the new image
manifest and so should again correctly trigger an update on zbi-only
changes.

Fixed: b/317524358
Change-Id: Ib6377bed206a6094b1a4b7c9750cacf25b9bda57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966533
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>