Reland "[aml-i2c] Implement FIDL instead of banjo"
This is a reland of commit 20cc939e39367a579b17fb0207e85c38f2853375

Original change resulted in the dfv1-version of the aml-i2c driver
trying to dereference junk pointers. This is because of two reasons.
The first reason is that the driver created an outgoing directory
without providing a pointer to a dispatcher.

The second reason is that the driver's outgoing directory was moved
after it started serving the i2c impl fidl service. The outgoing
directory creates callbacks that hold a reference to itself and when the
outgoing directory is moved then those references point to junk.

Also, nodes that bound to the dfv1-version of the aml-i2c driver were
unable to connect to the i2c impl fidl service that should be provided
by the aml-i2c driver

Original change's description:
> Make the `aml-i2c` drivers implement the FIDL version of the i2cimpl
> protocol instead of the banjo version.
>
> Remove the i2c testing from platform bus tests because the i2c driver no
> longer uses the i2c and i2cimpl banjo protocols.
>
> Bug: b/301004275
> Change-Id: I0dfaea0e8f9acbeb6203f3674799705720ea7d1e
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957853
> Commit-Queue: Nick Eaton <nickeaton@google.com>
> Reviewed-by: Garratt Gallagher <garratt@google.com>

Bug: b/301004275
Change-Id: Iec06c8ee5b135d4aecc8b82bc01e8173682a26db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964012
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>