[drivers] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I9109d5d7bdb73ecb6f4776dbe15cc44aec1dc26c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966774
Reviewed-by: Nick Eaton <nickeaton@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>