Reland "[physboot] Synthesize UART peripheral range"
This is a reland of commit 04553c50d166a8c619dc00d06cc938e1237dba9f

`extra_mem_config_ranges` was wrongly incremented unconditionally. This
means when serial is set to none(disabled) in user builds, there is an
extra uninitialized range handed off to the kernel.

This range can be anything, in particular, will be treated as a reserved
range, which cannot overlap with anything. Depending on the values, it
might overlap with specific ranges causing a kernel panic, which would
be observed as a boot loop and then rolling back to previous good build.

Original change's description:
> [physboot] Synthesize UART peripheral range
>
> Synthesize ARM Peripheral range for UART, since its not provided by the
> devicetree boot shim.
>
> Bug: 136174
> Change-Id: I47b8ec5363c74c67439acbf9284505eee8fcfa31
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941697
> Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
> Commit-Queue: Gianfranco Valentino <gevalentino@google.com>

Bug: 136174
Change-Id: I0e980dc8392048a0c2b8d48f219c229a1b7a452c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960100
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>