[build][rust] Match runtime deps against actual triple passed to rustc
In order to move to the canonical triple we need to supply a toolchain
with runtimes built for both the old and new triples, for a soft
transition. Currently this breaks the build because we always look for
the canonical triple (the one we pass to clang, with "-unknown-") in
runtime.json.

This works today because we supply both the canonical and non-canonical
triple as targets to match against in runtime.json. When performing the
soft transition, however, we need to include only the actual triple for
each runtime, so we need the build to match against the actual triple
it's passing to rustc.

http://fxrev.dev/968253 starts the soft transition once this change is
merged. The logic in this CL can be removed once we finish the migration
to the canonical triple.

Bug: 310271984
Change-Id: I190a6f823c5ef9f67dc5b99634aed82d778227d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/967421
Fuchsia-Auto-Submit: Tyler Mandry <tmandry@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>