[kernel][devicetree] Extend matcher API with OnDone()
This matcher method is called when a matcher reports kDone, and is a
useful place in which to put logic that finalizes the decoded
information (as will be convenient for the RISC-V CPU topology matcher).

Instead of extending the matcher API we could have indeed just included
a "finalize" method on any matcher that needed this logic, to be
manually called at the site orchestrating the matching. However, that
is more error-prone, as nothing enforces it actually being called.
Moreover, all matchers of interest inherit from a base class on which a
OnDone() stub may be defined, so no actual repeated boilerplate should
be required of matchers in the common case that have nothing to
finalize.

Now for the grammatically awkward choice of "OnDone": the "On"-prefix
matches every other part of the current matcher API, "Done" makes it
clear that this pertains only to when the matcher returns `kDone`, and
"OnDone" has been used in other projects (like gRPC's C++ bindings).

Bug: 124336
Change-Id: Icccba46ce930af987ca6830b1c7d84db03bda89c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959565
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>