[bt][avrcp] Workaround for short browsing packets
Some phones send us L2CAP packets in the browsing channel which are
missing the FCS bytes, even though we have not turned FCS off.  Those
bytes are chopped off in the host (unchecked) and we don't receive them.

Try a bit harder to be resilient to this situation until the issue is
fixed in either host or peer stacks.

Bug: b/316628374
Test: added tests
Change-Id: Ic1bdbf3a276d72f909bdbbabdc16dc46bf95fdb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965954
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>