[tests/dhcp] Fix address-removal-stops-client test
We have a test case checking that removing a DHCP-acquired address stops
the DHCP client, and that removing an unrelated address does not stop
it. The latter case relies on watching the DHCP client for renewal
traffic as an indication that it continues to operate after the
unrelated address is removed. However, this case set the test timeout
to be short enough that the client would not initiate renewal before
the timeout expired. To fix this, specify a shorter lease length to
prompt renewal sooner.

This was also causing the overall suite runtime to be long enough that
the dhcp-integration suite would occasionally time out entirely.

Bug: 42083184
Multiply: dhcp-integration
Change-Id: I2e8f190af299d208b630f95993b02b9e6e39351a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969412
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>