[sysmem] Remove sysmem syslog dep, for now.
We may essentially undo some of this after we've addressed b/299990391
and b/303308425, assuming the solution to those works for drivers also.

This CL is forced to pull in src/media files because of an implicit
indirect dependency on syslog via metrics_buffer - rather than making
that a direct dependency, just remove that dependency also.

BUG: b/317811365
Change-Id: I33d6e4a123aaf7013dc111745930c0a66b73184f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969932
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>