[tefmocheck] Add check for blob header timeouts
This is causing a number of tests to occasionally fail. Adding a
tefmocheck will both serve to better surface the underlying failure
mode, and also cause flake fetcher to start filing bugs for this failure
mode and not for each individual test that fails because of it.

Run-All-Tests: true
Bug: 318087737
Change-Id: Ibbd82ccd3445afc6e93f9d0075997d6fa250815a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970096
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>