[netstack] Change interface token to EVENT
It's not necessary to use an EVENTPAIR, since the object is only being
used as an opaque token, and not for any specific functionality.  It
also slightly simplifies the implementation, because the netstack will
only need to keep track of a single kernel object.

Bug: b/42069034, b/42084185
Change-Id: Ie797e3603815f3422da1e67ece51cef5f96373f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969612
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>