[starnix] impl ioctl() for action UI_DEV_CREATE for keyboard
In this CL we create a UinputDevice method `ui_dev_create` which is
called when ioctl receives a UI_DEV_CREATE request.

In `ui_dev_create` we register a fake device that we will use to inject
input events to Input Pipeline from Uinput. We store the client end of
the FIDL endpoints in UinputDevice so that we can access it later in
`write()` when we would like to inject events to Input Pipeline.

Bug: b/308156116
Change-Id: Ifcd06bc15a6a941e49b63d6c4b6e8727f4758f6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941916
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Justin Kobler <jkobler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>