[assembly] Support diffing GN and Bazel lists
Lists of paths in product configs can now be diffed across GN and Bazel.
To do this we now keep a list of nodes that are traversing and continue
to replace this list with the set of nodes at the next element.

For this json:
  root: {
    one: {
      two: [
        {
          three: path
        }
      ]
    }
  }

And this key: one.two[].three

We would walk down the elements one at time like so:
  element=one, config_nodes=[root], new_config_nodes=[one]
  element=two[], config_nodes=[one], new_config_nodes=[{three: path}]
  element=three, config_nodes=[{three: path}], exit

Bug: b/318763523
Change-Id: Idee8ad3f543e73fd9a063e0314b1797852f47d2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970701
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Iserovich <diserovich@google.com>