[rust][3p] Limit parking_lot to host-side targets.
parking_lot has a few disadvantages for Fuchsia:

* adaptive spin w/ thread yields is great for microbenchmarks
  but has poor system-wide interactions with interrupts and
  the scheduler
* the parking lot abstraction doesn't convey priority inheritance
* implicit heap allocations make it unsuitable for use cases
  like instrumentation code

By contrast, fuchsia-sync is built directly on top of our libsync
and Zircon futexes. It tends to perform slightly worse on
microbenchmarks because it lacks an adaptive spin but feedback
from the kernel team is that we should see better whole-system
performance with this implementation. It's also possible that
changing synchronization strategies will surface new test flakes
in code with latent timing bugs.

Limits visibility to fuchsia_sync so that in-tree code
goes through that API surface.

Bug: b/42169853
Change-Id: I2ce9be2978e8b61cbf852d6c7953c971c873facf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966916
Reviewed-by: David Koloski <dkoloski@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>