[fxl_cli] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ib3b892e59965fa480dedac79b0c974d4d7e547cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971445
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Adam Barth <abarth@google.com>