[build][cargo] Handle JSONDecoder error in generate_cargo.py
This adds handling json.decoder.JSONDecoderError to the
generate_cargo.py script. It follows how IOError is handled
in that it returns 0 so CQ is not broken by a failure.

It also adds the error message to the output to help diagnosing
what caused the error.

This change is motivated by the error encountered in
https://ci.chromium.org/ui/p/fuchsia/builders/try/core.x64-release-ffx_subtools/b8759271432932632401/overview
from fxr/971434

Change-Id: I0d37d0ef61297ac79eb06d3b8067ebcb5177cd78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972115
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>