[chunked-compression] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I48d602030f7973f8bc6643761024e5b36e4e70d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972124
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>