[fidl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ieb41c20e4681a4075d6b4e304ce6cb7fddfa4dfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972123
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>