[testing] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I38371a679f2f6069af51ab1f6a3d2cafaa1093d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972125
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>