[loadbench] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ieb6f0f0760499e3435406b7c27aba0fb1ad5da7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972126
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>