[zxdb] Make it possible to disable output
This CL makes it possible to disable output in the Console object. The
Console now starts with output disabled and clients must explicitly
enable output.

Not everything that appears on the console is considered "output" for
this purpose. Information that arrives via `LOGS` and the FD Steamer
is not disabled in this way.

This CL should not have any observable consequence. It's preparing us to
add a command line flag to control what output appears in the console.

Bug: b/319306239
Change-Id: Ia156d5661e1d65784af8ff4b6a1376fb475d3de6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972114
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>