[http_server] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ic9635dc57d3a520f1444388413fa5fdc8aec639a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972018
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>