[vfs][rust] Refactor `Connection::entry_not_found(..)`
`Connection::entry_not_found(..)` seems to be called when an entry is
not found, and it will create a new entry given the specified
`NewEntryType` if the caller indicates that a new entry is to be
created via the `create` boolean. If the caller calls `entry_not_found`
with `create` as false, `entry_not_found` always returns a `NOT_FOUND`
error.

This refactor involves changing `entry_not_found(..)` to be called only
when the caller wants to create the entry (and hence the name change
`create_entry(..)`). This allows the caller to be able to return early
with the `NOT_FOUND` error if entry is determined to not be created.

Bug: b/293947862
Change-Id: I4585008f062183841d9f85eab86e19551d908443
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966457
Reviewed-by: Brandon Castellano <bcastell@google.com>
Size-Review: James Sullivan <jfsulliv@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>