[botanist] Enforce using ffx and product bundles.
This should go in after https://fxrev.dev/970663 enforces that the ffx
tool and product bundle are provided to the botanist command. The only
place product bundles will still not be used is for cuckoo tests which
still use image overrides.

This also removes support for using the flash script since ffx flashing
is enabled on experiment level 0 so we only support ffx flashing where
the devices are idling in fastboot.

Bug: 42083611, 42172560, 42172906
Change-Id: I5307695c3abaf6139b90342b155cc83b4207f8eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970145
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Vinicius Felizardo <felizardo@google.com>