[callback] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.
Also remove an unused #include in the header file.

Bug: b/300157652
Change-Id: If8bcdb314d3063f9de87068320d1842eaa2b5927
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972020
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Brian Bosak <bbosak@google.com>