[audio] Log ERROR on renderer/capturer disconnect
In all cases where an AudioRenderer or AudioCapturer (or the
ultrasound equivalent) disconnects as a result of a client
call (usually because of bad params or incorrect sequencing
of method calls), log at ERROR level, not WARNING.

In addition, disconnect in one circumstance where we did not
previously, which was inconsistent with all similar cases.
Specifically, if a client called AudioCapturer::CaptureAt()
with an incorrect `payload_buffer_id`, previously we would
just ignore the call. In all other situations where a client
specifies an incorrect payload_buffer_id, audio_core aborts
the renderer/capturer connection, and that is now what it
does in this situation also. New test case has been added.

Multiply: audio-renderer-test
Multiply: audio-capturer-test

Change-Id: I0690b6a61fda7f1c51ba366002de1a4961f72ed1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966079
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>