[kernel] Remove unused logic for updating in-flight observe bits on packet
In an earlier iteration of the system ports would accumulate observed
signals from the time that a packet was queued until it was dispatched
to userspace. This was changed to "snapshot-early" behavior as part of
https://fxbug.dev/42128657 in early 2020. Now the initial packet for a
signal watch is not modified after it is queued in a port. This removes
the unused signal updating logic from the port dispatcher.

Bug: b/42128657

Change-Id: Ibb06cf77d08cc45b86739ac744c0d09c980c376d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965230
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>