[fuchsia-repo] Fallback to regular copying if FICLONE fails
On some devices the ioctl FICLONE is failing with
"Inappropriate ioctl for device (os error 25)". When converted to a
Rust std::io::Error, it has kind "Uncategorized" instead of
"Unsupported", so our fallback to regular copying is not triggering.

This commit updates the error handling to wrap the ioctl error in an
error of kind "Unsupported" so the fallback should trigger.

Change-Id: Iff73d08050706dd6b060f8b7504816778e4da6ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973173
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>