[bazel] allow driver verification check choose comparison type
The driver verification tool which checks to see if drivers are being
built the same between gn and bazel only checked for exact equality of
sizes. In debug builds the sizes don't agree but we only care that the
bazel build is not getting larger than GN. this lets us check for
equality in release mode but fallback to just making sure bazel doesn't
outgrow gn in debug builds.

In future changes we will look into why debug build sizes differ but
this allows us to bypass that for now.

Change-Id: I6bf5f6c5c678601178f54e9ca87b311029002aac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973057
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>