[wlan][policy] Allow selecting recovery behavior.
This change adds configurable fields to the wlancfg
component manifest that can be populated through platform
configuration.  These fields select the recovery profile to
use in determining whether or not recovery should trigger as
well as what to do with the recovery recommendations.  The
contents of this change allow for the recommendations to
either be ignored or logged.  Future changes will allow for
the recovery recommendations to be acted upon.

Bug: 60089
Test: Manually booted with defaults to ensure wlancfg picks
      up the configuration values and does not crash.
Change-Id: I758328082f2707cf52c3aec0a1f9b6e84d744a57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953871
Reviewed-by: Hayden Nix <haydennix@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>