[build] Use -ftemplate-backtrace-limit=0 everywhere
This was limited to the "Zircon" toolchains for no real reason.
This is a generic option that simply improves usability of
compiler diagnostics in arcane C++ template cases, which has
nothing to do with which kind of build environment C++ code is
being built in.

Change-Id: I285774fb472fed49636a4a918efe074ecda08de1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973513
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>