[sdk][vfs] Initial drop-in replacement for SDK VFS
Add new drop-in replacement for SDK VFS that uses the in-tree VFS as a
backend. Add test harness to verify the same behavior and interface.

Update in-tree VmoFile constants and type names for the default sharing
mode, and remove the `vmo_sharing()` getter (was only used to verify a
trivial property in unit tests).

Bug: b/293936429
Test: fx test sdkcpp-new-io1-conformance-test
Change-Id: I97d1a1d9576b7a3f48c4459b7a81d36b657a5623
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939897
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Chris Suter <csuter@google.com>