[starnix] Stub /proc/swaps
This implementation is enough to pass the remaining swapon LTP tests. In
order to improve this implementation, we either need to make
`swap_files` in `Kernel` into a regular `Mutex` or we need to plumb the
lock state through `FileOps` so that we can take the current
`OrderedMutex`.

Fixed: b/287121394
Change-Id: I70bea39b848ffdeca662079071f798f39b959c11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973914
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Maria Glukhova <mariagl@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>