[vfs][rust] Fix tracing across async boundaries
Trace duration events are rendered as an execution stack in trace
viewing UI so DurationScope objects should only be stored on the stack.
Holding a DurationScope across an await stores the DurationScope inside
of a Future on the heap. Trace durations used in this way may encompass
the execution of other tasks when rendered in the UI which quickly
becomes confusing. The rust vfs is also compatible with multithreaded
executors so a Future containing a DurationScope can be moved between
threads but the duration event will only be rendered in the UI on the
thread where the DurationScope was dropped.

This change replaces DurationScopes with TraceFutures where
DurationScopes were held across awaits. TraceFutures are designed for
tracing async code but do unfortunately clutter the code a bit more.

Bug: b/319162213
Change-Id: Ifef929f6b66ec8808fe06630dc4e7a7c3c5f2ac8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970894
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>