[ffx][flash]: flash uses event based reconnect
Now, in the flash process, when rebooting the bootloader (after flashing
the new bootloader) instead of using a configurable sleep to wait for
the target to reappear (and continue to flash the product), the tool
will use the discovery library to begin looking for the target to
reappear in fastboot mode.

This assumes:
  1. A target will never change fastboot connection mechanisms. For
     example, it is assumed that if a target is connected via USB, after
     the bootloader reboots it will continue to be available via USB
  2. A target's name (for fastboot over TCP/UDP) is consistent across
     reboots.
  3. A target's serial number (for fastboot over USB) is consistent
     across reboots.

Release-Notes: ffx no longer uses the daemon for flashing.
Change-Id: Ia8195092fc6fc4cadbcd0ca0a89c5b3721e67d77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953734
Commit-Queue: Colin Nelson <colnnelson@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>