[ld] RemoteAbiHeap
This adds the ld::RemoteAbiHeap class, which is the second piece
of the ABI remoting implementation.  It both gives access to the
stub dynamic linker's data segment where the ABI symbols' data
gets filled in, and extends that segment to hold the additional
dynamically-allocated "heap" arrays and strings that the ABI data
structures need to point to.

During testing it arose that an assertion failure and crash would
result when the TestPipeReader destructor ran in a test that
failed before its Init() had ever been called, so the destructor
is made more robust.

Bug: 318041873
Change-Id: Idea00f0e41933a2ca1549d4ef94ae04e41e5c3ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970293
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>