[netcfg] Modify bus path parsing strategy
Topological paths for devices aren't standardized in any form and the
position of a bus path isn't guaranteed to appear in any particular path
component.

The current implementation assumes the bus id appears immediately after
`/platform/` for SDIO devices and immediately after `/PCI0/bus/` for PCI
devices. That hasn't held true for platform-attached devices, and likely
won't be true for cases where there are multiple PCI buses. Further, the
code "matches" the path by filtering out any non-hexadecimal digits from
the path component, which means it will happily match non-numeric
inputs.

This changes the strategy to search for anything that looks like a bus
path in any of the path components (and to be stricter about what that
means) and it returns the first one it finds.

When nothing can be found, it returns an identifiable string to make it
clear that's what's happened, while also avoiding name collisions.

Fixed: b/319153091

Change-Id: Id15f51da8944e1e5f15366855372973d8967785b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972775
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>