[third_party] Add arrayref to the Cargo.toml
The arrayref crate is already present as a dep for another crate, it was
missing from the Cargo.toml previously and so was not listed in the
BUILD.gn either. Now it is properly tracked and can be imported to the
rest of the tree.

Change-Id: Ic4e38312a97bdd36479dde0e4266da14b3bccc3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975712
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>