[starnix][selinux] Use hard-coded fs labels in Enable mode
Currently SELinux is not yet implemented for the file subsystem, and
hardcoded labels are used in SELinux Fake mode. Also, the SELinux
testsuite container is currently running in Fake mode.

In order to be able to switch the test container to Enable mode, use
hard-coded labels in Enable mode as well until the file subsystem is
implemented in SELinux.

Bug: 304737892

Change-Id: Ic7514825882f104c48589cebd2a66711bdde46af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975713
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>
Reviewed-by: Wez <wez@google.com>