[bt] Remove bt-rfcomm from core group
`bt-rfcomm` is not required for core Bluetooth functionality. It is
declared as a static child of `bt-init`. However, `bt-init` can function
if RFCOMM is not available - `bredr.Profile` requests get forwarded
directly to bt-host.

`bt-rfcomm` will instead be included on a per-product basis. Currently,
`bt-hfp-audio-gateway` is the only BT profile that requires RFCOMM
support so we opt to include the RFCOMM package in configurations that
include HFP-AG.

Bug: b/42059641
Test: fx build, fx reboot device and verify `bt-rfcomm.cm` starts up OK.
Change-Id: I4b4f3f2656c841ac67393e740745048e58399202
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972954
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>