[amlogic-display][hdmi-host] Replace cea_timing with DisplayTiming.
cea_timing is almost identical to the display::DisplayTiming struct.
This change replaces the cea_timing with the DisplayTiming struct
to store timing information to configure the HDMI encoder.

This change also deletes the cea_timing struct now that it doesn't
have any client.

The only field that may be different is the `venc_pixel_repeat` field
which is used by some Amlogic-provided code to configure the pixel
repetition differently in the HDMI transmitter and the encoder ENCI.
None of the timings that need pixel repetition are currently
supported by Fuchsia, so we hardcode it to be false.

This change only replaces the field names in cea_timing with
orresponding field / method names in DisplayTiming mechanically.
It doesn't simplify any of the encoder configuration logic. This will
be done in a follow-up change.

Test: display-tool squares 1 on VIM3 with MIMO 10" display
Test: display-tool squares 1 on VIM3 with Surenoo SUR480480Y021A
Test: display-tool squares 1 on VIM3 with Amelin AML028-30MB-A1
Bug: 136952
Bug: 135377

Change-Id: Ida632b0422b20c21919630af0793863daefa097d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/951658
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>