[blobfs] Include blob in corrupt error message
Include the name of the blob in the error message that gets logged when
a blob was previously found to be corrupt. The name of the blob gets
logged with the original corruption error but that can get dropped if
the device has been powered on for a long time.

Bug: b/304770949
Change-Id: I9ef3c8cd0790d455835c13f7c64c33b8580864dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977156
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>