[netstack3] Repurpose SocketZonedIpAddr
Rename `SocketZonedIpAddr` to `StrictlyZonedAddr` and adjust the
invariants upheld by the type.

The original intent for this type was to update it to witness
`NonMappedAddr`, but in practice the places that need a zoned address
and a `NonMappedAddr` have been written to use
`ZonedAddr<SocketIpAddr<A>>` instead.

However, in a few instances, this type has organically evolved to
represent an address that is guaranteed to have a zone, when its scope
requires a zone. This change codifies this role, by renaming the type,
updating the doc comments, and removing the `From<ZonedAddr<_, _>>` impl
that circumvented the witness.

Any places where this new role is no longer relevant have stopped using
the `SocketZonedIpAddr` type and instead are using
`ZonedAddr<SpecifiedAddr<_>>`.

Bug: b/42082235, b/318875810

Change-Id: Idb33c5c2e01d68c1f19909bb16f933c037be5a10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969698
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>