[build] Update assert_no_dep restriction
The intent of assert_no_dep on //sdk/lib/syslog
was to prevent drivers from depending on syslog.
As currently written, this prevents drivers +
the tools used to compile them from depending on syslog.
This changes it to only perform this check when
compiling for target.

Fixed: b/321310121
Change-Id: Ifb3ab118392b24bb7d110f173e984b6b027d7e59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977150
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>