[devicetree] Update devicetree boot item check
Updated logs for when there is more than 1 boot item. The current
decision is to allow for multiple devicetrees to be part of the boot
items. The one built into the kernel image is always the first one. We
only parse that one. In future we may decide to consolidate both
bootloader and fuchsia specific devicetree.

Bug: b/319913179
Change-Id: I8c441cbdbfeb89de8a814b6fb7945b293e47b017
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976801
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>