[audio][test] Add pipeline error test
Previous CL created audio-core-api-error-tests, a suite to
execute test cases that are known to generate system logging
at ERROR level. This suite is allowed to run on DEBUG builds
and emulators.

Included in that set of tests was one case from the suite
audio-renderer-api-pipeline-tests, which cannot reliably
execute on non-realtime systems and thus is limited to only
release builds and non-emulators. This test case is now
flaking in CQ, so this CL moves the test case to a new suite
audio-core-api-pipeline-error-tests that does not execute on
debug builds or emulators.

Multiply: audio-core-api-error-tests
Multiply: audio-core-api-pipeline-error-tests

Change-Id: I2e3aac7b13dfa6a7180a2ae1958a160dc8353b61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978276
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>