[ffx] Simplify target address queries
This removes the distinction between queries with and without a port.
`<address>:0` now functions as a wildcard and is equivalent to `<address>`.
Before this, there was a tricky situation where AddrPort would always fail if the target address had a port.

SSH port handling has been improved.
_Technically_ Fastboot targets should set a different default port, but in practice the port isn't specified in those queries.

Fixes: b/319732734
Change-Id: I91fe8ee7c3f89c073d94104c4567a16442d4712a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975097
Commit-Queue: Dwayne Slater <dwayneslater@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>