[debugdata] Fix on-target profile merging for Rust
https://reviews.llvm.org/D138846 changed the layout of the profiles by
adding new fields into the profile header and data. This caused an issue
in on-target merging in Rust profiles because debugdata uses
InstrProfData.inc shipped by the Clang toolchain. This patch adds
support of merging profiles in the old layout as a workaround to recover
the coverage drop in Rust. It also cleans up the code that was added
as a soft transition of toolchain rolls, and regenerates the profiles
that use old profile versions.

Bug: 42086151
Change-Id: If49be38be6198eee2a44a9a2bff3df7fdc510c7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977320
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>