[emu] Stop reading flags template from product bundle
Currently, there is a template file that is included in the product
bundle to that is used to generate the command line flags for starting
the emulator for that product.

As things have evolved, it has turned out it is easier/better to
have a "universal" template that has conditional logic in it to
accommodate multiple cpu types and virtual device configurations, vs.
supporting the flexibility of having a template per product.

This CL stops reading the template path from the product bundle
and instead includes the template in the emulator subtool.

Follow-on CLs will remove the template path from the product bundle.

Bug: b/295031514
Change-Id: Ifcd80c394bbebdd80f25bb2ef38d2804f07f1d9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979192
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>