[bt] Define the host-watcher crate
Define a Fuchsia-specific crate for watching the status of the currently
active Bluetooth host on the system. This crate uses the
fuchsia.bluetooth.sys.HostWatcher protocol to monitor changes in the
active Host. The status changes are available as a stream of events. A
client of this library will instantiate the HostWatcher, poll the
stream, and dispatch event-specific work.

Refactor HostWatcher from the Fast Pair Provider implementation.

Test: fx test host-watcher-tests

Change-Id: I639ca525d412753b31bd75fbcbefb62eff5960db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959499
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>