[uart] AMLogic driver update.
(1) After setting the bits to reset the port (reset rx, tx, clear error)
they must be set back to zero or IRQ mode will misbehave. This could
have been done on `InitInterrupt` but seems better to complete the reset
in the same scope that is initiated. The observed behavior without
clearing this bits is that transmitted data was read as random bytes in
the host.

(2) IRQ Handler was rewritten to use rx_fifo_count which is the number
of elements in the RX fifo. Additionally this bounds the number of
operations in a single IRQ to UP to fifo characters.

(3)Enable Rx/Tx Interrupt were clearing the entire Uart Control Register
when called, and just setting the one bit. They must be initialized
from the current hardware state.

This drive was tested and stressed in a vim3.

Change-Id: I3c25611b3ad3fed39511f182416d11cbfa95d07d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979593
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>