[bt][avrcp] Adjust set volume value based on previous failure.
Sometimes, set volume request can get stuck in a bad state when the
controller keeps sending the request with the same volume value.
This happens because AVRCP and Fuchsia audio settings use two different
ranges for audio values. [0, 127] for AVRCP and [0.0, 1.0] for Fuchsia
audio settings.
Attempt to mitigate this bad state by remembering the requested value of the
latest failed request and adjusting the volume value when a request with
the same volume value comes in the future.

Bug: b/316961553
Test: bt-a2dp-tests
Change-Id: I5816fa5c85938ab0e4917f56c834b0883e9195b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979214
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Dayeong Lee <dayeonglee@google.com>