[fxfs] Hold lock during bytes_ready to maintain write ordering.
Holds the lock throughout the operation to include the read and the
update in a single locking. Also reworks the error logging and error
latching to to avoid repeating entries for the same error and moves the
message handling loop into the writer module.

Change-Id: Iba1f08713089f7c3bc61e37ae352785c178a9dd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978893
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>