[netstack3] for_any_device_id supports type param
Update the `for_any_device_id` macro so that it can be used with a
device type parameter. This is useful in cases where the compiler is
unable to infer the device type.

While here, make the macro public so that it can be used in bindings
code.

Bug: b/42051633

Change-Id: I1922b724fb28c3ba7d3ba52ff0bcf5e1e5e20016
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979857
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>