[aml-sdmmc] Use dedicated dispatcher for inlining FIDL calls.
This gives a small performance boost when the core sdmmc driver is
configured to use FIDL (not Banjo) to issue requests to the aml-sdmmc
driver.

Experiments show that a dedicated dispatcher gives an inlining rate
that is close to 100% (much higher than that of a default dispatcher
that is configured to disallow sync calls).

Bug: b/300145353
Test: fx test aml-sdmmc-test
Change-Id: I1c0a2795a97891ed75973748233e0e6bdd80e4e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981316
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>