[sdk][product_bundle]: canonicalize zip paths
The canonicalization code for PBv2 for zip files would create the wrong
path if a relative path on disk was a sibling to it.

Consider the pb.zip
```
pb.zip
-> product_bundle
  -> product-bundle.json
```

Which is in ~/work/ and is a sibling to a product_bundle directory like
so:

```
~/work
-> pb.zip
-> product_bundle
  -> product-bundle.json
  -> system_a
    -> fuchsia.zbi
```

If a system_a record for in the pb.zip's product-bundle.json indicated
the zbi could be found at product_bundle/system_a/fuchsia.zbi the
canonicalization path would expand that to
`~/work/product_bundle/system_a/fuchsia.zbi`.

Which may be an image entirely different form the one in the pb.zip.

This CL introduces the Canonicalizer trait which allows callers to the
ProductBundle::V2 to provide their own canonicalization methods.

This introduces a ZipCanonicalizer as well, which does no filesystem
checks for the canonicalization of paths.

Bug: b/322512730
Change-Id: I0df9b149edad546f04ae750105ba86bca1732a96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981547
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>