[bazel-sdk] fix extension stripping in fuchsia_cc_driver
The rstrip method in bazel does not strip a string but rather strips
a set of characters that are specified by that string. This was causing
issues when we were stripping extensions because if a name ended with
any letters in the extension they would be stripped as well.

Change-Id: Id3be750c2f161468aea91591dae252184fbe67b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982400
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>