[power-manager] Support CPU thermal client
Currently, there are two different methods for calculating the thermal
load for a given sensor, and both are used to obtain the thermal state
for different clients. The CPU thermal client will use the thermal load
calculated by the ThermalPolicy, and the conversion from thermal load to
thermal state is different from that used by existing thermal clients.
With this change, we will add a new mechanism to route the CPU thermal
load from the ThermalPolicy to the ThermalStateHandler. This will
decouple the CPU thermal load handling from the existing sensor-specific
thermal load handling. Additionally, this change introduces a new toggle
that allows the CPU thermal client to connect to the
`fuchsia.thermal.ClientStateConnector` service and retrieve thermal
state based on CPU thermal load.

Bug: b/295078019, b/320705983
Test: fx test //src/power/power-manager

Change-Id: Ibc1583eeb26e2274b59cd5d84501e6d6075ff902
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820628
Commit-Queue: Didi She <didis@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>