[sanitizers] Temporarily disable ASan container checking for outgoing
The bringup.riscv64-asan builder seems to be consistently failing with
a container-check in OutgoingDirectory::MakePath. After debugging for a
couple of days, I haven't been able to find anything that indicates this
is an actual bug, so this may likely be a false positive from something
else linked in that isn't built with ASan. This occurs after
https://github.com/llvm/llvm-project/pull/75845 landed which enables
container-checking for all std::string allocators.

For now, let's temporarily disable container checking for this
source_set to unblock migrating the bringup.riscv64-asan builder.

Bug: b/322528969
Change-Id: I374df6ba0d86e15a59a423f998768927273d13e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982654
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>