[bt][core] move LegacyLowEnergyScanner impl to base class
This change moves much of the LegacyLowEnergyScanner implementation up
to its base class, LowEnergyScanner. This change also updates tests to
be generic type parameterized tests with LegacyLowEnergyScanner being
the only current tested implementation. This is in preparation of adding
a new ExtendedLowEnergyScanner subclass to LowEnergyScanner. There are
no logic changes in this diff. The only changes are refactoring related.

Bug: b/305975969
Change-Id: I1d7962dde7818518e0fb05caf842094375850d29
GitOrigin-RevId: 676f1fdb0977d849b33091a37ca21f6e5cc8ea31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981276
Reviewed-by: Joshua Conner <joshconner@google.com>
Fuchsia-Auto-Submit: Faraaz Sareshwala <fsareshwala@google.com>