[bt][core] Advertise FCS Option
Not advertising this option causes bad behavior by some peer stacks,
they do not send FCS bits even though they should when this is the
default (16-bit FCS).

Add the FCS Option to the options parsed and understood, but do not
chaneg behavior.  Not sending the FCS option means the peers are
required to send us FCS as both sides must agree to turn it off and it
is non-negotiatiable.

Bug: b/318002648
Test: updated BrEdrDynamicChannelTest
Change-Id: Ic492c0bd2e2499dfb06cbd105a8bb0e050af1314
GitOrigin-RevId: f09da360b238871f9ae5b07ebb2fa92159968dda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981249
Fuchsia-Auto-Submit: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>