[kernel][oom] Trigger OOM response on "terminal" heap alloc failure
A while back, we implemented
https://fuchsia-review.googlesource.com/c/fuchsia/+/645109 to ensure
we trigger a system OOM response (when so configured) when an
allocation failure occurs.  We wired it up to pmm_alloc_page(s).
However, we did not wire it up to pmm_alloc_contiguous, since a
failure there is not necessarily OOM-worthy.  This change updates
cmpct_malloc to trigger a system OOM if it fails to allocate because
it was unable to obtain a large enough block via pmm_alloc_contiguous.

Change-Id: I77a60138da7634c41049308295f0f23431ac4515
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983085
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>