[fho] Introduce Connector adapter
A Connector implements TryFromEnv and simply holds on to the FhoEnvironment. It
can then be asked to produce an instance of some other type which implements
TryFromEnv. Basically this is a way for tools to keep the environment around and
instantiate new connections to services etc.

We'll need some other architectural changes in the injector to make this useful
but this should be the final API.

Bug: b/310035296
Change-Id: Ia79b99a5e9c7bcba6d9b53aa601fac1439bb72c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981472
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>