[brcmfmac] Correct shutdown on SDIO init failure
In the unlikely case that SDIO initialization failed there was a
possibility that the bus object created during initialization would be
kept alive and leaked. For the most part this was not an issue because
the driver would be non functional at this point and unloading it would
free up the resources.

With the introduction of dispatchers instead of threads this meant that
a dispatcher in a work queue was leaked and never shut down properly.
This means that if SDIO initialization failed the next reboot would be
significantly delayed by waiting for this dispatcher to shut down.

Correct this by calling brcmf_sdio_exit if SdioDevice::DeviceInit fails
after creating the bus object.

Bug: b/307600923
Test: Manually induced SDIO initialization failure and observe
successful reboots.

Change-Id: Id7becd4a9e11df240b8335fe6d7c4acd6c578a97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983417
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>