[bedrock] Function to open a bedrock capability and report errors
We now have a first instance of an error that is entirely handled in
bedrock code paths, without relying on legacy fallback. That is the Unit
capability/Void source.

This CL builds a little scaffolding so that more bedrock error handling
can be added easily in the future.

It turns out that we're not reporting errors to logs in the bedrock
paths, whereas we do in legacy routing. This CL adds error reporting as
best as possible.

Change-Id: I77351474b434d84f3963b5eae497d481f0425daf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983625
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>