[build] Specify single-letter RISC-V extensions separately
This change separates the specification of single-letter
mandatory/optional extensions within a profile from others. We do this
to directly construct the first 'token' of standard extensions in the
-march line, which eases future inclusions of optional, standard
extensions to our ABI (as we intend to do with "V").

Bug: 42075221
Change-Id: I99c8c3d40020a6054ad39d47408bbfb3477d0dc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981481
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>