[iob] Admit BlobIdAllocator::Allocate() variation to control copying
The initial implementation of `BlobIdAllocator::Allocate()` did a memcpy
of the supplied blob. While this is exactly what we want in userspace,
we need a kernel-friendly variation of this routine that abstracts out
the copy, as the kernel will need to take more care when reading from
the user-memory blob.

Bug: 319501447
Change-Id: Ie2aa83f1962cf78b12859d1b020960134c13d2e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982660
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>