[fdf] Update dispatcher in FakeGpio
The dispatcher in FakeGpio is set to the one
in libasync.  This change updates it to use
the dispatcher provided by the driver test
framework, if it exists.  Fallback behavior
is to use the libasync dispatcher still.

Testing: This changes test code.
Change-Id: Iec9ef4cdaa1d5bf2a935257457a2bd6ae8934b18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983392
Fuchsia-Auto-Submit: Garratt Gallagher <garratt@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>