[zxdb][debug-adapter] Fix breakpoint and frame issues
Some debug-adapter cliends needs breakpoints to be verified.
Updated code to set breakpoints as verified when they are
resolved in an attached process.

There was an update to zxdb to send update stack frame notification when
symbolization changes occurred. This broke debug-adapter support as it
result in invalid events related to thread invalidation. Fixed the
invalidation event to be sent only when the frame indexes are not valid.

Test: Manually tested with a debug adapter client.
Change-Id: I29b4446a20a11a4bd40073613b2348c7dbf38fc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983419
Reviewed-by: John Rosasco <rosasco@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>