[assembly] Remove package size check dedup by default
Unless the budget explicitly requires blob dedup, it will be turned off
by default. This is part of the process of making package size checks
product-agnostic.

Previously, when blob was seen in N different budgets, the size of the
blobs would be divided by N to compute a "proportional size". This
proportional size was added to each budget. This caused a lot of budget
churn when packages added/removed/updated blobs, and the share count
changed as a result.

Now, blobs are only counted inside budgets, and not across all budgets,
resulting in N=1 for the most part. This forces the budgets to have an
increased maximum (often 2x), which means they are no longer "relatable"
numbers. The budget maximums are now guardrails to help developers set
size goals/restrictions for themselves, and the product size checker is
a better signal of whether the combination of all the packages of a
product fit within a specific size.

Bug: b/321066699
Change-Id: Id8d358dfe513bf219a75f0a5bf05ff1e66b98a1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/980423
Size-Review: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>