[assembly][bt] Add config flag for BT Core packages
Add configuration options for including/excluding Bluetooth core
packages. This is defined as a toplevel enum instead of struct because
there are some packages that are valid with/without BT core and serde
tags provides a concise way to specify this.

This unblocks the effort of moving the #bluetooth-core realm from a
statically defined child of core to a product-variable child.

Bug: 94055
Test: fx build

Change-Id: I91b06c1029bcad3c708351dee815a9cc05a20090
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950658
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>