[fxfs] refactor open count flushing logic
Simplify the logic around flushing the file when the open count drops to
zero. We now subtract the open count immediately, and then purge if
required, and then, if we need to flush, only flush once.

The previous iteration of this function had a bug where dirty pages past
the end of the file would cause an infinite loop of tasks to be spawned,
holding active guards, which both actively consumed a core with the
spawning activities, and prevented filesystem shutdown.

It's not clear what race condition in the starnix tests caused this
particular issue to be hit, but it is pretty easy to replicate by
writing directly into the backing vmo of a file beyond it's truncated
size. A test with this behavior is added to guard against regressions.

Fixed: b/318434279
Fixed: b/319348464
Multiply: starnix_host_benchmark_e2e_boot
Test: fx test fxfs-platform-test
Change-Id: I5a5110e140af78583a42dbaf2f4e534970bd6609
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984112
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>