[components][dictionaries] Build program_output_dict in sandbox_construction
Move the construction of `program_output_dict` from model/component.rs
to sandbox_construction.rs.

Since we now add the Router directly to program_output_dict, this also
affords for a slight simplification of the trait to retrieve a Router
from a Dict.

Reasons for this change:

- Better organization; unifies more of the sandbox building code
- Prepares for followup change on dictionary extension, which involves a
  more complex procedure for creating the Dict

Somewhat unrelated, generalize the child router to accept an `impl
IterablePath` (new trait) instead of `SeparatedPath`, so we can use that
in the next change.

Bug: 323025721
Change-Id: I4fda5c1564c6a9f1c6c5d1a66f2d7178f63b21d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983936
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Gary Bressler <geb@google.com>