[debugger] Gracefully exit on fd_streamer close
This lets us nicely clean up the message loop and restore the terminal
when running in embedded mode.

If we ever need to survive longer than the lifetime of a fd_streamer, or
"no more debugging, but keep streaming this fd until it closes", we'll
have to take a closer look.

Bug: b/322418013
Change-Id: I38ca49f3c90ec52759ed621b1f8229fee1066a92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984174
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>