[fsverity] Write merkle attribute in multiple transactions if necessary
It's possible that the merkle tree might be too large to write in a
single attribute. Thus, if the size of the merkle tree surpasses a
certain threshold (defined in paged_object_handle.rs as
FLUSH_BATCH_SIZE), we break the write into multiple transactions. In
order to ensure that we do not waste storage if the power cuts before we
finish writing the entire attribute, this change also adds functionality
to the graveyard that enables tombstoning attributes.

Bug: b/323213914
Change-Id: Id4717690c70b3a23f973f622020b69c6d72a82a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976492
Commit-Queue: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>