[starnix] Send map information to the bpf verifier
This CL introduce a context for the bpf verifier and fills the context
with the position of all of the maps in the bpf code.
It also change the data sent to the code from the index in the local
object list to the address of the Map object in memory.

Change-Id: I141ab8aa73c682ae7326828b1ca6a51a07f27021
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984432
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>