[fxfs] Correct permissions on outgoing dir
We were overly strict on the permissions we give to the initial outgoing
connection to the Fxfs component.  This probably wasn't noticed because
of the custom protocols that we use between fshost and Fxfs.

This fixes tooling such as `ffx component copy`.

Change-Id: I6bfd83824382366eea976aebec3658eeb9640bc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984208
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>