[build][bazel] Fix @fuchsia_sdk and @internal_sdk incremental correctness
The content hash computed from SDK export directories was not computed
correctly. In the SDK export directory, meta/manifest.json and the meta.json
files it points to are actually symlinks to files that are located elsewhere
in the Ninja build directory.

Before this CL, update_workspace.py would create a content hash that only
depends on the symlink target values, not the content of their destination.
Hence the resulting hash would not change if the content of said files
changed between incremental builds.

This fixes the issue by using os.path.realpath() to point to the destination
files before returning the list of file paths to be fed to the content
hash function.

Note that this is only correct because there can not be any symlinks in
the final IDK archive (our build system doesn't know how to declare or
package these anyway).

Fixed: 323590606
Change-Id: I6e10cf872277d87f3a15084b023221331104f0be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985332
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>