[debugger] Look up addresses in DWO files.
Implement DwarfUnitImpl::FunctionDieOffsetForRelativeAddress for symbols
stored in .dwo files.

Previously this code used LLVM address -> DIE lookup
"getSubroutineForAddress" but this doesn't work for our Debug Fission
(.dwo file) implementation because the address generation requires
looking across DWARF binaries which LLVM doesn't know about.

This re-implementes FunctionDieOffsetForRelativeAddress using our own
code. It lazily constructs an address->DIE mapping for the unit the
first time this is needed. This new mapping is used for all lookups, not
just for DWO files.

Implements ModuleSymbolsImpl lookup in the correct unit for the address.

This implementation still does not provide correct symbols for DWO
files. The current issue is how FunctionDieOffsetForRelativeAddress is
used. Most uses end up constructing a symbol using the result of this
call using the main SymbolFactory, while to make this work properly, it
needs to use the DWO-specific symbol factory. The solution is to make
FunctionDieOffsetForRelativeAddress return a LazySymbol instead which is
unambiguous and cleaner from a caller perspective, which in turn
requires some nontrivial updates to the callers, especially line lookup.

Change-Id: I5f18691aa2a7e6f1954e0015f7f6f26cd4298b87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965063
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>