[sdk][vfs] Integrate new C++ VFS backend as shared library
Transition the internal SDK VFS library from a static to a shared
library called `libvfs_internal.so`. Publish as partner SDK in
preparation for use by the existing source set.

This new library uses the in-tree VFS for handling connections, does
not leak FIDL protocol details in the interface, and has better thread
safety overall. This serves a similar purpose to `libsvc.so` which is
used by `//sdk/lib/component`.

Add library preemptively to driver allow-lists, as some drivers still
use `//sdk/lib/sys`. This will cause a transitive dependency on this
library. As part of the development of this library, additional runtime
dependencies were eliminated for both `libvfs_internal.so` and
`libsvc.so`.

Reorganize //sdk/lib/vfs as part of the integration.

Bug: b/293936429
Test: fx test vfs_cpp_new_tests
Change-Id: I77509741ce393874e199cc4148f36efaa5ebdcc3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981288
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>