[gvisor_syscall_tests] Simplify loopback
Un-manually-shards the loopback syscall suites in favor of using
the automatic sharding. This allows us to simplify their expectations,
which will make it much easier to share them with Starnix.

As part of this, expand out a bunch of wildcarded fail/skip
expectations to make it more explicit which tests we expect to
pass/fail.

Bug: 318529127
Change-Id: Ib4ee9ccc269e37b07881c3cc277681b3b774b6e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983153
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>