[fidl] Allow breaking changes to in-dev
Since our partners no longer depend on the in-development API level,
it's not important for us to prevent breaking changes there.

This change was mentioned specifically in RFC-0239 as applying to
`NEXT`, but we may as well change how the checks work on in-dev so folks
don't have to make changes 2 API levels out.

Bug: 323889271
Change-Id: I89ae9036186a0e03be863a275f78f36d97f928e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986153
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>