[debugger] Use correct factory for functions.
Several things exposed a "function DIE offset" but this did not have an
associated binary or SymbolFactory. This meant that it was impossible to
know where it came from when it comes from a DWO instead of the main
binary.

Replace the uses of these in DwarfUnit and FoundLine with a LazySymbol.
LazySymbol contains the needed information and is also conveniently
higher-level and less error-prone to use.

To keep FindLine efficient, this allows comparing and sorting LazySymbol
objects. Some extra constructor code was added to populate the factory
and DIE offset for LazySymbols when constructed from a predefined symbol
object (used in tests).

This allows functions in DWOs to be resolved from addresses, but not the
line information for those functions. The line table is stored in the
main binary and current ModuleSymbolsImpl uses the unit's line table (in
the DWO) for this.

Change-Id: I519c254759a440f117e65cf7e88d805d2d6c455f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985307
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>