[kernel][phys] Introduce RISC-V feature set abstraction
This change introduces arch::RiscvFeatures, a container abstraction
around the set of RISC-V features of interest to the kernel. This is
populated in phys as the lowest common denominator of features across
all hart ISA strings and is included in the hand-off.

As a first use, the kernel's RISC-V feature detection constants -
previously hardcoded to true in wait of a programmatic means to query
this info - are updated to be defined in terms of the new handed-off
feature abstraction.

This will be soon be used to conditionalize support for select RISC-V
features in the kernel.

Bug: 124336
Change-Id: I289ee01beff1a3e17b2139d7d9ee568bec451c5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955485
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>