[caching-package-server] Support Inspect
Adds a method to register a lazy inspect node that will expose open
packages and their connection counts.

Also adds an option to provide the serve method with a RootDir to
speed up instances in which the client already has one and the package
is not open.

Bug: b/307977824
Change-Id: I1368c4f5dbbd9b9c24f14d37f138d18ed9ff1a94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979317
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sen Jiang <senj@google.com>