[starnix][fuse] Truncate writes to transfer size
The `fuse_out_header` has a `len` field that describes the length of the
payload data. If that length exceeds the total amount of data written in
this call to `write`, we should truncate the size of the payload to the
available memory rather than return an error.

Bug: b/323001474
Change-Id: Ibc0bbb62f39a8c245a4b82e007a81f122b7263f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986474
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Adam Barth <abarth@google.com>