[starnix][splice] Do not eagerly lock file offsets
Our implementation of `splice` was eagerly locking the offsets for its
operands. This locking is usually not necessary. For example, one of the
operands is always a pipe, which does not need to lock its file offset.
Similarly, if when the operand is a fuse device, the lock isn't needed
because those devices do not use offsets.

This CL resolves a deadlock which can occur when multiple threads are
reading and writing a fuse device using splice. Now, we take the offset
locks using the normal offset locking machinery, which is aware of when
various operations block.

Additionally, this CL removes the `read_raw` and `write_raw` operations
on files because these operations no longer have any clients.

Bug: b/323001474
Change-Id: If2ef6576d7655cda824dcf2a24970accb36d424a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986421
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>