[ffx][flash]: wait in a loop for Target to Reboot
When first executing the flash command, before the flashing code is
executed, we need to ensure the Target is rebooted to the bootloader.
Instead of waiting for a flat 10 seconds (brittle), wait in a loop.

If the target is not discovered in Bootloader mode in 20 seconds, warn the
user.

Change-Id: Iac07e5b69080f860f03e51f18535c90829a5e746
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986098
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>