[role_manager] Avoid unnecessary get_info call
We use zx_object_get_info calls to get information about the thread/vmar
we're setting a role on, but we only use this info to emit log lines to
the syslog at the DEBUG log level. This means that the results of the
syscall are functionally unused if the log level is higher than DEBUG.
This CL fixes the issue by making the syscall invocation conditional on
the log level.

Bug: 323262398
Change-Id: I1e40c62e9281b8e2cd2f2dffa83c74c688473ce2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987900
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>