[publish] Fix `fx publish`.
`fx publish cache` is broken because it tries to run
`fx build build/images/updates:prepare_publish`, but the GN build graph
is unaware of that target by default.
So, add it to `//:root_targets` for now.

This is intended as a quick fix, with a follow-up cleanup cl if
feasible.

Bug: 324385881
Change-Id: Ied1cdd97cb5e9f0cdbf085408e8155c9cd8b9f87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988152
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Darren Chan <chandarren@google.com>