[netdevice] Support binding to FIDL parents
The migration to FIDL added most of the infrastructure to bind to FIDL
parents but without any actual FIDL parents to test against, a couple of
things were left out.

Add the missing bind rule for FIDL parents and indicate that the network
device driver uses the network driver service. Without the use statement
the binding would begin but could not complete because the right
capability could not be found for the service.

Test: Verified network-device binds to both Banjo and FIDL parents.

Change-Id: Ie5aa15173e9e1ab0052ba2179cf92dd99529e0a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987220
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>