[power][broker] Add ElementInfoProvider protocol
Elements in the Power Broker topology can create a Status channel that
provides read-only access to the Element's PowerLevel to external
components. However, the Element is responsible for passing the client
endpoint to those components, and does not currently have a way to do
so. This CL implements a new FIDL protocol that enables the transfer of
endpoints and associated Element information.

Changes:
- Adds ElementInfoProvider protocol
  - Adds GetElementPowerLevelNames() to provide external components
    plaintext names for the PowerLevels of each element
  - Adds GetStatusEndpoints() to provide external components the Status
    client endpoints for each managed Element

Bug: b/311470566
Change-Id: I1d8dfa7796faf765d89a8aebced556cc5d197eab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984699
API-Review: Michael Brunson <mbrunson@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Alex Wong <alexdwong@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>