[scheduler] Copy role library to SDK directory
Copies the scheduler role library to the SDK directory. This library will
be used by clients to set roles on threads and vmars, and abstracts away
some of the boilerplate code associated with making FIDL requests to the
fuchsia_scheduler APIs (this is currently the `ProfileProvider` service,
but will be updated to use the `RoleManager` service once all of the
callers have been migrated to use this library).

This change also migrates all in-tree users of the scheduler library to
use the new SDK based library.

This CL does not delete the role library from `//src/lib/scheduler` as
there are existing OOT users that must be migrated first.

Note that this change keeps the scheduler library internal only. Future
CLs may change that after an API review.

Bug: 323262398
Change-Id: Icbda5920a126fedefa57cbcc4a5c0876a69d2ac1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988552
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>