[virtio-guest] Extract VirtioPciDevice class in v1 driver.
The VirtioPciDevice class encapsulates virtio device implementation
details that are not specific to the virtio-gpu device type.

This CL is focused on moving the code out of the Gpu class. Some
details specific to virtio-gpu, such as the virtqueue setup, are moved
to the VirtioPciDevice class. Follow-up work will revise the API
boundary.

Bug: b/42182631
Test: Booted core.qemu-x64 using ffx emu start --engine qemu
Test: ffx target ssh display-tool squares
Change-Id: Ia22d61af9cc9f4d7bd178f5aa52fc2722ab23cd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988037
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>