[bedrock] Add a Capability enum
Use a Capability enum everywhere instead of AnyCapability.
This gives type safety when converting back from the types
within a Dictionary and from within the FIDL registry.

AnyCapability still exists because Router is defined
outside of the bedrock library. Once it is moved into the
bedrock library the type can be deleted completely.

The Opaque and Lazy capability types are removed because
they are unused and they require templating so it is harder
to place them in Capability.

Bug: b/324606473
Change-Id: Icfc8c2085e7b6877454039bba6f9b86bf912a8c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987774
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>