[kernel][vm] Fix ZeroPages for SnapshotModified
Fix the bug in Snapshot-Modified in which ZeroPagesLocked applies an
inappropriate optimisation. Previously, ZeroPagesLocked would check if a
clone's parent was hidden, and if it was, a gap was inserted in lieu of
a marker or zero page.

This doesn't work for Snapshot-Modified clones as a hidden cow pages can
now have a pager-backed parent. This can result in pages from the
pager-backed root being leaked on reads to a range that has been zeroed
with a gap.

Bug: 323447923
Run-All-Tests: true
Change-Id: Ic009fdc85d235a6cc478efedf338bcbc64819436
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984675
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Sage Barreda <sagebarreda@google.com>