[driver][sdk] Use SDK to build Example Board Driver
Fix the dependency graph of the devicetree libraries. Now the example board driver can compile, but the test for it is not wired up yet. This will happen in next CL.

Also included in the change is the support of cc_source_library depends
on a bind library in SDK.

Change-Id: I09db218e0deffa02fe77daae54e4c59ba48b0f75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988378
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Jiaming Li <lijiaming@google.com>