[netstack3] Allow configuring DAD transmits
Introduce a configuration to fuchsia.net.interfaces.admin allowing the
number of DAD transmits to be changed, this can help deflake some tests.

We don't implement that in Netstack2 because a) we can just disable
flaking tests in NS2 if they become too much of a nuisance, b) we're
experiencing most flakes on RISC-V builders, which run Netstack3, c)
gVisor is not set up to configure DAD transmits per-interface.

In Netstack3, we have an impedance mismatch between the FIDL API
organization and the core API because `dad_transmits` already exists
there and is inside the IPv6 device configuration lock. The natural
follow from the FIDL form would be to put DadConfiguration inside core's
NdpConfiguration, but that is held by the device layer and far away. We
can revisit this if we add more configuration and a clear and obvious
improvement becomes available. Part of desire to split that is already
captured in https://fxbug.dev/42173351.

Fixed: 323453776
Bug: 42173351, 42182861, 42077260
Change-Id: Icbad757aa76e7c45a44388acb0c0ebd8954d8841
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990262
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>