[kernel] Clarify Dispatcher <-> PortObserver ownership relationship
PortObservers have a complex relationship with the Dispatcher
representing the object under observation. When an observer is live,
callbacks on the observer are made with the Dispatcher's lock held. This
implies that something outside the observer must be retaining the
dispatcher so that its lock can be released once the observer callback
returns. The SignalObserver callbacks can never directly call the
destructor of Dispatcher.

The observer can also be torn down by the PortDispatcher's
on_zero_handles handler. In this case, the Dispatcher lock is not held
initially and the PortDispatcher's lock is. The Dispatcher lock cannot be
acquired while holding the PortDispatcher lock so on_zero_handles must
retain a reference to the Dispatcher on the stack while it drops the
PortDispatcher lock and requests that the Dispatcher remove the
observer.

This changes the PortObserver's pointer to the Dispatcher from a
fbl::RefPtr to a raw pointer to reflect that it is not an owning
relationship. This avoids refcount churn in observer callbacks.

Change-Id: I6a5521e84fb754cb5aaf31cb26aecd1015ec375e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989039
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>