[starnix][selinux] Initial file-create security context algorithm
This change implements a (somewhat incomplete) algorithm for computing
the security context of newly created files.

The algorithm does not know how to interpret one of the default range
enum values (`ClassDefaultRange::DEFAULT_UNKNOWN_USED_VALUE` appears in
known policies, but we have yet to determine its interpretation). It
also does not account for `*_transition` statements that should cause a
transition to a new context when the object is created.

Bug: b/322353836
Change-Id: I3434612b43352b98814b931bfc8b7d3bbff03b7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986275
Commit-Queue: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Marina Ciocea <marinaciocea@google.com>