[virtualization] Switch to setting profile by role
The virtio_sound driver was explicitly constructing and setting a
deadline profile using the deprecated `GetDeadlineProfile` method of
`ProfileProvider`. This change does a couple things to remedy this:
1. Creates an explicit role for this deadline profile.
2. Modifies the `virtio_sound` driver to set this profile using the
   new SDK scheduler library that calls `SetProfileByRole`
Unfortunately, the virtio_sound_driver will still need to know the
deadline profile it sets because it uses it to compute some other
things. This can be replace with output parameters once we switch over
to RoleManager, but I'd like to keep this change as mechanical as
possible, so that's left for a future CL with a TODO in this one.

Bug: 323262398
Change-Id: Ia24117098743eebaf08e4e30ef3989c53e9ef2fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990516
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Tim Kilbourn <tkilbourn@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>