[wlan][policy] Retain recovery events for one day.
This change causes defects that are observed and recovery
actions that are taken to be preserved for 24 hours.
Previously, these events were immediately dropped as their
historical records were not actually used in making any
decisions.  Recovery is still logging-only.  The effect of
this change is to ensure that wlancfg logs when recovery
*would* occur if it were enabled.  Some small amount of
memory will be consumed in storing this extra information,
but it is not expected to be overly burdensome as the
events that are tracked are extremely infrequent.

Test: fx test wlancfg-tests
Test: fx test hw-sim-tests
Test: IOU - fxr/991646
Change-Id: Iae7378289b10c67e25e2ce3b23f3b4af7bd8c8d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991645
Reviewed-by: Naomi Mccracken <nmccracken@google.com>