[kernel][riscv] access user counter register in user space if zicntr is present
Instead of always hard coding that current tick is only available via a
syscall, use the presence of the zicntr extension as a sign that user
space can directly access.

The machinery to do this is already present in the VDSO, it just
required a true or false from
platform_usermode_can_access_tick_registers().

Change-Id: I6fd65f09f4dd2bac922c370879da5d9472f99910
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989452
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>