[bt][snoop] Open bt-hci as Vendor
We moved from publishing devices as Hci FIDL and instead
publishing as a Vendor to enable using Vendor FIDL in bt-host
component.

We switched this recently, and now bt-hci-broadcom panics if
it gets an unrecognized message.

Open the device as a Vendor FIDL instead and get the Hci FIDL
from it.

Bug: 325329669
Bug: 325302214
Test: build nelson and use A2DP sink
Change-Id: I234958e026a0869e56a500e02a6a87b70a47391f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992712
Reviewed-by: Lulu Wang <luluwang@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>