[netstack3] Validate and install filters in Core
Add a witness type to netstack3_filter, ValidState, which witnesses to
the invariant that the contained State has been validated to uphold
properties that are not guaranteed by the type definition: namely, that
no rule has a matcher or action that is invalid for the context in which
it will be evaluated.

For now, the only possible case of this is a rule matching on the output
interface in the INGRESS or LOCAL_INGRESS hooks, or a rule matching on
the input interface in the EGRESS or LOCAL_EGRESS hooks. Once we add
actions like TPROXY and NAT, those will also have to be validated.

To allow Core to provide a specific error to Bindings on validation
failure, add a ValidationInfo type parameter to netstack3_filter::State
that specifies ancillary state to be stored for each Rule. This type is
zero-sized for validated state at rest in Core, but Bindings provides a
RuleId with each rule when installing the state so that Core can report
which rule was invalid when validation fails, and Bindings can in turn
report that to the caller.

Fixed: 318738286
Test: netstack-filter-fidl-test
      netstack3-filter-test
Change-Id: Ib96d88d7dbfd55e2e2db52316733e1f465242b02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986254
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>