[build] Additional fast-math configuration
Add a configuration to enable `-ffast-math` without the
accompanying flags `-fhonor-infinities`, `-fhonor-nans` and
`-fno-finite-math-only`.

In order to do this, a third "baseline" config is created
that sets the common flags, but leaves the finite-math-only
flag to be explicitly set -- or explicitly cleared -- by the
more specific config.

Change-Id: I10f4d291833b862dba174e532a2800ca994e449b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988786
Reviewed-by: Paul Kirth <paulkirth@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>