[loadbench] Switch to using the ProfileResource
This is part of the ProfileProvider deprecation effort. The
loadbench shell package currently uses the ProfileProvider service
to create a bunch of threads with varying deadlines and priorities.
We could switch this over to using the `RoleManager` and define a bunch
of roles for this use case, but that's a very unwieldy setup for a shell
package stress test.

Since this binary is only used in an SSH shell, we can just create
profiles directly with the Profile resource. Note that I did have to
modify the security policy to allow the `ProfileResource` to be routed
to sshd, but seeing as this is a very small subset of the privileges
granted to the sshd component already (e.g. the root job and root
resource), I don't think this increases our attack surface.

Bug: 323262398
Change-Id: I00e402c17b7c087a7a5b22f794e102556c5a6c56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991681
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>