[netstack3] Lookup Device to Join Multicast Group
When joining a multicast group via the IP_ADD_MEMBERSHIP (Or
IPV6_ADD_MEMBERSHIP) socket option, callers may provide the device with
which they'd like to join the group. In the case that no selection is
made by the caller, the netstack should select a suitable default device
(i.e. a device with a route to the corresponding multicast address).

Prior to this CL, this case was unhandled, and resulted in callers
observing ENODEV. This CL adds support for this case by consulting the
routing table and selecting an appropriate default device.

Fixed: b/323355372
Change-Id: I1037b536ac7dd46fde0016309926af9133170118
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984652
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>