[zxdb][line_input] Ignore Ctrl-C in embedded mode
This CL teaches zxdb not to handle Ctrl-C (and other similar keystrokes,
like Ctrl-D) in the embedded console mode.

The key is to put the terminal back into "normal" mode, which cases the
pty to intercept these keystrokes and deliver them to the controller of
the pty, which is typically the shell. When zxdb runs inside `fx test`,
this causes a Python exception to be delivered to `fx test`, which is
what it normally expects.

Even though zxdb is responsible for watching stdin for input, the
`line_input` library is responsible for controlling the terminal mode.
For this reason, we plumb the information about whether we want to
listen for these keystrokes through the `line_input` library.

Fixed: b/324926524
Change-Id: Ibb8e41f3d124e8a1bd7e9993301313ac7a34bd12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993065
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>