[debugger] Add AnalyticsReporter and Session events
This adds a way for the Session object to provide analytics events to
the rest of the client. Basic events are added for the Session object,
which report what the Session is connected to and how long the session
lasts.

Users of the zxdb client library will have analytics disabled by
default, but may enable them if they so choose.

Change-Id: I17c39002a4463d8c23c82fdd1f1b1c65a97b2759
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993059
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>