[bazel_sdk] Fix remote link operations with Asan.
Expose the Clang Asan runtime link libraries in the sandbox
to support remote linking operations. See associated bug for
details.

Note that it is currently not possible to use a select() statement
that targets the `:asan_variant` config_setting() as this creates
a circular dependency.

Bug: 325505298
Change-Id: I1906738d6a5bd8631ffa7a23fe403c372a6ff55f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993654
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>