[honeydew] Fix tracing implementation.
The loop attempting to read the channel does not register for updates
when ZX_SHOULD_WAIT is returned from the channel, leading to a flurry of
missed channel updates while the channel is held that are never read
(and lots of log spam). Uses the AsyncChannel implementation instead so
that this is read/downloaded properly.

This also adds additional verification testing to the functional test to
ensure that tracing is actually happening and valid data is being
recorded. So far this only verifies that the trace can be converted into
JSON and that at least one "kernel:syscall" category (a default if no
categories are listed) is included in the trace.

Multiply: tracing_test_fc
Bug: b/324123125
Change-Id: Iee6ff4f437525d49e68d0ed1d36f34329c410296
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991220
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>