[dsi-dw] Move the driver implementation to a standalone library.
This change moves the dsi-dw driver implementation to a standalone
class (designware_dsi::DsiHostController) in a standalone library.
The DsiHostController class uses the same interface and banjo types
as DsiDw.

In a future change, we'll add a DsiHostController class instance to
the amlogic-display driver so that the amlogic-display driver won't
need to do inter-driver banjo calls.

After that, we'll be able to delete the DsiDw driver safely, and then
we can replace the banjo types used within the amlogic-display driver.

Test: no-op. panel power off/on works on core.nelson.
Bug: 319379442
Change-Id: Ib3168e7a2143da9cb3ea7718764f90a64462122d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993338
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>