[component_manager] Remove Router path field
This CL simplifies the Router type by removing the path field.
This has the nice property that calling Route will always return the
same type of capability.

This prevents the error-case of using a path field to index into
a Sender/Receiver.

Instead of feeding the path forward, the code has been updated
to call Route, check the capability that was returned, and then
index into that capability.

Fixed: b/325096862
Change-Id: I186cc57027e75e61c04a010395203ba4bef4cb07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993133
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>