[netdevice] Make tests correctly wait for startup
With the move to FIDL there was a change in how the
NetworkDeviceImpl::Start sequence occurred. Previously the completion of
Start (through callback(cookie, status)) processed the remaining part of
the Start sequence inline. Now the completion of Start (through
completer.Reply(status)) is asynchronous and the effects of the
remaining part of the Start sequence are not necessarily observed just
because reply is called. This could cause problems in tests that relied
on the old behavior. If the remaining Start sequence was delayed enough
it could interfere with other parts of the test by send RX queue trigger
packets at inopportune times.

Fix this by adding a callback in FakeNetworkDeviceImpl::Start that tests
can use to install an RX queue trigger handler before the rest of the
Start sequence is about to run. This ensures that the RX queue trigger
from the Start sequence is consumed before tests proceed and that all
other setup is done as well.

Bug: b/323601448
Multiply: network-device-lib-test
Change-Id: Ibd0d12b7808018edcaa4b66a48131620e3981de0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986492
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>