[netfilter] Refactor parser structure
Move the grammar definition for netfilter into a separate file so that
it can be used by the current parser (that supports
fuchsia.net.filter.deprecated) and the new parser that will support
the new filtering library.

The uses of these external parsing functions has been updated in netcfg
and net-cli. The new parser for the filtering library will take over the
main `parsing` module in a future CL.

Bug: 42083796
Bug: 323950171

Change-Id: I01beca3027f354d1359626b639b471756db15dcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989052
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>