[bazel][sdk] Fail if an API level is not set
If a user does not set an API level when building a fuchsia package with
the bazel sdk we will fail.

We do not fail if the attribute is not set because people can set an api
level on the command line. In the future we might want to revisit this
decision if the errors are not clear.

BUG: b/317417932
BUG: b/303683945
Change-Id: I35bb77a272c8fd516593a48ec70d55e912b81136
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993240
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>