[trace_manager] Use async::Executor
As part of an effort to allow better reasoning about async control flow
in trace_manger and reducing blocking on the async loop, it will be
helpful to add the ability to rely more heavily on the Fuchsia async
promise libraries.

In this patch, we lift the `async_dispatcher_t*` held by the various
objects in trace_manager into an `async::Executor&`. When
async::Executors are destroyed they shutdown the async loop, so we need
to keep a single global one in `main` and pass around references to it.

Bug: 295183613
Change-Id: I7be74601c4fec61d701451624048c24b95509fc6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995083
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>