[bazel_sdk] Don't set soname in fuchsia_cc_driver
If a soname is set, the loader looks at the soname to determine if the
shared library is already loaded, and if so returns the old copy. This
prevents us from loading a new version of a driver if an old version
already exists.

Bug: b/326445083
Change-Id: I73816fd0958e23c9826955bde38d08a43915b37d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995435
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>