[bazel_sdk][tests] bazel_test.py: Incrementality fix.
The script tries to create a symlink to a specific target file,
but the logic used failed when an existing dangling symlink already
existed at the current destination location, even though the
original code tried to deal (incorrectly) with this situation,
due to the subtle differences between os.path.exists(),
os.path.lexists() and pathlib.Path.exists().

To fix this, implement a `force_symlink()` that always removes
the destination link in case of EEXIST. This is similar to what
is done in other Python scripts of our build system (but
bazel_test.py cannot rely on them, since it is supposed to be
usable out-of-tree).

Bug: 326567616
Change-Id: Ifdad36e44478772dcc33654570d8b09ceb271990
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996052
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>