[wlan][policy] Fix e2e test dependency on async ordering
After receiving a client state update, the tests did not _ensure_ that
the ack was processed. This CL fixes that, by running the main future
after the ack is sent. Prior to http://fxrev.dev/990573, the timing was
different, and the ack was processed as a side effect of some other
`run_while()` in the test.

In making this change, another minor ordering assumption was revealed in
`prepare_client_interface()`, which previously did not ensure that the
state machine fully processed its initial setup before the end of the
`prepare_client_interface()` function.

Test: fx test wlancfg-tests
Bug: 324986229
Change-Id: I1bb47d863d3e6da8b899802021ae8bd481c979f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995393
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>