[debugger] Print the real filter pattern in attach
When using process names, we trim the string to zircon process name max
length, which modifies the string. We print a warning that we are doing
this, but then print the unmodified string we were given, which is
confusing.

Make sure to update the pattern we print (which is also used for things
like job koids) to the trimmed string.

Bug: b/193336765
Change-Id: I6d12a8c8ae258e67e8364d16cbc127eea28a9b69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995744
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>