[fxfs] Only deserialize keys while seeking
Defers deserializing the full item (i.e. the value) until we're done
seeking through the SimplePersistentLayer.

This is an adapted and simplified version of fxrev.dev/990832, which
extended the lazy deserialization further (via letting clients partially
deserialize keys and later access the value), which turned out to be a
great deal of work for marginal benefit since frequently values need to
be accessed anyways (e.g. for filtering and merging).

Bug: 323571978
Change-Id: Ic60b6a0d87dc8215d95fdb15e070e71a0fdc342b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995796
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>