[ddk-environment-test] Delete test
This test made more sense in DFv1 when we manually launching the DFv1
driver host and providing it with it's incoming namespace. Now that we
are not doing that, this test would simply serve to validate CF is
actually providing the protocols specified in the driver host's
component manifest, which isn't very useful.

Bug: 316176095
Change-Id: I40bebe13d6e688940ad58287bb88afa4619360e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995733
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>