[netstack3] Put datagram sockets behind references
Change the datagram module to expect datagram sockets to be behind
references and propagate the changes to Udp, Icmp sockets.

A follow-up change will move the bindings HashMap on socket IDs to
opaque data in core.

This follows the exact same model as TCP, including the lock ordering
layout. Core keeps a set of all the sockets it knows of, which is used
for inspect (not yet implemented for datagram) and nursing of Primary
references.

Bug: 42076297

Change-Id: Idd7aa85e4a449ff3642ef5d4edfe54d668a18e81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994097
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>