[seriallistener] Read from botanist's serial log.
In the hardware case, the serial output is forwarded from the serial mux
which only forwards logs after the client connects so there's a
possibility that seriallistener may start reading logs after the success
string has already been written. In the arm64 emulator case, we see
flakes due to what seems like the emulator being slow to start, which
may be due to enabling the emulator as a serial device which we can
write to over serial. The boot tests don't require writing to serial, so
we can just disable this feature and just read from the serial log
already captured by botanist instead. After this goes in, we can disable
serial for boot tests in fxrev.dev/996088.

Bug: 116559, 121919, 318507750
Change-Id: I25f0f3f506c3fb87d2a65b6dde0c62069ac871e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807729
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>