[assembly] Use new versioning library.
The only behavior change is to the error messages printed by assembly
when an ABI revision isn't supported. The new errors are slightly less
pretty, in that they don't do the nested indentation as well, but are
more informative, giving the API level for each ABI revision - as well
as being consistent with the analogous errors generated by
component_manager.

See bug for justification of the versioning library changes.

Bug: 326104955

Change-Id: I329e522d4fbdaf7a1b28d6578d8aab2998e80281
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994137
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>