[fidl] Remove outdated tests and docs
Before we had versioning and an SDK it was extremely important to know
that we could do "soft transitions" when changing APIs whose
declarations and uses were split across different repositories. We built
detailed documentation, coupled with tests that validated that the
documentation was correct.

Now we have API levels so if we want to change the platform interface we
can do that simply and cleanly without breaking the build of any
repositories. The advice that used to be critical to keeping builds
green now causes errors.

It would be nice to have a replacement for this documentation and these
tests ready to go, but it's more important to stop telling people the
wrong thing.

Bug: 42078310
Change-Id: I03539de875a450cc4720523a73c15c6e209d08c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995757
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Seth Ladd <sethladd@google.com>