[fuchsia.component.config] Add legacy=true
This adds legacy=true on the fuchsia.component.config library, which was
removed in API level 14, because it still exists in a supported API
level (12).

This library was removed in If825f4535a43d1c5fa59a60cd63aac5764100aa4
without legacy=true.

The motivation for this is to line up legacy=true with the current set
of supported API levels for the implementation of RFC-0232.

Bug: 42085274
Change-Id: I892c72b88e3113ae32006ab0c827fdc1b84457b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997856
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>
API-Review: Ian McKellar <ianloic@google.com>