[wlansoftmac] Add async Rx path tracing
In general, when tracing is off, each tracing call in the code is
"about the cost of a branch instruction" which is on the order of
nanoseconds. When tracing is on, it should cost a hashtable lookup
plus some small memcpy's which is on the order of <200 nanoseconds. In
other words, with tracing on or off, we would not expect tracing to
bottleneck data path performance until we're receiving at very high
data rates.

Test: Manually recorded a trace
Change-Id: I4f3c076fb86c198033477a1954e53a5008672af2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979817
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>