[fastboot]: upload takes Reader
Previously upload function would take a slice of u8 to upload. While
this worked for small files this would require large files to be read
entirely into memory in order to be uploaded to the Device. This lead to
inneficient memory usage.

Now, the upload function takes an impl Reader and a size to inform the
Device how much to be uploaded.

Bytes are uploaded to the target in 4096 byte chunks.

This also allows for the on_progress option for the
UploadProgressListener to be called as preivously there was no progress.
Files were uploaded in their entirety or they failed.

Bug: b/325501917
Bug: b/326990810
Fix: b/325501917
Fix: b/326990810
Change-Id: I43e5461348b9317b5dfa153ee5f4e33c7f93ee0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997672
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>