[driver_runtime] Migrate to Role Manager
Migrate the driver runtime and compatibility shim to using RoleManager.
Note that the fuchsia.scheduler.RoleManager capability was already
routed to all the relevant places in https://fxrev.dev/995272, this
change just modifies the code to use that capability.

A notable exception is the SPI code, which was committed after the
aforementioned CL, so this CL adds the necessary route.

This CL also removes an API level check that is no longer needed.

Bug: 323262398
Change-Id: I1166fffed2e8024267d5ca375addebd31590a2d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997934
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>