[starnix] Fix ptrace01 flake
Root cause was that a tracee notified its tracer of its exit before
it removed itself from the set of children.  If the tracer made two
calls to wait in a row, it could wait after the tracee sent the
notification, but before it told the tracer not to wait on it
anymore.

This change modifies the code so that a lock is held during the
notification / removal process.

Bug: b/325437151
Change-Id: I701f3c600cc16fadfbe8f55194cc74d8c1e81683
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/998772
Commit-Queue: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Adam Barth <abarth@google.com>