[devfs][tools] Removing multiplexing from iotime
As part of removing Devfs, this CL removes an
instance of fidl multiplexing, where a channel
is used for the the fuchsia_hardware_block::Block
an fuchsia_io::Node protocols.  This behavior is
hidden by the use of file descriptors.

Changed to now explicitly use typed ClientEnds.

Bug: 42063787
Change-Id: I6925e9df9964e8971a75af78dc5b51ed86d8884b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995412
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Garratt Gallagher <garratt@google.com>