[devfs][storage] Removing multiplexing in block-verity
As part of removing Devfs, this CL removes an
instance of fidl multiplexing, where a channel
is used for the the fuchsia_hardware_block::Block
an fuchsia_io::Node protocols.  This behavior is
hidden by the use of file descriptors.

Changed to now explicitly use typed ClientEnds.

The only use of the affected functionality of this driver is
in tests.

Bug: 42063787

Testing: Updated relevant tests.
Change-Id: I8bb72d3b9863a27211fc8462129efac68fb3513c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994963
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>