[starnix] Implement first version of the ebpf verifier
The verifier ensures that an ebpf program cannot:
- Do any out of bound access
- Leak kernel address to userspace
- Use any uninitialized data

This first version is extremely conservative. A more subtle
implementation will require a more refined type system to be able to
keep track of comparison between instances.

This first version can also be improved from a performance point of
view. By keeping track of which state is correct at a given pc, it is
possible to cute computation when the same pc which a smaller state is
reached again.

Change-Id: I83811cabbc428dcd3da08e62806f804c96188f60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999353
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>