[ffx] Daemonless knock implementation
This change introduces FIDL-pipe, which is a simplified re-write of
`ffx/daemon/target/src/overnet/host_pipe.rs`, which is the backbone of
the daemonless discovery function. It has three streams, one being
input, one being output, and the last being stderr from the SSH command.

The daemonless knock function is implemented essentially as:

* Resolve target address (IP address or nodename are acceptable).
* SSH to target.
* Establish overnet node.
* Knock RCS.

An example of this being used is in the follow-up CL in which this is
being used for ffx emulator start.

Most functionality from host_pipe.rs (regarding ssh message parsing) is
retained and untouched, moved over to ffx/lib/ssh/parse.rs

And other code touched is either kept identical or slightly altered in
order for it to be usable in a library outside of the daemon.

Bug: b/324322258

Change-Id: I18e442474a28d35d8b647c5978ab4e50c0b5f2fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982392
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Dwayne Slater <dwayneslater@google.com>