[kernel][cmpctmalloc] Clean up build rules and header use
This cleans up the build rules for //zircon/kernel/lib/heap
and its subdirectories, which uncovered some missing deps
in other kernel code that is using `<lib/heap.h>`.

It narrows down which headers are used by what to what's actually
needed, which simplifies the picture for the host test build of
the cmpctmalloc code.  It also makes `<lib/heap.h>` not declare
the kernel functions when used outside the kernel, i.e. in the
host test program.  That test code really shouldn't interact with
that layer at all, but currently needs the HEAP_DEFAULT_ALIGNMENT
constant because the cmpctmalloc doesn't provide it though it
exists to match the hard-coded 8 in the cmpctmalloc code.

This all fixed a build regression when upstream libc++ headers
changed so that <stdlib.h> is sometimes included in places where
it wasn't before, and its declarations of the standard `malloc`
et al can conflict with the kernel-oriented declarations in
<lib/heap.h>.

Fixed: 327668937
Change-Id: I367bceda056c28054c454b5cc11a47c1f7ad41e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000493
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>