[sdmmc] Merge the Banjo and FIDL unit tests.
sdmmc-block-device-fidl-test.cc is almost the same as
sdmmc-block-device-test.cc, except that the SDMMC protocol opreates
in FIDL in the prior and in Banjo in the latter. This change merges
the two tests into one in the form of a parameterized unit test.

The contents of sdmmc-block-device-fidl-test.cc is kept under the
name of sdmmc-block-device-test.cc. An additional few lines of changes
introduces the test parameterization (GetParam == true for FIDL, false
for Banjo).

The motivation for this change is to remove the burden of maintaining
what is essentially a same copy of the test.

Test: fx test sdmmc-block-test
Change-Id: Icdf3c1ba5d7e100d2a24f723f8b584e15a7c7415
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000253
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>