[rust][build] Temporarily allow unused_imports
Just for the sake of rolling the toolchain reducing the risk of races
for this high potential to race lint.

Expect this to be very short lived, as soon as the toolchain rolls this
lint will be re-enabled in a change that fixes any unused import that
was added in the meantime.

Bug: 326330182
Change-Id: Icb7c8ad7e9fb636445448e752758f86cbfe9e165
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000279
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>