[debugger] Set the arch during local init.
The "arch" on Session is documented to never be null but there is a
window it was null during initialization with a local connection.
The commands run with a script file would execute during that window,
and anything that depended on the architecture (like setting a
breakpoint) would crash.

Sets the architecture to "unknown" during that window. Since no binary
could have been loaded yet anyway, any architecture-dependent code can't
have an effect.

Change-Id: Ide27b877e1617b8f2070693e78aac3dd1ffe905d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001397
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>