[network/test] Configure DAD before iface up
I1b07bf07b776e4701bae69ba59d4843754f7739e updated the inspect ns3
integration test to configure the interface to use u16::MAX DAD
transmits in order to avoid the link-local address reaching Assigned
state flakily before the inspect data is collected. However, because
we update the DAD configuration after the interface is installed on
the stack, it's possible for the link-local address to complete DAD
before we get a chance to update the config.

Address this by adding a dad_transmits option to `InterfaceConfig` so
that it can be set before the interface is enabled.

Fixed: 326940678
Multiply: netstack3-inspect-integration
Change-Id: I5e28a484b04e67f8f64ba563920042428079a81b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999378
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>