[bt][core] refactor LowEnergyScanResult to its own class
The LowEnergyScanResult class was previously a struct with all public
members used to track the elements related to a scan result. A bit
awkwardly, the advertising data and scan response data related to the
scan result was stored in LowEnergyScanner::PendingScanResult.

This change updates LowEnergyScanResult to store the advertising data
and scan response data within itself. To present a better API to users,
and to prevent accidental modification, LowEnergyScanResult is now a
class with getters and setters instead of a struct with all public
members.

Since LowEnergyScanResult stores the advertising data and scan response
data itself, we can also modify the
LowEnergyScanner::Delegate::OnPeerFound API to simply take the
LowEnergyScanResult alone instead of also taking in the data as a
separate argument. We only did that because the data was stored
separately.

Change-Id: I476bfa2c117e3fd717d7851a27b517725adff68e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987909
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Lulu Wang <luluwang@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>