[audio][device_registry] DeviceInfo fields
Add three fields to DeviceInfo, and rename another.

Rename `supported_formats` to `ring_buffer_format_sets`, to
clarify what type of formats this references. Note that this
is ONLY in the Info table now (not also retrieved by method)
since this is confirmed to be constant over the lifetime of
the driver instance.

Add `is_input` as a static field, even though this can currently
be deduced from DeviceType. This will not be possible for Codec
or Dai devices (in subsequent CLs).

Also add signalprocessing topologies and elements. These are
static properties that we can offer upfront in Info, rather
than requiring the user to explicitly retrieve them.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

BUG: b/323273941

Change-Id: I4f25f94f42fa10492da74f08c75932a14b52f697
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999196
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>