[bt][core] update LowEnergyAdvertisers to use EmbossCommandPacket return type
This change also fixes the
LowEnergyMultipleAdvertisingTest.AdvertisingHandlesExhausted test. This
test is meant to advertise up to the maximum number of advertisements
and then try once more to make sure the next one is rejected due to
exhausted advertising handles. However, in the case of
ExtendedLowEnergyAdvertiser, we use the default maximum number of
advertisements for its advertising handle map:
hci_spec::kMaxAdvertisingHandle + 1. In the test, we were going up to
only two so the third attempt would succeed. It was supposed to
fail. This change fixes the test to use MaxAdvertisements() instead so
as to ask the advertiser what its maximum number of advertisements is
instead of assuming it.

Fixed: b/312896684
Change-Id: Ia2350243072fc0aef751fe3244a8be774001f6c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984735
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Lulu Wang <luluwang@google.com>