[fuchsia-async] Always attempt to read channels when polled
In b/303919602, there's a need to drain channels when shut down.  I
believe this is the only way that this will be possible.  Prior to this
change, it would be possible for a port packet indicating a channel is
readable to not have been processed, which would mean we'd miss a
message.  With this change, so long as all futures with channels are
polled, the fact that a port packet hasn't been processed yet won't
matter.

The same change is made to OnSignals: polling the OnSignals future will
always poll the kernel to see if the signals are set.

This might have a performance impact where futures get polled when they
haven't been woken (which can be the case with some future combinators
that don't track wake-up notifications separately), but that should be
rare.

Bug: b/303919602
Change-Id: I6c51386f2d9c525a418b936145ce3c69d7b7160a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990573
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>