[fx][ssh] Updating fx-test rewrite to use ffx for ssh keys
This replaces the usage of .fx-ssh-path with calling
`ffx config get ssh.priv/pub` to get the path of ssh key.

This converges on a single source of truth of ssh key management.

This is part of a larger effort to unify ssh key configuration:
see: go/go/fuchsia-ssh-unification

Bug: b/328246503

Change-Id: Ibb29ef2c511e5b6b15d4c7c4129a89089891d80e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001367
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>