[fxtest] Don't hide the cursor
Previously, hiding the cursor was important because it would flash when
the status area updated. Now that the status updating code works better,
we do not need to hide cursor. Instead, fxtest "parks" the cursor in the
bottom right corner of the terminal.

This change makes the integration with zxdb easier because when zxdb
catches an error, the cursor is now visible at the zxdb prompt.

Fixed: b/322420507
Change-Id: Icedcafb67c1b90333a43fe249361960183b16544
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001901
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>