[power][broker] Update documentation for ElementInfoProvider
ElementInfoProvider serves Status endpoints and Power Level information,
including levels and plaintext names, to clients that are mapped to a
specific Power Element via the Element's plaintext name. These names are
constants within the Elements and/or their managing components and
should remain constant between reboots of the same build, but may change
between boots of different builds due to changes in the codebase and
consistency is not guaranteed in that case. This CL updates the
documentation for the ElementInfoProvider tables to clarify this
information.

Changes:
- Update documentation for ElementStatusEndpoint, PowerLevelName, and
  ElementPowerLevelNames to clarify names/identifiers are unique between
  element and only guaranteed to be consistent between boots of the
  same build

Bug: b/311470566
Bug: b/281900115

Change-Id: Ib519bb14dfe8f1839ba11de37abd33ac3a05ddf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996169
API-Review: Michael Brunson <mbrunson@google.com>
Commit-Queue: Alex Wong <alexdwong@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>