[diagnostics][sampler] Increase test Reader timeout
Reader has a default timeout of 10 sec for a component to
supply Inspect data. For tests in emu, this is sometimes
exceeded and causes flakes.

This CL adds a for-test config to Sampler that tells the
Reader to have an infinite timeout.

As part of this, SamplerConfig is changed to a builder
pattern so that other code which uses SamplerConfig won't
have to change in the future every time we add a config option.

This CL also adds a check to the tests to make sure the
"single-counter" test component is publishing Inspect data
before Sampler tries to read it.

Fixed: b/320697654

Multiply: sampler-integration-tests

Change-Id: If8958bdbab5cfbb3d542f3297d5f561426cf680a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002323
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Chris Phoenix <cphoenix@google.com>