[fuchsia-async] Always attempt to read channels when polled
This is a reland of the original: this addresses a race in OnSignals
where notifications could be received before the waker was registered.

In b/303919602, there's a need to drain channels when shut down.  I
believe this is the only way that this will be possible.  Prior to this
change, it would be possible for a port packet indicating a channel is
readable to not have been processed, which would mean we'd miss a
message.  With this change, so long as all futures with channels are
polled, the fact that a port packet hasn't been processed yet won't
matter.

The same change is made to OnSignals: polling the OnSignals future will
always poll the kernel to see if the signals are set.

This might have a performance impact where futures get polled when they
haven't been woken (which can be the case with some future combinators
that don't track wake-up notifications separately), but that should be
rare.

Bug: b/303919602
Change-Id: I7f298efc615faf2982f8fa9e7b5ede4ec03ed037
Multiply: fshost_integration_tests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002472
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Chris Suter <csuter@google.com>