[display][banjo] New struct for SetBufferCollectionConstraints() args.
DisplayController.SetBufferCollectionConstraints() currently reuses the
Image structure to convey the expected usage of the buffers in a sysmem
BufferCollection. This is misleading, because one of the fields (the
imported image handle) is not meaningful, and other fields (width,
height) are not necessary.

This CL reduces the confusion by adding a dedicated struct for
SetBufferCollectionConstraints(), and by updating the method comment to
clarify the implementation's responsibility.

Bug: 316631158
Change-Id: Ifbe8a47a60af1723b883410800735c594486aba5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003294
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
API-Review: John Bauman <jbauman@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>