[testrunner] Change "unable to record output files" to debug log level.
The log appears when trying to collect test outputs, which would not be
a reason for a test failure, so we should change the log level to be
less confusing. Also, this log will appear if there are no output files
to collect too, which is not an error.

Change-Id: Iff735956d4f9986dbe170da1ca37035d92fa1539
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004237
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Danielle Kay <danikay@google.com>