[starnix] Fix Perfetto FD leak
Previously, we were using the sys_socket syscall to create a socket, but
that syscall also adds an FdNumber to the current task. We were not
closing that FdNumber. Also, the code that obtained the FileHandle from
the FdNumber was racy because some other thread could change that
FdNumber in the FdTable between the sys_socket call and this thread
grabbing the FileHandle

After this CL, we use the new_socket_file function to create the socket,
which returns the FileHandle. The FileHandle is reference-counted, which
means the socket will be cleaned up automatically once there are no
longer any references to it.

Change-Id: I9aba885092cf401652885ba9df064e824556ddcc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004682
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>