[component] Properly wait for component start in tests
Currently these tests just happen to work because the async executor ran
things in the "right" order, but will fail as soon as that is changed.

The lifecycle tests need to wait for the component to be started before
calling `abort_controller`. Otherwise the MockRunner can't lookup the
controller KOID.

The destruction related tests might recreate a component with the same
URL. That causes `wait_for_url` to assume that the component is already
started so now the start functions call `reset_wait_for_url` to
forget the previous component.

Change-Id: I8a680c352302514774066be3a961a57f4b72222c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002576
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>