[pkg-cache] Use open packages to short-circuit resolves
This behavior will be enabled when dynamic package protection is
disabled (both of which are controlled by the protect_dynamic_packages
structured config flag).

During Get, pkg-cache does not check that all blobs are resident for
packages it knows must be cached, which saves reading the meta.far and
listing blobfs contents (repeated for each subpackage). This CL
replaces the dynamic index with the open package cache for Get
short-circuiting. This will slightly slow down resolves of packages
that would have been in the dynamic index but are not open.

This CL also adds the use of the cache_packages index to
short-circuiting.

Bug: 327208811
Change-Id: I95725d83ccf3103d667c9514c5ad686db17808df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004550
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>