[ffx][writer] Adding VerifiedMachineWriter
This introduces VerifiedMachineWriter as a struct
that effectively extends MachineWriter to allow
the production of a JSON schema describing the machine
output.

A partial implementation of a bespoke schema is partially
replaced in this CL due to name collision with VerifiedMachineWriter.
The rest of that implementation will be removed in a follow-up CL.

Bug: b/326125002
Change-Id: I8d93bab94c76811ab9a06f6611e0c9d85ee1fe6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006493
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>