[fuchsia-fs] Add another way to read files
The fastest way to read the contents of a file is to:
 - open the file with DESCRIBE
 - get the stream from the OnOpen event
 - seek to the end of the stream to determine the file size
 - read that many bytes from the stream

The problem with the existing fuchsia_fs::file::read method is that the
file has already been opened and now Describe needs to be explicitly
called to get the stream. Most clients get the FileProxy by calling
fuchsia_fs::directory::open_file and then only use the proxy to read the
contents of the file. Combining open and read into one method gives the
best performance.

Bug: b/324239375
Change-Id: Ie2699005ac433e569bbc6caa0dc48f0732cc83d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999712
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>