[display][banjo] Merge clamprgb protocol to display controller protocol.
The fuchsia.hardware.display.clamprgb library had an optional
ClampRgbImpl protocol for display engine drivers to set the minimum
RGB value to be shown on the panel.

This was designed to address a hardware issue on a specific board
so the protocol implementation was optional.

Having multiple banjo APIs makes it difficult to migrate the driver
to DFv2. To make it easier for the DFv2 transition, this change merges
the ClampRgb protocol to the display Controller banjo protocol;
SetMinimumRgb() now must be implemented for all display engine drivers.

This doesn't change the FIDL fuchsia.hardware.display.engine API. The
design of the ClampRgb method in the FIDL API will be addressed in a
separate change.

Test: no-op.
Test: On Astro, "display-test --clamp-rgb 128"
Test: internal product on Astro works.
Bug: 328903017
Change-Id: I95bfee0e5cabb09c6a0f82bc93d002249c4c6a82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004583
API-Review: John Bauman <jbauman@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>