[tee] Add scaffolding for TEE implementation
This adds generated Rust bindings and an implementation of single TEE
entry point (TEE_Panic) as well as an example TA that panics on each
entry point. The implementation is currently factored as a Rust shared
library, but this may evolve over time. The TA links as a shared library
with a dependency on `libtee_internal_api.so` and undefined references
on the TEE_ entry points.

There are 3 layers in the implementation of TEE bindings:

- bindings.rs contains the bindgen generate Rust definitions of the TEE
  internal APIs
- binding_stubs.rs contains hand written mechanical forwarding stubs
  that implement the bindgen extern "C" functions by calling a
  (non-extern) Rust function in the tee_impl crate.
- tee_impl.rs has our implementations of the entry points as regular
  Rust functions. Over time these will likely call in to other crates to
  implement more complex functionality.

The stubs layer exists so that we can get Rust type checking on our
implementations as the generated extern "C" functions do no type
checking on parameters and return values.

Change-Id: Ifc62d7cd1cf52924328a2349880184e13861b5df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002113
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>