[bt][obex] Allow duplicate headers in multi-step requests
Update the HeaderSet to allow for the adding of duplicate headers to the
set. A duplicate header is a Header that exists in the collection with
the same header value. If the value differs, then adding the header to
the collection is still an Error.

Bug: b/42076096
Test: fx test bt-obex-tests
Change-Id: I337e610a7f47097df8196403532ee331fc66c338
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006716
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>