[sysmem2] Add R8G8B8X8 and B8G8R8X8 formats.
Adding R8G8B8X8 and B8G8R8X8 to make X vs. A explicit, instead of
relying on participants to communicate X vs. A separately.

In some cases this will end up mapping to "A" in vulkan since vulkan
doesn't have X vs. A at format level, but at least this way we can
complain to the log (or similar) if the format says X but a client is
trying to treat as A.

BUG: 327498161
Change-Id: I32d2e5a6836d36c8e805a8cc4a884395adde1297
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005645
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>
API-Review: John Bauman <jbauman@google.com>