[ffx] Fix discoveryless partial-matching
When doing local discovery, there was a path for matching against
a specific name. But when using daemon-based discovery, partial
matching is supported. This CL implements more general matching
in local discovery.

There was a kink in the implementation: when specifying a string,
we still want to return immediately if we find that target name,
but if it's just a substring, we still want to capture all the
matching targets. This required re-working the stream logic
significantly.

Change-Id: I7741f69f2143499b906239615c5474e462e4714e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003954
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>