[bazel][docs] add tool to extract bazel docs
We cannot use stardoc directly because it uses rules_java which our
build does not support. Instead, we use the new rule that is introduced
in bazel 7.0 which extracts docs into a protobuf. This is what stardoc
is doing under the hood.

This CL has a stubbed out tool which will parse the protobuf and render
markdown to be consumed by our docs site. The implementation of the tool
will come in a later CL.

To extract the testdata proto run the following command:
`fx bazel build
//tools/bazel-docgen/testdata:extract_foo_docs.textproto`

BUG: b/306761829
Change-Id: Iefd54be91f683832d1fd90d06f58e0a7ea5ee441
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004838
Reviewed-by: Jerry Belton <jcecil@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>