[bedrock] Unify RouteFn and Routable
Routable is a trait that lets you route capabilities, so we can
implement Routable for all RouteFns. This matches the idea that Router
is the object-capability representation of Routables.

Also reorders decls a bit so that the most important types float to the
top.

Change-Id: Ie04133ac9f973c9689b736423b449a1c4501bef7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007794
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>