[botanist] Close ssh client if not using it for capturing syslog.
We create a new client for adding the package repository and then
capturing syslogs. However, now that we're using ffx log to capture
syslogs, that client doesn't do anything for the rest of the task unless
the target reboots and we need to readd the package repo. However, it'll
continue to keep sending keepalives to keep the connection alive. We
should just close it from the start so we don't keep sending keepalives
and adding unnecessary logging for a connection that mostly doesn't do
anything.

Bug: 324285824
Change-Id: I2e56ec9d347339743514442e3fb932dfa9916ebb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007585
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>