[kernel][vm] Prefer size_locked as virtual method
Users of VMOs sometimes need to know the size while they hold the lock
to prevent races. This change means that the lock is needlessly grabbed
when querying the size of physical VMOs, but these are a very small
amount of VMOs so this overhead not very relevant.

Change-Id: I35026667b82b41fee9e4903f97402faa8eead551
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007912
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Reviewed-by: David Worsham <dworsham@google.com>