[bazel][boringssl] generate bazel rules during roll
Updates the roll_boringssl.go script to also generate the build files
for bazel. These files are used in the boringssl bazel repository. We
were generating these on demand but the python script spawns a process
which calls out to the go binary on the user's path which is not
hermetic in our build system.

BUG: b/327257847
Change-Id: I168ddeb3fa3495ff9ec50811eca3de122b45b027
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008842
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>