[netstack3] Remove OnceCell from netstack3_filter
Currently netstack3_filter stores a OnceCell on each uninstalled routine
to cache the converted version of the routine -- the version with its
"debug info" stripped. Instead of using a OnceCell, which has
unspecified behavior on reentrant initialization, eliminate locking and
use a regular HashMap to incrementally build an index of converted
routines. This has the benefit of allowing us to define the behavior
when encountering a routine cycle; currently we define this as a panic
because Bindings checks for this, but we could also choose to return an
error to the caller.

Fixed: 325492760
Test: netstack3-filter-test
      netstack-filter-fidl-test
Change-Id: I16c5b6981336edb75fd258b1c85a55dd9c51a01b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007957
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>