[bedrock] Receivers are just client types
Update the bedrock API so that Receiver is something that clients
implement so they can create Senders.

This means that Receivers cannot be placed into Dictionaries.

In order to add something to a Dictionary, a client must register
their Receiver and create a Sender.

In the process, Receiver does not need to be clonable, meaning
client's implement less.

Bug: b/328087055
Change-Id: Ic8063cd40d792f33901f74d0c6e81d090a09f3a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008473
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>