[network/test] Shard filter FIDL test
Use the automatic test sharded to break the filter FIDL test up into 2
components so that each one is less likely to time out; the test already
has a long duration and there is a lot more coverage that will be added
to it.

Also, use a shorter timeout when waiting for an event we expect not to
occur. We're currently using
netstack_testing_common::ASYNC_EVENT_NEGATIVE_CHECK_TIMEOUT, which is 5
seconds, and in the filtering tests we have *lots* of negative checks as
we add filtering and verify that packets get dropped (and therefore,
e.g., a socket connection fails). Reduce the timeout to 1 second. This
has the risk of increasing the chance of a "false pass", where the event
was delayed by up to a second and therefore the test passed when it
shouldn't have, but it will still fail often enough to catch issues.

Fixed: 330243476
Multiply: netstack-filter-fidl-test
Change-Id: Ibd28d80434367515909d42462f2fbf636377b377
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009375
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>