[phys] Ensure physboot-elf-kernel-test has a large kernel package
As an interim measure, physload reuses the space previously occupied by
the kernel package (i.e., the STORAGE_KERNEL ZBI item) for the hand-off
to the kernel. With profile instrumentation, this hand-off further
encodes profiles which can take up a fair amount of space. When the
kernel package is compressed, it is too small to fit a hand-off with
profiles, so we reserve more space in there for now by including the
kernel proper.

Fixed: 325513615
Change-Id: I74446b77941062140d52ae7955f848976c3bd927
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011132
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>