[netstack3] Support MTU updates on Pure IP devices
Because the MTU can now be updated, it is dynamic state that needs to be
held behind a lock, with a corresponding lock level. While here, encode
accesses to the MTU through a `PureIpDeviceStateContext` trait which
will make it easier in the future to 1) add more state to Pure IP
devices, and 2) write test mocks.

Bug: b/42051633
Change-Id: Ic0d7146d5fa4a269d3276cdfd50369fc46515c95
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007962
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>