[ffx][product] Removing Sync trait on structured_ui
This removes the Send + Sync trait constraints on the structured_ui
interface. Close inspection of the code did not find any locations
that actually share the ui across threads. Generally speaking, having
multi-thread access to the UI, especially a serial interface such as
stdio can lead to problems.

The reason this change is being made is to support using the
fho::MachineWriter interface for supporting JSON encoding output
including schema support. This is desirable so the ffx frameworks
related to compatibility can be used to monitor compatibility of
the machine readable output.

Follow-on CLs will implement a more maintainable schema for the
machine output.

Bug: b/328618096
Change-Id: Iba3f57f0c6ad460887f7531b5467599b7eabcca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010977
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>