[kernel] do not sort text/data/bss sections by name
Keep sorting by alignment but don't sort by name after alignment.

Sorting by name is not particularly productive since most of the symbols
in the system are C++ name mangled, so the net result is largely sorting
by the length of the first mangled field, and even then it's an alphabetic
interpretation of a numeric field in the mangling.

This keeps symbols roughly in the same order they are in the particular
object files so in general keeps related code near each other in the
in the binary. Also makes it a little easier to read the disassembly.

Change-Id: I0108d9d2212d3505803c7d318c53a62f2c6897ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009328
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>