[check-licenses] Support landing fuchsia-infra-bazel-rules.
The fuchsia-infra-bazel-rules repo does not come with its own
README.fuchsia file, so one needs to be created for it in fuchsia.git.

The readme file can't be added to fuchsia.git in the same CL as the
project itself, since they exist in two different repositories.

This CL puts a temporary skip for the new project. It will be removed
shortly, after the project lands and a README.fuchsia file is
created in the //third_party/fuchsia-infra-bazel-rules folder.

Bug: b/327010755
Change-Id: I01d0ac19b5fc1f5aad1ec5006266d491022722a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011933
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Darren Chan <chandarren@google.com>