[debugger] Add recursive component filters
Recursive filters will attach to all child components spawned under the
moniker of the given filter. The filter can be any variation of
component filters, for either component URLs or monikers. When the
filter is matched, the full moniker will be known by DebugAgent and a
special prefix filter installed (but not visible to users) will match
child components that are started.

This also works for attaching to components that do not have an ELF
program block, which can be the case for tests that are rooted with a
routing component.

Bug: b/329916323
Change-Id: Ifaf27760910c2e6c9cbaeb3f49f78604b7bcade1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011476
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>