[bt][virtual] Migrate LoopbackDevice Driver from Banjo to FIDL
RootCanal and PTS-Bot use the new bt-host component to connect to the
LoopbackDevice driver using FIDL. This change updates the
LoopbackDevice to use the FIDL server end of the HCI protocol instead
of the Banjo version.

This is also part of the larger effort to migrate drivers to DFv2 and
move bt-host out of Fuchsia.

Bug: b/329506134, b/303116559
Test:
- fx test bt-hci-virtual-tests
- Manual test with RootCanal, using ffx bluetooth subtool to start
  Pandora and send HCI packet between RootCanal and bt-host

Change-Id: I1a42ff55ccba4f1db70122ad5b2505fb2762863f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010218
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Lulu Wang <luluwang@google.com>