[starnix][selinux] Use policy-defined Ids in SecurityContext fields
Replace strings in SecurityContext field values with policy-defined
(numeric) identifiers.  This removes the need to map from strings
to Ids when looking up rules, and ensures that if a SID can be
resolved to a Context then the fields are guaranteed to have values
actually defined by the policy.

Consistency of fields e.g. range levels, whether the role is
valid for the user, will be landed separately.

Bug: 328375715, 319232900, 330677360
Change-Id: I8bebbbf0e9b63d0aee037f23a133273d9c29c121
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011912
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marina Ciocea <marinaciocea@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>