[doc] ffx component start doesn't update package
The component start process attempts to resolve the component
if not already resolved. However, if the component is already
resolved, it is not re-resolved again. Thus one cannot update
a component just by starting it. This incorrect documentation
is spotted in chat threads so we correct it.

The original table also doesn't make sense that you can update
the package without updating the manifest. The manifest is
contained in the package.

Change-Id: I648d8df4b42d5b0fd0338244f1d49acb46faef11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011238
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>