[fshost] Add 2 minute ramdisk client timeout to integration tests.
Some of these tests are flaking with timeouts that seem to be nearby
this call, which has historically been a source of slowness. Currently
the timeouts don't give any stack information, but it would be good to
be able to implicate what we're actually hanging on.

Bug: b/326397893
Change-Id: I7c77d6056e604c4da2a4af98c81c00155a9e69bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011220
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>