[ld] Add partial-page-bss test case
The existing modules don't explicitly exercise the partial-page
bss zeroing code paths.  It turned out they were being tested
only in coverage and profile builds where the test modules happen
to come out that way.  This adds a test tailor-mode to ensure it
will exercise those code paths in the loading implementations.
This is both the {Vmar,Mmap}Loader, which may already be tested
for these things, but also SegmentWithVmo::AlignSegments, which
was not tested in all paths in the elfldltl tests but is used in
the remote cases.  The new test exercises the missed paths there.

Bug: 326524302
Change-Id: I9fa7672ecababf5b39c3196172eab16d58876e14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012253
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>