[pkg][blobfs-client] Don't log in open when blob is missing
This log is emitted when the retained index is set, which is generally
expected and can be misleading. The log was more useful when support
for the new FIDL api was being added. The logged status is still
available in the epitaph.

Change-Id: I65df9adba1d18a1023dc881e8ee9b13b6b4b4c19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013153
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sen Jiang <senj@google.com>