[vfs] Add array-backed simple files
This is a way to create a simple file that works on both host and target
(i.e. with or without VMOs).

On host we present the `read_only` function in the top of the file
module, and on target we re-export the vmo-based `read_only` in the same
position.

Change-Id: I3130033e1432d59ec6340e581cd20979a4e78aab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012880
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>