[debugger] Notify when installing recursive filter
Adds a new notification to be sent from the Agent to the client,
notifying that a ComponentDiscovered event was received by the
ComponentManager and a recursive filter has been applied to match other
components being launched in this realm.

Currently, only the session cares about this notification, so there is
no modification to the SessionObserver interface, but can be easily
expanded to include an observer notification.

Fixed: b/330571289
Change-Id: I40ccfcecaabdc6135ff2a94c3ed4b0306c53acbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013624
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>