[sdk][vfs] Fix PseudoFile write handler
If a write handler is not specified, the file should be read only. The
write handler was unconditionally initialized even if one was not
specified, however. This would lead to a nullptr dereference if the file
was opened as writable.

Test: fx test vfs_cpp_tests
Change-Id: I7c6382464894c8a60bf997fc390b8af622b1b578
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014612
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>