[test][intl] Fixes change detector tests for date l10n
The about-to-be-rolled version of the ICU library changes the way
localized times are formatted for Spanish. It turns out that
such changes happen all the time, which makes this test brittle
in general.

And with a diminished interest in l10n on Fuchsia, I am not quite
sure how useful this example test is.

Let's patch it up now (by replacing the changed part of the string
with a permissive regex), and if it continues to generate churn,
consider removing it completely.

Bug: b/330351469
Change-Id: I67ff3d1225c97da37ae5073d965a49bc17f93daf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014633
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Justin Kobler <jkobler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>