[ffx] More flexible Error handling
When deciding whether an `Error` is an `FfxError` or not in order
to convert it into `command` error, first check whether it is
_already_ a command error.

This may reduce the occurrence of some `BUG` messages for user errors
(or alternatively, make it easier to generate user errors).

The error layering code in `ffx` is needlessly complex, so this is
really a hack to avoid dealing with some of the confusing code. But
it shouldn't harm anything.

Change-Id: I737c25a344b596a730a6ac4d84cffda609bf3177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012792
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>