[debugger] Don't crash for invalid thread KOIDs.
I saw a crash where an invalid KOID was sent (probably from a thread
that had exited). This prevents that crash on the principle that we
shouldn't crash based in bad IPC messages, although the agent shouldn't
be sending such bad messages. I could not reproduce to get to the bottom
about the bad message.

Change-Id: Ibef33782141388a755d4e489a1a68a73192a6012
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016272
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>