[tee] Runtime binary and TA loader
This adds a binary to host the runtime support and load Trusted
Applications from dynamic libraries. The runtime binary exports TEE_...
symbols to the global symbol table for the TA to resolve when it loads.
The runtime binary expects to be included in a package that includes a
ta_name file from which it can infer the name of the TA shared library.
In this change, the binary doesn't do anything with the TA once it is
loaded.

Change-Id: Iee3230e8aa3aeeb61dd226ac07462e96f29e9ac8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015693
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: James Robinson <jamesr@google.com>