[kernel][sched] Rename "capacity" as "firm capacity"
This change eases the full introduction of the concepts of firm and
flexible capacities, the latter of which will be introduced in a
follow-on change. That introduction will also bring with it a new
"flexible" bandwidth parameter, so it also seemed sensible to reverse
the ordering of sched::BandwidthParameter fields as first period and
then firm capacity. (So we end up with (period, firm, flexible) and not
(firm, period, flexible)).

Bug: 328641440
Change-Id: I77dbe3b2433a78103cf317ef359779f8723cd3ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015695
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>