[ffx][agis] Remove from being built-in
Removes ffx-agis from being built in to the main ffx binary, allowing it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes when `ffx agis` is invoked, but makes it so that both the tool and ffx can build faster and with fewer resources making it easier to iterate on them.

Bug: b/329326263
Change-Id: Idce1cba30527b8e9fbb591a2b60575a0c432123e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003649
Reviewed-by: John Rosasco <rosasco@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>