[ffx][playground] Fix some weird grammar gotchas
We have ranges in playground, which meant "cd .." would parse the ".."
as an open range. That's uncomfortable (especially since open ranges
don't have any semantics yet), so we'll add the slightly arbitrary rule
that we always choose to parse as a bare string if we see
a dot at the start of the argument. You could still pass a range with
"cd (..)" if you needed to.

Also, integers don't clear their surroundings when they parse, so
"foocmd 123abc" parses like "foocmd 123 abc" with the 123 appearing in
the parse tree as an integer. We'll be a bit more forceful about the
whitespace at the end of the string to fix that.

Fixed: b/325486174
Fixed: b/325989896
Change-Id: I196122a5b6d75923dfcc89ff15cf6cf968158584
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995237
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>