[fx][build] Fix to support more than one target.
A recent CL broke the ability to pass multiple targets to
`fx build`, as only the first one was taken. The root cause
is that `read` can only read one line at a time, and must
be called in a loop.

Define a command-stdout-to-array function that uses `read`
in a loop with Bash 3.x (for MacOS), and the much more
efficient `readarray -t` for older Bash versions (Linux).

Bug: 42084668
Change-Id: I1051db321472c6aac59e07b7b1c18014f6736522
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017033
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>