[debugger] Parallelize loading .dwo files.
LLVM can't handle accessing a binary on more than one thread, but with
debug fission compilation we have multiple binaries that can be loaded
in parallel. This improves indexing performance, taking about 60% of the
original time to load Chrome.

Adds a simple WorkerPool implementation from GN.

Adds a new file to manage the indexing to keep the threading logic out
of the already-complex ModuleSymbolsImpl.

Converts the Index member variable to a pointer so it can be moved from
the indexer to ModuleSymbolsImpl.

TEST=existing fission tests (there should be no behavior change).

Change-Id: Ibd6b1cdfaed5c3bfb1e7f26647a683c63668f91f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016178
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>