[shush] Avoid parsing files with c-string literals
c-string literals were stabilized in rust 1.77, and are in use in
fuchsia source code. Unfortunately, proc-macro support was not
added for c-string literals, so syn is unable to process these
leading to a crash in syn if we try to parse any files with
c-string literals. Skip over any files and emit an error requesting
the affected lints to be manually updated in those files.

Bug: 331979452
Change-Id: Ib32661a66c36e7fc7a5db423502e05130f984c69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017660
Fuchsia-Auto-Submit: Paul Faria <paulfaria@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>