[profiler] Properly Clean Up Launched Components
If we launched a component that failed to resolve we should still clean it up.

Rather than setting that we need to destroy after we start the
component, we instead immediately set the destroy requirement on
creation.

Fixed: 329497443

Change-Id: I982ce326c7faa7b103046996617b9b07cfeca0d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007693
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>