[firmware] Fix `fx build sdk:firmware`
After https://fxrev.dev/1017115, the build was still failing due to this
collection with the SDK category "experimental" depending on SDK atoms
in the "excluded" category.

Move all SDK atoms, including the collection, that are specific to the
firmware SDK to the "excluded" category.
//third_party/android/platform/external/avb, which is in a separate
repository, is the remaining exception.

Also remove two of the atoms from the sdk_source_sets_allowlist since
they are not in the relevant SDK categories.

Bug: 331832768, 331962496
Change-Id: I218dd68bcc221bb61a5f8dbf2ccd3ec15833f552
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017794
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>