[kernel][x86] Move x86 PT code into header
This is a minimal cut+paste to move the X86PageTableBase methods to be
implemented in the header. Purpose of this change is to allow the class
to be templated without needing to move the implementation in the same
commit.

Change-Id: I8350439881d9ac027fb693f14b922a4f4c4f7685
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013173
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>