[libc] Don't use a system call in abort
Remove the "fallback" code in abort that called zx_process_exit
in the event the trap instruction wasn't fatal.  In fact, the
compiler was already eliding the call because it "knows" that
it's "impossible" for `__builtin_trap()` to return.  It also
elides the loop.  So really there is no point in having anything
in abort other than the trap itself (and its ICF avoidance).
Removing it also makes it easy to eliminate the link-time
dependency on the vDSO (which per above was already not really
being used).  This was the source of the unexpected vDSO
dependency in the user.basic //sdk/lib/ld/test/modules modules.

Some of the test expectations had been set based on the extra
dependency being there, so fix those.  One of those was being set
that way inadvertently because the IN_PROCESS_TEST define had not
been plumbed to be visible in each test executable's own code,
only in the common test-start code.

Bug: 42085436
Change-Id: I721c3662b89a7980ed17bcfcc11b19504997ea5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019612
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>