[inspect] A Graph API for Inspect
This library intended for usage in Power broker and potentially in
component manager. It provides a Digraph abstraction with a common
schema. This provides the benefit that a single tool could be used to
convert this schema to, for example, dot and easily write tools and
visualizations around graphs recorded to Inspect.

The Digraph API follows Inspect RAII semantics.

Each vertex and edge of the graph can contain key-value metadata
properties representing state of the vertex/edge or metadata associated
with it. The values can be updated and overwritten.

The follow-up will introduce tracking of events representing changes to
the graph topology or the metadata associated with each node and edge.

Bug: 331657401
Change-Id: Ib8612bb7f3c9aec674add2e26d56d0dbf8fb8437
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017412
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>