[ld] Avoid extra vDSO link-time dependency in variant test modules
The coverage et al configurations need to inject vDSO dependency
and other link-time settings for the default Fuchsia toolchain.
But in no-compiler-abi toolchains as in is_kernel toolchains, it
should not inject any dependencies or settings because those are
meant to match the needs of using the prebuilt runtimes from the
toolchain, which are not used in no-compiler-abi or kernel cases.

Bug: 42085436
Fixed: 332787312
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I0edcc47b3ae9ae7aeb27138e93c1705adee775f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020515
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>