[starnix] Cache a task's /proc/pid directory
The cost of reading /proc/<pid>/status dominated the cost of running
`ps`.  Each read generates a /proc/<pid> directory on open(), and
then drops it on close().  About 40% of the cost of the read
(verified by ubenchmarking) is attributable to the drop associated
with the close() operation.

This change constructs the directory once per Task, and then drops it
when the Task is released (after it exits).

Rough benchmarking shows that the time to execute `ps` after this
change drops by about 25-30%.

This change also includes a small optimization (2-3%) that comes from
removing a clone of a Task's info.

(slightly simplified) ubenchmark used:

```
 std::vector<std::string> vec;
 for (const auto &entry : std::filesystem::directory_iterator("/proc/")) {
   std::string s = entry.path().filename().string();
   if (s != "self" && s != "thread-self") {
     std::string fname = entry.path().string() + "/status";
     std::filesystem::path p = fname;
     if (std::filesystem::exists(p)) {
       vec.push_back(fname);
     }
   }
 }

 char buf[2048];
 struct stat statbuf;
 int i = 0;
 struct timespec start, end;

 clock_gettime(CLOCK_MONOTONIC, &start);
 while (true) {
   for (const auto &entry : vec) {
     if (i++ == 100) {
       goto outer;
     }
     int fd = openat(AT_FDCWD, entry.c_str(), O_RDONLY);
     fstatat(fd, "", &statbuf, AT_EMPTY_PATH);
     read(fd, buf, 2048);
     lseek(fd, 0, SEEK_CUR);
     close(fd);
   }
 }
outer:
 clock_gettime(CLOCK_MONOTONIC, &end);
 auto start_time = start.tv_sec * 1000000000 + start.tv_nsec;
 auto end_time = end.tv_sec * 1000000000 + end.tv_nsec;
 std::cerr << "Time = " << end_time - start_time;
```

Multiply: verity-benchmarks: 100
Bug: b/291962828
Change-Id: Idc1010a1b9893954a494043baf83b7b669294521
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009609
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>