[igc] Keep a counter for descriptor ring lengths
When the tx and rx descriptor rings become full their header and tail
indices will have the same value. This causes problems when
NetworkDeviceImpl::Stop is called because the driver can't use use the
comparison header == tail to determine when it has returned all buffers.
That comparison is immediately true without having returned any buffers
at all. An example of when this can happen is if a device is powered on
without an ethernet cable plugged in and then rebooted. The rx
descriptor ring will be full because there has never been anything to
receive. This caused reboots to be significantly delayed because
netdevice is stuck waiting for all buffers to be returned before
completing shutdown.

Fix this by adding a length variable to track the length of the
descriptor rings and use that to determine how many buffers to return on
the Stop call.

Also update the tx and rx ring tail index for each buffer in QueueTx and
QueueRxSpace to avoid firmware being confused by writing the tail value
to be equal to the head value when a single call to QueueTx or
QueueRxSpace fills up the entire ring.

Test: Manual tests to trigger these situations. Updated unit tests to
verify ring lengths.

Change-Id: I8e2c18d42f74120423dad1209e30d59c9bc676c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021141
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>