[starnix] Drop `Arc` around `FuseOperation`
There is no need to have a shared reference to the FUSE operation.
We only need minimal information about the requested operation to
handle the response/error.

This reduces allocations when making requests against FUSE.

This also allows us to free memory allocated to hold a `FuseOperation`
as soon as the operation has been read by the FUSE user-space
implementation. This is especially important for `FuseOperation::Write`
which could in theory hold a very large buffer.

Bug: b/324955805
Change-Id: I4decfb848fae9d90cc7bc906801a8011a551f24b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016186
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>