[netstack2] Fix Flake in interfaces watcher tests
The interfaces watcher tests asserted that `Zx.ErrCanceled` is returned
when a watcher client is closed with a pending call to `Watch`. This
assertion is flaky because the call to `Watch` is run in a separate
goroutine, and it's impossible to delay the assertion until we know that
`Watch` has been called.

Instead, remove the assertion, as it's really just verifying a property
of the golang FIDL bindings, rather than the correctness of netstack2.

Fixed: b/332987123
Change-Id: If20cffccc39a5c9787819fb230ef7aea67980345
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021832
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>