[debugger] Add more fine grained control of fd streamers
..to the Console. Now there is a similar counter as for input and
output, but applied for streamers. The Console::Stream method replaces
Console::WriteIfEmbedded, and is tied to the counter being enabled. Now
we are able to disable streaming where we need to without changing the
transition points for embedded mode.

The first use case of this is when we load symbols, so we can show a
helpful message so the user knows their test isn't stuck.

Fixed: b/330771902
Change-Id: I9f74e15cd6dba3109e934d802a1b5fb962524128
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013092
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>