[debugger][e2e] Pass target address verbatim
When dealing with ipv6 addresses, our logic was not doing the right
thing to encompass an address in square brackets. I think at some point
in the past this was important since we didn't always get the ipv6
address in the format that ffx wanted, but that's not the case anymore.

Fixed: 333083347
Change-Id: Ib42f244fc864bf4da27ce7307d1bc019cdb9ee8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021856
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>