[tee] Add session lifecycle commands and parameter support
This adds support for tracking sessions in the TA runtime and adds
support for translating parameters for calls into the TA from the FIDL
representation to the in-memory representation specified by the TEE
Internal Core API. Value parameters are copied to their in-memory
representation and buffer parameters are mapped into the runtime's
address space with protection corresponding to the parameter's
"direction" specification.

This change does not add support for interpreting the parameters after
the TA returns.

Bug: fxbug.dev/332956721

Change-Id: I2e2e662ebee9f4462e5eabcfbd2486116bacaa3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021394
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>