[inspect] Open validator protocol and remove unused methods
Go FIDL does not support open and flexible protocols. Therefore,
the VMO-based Go Inspect validator is moved behind a Rust intermediator.

The protocol between the intermediator and the internal puppet is
as minimal as possible. The intermediator now intercepts Unimplemented
actions on behalf of the Go component, as this doesn't have any real
bearing on what the tests are testing (the format).

The Validator test also queries the Archivist for the Inspect from
the component; the Go component does not export any Inspect. Instead,
the intermediator does that on its behalf after a call to
InspectPuppetRequest::Initialize. Again, this is because who serves the
VMO is not really relevant to what is being tested: the format of the
VMO.

The intermediator does nothing upon receipt of an
InspectPuppetRequest::_UnknownMethod. This means the protocol used by
the full-featured libraries (C++ and Rust) can continue being extended
without harming the coverage of the Go library.

Fixed: b/330901977

Change-Id: I21250ed57231306d9ad1d7f119d244fe05d93705
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013033
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>