[fidlc] Simplify resourceness validation
This makes fidlc only report errors on value types that directly contain
a resource type. This was my original intention in the RFC:

https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0057_default_no_handles?hl=en#language

> It only needs to check direct fields: if A contains B, neither are
> marked as resources, and B contains a handle, then compilation will
> fail due to B and there is no need for a separate error message about
> A transitively containing a handle.

However, for some reason in Ide6bdd48258cd94b8014284d51b341354463c4aa I
went against that and reported transitive errors. I no longer think this
is a good idea, because it assumes the fix is to add `resource`. The fix
might well be to remove handle fields instead, in which case the extra
errors are harmful, masking where the real problem is. Only reporting
direct errors is also more in line with what other compilers do for
similar concepts such as constexpr or unsafe.

The other motivation for this is it simplifies the compiler, making it
easier to maintain.

Test: fx test fidlc-test
Change-Id: Ib3929a3170897f4168e7d2b90f4b136401aeb3ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020356
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>