[kernel][lockdep] Improve circular dependency reporting.
Improve circular lock dependency reporting by displaying each direct
dependency of a lock that is in the same circular dependency set. This
makes it easier to determine how the circular dependency is formed and
which dependencies may be the easiest to remove.

Bug: 333077530
Test: Run: 'k ut lock_dep_tests' and observe the correct dependencies
listed for the circular dependency tests.

Change-Id: I9b52c1fdadeec242c80a922c67f0b79e18062ff9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021879
Fuchsia-Auto-Submit: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>