[kernel][sched] Have RunQueue hold onto the current thread
This change updates RunQueue to hold onto the "current" thread as state.
This is important as it will soon need to manage state relating to both
the running(/current) thread and the ready ones (e.g., for determining
the total 'flexible demand') - and it is easiest to do this when the
current thread is on hand.

This change also allows the current thread to be null, as there are
indeed cases where there is no eligible thread to run.

Two smaller consequences:
* EvaluateNextThread() no longer needs the current thread passed it;
* EvaluateNextThread() - having a name with tentative/'const'
  connotation - becomes a bit of misnomer with the introduced
  side-effect of updating the current thread: as a result, we rename it
  to `SelectNextThread()`.

Bug: 328641440
Change-Id: I9bdb92f79d52cd348f1e422ee9d877bcca103f9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016493
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>