[audio][adr] Control for Composite devices
Implement ControlServer for the Composite device type.
This includes SetDaiFormat, CreateRingBuffer, Reset and
ControlNotify methods, the first two of which target a
specific element_id.

Although some level of Device-level signalprocessing
support is present (including Notify methods), this CL
does not add support for the signalprocessing-related
client-facing FIDL methods in Observer/Control.

Additionally, the AudioDeviceRegistry FIDL servers use
the ControlNotify interface to subscribe to state-update
notifications from the device layer. The `DaiFormat was
not changed` message previously used a zx_status_t to
denote the reason for the failure to set the format.
Now it directly uses the error type that the FIDL server
can return to the ADR client.

A similar reworking was also done, with the internal
return codes used within `CreateRingBuffer`, in order
to enable the more nuanced FIDL error return codes.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

Bug: 321292607

Change-Id: I38bafb7655c9b392b14265db3dfed8a1a483a87c

[audio][adr] Convert SetDaiFormat error codes

The AudioDeviceRegistry FIDL servers use the ControlNotify
interface to subscribe to state update notifications from
the device layer. The `DaiFormat was not changed` message
previously used a zx_status_t to denote the reason for the
failure to set the format. Now it directly uses the error
type that the FIDL server can return to the ADR client.

Bug: 321292607

Change-Id: I38bafb7655c9b392b14265db3dfed8a1a483a87c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016525
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>