[fuchsia-component] Support opening service instances directly
Existing methods allow clients to connect to default or named instances
of a service capability by providing a DirectoryProxy to a directory.
However, these methods expect the service instances to reside in the
service name subdirectory of the DirectoryProxy rather than the proxy
itself. This CL adds two methods to connect to a default or named
instance of a service located in a given DirectoryProxy.

Changes:
- Adds connect_to_service_in_dir() and
  connect_to_service_instance_in_dir() to allow connecting to instances
  of a service located directly in a given DirectoryProxy instead of the
  service subdirectory rooted at that proxy

Bug: b/333433673
Change-Id: I5713a273c1bab001748c4db23020e5619240e971
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023401
Commit-Queue: Alex Wong <alexdwong@google.com>
Reviewed-by: Gary Bressler <geb@google.com>