[kernel][pmm] Remove disarm from checker
Removes the ability for the PmmChecker to be disarmed, and as a
consequence the ability to reconfigure it with different settings
during a single kernel execution. Doing this will allow for later CLs
to relax the locking requirements without losing correctness.

As the checker cannot be reconfigured, the existing host test is split
into two tests.

Change-Id: I89c52e060244eeca7c06b3f54979bc492ac52aaf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011452
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>