[netdevice] Use synchronized port shim dispatcher
It seems that using an unsynchronized dispatcher for the port and mac
servers in NetworkPortShim can cause a race during removal of the port.
The NetworkPortShim lifetime is tied to its server binding, when the
binding is unbound the NetworkPortShim object is destroyed.

The DevicePort class which owns the client end of the binding will call
NetworkPort::Removed on the client in its destructor, then immediately
after the client is destroyed and the binding unbound. There is some
evidence that indicates that doing this with an unsynchronized
dispatcher can lead to the unbind happening before or during the call to
NetworkPort::Removed. The Removed call would then attempt to use a
destructed NetworkPortShim object.

Change to a synchronized dispatcher for the port shim instead as there
is no reason to use an unsynchronized dispatcher here. Also remove the
storing of the binding refs in both NetworkPortShim and MacAddrShim.
These were never used and served no purpose.

Test: igc_tests netdevice-migration-tests gvnic-tests
      usb-cdc-ecm-unittest usb-cdc-ecm-integration-test
      virtio-netdevice-test network-device-lib-test
      network-device-microbenchmarks network-device-mac-lib-test
      network-tun-test wlan-drivers-components-tests
      virtio_net_unittests virtio_net_cpp_unittests network-tun-example
      netstack-integration-tests netdev-rust-client-tests
      netdev-cpp-client-tests network-device-test
Bug: b/331057379
Multiply: netstack-socket-integration-test-no-err-logs
Change-Id: I4d9ece464109c859506d3887f31d928864c2f8e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024114
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>