[inspect][contrib] Remove dependency on fuchsia_async executor
There was a dependency on the fuchsia_async Executor that caused this
library to not be usable in syncrhonous-only contexts. This change
removes this dependency.

As part of this change, the tests are now written using the FakeClock.
However, to use this fake clock we need to disable the parallel
execution of the test cases since there's one instance of the fake clock
shared across all tests.

Multiply: fuchsia-inspect-contrib-tests
Fixed: 333603664

Change-Id: I11e058f3ac8534eb7719a27629ca229985b979a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024372
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>