[ffx][process] Remove from being built-in
Removes ffx-process from being built in to the main ffx binary, allowing it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes when `ffx process` is invoked, but makes it so that both the tool and ffx can build faster and with fewer resources making it easier to iterate on them.

Bug: b/329326263
Change-Id: I57f66c61b02e0584275ed703e330d568bdc7c24c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003813
Reviewed-by: Przemek Pietrzkiewicz <ppi@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>