[time][rtc] Add an option `rtc_is_read_only`
This options allows us to configure a Timekeeper instance that
never writes to the RTC.

It may be helpful in some testing configurations such as the
one in the attached bug.

I also think props are in order to the original authors of this code,
whose design decisions way back when made this change very easy to do
today.

Bug: b/333581050
Change-Id: Id33f738111a332f74d33419d12a3d61ab01d0f73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024131
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>