[fidlcat] Pass string_view.data() to ifstream ctor
The ifstream ctor does not explicitly take a string_view as an argument
for its ctor. Prior to https://github.com/llvm/llvm-project/pull/85079,
the argument for ifstream was a `const std::filesystem::path &` so a
path is implicitly constructed. After that change, the ctor that
would've taken the string_view only participates in resolution if the
type is explicitly a `path` object.

This just instead passes the raw char ptr from the string_view which is
a valid ctor argument.

Change-Id: I0627139315a915d7d66c42a15f92505f0ce27065
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024987
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>