[kernel][arm64] dont map the kernel with > 4K pages
In specific situations, the boot mmu code may have mapped the kernel
with 2MB pages, which causes a failure later on during VM init when it
goes to change the protection bits on the kernel sections, causing a
large page split. With the new, proper break-before-make code in the
arm64 mmu, this can cause the kernel to temporarily unmap itself while
running.

Add an argument to the arm64_boot_map routine to specify if its allowed
to use large pages, and set to false when mapping the kernel itself.
This will make sure the kernel is mapped with 4KB pages.

This should fix a problem that was showing up frequently on CI/CQ only
on KASAN builds. The reason it only showed up then was it was the only
time the kernel was large enough to cross 2 2MB pages and having the
text segment end within the second one, causing the second 2MB page to
need to be split later.

Bug: 333424117

Change-Id: I18664e28da934be910c1a8d95f39275896d4ce80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024532
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>