[netstack3] Fix bug in LocalTimerHeap
Caught this while migrating fragmentation unit tests to local timer
heap. If we have two timers at the top of the heap with the same time,
the heap would not reschedule the next wakeup because of the
optimization to not touch the timer.

Change the rescheduling strategy to be locally tuned to each of cancel,
schedule, and pop.

Change-Id: Idc350cc5e09d1ec4270337e97fd6edd778850419
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024982
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>