[scrutiny] Pre-signing golden dirs
This change allows scrutiny pre-signing invocations to specify a path to
a directory containing golden files. The policy spec will update to
specify the name of the golden file within the dir.

This is needed to solve the issue of resolving the golden file location
at build time.

Bug: b/303923440
Test: fx test //src/security/lib/scrutiny/utils:utils_test
Test: Manual local test of scrutiny invocation with goldens dir
Change-Id: I3031ae14bf5465fca7da1602f65af7699c6e05f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024916
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>