[kernel][phys] Fix BootZbi::Error formatting
Make sure that the `.write_error` is not `nullopt` when
`.write_offset` is set, and same for `.read_error` with
`.read_offset.  This is necessary for the generic printing code
to actually show the offset in the error messages.

Bug: 333770144
Change-Id: Ie2a6d5cbb60561bd8f67e0faedccb65c37b268a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025834
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>