[display-coordinator] Clarify Client::SetLayerImage() logic.
This CL clarifies the use of an Image instance, vs the use of an
iterator in a HashTable that contains image. The CL also cleans up the
SetLayerImage error messages. This is a pre-factoring for future work
that reduces image_t usage in the display coordinator.

This CL does not introduce any behavior changes.

Bug: 42078780
Test: fx test //src/graphics/display/drivers/coordinator/tests
Change-Id: I314be39df7e4b581fb08797542bc003b92bce58d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026052
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>