[kernel][phys] Fix exception-case backtrace formatting
This corrects the backtrace printing in the exception case so
that the frames are numbered correctly: frame #0 is actually the
interrupted state, not a collected backtrace frame; frame #1 is
the first collected backtrace frame.  It also ensures that the
header line before the backtrace is before frame #0 in the
exception case (when frame #0 is a special case rather than the
first of the normal cases), and further unifies this printing
code between the frame-pointer and shadow-call-stack cases.

Bug: 333770144
Change-Id: Iceef867fb65a2f9f6ed9ade6aec4e0c8a715c9e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025974
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>