[ffx][log] Remove from being built-in
Removes ffx-log from being built in to the main ffx binary, allowing it
to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes
when `ffx log` is invoked, but makes it so that both the tool and ffx
can build faster and with fewer resources making it easier to iterate on
them.

Bug: b/329326263
Change-Id: I330adae02d0b67075051f339b39240f8a5899506
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003812
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>