[usb] Introduce usb-dfv2-compat-logging target for use in DFv2.
The existing usb-request.c file uses zxlogf. This change introduces an
alternative flavour of the existing target that allows for use in DFv2.

This is a non-functional change. The usb-request.c file has been
changed to a .cc file to allow for the DFv2 compat logging (using
zxlogf in DFv2).

Bug: b/333949904
Change-Id: Ib9e62d77013ffc325c3053539acb29b2e40ff513
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025659
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>