[component_manager] all ActionSet functions are async
In a future commit all of the functions in ActionSet will need to
communicate with a separate task, which necessitates the use of async
code. To reduce the number of lines changed in that future commit, this
commit makes these functions async and updates their callers as needed.

Bug: 319095979
Change-Id: I8e80f97d4376daa8f4b7e455a77ede05f0403929
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023952
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>