[netdevice] Improve shim ownership and teardown
There were situations when a NetworkPortShim object could be destroyed
and then have a FIDL Removed call attempted on it. This would cause the
FIDL wire messaging code to crash as there was no longer a server for
the call.

Change the ownership of NetworkPortShim and MacAddrShim such that each
object of these classes are owned by their respective parents instead of
being owned by the server binding they serve. NetworkPortShim objects
are now owned by NetworkDeviceShim and MacAddrShim objects are now owned
by NetworkPortShim. When one server unbinds it will signal its parent
through a callback and the parent will destroy the object.

This change creates the need for a modified teardown process for
NetworkDeviceShim as it now must correctly wait for NetworkPortShim
objects to be destroyed. Add a multi-stage teardown process to handle
be able to complete teardowns either as a result of a port being unbound
or the NetworkDeviceShim's own server being unbound.

Bug: 331057379
Test: netstack-socket-integration-test-no-err-logs on riscv64
Multiply: netstack-socket-integration-test-no-err-logs
Change-Id: Icee287661357ed2e2b2802572b908ad9382741fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027059
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>