[component_manager] unresolve is no longer recursive
The unresolve action is no longer recursive. Before this commit the
unresolve action would recursively unresolve all children after shutting
them down. This work was unnecessary because these children would then
be immediately dropped by the parent when the parent was moved back to
an unresolved state, and this work was also undesirable because shutdown
is a terminal state whereas discovered is not.

Additionally, unresolving the shutdown children won't be possible in a
future change, as it won't be possible to move a shutdown component to
any state other than destroyed.

Fixed: 334867885
Change-Id: I7407968f6b7357c51f3fc2d50f3963b2c6c5675a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028293
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>