[tee] Translate results of TA calls back to FIDL
This adds logic to interpret the results of calling a Trusted
Application's OpenSession and InvokeCommand entry points and translate
them to FIDL types to return to the client. There is now a ParamAdapter
type that consumes a set of FIDL parameters and produces the TEE_Param
array and param_types value that the TA needs as well as remember the
types and details about the parameters in order to interpret the
results.

Bug: fxbug.dev/332956721
Change-Id: I80bbbff67acfe93bb2d168042a794767e09f7c48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023617
Reviewed-by: Eric Stone <ecstone@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>