[package] Create a bootfs binary for exploring packages
This binary is like "ffx target-package" except that it is in bootfs
and so is available for debugging over serial. This will be useful
when the pkgfs shell directory that currently exposes all base
packages is removed.

To save space in bootfs, this tool and the "component" binary that
similarly mirrors "ffx component" are now implemented as libraries
that are linked into the same multi-call, busy-box style binary.
Copies of this binary named "package" and "component" are added to
bootfs, which deduplicates the files on disk.

Fixed: 328455656
Change-Id: If4d3fdd30aa4f9002f1ed807f2a793b950f04f0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027033
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Gary Bressler <geb@google.com>