[kernel][sched] Introduce hybrid scheduling support
This change introduces the notion of flexible work: that which is done
when the firm work leaves bandwidth for it. Unlike in the kernel today
where a thread having flexible work to do is mutually exclusive with
having firm work, we introduce 'hybrid scheduling' in which a thread may
have both.

Bug: 328641440
Change-Id: Ie9120d7d86bd3fbbd22621356a3ae8c164a18ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024515
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>