[zircon][sdk] Require explicit sdk category in zx_library
Previously, the zx_library() template allowed the `sdk_publishable`
parameter to be a Boolean or string containing an SDK category name. A
value of `true` was converted to "partner", and `false` was the same as
not being specified.

To simplify the logic and make the previously implicit "partner" SDK
category more visible, eliminate the Boolean options and require that
the parameter be set to an SDK category name if and only if it should be
"sdkable."

The `sdk_publishable` values added to individual targets do not change
the behavior and are not a judgement on whether that is an appropriate
category or even whether the target should be "sdkable." They simply
reflect the formerly implicit category.

Bug: 333125197
Change-Id: Ia829a37b1d0dac5d184c9e58927e27b0c8c04d97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021875
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>