[netstack3] Add IP-layer sidecar packet metadata
In the short term, this will allow us to associate a conntrack
connection with each packet.

The change was generated by adding a new parameter to the filtering
hooks and going where the compiler complained. It just so happens that
this manages to touch the majority of the IP layer and a small part of
the device layer as well.

In the long-term, we'd like to be able to pass information across layers
(e.g. from transport to device).

Bug: 328063820
Change-Id: Ic54d1942183cdec51f64106f5cab74231c814196
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021352
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>