[Honeydew]: Wlan typing and wlan affordances fix
This change fixes KeyErrors in the wlan and wlan_policy affordances by
switching over to using get(). It also fixes two typing issues. One with
the ProtectionDef returned by the wlan facade which is converted to a
string by serde in Rust. The other fix is changing ClientStatusResponse
to a protocol to better represent the fact that only one type of
ClientStatusResponse exists at a time. Finally, it updates and fixes
some of the unit tests for wlan.

Verified the following on Patchset: 5
* HoneyDew code guidelines
* Successfully ran the impacted functional tests using InfraRun

Change-Id: If24afaf45193688f6c26f37340cc913f6a0d501b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024592
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Patrick Lu <patricklu@google.com>