[kernel][vm] Do not Resume a thread while holding spinlocks.
Do no call resume on a thread while holding spinlocks, unless local
preemption has been disabled. If we are holding spinlocks when we call
resume, and the thread we are resuming is assigned to our current CPU,
it will trigger a preemption event.  It is illegal to allow a thread
to become preempted while holding a spinlock (otherwise we might end
up with non-running threads holding spinlocks, which would be Very
Bad).

This instance was missed, probably because the code in question only
runs in builds with the debug compressor explicitly enabled (which,
AIUI, is not many builds).  Even when this happens, the odds of
hitting the failure are relatively small as they require the system to
be sufficiently loaded to the point that the best CPU to put the
resumed thread on is the current CPU.

So, add an ASSERT to Thread::Resume meant to catch any attempts to do
this in the future (regardless of whether the operation would have
resulted in a local or a remote reschedule), and add an APD around the
spinlock scope to insure that any local reschedule operations are
postponed until after the spinlock has been dropped.

Fixed: 335213526
Change-Id: I7bd0aa915b0c3a69a4ae452a47df29230ccb1f7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029435
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: John Grossman <johngro@google.com>