[component_manager] resolve waits on discover, refactor lifecycle test
The test model::tests::lifecycle::bind_action_sequence was asserting
that a parent's resolve event must arrive before a child's discover
event, when there was no such guarantee in the events system.

The child's discovered action is started as part of the parent's resolve
action, and before this comment which action triggers an event first was
not deterministic.

This commit causes a parent's resolve action to block on a child's
discover action, which guarantees that the child's discover event is
emitted before the parent's resolve event.

Fixed: 335118565
Change-Id: If25a09d8ea14a25670445620579807ffe3638871
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029272
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>