[rfc] Explicit External / Platform Split in SDK Interfaces
The Fuchsia IDK doesn't specify whether products should implement clients or
servers for the various FIDL protocols that it defines. This makes it difficult
to correctly evaluate the backwards compatibility implications of changes to
Fuchsia APIs. This proposes a way to express that information in a both human
and machine readable way in FIDL. This unlocks opportunities to improve the
correctness and stability of Fuchsia-based products.

Change-Id: I52aed82b40c4c40f98da5a0ac17a1b0360961133
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982353
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>