[netstack3/conntrack] Basic generic state tracking
This implements state tracking that isn't specific to any particular L4
protocol.  For now, this means connection state (unestablished or
established) and the time the last packet was seen on the connection.
This will be especially useful for the table size limit
(https://fxbug.dev/328064032) and connection GC
(https://fxbug.dev/328063959).

Because this requires getting the current time, there's a lot of
plumbing to get FilterBindingsTypes and the new FilterBindingsContext
into the conntrack module.

Fixed: 328064022
Change-Id: Ic2d2052d4aaa8444a2cc4fced9d2819361a2a327
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031172
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>