[aml-pwm] Allow arbitrary number of pwm id metadata
The existing code expects a specific sequence of pwm id to be passed in
as metadata. Updated the code to move the assumption on pwm id into the
driver and only validate the incoming pwm id metadata on whether it is
supported on the current board.

Test: aml-pwm-test, device enumeration test
Change-Id: Iaacdba2bd29bcb754b8239a432f10813bc616dde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030498
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>