[fxfs] Filter paged ranges for unopened files from profiles.
Adds a capture of file opens, and filters out paged in ranges for files
that were never opened (and therefore were not actually used) when
writing out the recording.

This enables recording only files that are actually used by the system
and will drop ranges for files that are opened internally for a profile
replay.

Bug: b/333552162

Change-Id: Id661c2b03e2a8157b82e3bb49eb86a537a89c6ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031180
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>