[assembly] Introduce a new feature set level: Embeddable
This new feature set level is aimed at use cases where we don't expect
fuchsia to be self-updateable, but rather, externally updated. This
affords us the ability to drastically shrink the set of dependencies
necessary.

It is also meant to be the basis for products which are as small as
possible. It strips out the core realm entirely and will only bring in a
subset of the bootstrap realm. Features like power management, thermal
control, diagnostics, etc, are all considered optional add-ons.

Embeddable is currently a complete subset of the bootstrap feature set
level at present, although that may change in the future.

Bug: 325104472
NO_IFTTT=Lint is incorrect
Change-Id: Ide3f7d439d97c2a68d5fa80af4a1eab3e7e7872c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027172
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>