[ffx][schema] Adding product download/lookup to golden test
This adds the schema for `ffx product download` and
`ffx product lookup` to the golden file test for output schema.

These commands use the golden_file_test in
//src/developer/ffx/tests/mw-goldens/BUILD.gn

to generate the schema from the command line tool and compare it
to a checked-in golden file.

Bug: b/336385427
Change-Id: I6a46bb16204b8c1f02fe4e26acaf5e9657350214
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034313
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>