[build][info] Use Jiri hook generated values.
This is step 3 of the associated bug. This requires tqr/832170
to be submitted in integration.git first though.

Only use the values generated by the Jiri hook to compute the
final last commit hash/date/timestamp values. This removes any
invocation of `git` as a build action.

Bug: 336443040
Change-Id: Iac572b4ebc3ed0de1d3c520777d8817c4910ca89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1033936
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>