[fuchsia.io] Downscope connection rights based on protocol
Update protocol to document that the rights granted to a connection
can be downscoped to the operations which the target node can possibly
support. These rights can still be requested if the parent connection
has them to support protocol resolution.

This makes the resulting behavior more consistent with the current
implementation. Previously, opening a file with OpenFlags.RIGHT_WRITABLE
would only grant Rights.WRITE_BYTES + Rights.UPDATE_ATTRIBUTES to the
connection. However it was possible to obtain Rights.MODIFY_DIRECTORY by
specifying that right in an Open2 request (or by specifying no rights).

Report which rights a directory supports opening child entries with as
part of the node's abilities, and document that these rights apply to
both files as well as children of the current connection. For now we
update the conformance tests to assume directories support all rights,
but in the future this may have to be moved to the test harness (e.g.
the C++ VFS does not support client-side mutable directories).

Add TODOs to track disallowing specifying no rights with Open2.

Bug: b/324932108
Test: fx test rustvfs-io1-conformance-test fxfs-conformance-test
Change-Id: I176a29d4ecdd4a1139d4ba18bf3dd6628e5345b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031188
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>