[fxfs] Makes profile replays and recordings happen together.
Previously either a recording or a replay would begin exclusively
but now a recording is always begun and if a replay is available then
that is also replayed at the same time. The recording will start in the
graveyard and only be placed in the profile directory when properly
written and completed.

Bug: b/333552162

Change-Id: I25093b1c73b7a69925bd8c50bf2549b2d13314c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031182
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>