[starnix] Treat kernel processes as critical to the job.
Only terminate the kernel's job if a process exits with a non-zero code
so that normal termination of a process won't kill the job. The only
time that Starnix should have a Zircon process exit with a non-zero
code is when there's an unhandled exception that bubbles out of the
kernel's job. Regular process termination and Linux signals that
kill a process should not appear as non-zero Zircon exits.

This replaces the use of the kill_job_on_panic library to ensure that the
entire kernel is torn down if one of the processes panics. In turn
this should allow the feedback stack to collect a real stacktrace from
the crashing process.

Fixed: b/333891263
Change-Id: I7e0b400b80a78dff30fb0342b85bb684efa4c0f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028894
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>