[starnix] Add tests for various `brk()` edge-cases
Verify that pre-existing mappings prevent `brk()` being grown.
Verify that mappings over the program break do not prevent it
being reduced.
Verify that the program break address is the first byte after the
program break, not part of it, such that if the address is aligned
then the page containing it is not accessible.
Verify that shrinking to an aligned break, and re-growing,
correctly allocates fresh program break pages.
Verify that mappings over the existing break region do not prevent
the break from being extended, and are not overwritten.

Bug: 334365844, 333435783, 333450691
Change-Id: I6e4830805b94671fe0825392c581f19c6808f67b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034794
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Wez <wez@google.com>