[debugger] Make PC relative disassembly easier
Before this change you'd have to know what architecture you were on and
that you had to specify "*" on the disassemble command to tell it to do
some math to the PC before fetching the memory. Now just putting a '+'
or '-' in front of a raw number will be all it takes to get a PC
relative offset.

Change-Id: I25f4710b4655e17a646cdf090279b7b017f2dab1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1022000
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>