[devicetree] Preserve the devicetree node name
For use in driver framework, the devicetree node name was modified to
replace some of the special characters not supported by fdf. But the
name of the node is crucial for visitors while filtering out nodes that
they need to act upon. For this purpose, this CL adds a new name field
fdf_name() that holds the modified name of the node and the original
name is stored in the name() field as it is.

Test: *visitor-tests
Change-Id: Ic297b801a5b6722a5e986ae7c8c11ae68481fa23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035278
Reviewed-by: Braden Kell <bradenkell@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>