[lib/routing] Ensure dictionaries are not supported in legacy
We previously had a check in legacy routing that threw an error if a
component tried to offer or expose a protocol from a dictionary to
another component. However, it didn't fire unless the dictionary was
being used directly. If the protocol flows from a component that is
multiple hops away, it would not work.

Fix that issue, by throwing a dictionary not supported error whenever
a protocol is sourced from a dictionary, instead of a NOT_FOUND.

Bug: 326086827
Test: Added unit tests
Change-Id: Ife9435d72fd7228c7a9028c6be1b8ebfadcdebef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036113
Commit-Queue: John Wittrock <wittrock@google.com>
Reviewed-by: Gary Bressler <geb@google.com>