[graphics] Change gpu_vendor_id to uint32 and expose as config
We can use hex in cml files to specify the gpu vendor ID, so it makes
sense to use it as a uint32, rather than have a string that needs to be
converted to a uint32. The value 0 is now treated as undefined, and an
arbitrary GPU will be selected with that value.

Also expose this value through the fuchsia.gpu.test.GpuVendorId config
capability.

These changes modify the interface of the code in the SDK, but no
clients were using the old config value, so it won't cause problems.

Change-Id: I79aa19b9af6b9ad035615fb600d5c21ef677feb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036112
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>