[starnix/tests] Refactor FUSE tests
This introduces a lightweight implementation of a filesystem that can
store directories and files. This is needed so that we can make sure
that filesystem nodes have persistent node IDs. This also cleans up
how we differentiate between file types, improving readability.

Later changes introduce fixes to the in-kernel FUSE implementation
that require that nodeids not change for a node.

Bug: b/324955805, b/331965426, b/331965660
Change-Id: I929d9c0b345ccea2ad2ecb13086916ec74a092af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037013
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>