[graphics][vkproto] Change gpu_vendor_id to uint32 and expose as config
We can use hex in cml files to specify the gpu vendor ID, so it makes
sense to use it as a uint32, rather than have a string that needs to be
converted to a uint32. The value 0 is now treated as undefined, and an
arbitrary GPU will be selected with that value.

Also expose this value through the fuchsia.gpu.test.GpuVendorId config
capability.

Change-Id: I8ad216a10a37ff07e3465b8ff6778dcd99ef4201
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037232
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>