[superproject] Roll third_party/pigweed pw_transfer: Attempt to recover when receiving invalid size
This updates the C++ transfer code to not immediately terminate a
receive transfer if the transmitter send more data than was requested.
Instead, it will resend transfer parameters and attempt to recover.

Previously, this state was assumed to indicate a bug with the
transmitter, but with the addition of adaptive windowing, there are
legitimate cases where a window may shrink in size with in-flight
packets from the previous window.

To avoid an infinite retry loop if the transmitter is actually
functioning improperly, recovery attempts are limited to the transfer's
lifetime retry count.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/206632
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Original-Revision: b2c733596a777c5673ad1ca0983bd922d68ab619
GitOrigin-RevId: 2f6e5441e84cffe631a6f4173efe8e6c23297b2a
Roller-URL: https://ci.chromium.org/b/8749571569131246113
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I578c0a0c2be0920d1d19a4c4bbb0f725fd1267ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037732