[netstack3] Define transparent proxy filter action
Define a new fuchsia.net.filter/Action, TransparentProxy, that can be
added to rules in IP and NAT filter routines. When a packet matches a
rule with a TransparentProxy action, routing is skipped and the packet
is transparently redirected to a local socket bound to the specified
address and/or port *without* changing the packet header. (The receiving
socket can observe the original destination of the packet if the
IP_TRANSPARENT socket option is enabled.)

The TransparentProxy action is only valid in the INGRESS hooks, and must
be accompanied by a matcher that ensures the presence of a TCP or UDP
header. It is also a terminal action: if a packet matches a
TransparentProxy rule, no further rules will be evaluated for the
packet.

For now, this change defines the action, adds validation logic and
corresponding tests, and adds the correct processing logic to the
filtering crate, but does not actually perform the proxying in Core.
That will come (along with tests) in a future change.

Bug: 331819771
Test: netstack3-filter-test
      netstack3-core-test
      netstack-filter-fidl-tests
Change-Id: I38c41d0249bf4882190917c7a3d6fc09044c0b27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017532
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>