[ffx] Determine RCS state in target list (reland)
When running `ffx target list` when discovery is turned off
in the daemon, we discover devices locally. This CL adds the
next requirement, which is to try to connect to RCS, and get
get the product/board information.

(Added in reland) Timeout (default: 10 seconds; configurable
via "target.host_pipe_ssh_timeout") if there is no response from a
target, in which case report "unknown" for RCS and product.

Change-Id: I3aa8097f2e1e3e14a9846d761967f69af62eb461
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036333
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>