[assembly] Introduce a new AIB for ffx over vsock
This new AIB only makes sense for a product that targets virtual
machines and lack a network stack, such as microfuchsia. It's possible
to also enable it for use on other products that run on virtual machines
if we find a desire to avoid depending on the network stack, but that
will likely require another AIB with a core shard instead of a bootstrap
shard.

Change-Id: Id6b8898cdc85f484b75c5bbedd03b263c419cb65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030782
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>