[debugger] Automatically add source-map for rust stdlib
The rust stdlib sources are distributed as prebuilts, but do not seem to
adjust the DW_AT_comp_dir appropriately.

We can work around this for now with a default source-map setting while
the rust toolchain build is corrected.

Modifies the verbs_settings_unittest to not be dependent on path
resolutions for the new default value.

Bug: https://fxbug.dev/336868560
Change-Id: I20841cf512436528f3e8d8dca0436a0b0e70a15e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036353
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>