[dhcpv6/client] Add configurable DUID over FIDL
Add the ability to configure the DHCPv6 client's DUID over FIDL. In
particular, this allows using the DUID-LL and DUID-LLT methods of
generating a DUID using the link-layer address of a device, which
allows the DUID to be the same across restart of the client and
across reboots.

Netcfg has been changed to use DUID-LL with the MAC address of the
interface the DHCPv6 client is started on. This is good enough for
now, but the plan is eventually to use DUID-LLT and persist the value
to disk.

Tests were changed where needed. Also added a unit test for
constructing the DUID from FIDL types as the usage of `ArrayVec`
introduces panic opportunities.

Fixed: 330190503
Change-Id: Ibf826d8b10c69c88dbb486639fe5370d60d95ee8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028896
Commit-Queue: Tony Gong <gongt@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>