[starnix] send flatland frames after GP connection
This change merges the `start_flatland_presentation_loop` and
`send_view_to_graphical_presenter` threads into a single, ordered set of
operations in which we (1) first connect to GraphicalPresenter and
PresentView to the server, then (2) start sending frames to Scenic Flatland.

Without this change, it's possible to start sending frames to Scenic
before it is ready, e.g. if Scene Manager has not set a root view yet,
causing frames to be dropped. Since we know that when PresentView is done,
Scenic and Scene Manager will both be ready, this is a better signal
after which we can start sending frames that we know will not get dropped.

Bug: 333115991
Change-Id: Ic6f1c5fa81b9408360b0782ecbed37291b515fbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036081
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>