[aml][pwm] Apply channel config during init
This is a bug where the channel config passed in from the board driver
is not applied during init and therefore incorrect config was passed on
to the pwm clients as well. This resulted in pwm period being set to 0
for both voltage regulators in vim3.

Also added logs in aml-pwm-regulator driver to fail early when invalid
period is passed.

Bug: 338041521
Test: Manually checked the config values.
Change-Id: Iafe8d321608c9ce0ed3a5a0b8ceb87f8c0f0d18a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039915
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>