[devshell][serve] Fix setting the port
The change in fxr/1038492 changed which positional arg the arg parsing
was reading to find the port in a way that resulted in it usually being
unset. This restores the functionality by making the behaviour more like
the other arguments that take a second arg.

Change-Id: I9c7e01324bda7d3bcfa153ab1b9d2c6ebcd53056
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040172
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>