[audio][adr] DSP type-specific validation/logging
Strict validation and logging of signalprocessing Elements
and ElementStates, including all type-specific Element and
ElementState aspects that have been defined to date.

Depending on how 1-2 currently open questions are resolved,
this CL will need to be revised, but at the moment this CL
is valid as-is.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

Fixed: b/42069012

Change-Id: I982dbc4140dd09db99065349bb2ab10dc1b36c20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038018
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>