[components][dictionaries] Extension is on demand
Currently, when a dictionary that has `extends` is routed, dictionary
extension is performed once, and that extended dictionary is cached and
reused on subsequent routing requests.

Instead, this change makes it so that the extension is performed anew on
every route.

The reason for this change is to prepare for `extends program`, which
needs these semantics because the dictionary being extended will change
if the program is restarted. With this, `extends <other>` will be
consistent with `extends program`. It's also simpler to implement; the
only downside is the minor performance cost of creating a new dictionary
and routing the source dictionary on every routing request.

Bug: 336363726
Change-Id: I208ddd0700e52380398e52ec8b3dc794d65eace0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036512
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>