[brcmfmac] Fix interface destruction error code
The recent change to use FIDL for netdevice accidentally changed the
error code returned by DestroyIface when the interface didn't exist.
Upper layers in the WLAN stack depended on a specific error code to
re-create the interface after a crash recovery. With the accidental
change the crash recovery no longer worked.

Fix this by returning the correct error code and change the log message
to match what it was before the netdevice FIDL migration. Also add a
crash recovery test to ensure that the correct error code is returned.

Test: Manually induced crash test, updated unit tests
Change-Id: I8e824a4a0410069456524b45037d4f55b8877867
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040794
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>