[netstack3] Prepare for deferred address removal
We can't assume that the primary reference for addresses can always be
dropped with no strong references held anywhere. Prepare the world for
that by allowing addresses to enter a "pre-removal" state where the
config has been extracted out.

This is somewhat an artifact of how SLAAC works - it wants to be
notified *after* the address was removed and it wants to own the
configuration.

Bug: 336291808
Change-Id: Ifb7a7478393f53167d5d80b6f880b27315600602
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040180
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>