[fx-serve-remote] Fix `fx serve-remote`
This corrects two issues that was breaking the `fx serve-remote`
workflow. First, `fx` might not be available in the current $PATH, and
so we need to use `fx-command-run` to run fx commands. Second, parsing
out the port from the `-l` was not stripping the `:` character.

Change-Id: I3c2c05942ef8c029cf0317cc78b52b4a7764af67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040952
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>