[swd-e2e-tests] Look up ssh port in ffx
This queries ffx to figure out which ssh port we need to use in order to
communicate with the test device. This partially fixes the use of the
OTA tests with `funnel`.

Note: This unreverts http://fxrev.dev/1034317, with the following
corrections:

* Don't try to serve packages on the ssh port 22.
* Switches to `ffx target get-ssh-address`, which avoids the json
  schema change with `ffx target show`.

Change-Id: I3b33d59aea02bfaed626ccc88e416435f13c7302
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040333
Reviewed-by: Sen Jiang <senj@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>