[starnix] Simplify `set_brk()` implementation
Simplify the `set_brk()` implementation backing the `brk()` syscall
to rely on `map_anonymous()`, rather than attempting to grow/shrink
a dedicated backing VMO.

In the `alternate_anon_allocs` configuration this provides
automatic coalescing of newly-mapped BRK pages into the BRK mapping
and in the default configuration `try_remap_in_place()` is used to
extend an existing `Heap` mapping, rather than mapping a new
anonymous VMO.

The program break region is reserved via a placeholder VMO mapping
into the Zircon address-space, when `brk()` is first called. This
is overwritten with read/writable mapped anonymous pages as the
BRK is grown, and re-mapped into place if the BRK is shrunk.

The BRK is also fixed to treat the current program break as the
first byte after the break, rather than as the final byte of the
break. This is required in order for BRK pages to be correctly
zeroed when the BRK is shrunk and regrown.

Bug: 334365844, 333435783, 333450691
Change-Id: I4535cd4aab93537e3338e16f3b8c0eafe75776ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035852
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Wez <wez@google.com>