[serial][usb-cdc-acm] Convert to serialimpl FIDL
Implement the FIDL version of serialimpl instead of Banjo. There are a
number of behavioral changes included:
- Write() requests write the entire buffer, as mandated by the
  serialimpl FIDL protocol
- Write() requests wait for all USB requests to complete before
  replying, and surface USB errors
- Read() requests wait for some data to be received before replying, and
  surface USB errors

Also ignore USB requests that are completed with zero bytes, and a
remove a test case that was rendered obsolete as a result.

Bug: 333094948
Test: usb-cdc-acm-integration-test
Change-Id: I2064eee4b30b3c19e2897c3471cec0b34dffe4a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034096
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>