[swd-e2e-tests] Use the ffx that corresponds to the build
We need to use the correct version of ffx that can communicate with the
device. Initially this will be the ffx that's configured on the CLI, but
after we OTA the device we need to use the ffx from that build.

While here, this also makes sure to shut down the ffx daemon when the
test is complete, optionally support an external isolate directory, and
to delete any temporarily created isolate directories.

Bug: 336897946
Change-Id: I8405412ec50ceca4ea667ff98c6a8c5bcbc0b4a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989203
Reviewed-by: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>