[netstack3] Extract EventContext to base
Extract EventContext and `FakeEventCtx` to the base crate.

We hit orphan rule/coherence problems here because of the blanket impl
on DadEvent, which is transformed in core before it's exposed to
bindings.

Solve that by adopting the same pattern used for timers: Events that
need to be converted before they're exposed to bindings do so via the
core context, and integration takes care of the type conversion.

Bug: 338448790
Change-Id: I51419229664a7e80c794a74b0d3280ea0d092176
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041943
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>