[netstack3/conntrack] Implement connection GC
The GC here is very simple.  Every N seconds, it scans over the entire
table and removes any connections that are expired.  For now, the time
limit is the same for all protocols, but this will change once we add
TCP state tracking.

Fixed: 328064032
Change-Id: I26b434c9b75578d41a529a84ae666139ea8d22d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040236
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>