[netstack3] Extract NonTestCtxMarker to base crate
This marker is load bearing to barrier handler blanket implementations
for fake contexts and is expected to be used in different crates as part
of the split. Given that, it should live in the base crate.

Bug: 338448790
Change-Id: I3e741650e1a395ef5047801721b58223d60906b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042001
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>