[ffx] Remove discovery.mdns.enabled
The default is also hard-coded, and removing it here allows
`ffx.isolated` to act as an alias. Due to the way the `ffx.isolated`
alias is implemented, for it to work when `--isolate-dir` is set, there
can't be an explicit setting of `mdns.discovery.enabled`, even as a
config.json default.

Note: this doesn't change the functionality at all. If it's not set,
the hard-coded value is "true" (see `is_mdns_discovery_disabled()`
in `config/src/aliases.rs`. It just addresses a (ideally temporary)
situation in which the "alias" `ffx.isolated` doesn't actually affect
the mDNS discovery.

This change also makes it consistent with `fastboot.usb.disabled`, which
is not mentioned in any default `config.json`

Bug: b/331643328
Change-Id: I3eb8097c213e75818ae9c6bceb13b42644cce979
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040727
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>