[vfs][cpp] Downscope connection rights to abilities
Implements the changes described with the protocol changes made in
I176a29d4ecdd4a1139d4ba18bf3dd6628e5345b9. Move protocol negotiation to
internal namespace as this is mostly a VFS implementation detail.

Rights are downscoped before a connection is served based on the
protocol which was resolved. We also must perform this operation when
cloning a node, as fuchsia.io/Node.Clone still uses io1 OpenFlags, which
may include rights not relevant to a particular protocol (and thus the
parent connection is lacking).

Bug: b/324932108
Test: cppvfs-io1-conformance-tests
Change-Id: Ied3bb6d2c90f8ea81338a2adc95afd51afe2f385
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041693
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Rui Qi Sim <rqsim@google.com>