[analytics] Adding user_first_touch_timestamp
This populates the GA4 field `user_first_touch_timestamp` with
the last modified time of the uuid file. In most cases, this file
is created once and not modified unless the user opt outs/in
multiple times.

The goal is to use this to segment users into 2 groups,
"first day users" and everyone else. The hypothesis being CI systems
start new for each build.

Bug: b/322234459
Change-Id: I8d7d7a0bdf0571aae427154c11018f8da87d02ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042792
Reviewed-by: Bob Evans <bobevans@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>