[netstack3] Extract FakeBindingsCtx to base
Core has 2 FakeBindingsCtx types: One is meant for small local unit
testing and the other is multithreading-capable and used for integration
testing. This CL extracts the small one to the base crate.

The plan is that the larger one will remain as part of testutils for the
final core crate.

As part of this extraction, clean up some of the fns in FakeBindingsCtx,
like removing unnecessary accessor for fields that are safely public or
are not meaningful helpers.

Note that to satisfy orphan rules, the `FilterBindingsTypes`
implementation has been moved to the filtering crate, showing how we'll
accumulate other trait implementations on this types as we proceed with
crating.

Bug: 338448790
Change-Id: I51ef83fb55175b6c04ee650775a0ebb37be93492
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042756
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>