[kernel][x86][phys] Start compiling linuxboot support code for 64-bit
This starts the process of getting the linuxboot and legacy-boot
code built for the 64-bit environment.  This is mostly just build
wiring for a new "linuxboot64" flavor of "qemu" phys test to
enable experimentation.  This causes new builds of various phys
boot tests, but they are marked disabled so bots won't run them.

The main code change here is to start handling the full 64-bit
addresses for the command line and ramdisk (initrd) image, and
for the ramdisk size.  In the load-bearing 32-bit build, these
just now check for the high bits in the Linux boot_params that
were ignored before.  Where previously an address or size above
4GiB would be silently truncated to its lower 32 bits, now such
values will not be used but instead very early logging will be
attempted (and might have found a UART to log to) to complain
about the too-large values.

The other change only has any effect at all in the new 64-bit
build, which is to do early address-space setup in InitMemory.
The call was skipped before because it's a no-op for 32-bit.

Bug: 335547883
Change-Id: I91f8936001662d79c7ddab4c959aa20dc202ba26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037414
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>