[arm64-boot-shim] Periph ranges are represented GB aligned ranges
This CL separates how the periph ranges are generate into two steps,
the first step is accumulation and the last step is coalescing and
inflation.

We first accumulate all page aligned peripheral ranges into the
memalloc::Pool, when all ranges have been accounted for, we proceed to
coalesce them.

The coalescing involve looking for consecutive ranges in the pool, such
that there are no non-peripheral ranges in between them. After the
maximal run has been identified an uninflated range is calculated.

The uninflated range consist of a range starting at the physical address
of the first range of the run, and the length is the difference between
the "end" address of the last range of the run and the start of the
uninflated range.

The last step is inflating the range. This achieved by an ordered list
of alignments to try, in order to properly fallback to less optimal
alignments if the layout seems problematic. While unlikely is nice not
falling over if this happens.

Now the inflated range is aligned AND contains all the peripheral ranges
in the run, and is marked as peripheral, which coalesces all the smaller
ranges into it.

This is done for every run of peripheral ranges in the pool.

Fixed: 337332460
Change-Id: I66d422d949da3fb37e9c14e830770588c24c5d99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038877
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>