[bazel-sdk] Allow fuchsia_package to be testonly
When a fuchsia_package is marked testonly it fails to compile because
the tasks don't get marked as testonly. This is a quick change that
allows these packages to be testonly. In the future we should change how
we create the tasks for these types of packages.

Change-Id: I1167da2461d02171437e8db8d634a843a58b167e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043388
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>