[kernel][x86][phys] Add linuxbiboot_executable()
This implements a "fat binary" scheme for Linux/x86 "bzImage"
format compatible bootable images.  This combines the existing
support for a "linuxboot" phys_executable(), which runs phys32
(x86-32) code and can be booted in 16-bit or 32-bit mode, but not
64-bit mode; and the support for a "linuxboot64" phys_executable(),
which runs phys (x86-64) code and must be booted in 64-bit mode.

The combination is called a "linuxbiboot" phys_executable(), and in
fact compiles the same program twice, once as kernel.phys32 and once
as kernel.phys (64-bit).  The latter is packed as unused data in the
former such that the whole image is compatible with Linux/x86 boot
loaders using 16-bit, 32-bit, or 64-bit entry paths.  If the boot
loader chooses the 64-bit entry path, the kernel.phys program runs.
If it chooses 16-bit or 32-bit paths, the kernel.phys32 program runs.

Bug: 335547883
Change-Id: I6f0f4b8d843ae19f45b21f28c4f0876c24219b67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037955
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>