[superproject] Roll llvm-project/libc [libc][math] fmod: clear exceptions before the test instead of after
The test has no control over the CPU state before the test runs.

This test checks whether no exception flags are set, which may not be
true at the start of the test. This used to be not a problem because the
check was broken but that was fixed in ecfb5d9951554d8bdb6a499c958f48cc35f78a88

GitOrigin-RevId: 303447617712b6c47f9c207289ba829c053c3440
Original-Revision: 49c2af031e72d9f5510003210842f51c56cf447a
Roller-URL: https://ci.chromium.org/b/8748585043003475041
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I8d1285d5b468f97b4ca78ee8d71136eca6cc94a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043973