[fuchsia.io] Stop inheriting rights in Open2 when rights are not set
Previously, when rights were not set in Open2, the rights would be
inherited from the parent connection. This would lead to potential
rights escalation across remotes. This change stops the rights
inheritance when rights are not set. Now, when the rights field is not
set in an Open2 request, it would be treated as if empty rights were
requested.

Bug: b/336285069
Change-Id: I30c3d3e7f735c967e5041346d7265e68a891d5bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039512
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>