[kernel][arm] fix a typo in the boot allocator
The code is trying to compute a bitmask for a page size, but
accidentally uses -2 instead of -1, which would result in a mask that
has the bottom bit unset. It should mean it only matches on every other
large page, so it's been mostly harmless except it doesn't utilize large
pages as well as it should have been able to.

Change-Id: Ic83eaf0f98eeba37d8a370d61e66309d19f5c9ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043712
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>