[wlanix] Adjust scoring for BSS selection
Remove 5GHz bonus, as we no longer intend to support connecting to 5GHz
networks.

Apply scoring penalty for BSS's with recent connect failures. This helps
with the case where a BSS is problematic for some reason, or the AP is
trying to steer the client device to a different one.

Bug: 335462721
Test:
1. Added unit tests
2. Manually validate that connect still works

Change-Id: Ife9493c7953e60a7d8dce96c9873b8c8f95af054
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037941
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>