[bt][core] clean up legacy advertising report parsing API
The `bt::hci::AdvertisingReportParser` API, as currently implemented,
resembles an iterator API whereby after calling `GetNextReport(...)`,
users can check whether or not there is another report via
`HasMoreReports(...)`. Users can check whether the parsing encountered
an error via `encountered_error()` as well. This API is very brittle,
low level, and cumbersome to use. Partially, this was because it existed
before our move to Emboss. With Emboss, we can clean up some of the APIs
and make it easier and more consistent to use this code.

This API is needlessly complex. Rather, we can do all this in one
method and return back an std::vector with all parsed reports, stopping
the parsing the first time an error is encountered. This change makes
that update.

In short, we preserve current functionality with a much cleaner, easy to
use, and easy to maintain API.

Fixed: b/308500308
Change-Id: I7822f5b5025ee74c5544deb551ebf61f4ccac72a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034692
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Faraaz Sareshwala <fsareshwala@google.com>