[swd-e2e-tests] Use the ffx that corresponds to the build
This is an unrevert of http://fxrev.dev/1042292. The main change is that
we download all the blobs up-front, rather than lazily downloading the
blobs on demand.

We need to use the correct version of ffx that can communicate with the
device. Initially this will be the ffx that's configured on the CLI, but
after we OTA the device we need to use the ffx from that build.

While here, this also makes sure to shut down the ffx daemon when the
test is complete, optionally support an external isolate directory, and
to delete any temporarily created isolate directories.

Bug: 336897946
Change-Id: Ide46de1ec2ae318306dd90bd254711e7887a35b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042612
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>