[debugger] RootJob should not be used in unittests
This unittest was erroneously depending on notifications from the system
state after cloning itself.

This kind of behavior is improper for a unit test situation and should
instead be left for integration testing.

Allow the respective test access to the SystemInterface and
ComponentManager internals so that the test can be performed with the
necessary set up appropriate for a unit test.

Change-Id: I85d57b79d7230fd12255695df089b62c16b98e79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045222
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>