[ffx] Improve RCS knock timeout impl.
This updates RCS knock timeouts so that they are more consistently used.
The defaults are now derived from the amount of time it takes for RCS to
knock in general, and there is an error check if knock timeouts are
shorter than allowed.

Multiply: ffx_wait

Bug: b/339516247

Change-Id: I04095ce217f182283d33ab60df1552b975d54f77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045042
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>