[kernel][chainlock] Don't touch lock contents after release.
Make absolutely sure that we never touch the internals of a ChainLock
after its internal state has been set to Unlocked.  This aligns the
ChainLock implementation with special requirements defined for
std::mutex; basically that as soon as a lock is free and can be
obtained by someone else, the code releasing the lock has to consider
the object to be dead and cannot touch any internal lock state.

Prior to this, we were reading the ChainLock's contention mask
immediately after setting its state to Unlocked, which sets up UAF
potential.  See comments in the CL and http://fxbug.dev/339437199 for
more details.

Fixed: 339437199
Change-Id: I17ac50a6da0a3eb7e23851b715b0018e955e07d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045611
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>