[ffx][fho] Change verify_schema to associated func.
In VerifiedMachineWriter there is a function that takes `&self` that is
unused. As shown in this CL, requiring this makes it difficult to test
things when a writer is passed by value into a function.

This changes the verify_schema function to be an associated function
instead of a method to rectify this.

Bug: b/339725104
Change-Id: I2d7b53964cb615b889a48440ae82bb38c2d1087e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046093
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>