[fdf] Pass properties from composite node spec
Node properties of composite nodes contained all the properties of the
parent. This is different from the properties used for matching against
bind rules. This CL publishes only the properties of the parents exposed
from the composite node spec. We may have to follow up this change to
merge the properties of the parent as well.

Bug: 338667323
Test: composite_node_spec_v2_test
Change-Id: I6532442a519566397e866ab7092964ed5523007b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042464
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>