[kernel][chainlock] Don't evaluate types during if constexpr (false)
Make some of the internal chainlock methods templated in order to
prevent type expansion from happening in the false half of an `if
constexpr` statement.  If an `if constexpr` occurs in a non-templated
entity, both halves of the statement are fully checked, which is
apparently something that has always been the rule, but which is just
now starting to be enforced in compilers.

See the (large) new comment included in the CL for details.

Fixed: 339652126
Change-Id: I55710cf45f51328eb6463cdab65082679cb9fa9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046152
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>